Add GPU watchdog V2 framework and the finch flag
Introduced a new finch flag kGpuWatchdogV2 disabled by default. All V2 functions
added in this CL are the same as V1 for now.

The purpose of GPU watchdog Version 2 is to simplify the logic and reduce
the number of GPU crash reports if possible

Bug:949839

Change-Id: I7b6cd58425a91d379b7abc5c5452c974def905a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1652561
Commit-Queue: Maggie Chen <magchen@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669060}