tablet: Fix two bugs caused by adding screenshot.
One is the screenshot was showing weird on actual device.

The other is the screenshot is not deleted when using
SetBoundsDirect. That was because it was not animating, but
still managed by TabletModeWindowManager. The screenshot
should have been deleted if it is either not animating or
not tracked, not require both, which this CL fixes.

Test: manual
Bug: 948706
Change-Id: I87add8592a8f1e6bbec55dbe44e93eb52844c5a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1659020
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669099}