Remove extra attribute from Supervision Onboarding.
This was causing a JS exception by the end of the login flow because
the attribute adds the onboarding screen to the OOBE screen group[1].
While trying to adjust screen sizes for login screens[2], the OOBE
screens are null, so we throw an exception.

This attribute is not necessary to define the size of this screen
since we always take 100% of our parent container's width and height.

[1]: https://cs.chromium.org/chromium/src/ui/login/display_manager.js?rcl=7dad65d92a112624fe5706d97142c5325f163b94&l=744
[2]: https://cs.chromium.org/chromium/src/ui/login/display_manager.js?rcl=7dad65d92a112624fe5706d97142c5325f163b94&l=776

Bug: 958995
Change-Id: I541520f5e7ada60b24f349b980f23769d9570a6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1659469
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>
Auto-Submit: Lucas Tenório <ltenorio@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Michael Giuffrida <michaelpg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669271}