Fix UAF when nested ScopedFeatureLists are used.
Previously, multiple nested use of ScopedFeatureList did not
create a separate FieldTrialList for each ScopedFeatureList
instance, which could cause a feature to be associated with
a field trial that has been freed. This change fixes this by
making it so a separate FieldTrialList is created at each
nesting. Nested instances are allowed, but interleaving their
creation and destruction is not, which is covered by DCHECKs
and multiple tests are being fixed by this CL to not do that.

Parameters are also now overridden and restored properly in
nested use.

A DCHECK is added to FeatureList to detect the UAF scenario,
which would previously happen with nested ScopedFeatureLists,
but is now fixed by this CL.

Also updates tests that were previously managing their own
FieldTrialLists but now don't have to, since ScopedFeatureList
does this for them.

For some browser tests, there was a problem with the instance
of FieldTrialList created in content_main_runner_impl.cc being
deleted before the test's ScopedFeatureList is destroyed. To
fix this, this CL makes that FieldTrialList leaked, just like
the other per-process instances.

Also cleans up some unnecessary base:: prefixes in base:: code.

TBR=skyostil@chromium.org

Bug: 955643
Change-Id: Id89e5c34a605a32f84a5930492dab1480bbd07d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1624805
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Reviewed-by: Jun Cai <juncai@chromium.org>
Reviewed-by: Parastoo Geranmayeh <parastoog@google.com>
Reviewed-by: Carlos Knippschild <carlosk@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Reviewed-by: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669390}