Clean up code in chrome/browser/ui/views/location_bar.
This mostly entailed running `git cl format` over the directory. A
couple missing includes were also added, and one enum was converted to
an enum class.

The intent is to land changes to the code format before landing more
substantive changes to the code in another CL. Most of these files are
not currently planned to be touched by that CL, but I figured it made
sense to apply changes to related files since this is mostly automated.

Change-Id: Iad1d14afce8a5283ea4c927ee0bda75b0f029c3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1653890
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Aran Gilman <gilmanmh@google.com>
Cr-Commit-Position: refs/heads/master@{#669769}