desktop-pwas: Enable overscroll for app windows
Overscrolling for app windows was disabled in [1] because it was making
the Files video player slow. (old overscroll code used to take a
screenshot after each navigation and the video player was navigating
every 5 seconds). The code was later optimized to reduce the number of
screenshots taken, and eventually the code to take screenshots was
removed[2].

There is no reason why the feature has to remain disabled now that we
don't take screenshots.

[1] https://codereview.chromium.org/12082083
[2] https://chromium.googlesource.com/chromium/src/+/6e6c40b4697e889533a480fa04ffb2eb96dcda1d

Bug: 747587
Change-Id: I24b499b7f4c2f412d5efa2f687c9d96dd4cdd5f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1659650
Commit-Queue: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: Michael Wasserman <msw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669923}