Onion soup webrtc_local_audio_source_provider.cc|h
This CL is part of the phase 3 on the document [1].
It includes moving to Blink files and content/renderer/media/stream
dependencies from media/audio or media/webrtc that block Onion souping
more entangled files.

[1] https://docs.google.com/document/d/1rHJGi1U72qZsOGIctIf7GKRaY8d7BgtVW8MaMYumzYY/edit#

As a follow up to the CL, this class will be renamed to
WebAudioMediaStreamAudioSink.

BUG=704136
R=guidou@chromium.org, haraken@chromium.org

Change-Id: Id86386441212d7909e20cb2be90b2ca8c4c1f409
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1659474
Commit-Queue: Antonio Gomes <tonikitoo@igalia.com>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669938}