Deprecated wayland_aux
We originally created the wayland aux dir to store the tools used to
interface with wayland, however later on we moved to Jinja which made it
more convenient to keep the wayland tools closer to the source.

This CL replicates the historic functionality of wayland_aux under the
fuzzer/ directory and removes the old aux tools.

Bug: 961564
Change-Id: I60394f66fb6399fd1e2bd3dfe2d6d051006f3409
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1662091
Commit-Queue: Nic Hollingum <hollingum@google.com>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Timothy Loh <timloh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#669984}