Disable 2 WebRTC browsertests when network service is in-process
This CL disables two of the WebRTC browsertests when network
service is in-process. These specific tests verify that the
network service records the count of connections correctly.

This is similar to other browsertests that are disabled
when NS is in-process. See:
https://cs.chromium.org/search/?q=f:browsertest+IsInProcessNetworkService&sq=package:chromium&type=cs

Change-Id: I8a31c9a3601e2dadc24755dc96315c22c7a20ef2
Bug: 976186
TBR: sergeyu@chromium.org
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1663888
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Commit-Queue: Tarun Bansal <tbansal@chromium.org>
Auto-Submit: Tarun Bansal <tbansal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670186}