Fix NPE while recording AnswerType for Calculation suggestion.
Calculation suggestion is not like the others - it is not strictly an answer and does not
have a SuggestionAnswer attached. It is, however, processed by the AnswerSuggestionProcessor.

This change ensures no access to Answer structure is made by AnswerSuggestionProcessor
if any (currently only Calculator) of the managed Suggestion entries does not have an
associated SuggestionAnswer structure.

Bug: 976239
Change-Id: Iddfd0e90b901847a47628038e96851f3741d4ae5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1664039
Reviewed-by: Theresa <twellington@chromium.org>
Commit-Queue: Ender <ender@google.com>
Cr-Commit-Position: refs/heads/master@{#670199}