[Previews] Adding hosts to TopHostBlacklist on initialization.
When the HintsFetcherTopHostBlacklist is initialized, the top hosts
currently in the Site Engagement Service are added to the blacklist.
The number of hosts is limited to two times the number of hosts that
hints can be fetched for by the HintsFetcher. Metrics are also added
for the number of hosts added to the blacklist on initialization
and the number of hosts in the blacklist when top hosts are requested.

The next change will remove hosts from the TopHostBlacklist if the
blacklist has been initialized and the user navigates to the host.
Browser tests will also be added in a future change along with
functionality to force a request for top hosts to be made without a
delay.


Bug: 943846
Change-Id: I129a90985b318815361d17a25361a75eedd21c3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1650080
Commit-Queue: Michael Crouse <mcrouse@chromium.org>
Auto-Submit: Michael Crouse <mcrouse@chromium.org>
Reviewed-by: Doug Arnett <dougarnett@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670236}