Have DeviceO2TS return empty accounts if robot account ID is empty.
This CL changes DeviceOAuth2TokenService::GetAccounts() to return an
empty list if the robot account ID is not known. This is consistent
with how DeviceO2TS generally treats the robot account ID (e.g., only
firing refresh token is available notifications if the robot account ID
is set). It will concretely be necessary for an upcoming
OAuth2TokenService change that fires OnRefreshTokenAvailable() for all
accounts in GetAccounts() when refresh tokens are loaded;
FireOnRefreshTokenAvailable() DCHECKs that the passed-in account ID is
not empty.

Bug: 919793
Change-Id: Ia91fdd939e5840c0384bf7bde666db459029dc94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1663411
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670422}