[LayoutNG] Don't inline DetermineAlgorithmAndRun().
CL:1665928 caused a slight stack usage increase. Not inlining
DetermineAlgorithmAndRun() solves that problem, and even gives a tiny
stack usage reduction, compared to before the mentioned CL. Be explicit
about not inlining the function. I tried to just drop the "inline"
keyword and let the compiler decide, but the compiler then decided to
inline it, so ...

Bug: 976635, 976440
Change-Id: Ie4e0414b843e342b8d04c4044877e3d1f1e6a961
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1666635
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670479}