DevTools: Shortcut to open a breakpoint input dialog
The shortcut is Ctrl+Alt+B when focus is on some line in
the sources pane text editor. It opens the dialog used for editing a
breakpoint or creating/editing a conditional breakpoint or logpoint.
By default, it opens to a dialog for a conditional breakpoint, unless
there is an existing logpoint on that line. I also added this shortcut
to the shortcuts page in Settings.

Change-Id: Ie082fde1f01ecb1ba6cf9c2025e4207e54ca1d64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1654070
Reviewed-by: Joel Einbinder <einbinder@chromium.org>
Commit-Queue: Olivia Flynn <t-olflyn@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#670533}