Print Preview: Fix focus ring in advanced settings dialog
The focus rings around the selects in the advanced settings dialog are
being cut off on the left. Add margin on the start of the select
rather than on the end of the label to keep the same spacing but allow
the focus ring to be seen.

Change-Id: If60ab1d7f9ba5ad374fdf596ed2d82277033e9a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1662871
Reviewed-by: Esmael El-Moslimany <aee@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670572}