WebUI: Remove --cr-dialog-body mixin
The --cr-dialog-body mixin was used to style slotted content. Because
slotted content is introduced outside of the <cr-dialog> element and
is reachable using standard CSS selectors from where they are provided,
it is not necessary to use a mixin or expose a separate shadow part.

Bug: 973674
Change-Id: I0dfcf1bf896c11cc47bf845ab3451fdf0ba84632
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1663370
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670575}