CrOS Shelf: Remove dependency on back and home as ShelfItems.
This is a step in the direction of separating the navigation buttons
(back and home) into their own widget. The change is rather large
because of all the places that assume these two elements are part of
the model, even though that isn't strictly necessary.

Bug: 971426
Change-Id: I6cb3b02d8ccd69bf7119c294135eea94b3a51901
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1648821
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Commit-Queue: Manu Cornet <manucornet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670632}