Reset link highlight when the containing frame is throttled
We don't set layer state for GraphicsLayers under throttled frames,
so if a link highlight is created before the frame was throttled, we
should reset it when we find the frame is throttled in
LinkHighlightImpl::UpdateGeometry() and LinkHighlightImpl::Paint().

This is a speculative fix.

Bug: 967281
Change-Id: I98ce93a48ffd88bbfd184a1fc15d891af077568c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1663317
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670673}