Make App Service conditionally sync app removal
When the AppServiceAppModelBuilder removes an ARC++ app, whether or not
this is a sync'ed change is now conditional on whether Play Store as a
whole was enabled or disabled.

This matches the existing behavior (with the App Service disabled), in
ArcAppModelBuilder::OnAppRemoved.

This commit also changes ChromeLauncherControllerWithArcTest, adding
RunUntilIdle calls so that it works with the Mojo IPC based App Service.

Together, this fixes `unit_tests --enable-features=AppServiceAsh
--gtest_filter="ChromeLauncherControllerWithArcTest.ArcAppPin"`. It
fails before and passes after this commit.

Note that `unit_tests
--gtest_filter="ChromeLauncherControllerWithArcTest.ArcAppPin"`, without
AppServiceAsh enabled, passes both before and after this commit.

BUG=826982

Change-Id: I418ee223ba40ae0099a173f06d6a06217cbae7d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1626206
Commit-Queue: Nigel Tao <nigeltao@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Yury Khmel <khmel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670805}