Enable PauseExecutionContextOnBackgroundFreeze
Thie change enables (experimentally) pausing the execution context
when a document is frozen.

The WPT tests written required that promise resolution continued after
the page was frozen. This isn't the case when
PauseExecutionContextOnBackgroundFreeze is enabled so rewrite the tests
to poll for the correct state.

BUG=957589

Change-Id: If104235fd7281d88b5d86a3f0b89337828e7f4e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1662633
Reviewed-by: Shubhie Panicker <panicker@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670883}