Implement the average character width for PlatformFontSkia
This CL is adding a way to estimate the size of an average
character when the SkiaMetrics doesn't provide it.

Some SkiaFontMgr don't provide the average char width
since the metric is not always available. As an example,
the DWrite doesn't provide the average char widths
(see structure DWRITE_FONT_METRICS).

The blink code is using an approximation by looking for the
glyph 'x' and getting the size of that glyph.

We are proposing to use the same solution and fallback to
the max char width if the glyph is not available.

R=robliao@chromium.org, asvitkine@chromium.org
CC=​​msw@chromium.org

Bug: 969687, 944227

Change-Id: I8cf23b5e02fba211edb321b2ac66306ff458993b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1666720
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Michael Wasserman <msw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670892}