Verify if model was updated ahead of handling omnibox suggestion.
Omnibox suggestion interactions are handled in a non-sequential manner:
each interaction can - and typically is - scheduled for execution.

Before that execution takes place, a number of asynchronous events can
take place, for example - a new suggestion list can be collected from the
autocomplete mechanism.

In such situation users may interact with suggestion that is about to be
removed from the model; in extreme situations this model may appear empty
by the time we reach this code, causing the application to crash.

This change extracts the portion of the code performing this action from
`updateSuggestionUrlIfNeeded` method and reuses it to locate suggestion
in `onSelect`, too.

Bug: 976595
Change-Id: I8321767cbff01f53616ace4ee92c6984e4a9a49b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1668052
Commit-Queue: Ender <ender@google.com>
Reviewed-by: Theresa <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#670970}