AppCache: Mitigate bad traffic caused by invalid cached data.
This CL works around the following confluence of problems.

1) If a cached response contains a Last-Modified header,
AppCacheUpdateJob::URLFetcher::AddConditionalHeaders() adds an
If-Modified-Since header, so the server may return an HTTP 304 Not
Modified response. AppCacheUpdateJob::HandleUrlFetchCompleted() reuses
the existing cache entry when a 304 is received, even though the HTTP
specification mandates updating the cached headers with the headers in
the 304 response.

This deviation from the HTTP specification is Web-observable when
AppCache resources are served with Last-Modified and Cache-Control:
max-age headers. Specifically, if a server returns a 304 with a
Cache-Control: max-age header, the response stored in AppCache should be
updated to reflect the new cache expiration time. Instead, Chrome
ignores all the headers in the 304 response, so the Cache-Control:
max-age directive is discarded.

In other words, once a cached resource's lifetime expires, 304 responses
won't refresh its lifetime. Chrome gets stuck in a cycle where it sends
If-Modified-Since requests, the server responds with 304, and the
response headers are discarded.

2) The implementation of
AppCacheUpdateJob::UpdateURLLoaderRequest::OnReceiveResponse()
introduced in https://crrev.com/c/599359 did not populate |request_time|
and |response_time|. When the Network Service was enabled, caches got
populated with the default value of base::Time, which is the Windows
epoch. So, cached entries with max-age values below ~40 years will
require re-validation. https://crrev.com/c/1636266 fixed the cache
population bug, but did not address the incorrect times that have
already been written to users' disks.

The 1st problem, on its own, hasn't had a large impact. This is likely
because we have been advising sites to set max-age=31536000 (~1 year)
for immutable resources, and most AppCache caches have been getting
evicted before the entries' max-age expired. However, the 2nd problem
caused us to create a large number of expired cache entries, and the
unnecessary If-Modified-Since requests are causing noticeable levels of
traffic.

This CL introduces a workaround while a longer-term fix gets developed
and deployed. We'll consider all cache entries with invalid times to
have been created on Tue, Jan 29 2019. This is the day when M72 was
released to Chrome's Stable channel, and was chosen be cause M72 had the
first large Network Service deployment.

This CL was tested manually because the cost of developing an automated
test outweighs the benefits.
* Costs: AppCache doesn't support injecting a base::Clock, so the test would
have either required a lot of plumbing, or would not have tested the
exact bug scenario.
* Benefits: The code added here is a workaround. It will be replaced by
logic that updates the cache entry's headers when a 304 response is
received. Since the code has a short expected lifetime, it's unlikely
that an automated test will catch any regressions.

Tested: Manually, using the repro steps in https://crbug.com/972983#c3
Bug: 972983, 968179

(cherry picked from commit 20234641d9b26cc60123f5e9b5d2691fb311109f)

Change-Id: I68860ed28c4980e0a16297a243181995f5993ae1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1661472
Commit-Queue: Victor Costan <pwnall@chromium.org>
Reviewed-by: Staphany Park <staphany@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#669559}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1669178
Reviewed-by: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/branch-heads/3770@{#1051}
Cr-Branched-From: a9eee1c7c727ef42a15d86e9fa7b77ff0e63840a-refs/heads/master@{#652427}