[Autofill] Adds parameterized tests for checking suggestions.
The tests are parameterized to account for cases in which the feature
kAutofillUseImprovedLabelDisambiguation is and is not enabled.

The test class is structured in a way that will also facilitate
checking suggestions created with a mobile label formatter, whose
implementation is a work in progress.

Bug: 955513
Change-Id: I03d07591bbfc4041ab42f9811b4eedbfc438cb7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1578987
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Commit-Queue: Caitlin Fischer <caitlinfischer@google.com>
Cr-Commit-Position: refs/heads/master@{#671043}