Adding CORB protection UMA metrics (part 1).
Adding UMA logging that checks if a resource "seems sensitive"
according to either the CORS heuristic (i.e. has access-control-allow-
origin with a value other than * or null) or the Cache heuristic (i.e.
has cache-control: private and vary: origin headers). If the resource
satisfies either of these conditions we report if CORB would have
protected it, were it requested cross-origin.

This CL does not include sniffing logic, so we can only report if CORB
would have blocked the resource based on headers, or if it wanted to
sniff.

Bug: 815290
Change-Id: Ieb09218c309bbc27b4473e8a7eacd231318a0e49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1648903
Commit-Queue: Kristina Nelson <krstnmnlsn@google.com>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671145}