Fixed use-after-free in zcr_stylus
It is possible to trigger a use-after-free by binding to the touch and
stylus interfaces, and then releasing the touch handle first.

Likely this was not triggered, as Arc might release the stylus first,
but with crostini it would be easy to cause this.

Bug: b:135644583
Change-Id: I1fc72b63117738f6d9c88477ffce7e74ebb0defc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1668732
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Nic Hollingum <hollingum@google.com>
Cr-Commit-Position: refs/heads/master@{#671157}