Add format dialog ui element to file manager
This CL adds the format dialog polymer element, and hooks it up to the
file manager's context menu. It currently does not do anything yet as
the backend is not yet finished.

Also adds tests to ensure the dialog appears and disappears as intended.

Bug: 632988
Change-Id: I047e27d308eba67b0adf118b3414f7c4777757ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1663175
Reviewed-by: Noel Gordon <noel@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671273}