Fixed use-after-free in zcr_remote_shell
It is possible to force a UAF in the remote shell by pinning the surface
and then deleting it (without unpinning).  Eventually ash will try to
keep the window on top which involves using the deleted surface.

This cl fixes that by unpinning the surface on deletion.

Bug: b:135707944
Change-Id: I4cac53620384958dcee1f36366a500357d97c380
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1670172
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671334}