Pass the testing mock clock to IdlenessDetector via DummyPageHolder
IdlenessDetector was not using the mock clock coming from TestMockTimeTaskRunner
in some tests. That can be really problematic sometimes (like when the monotonic
clock has a really high value as a result of not rebooting a machine for quite a
while for example), as it might consistently schedule timers forever, making
calls like TestMockTimeTaskRunner::FastForwardUntilNoTasksRemain() to go on
indefinitely.

Instead of fixing the specific tests that were failing we're modifying
DummyPageHolder so it creates the IdlenessDetector with the passed in clock.
IdlenessDetector does no longer need a statically set clock. All the failing
test cases were either creating the DummyPageHolder directly or indirectly as
descendents of PageTestBase.

In the particular case of PageTestBase tests, they will automatically get the
mock clock set whenever EnablePlatform() is called. The API contract of
EnablePlatform() had to be slightly changed and from now on it must be always be
called before SetUp(). For those tests not wanting to use the
TestingPlatformSupportWithMockScheduler provided by PageTestBase a new virtual
method GetTickClock() was added so they can still hook up and provide their own
mock clock.

Bug: 976539
Change-Id: Id0af3424a60787a642023963de8698314e87bad1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1667311
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Sergio Villar <svillar@igalia.com>
Cr-Commit-Position: refs/heads/master@{#671339}