Fix gradient calculations for cubic Beziers.
The start and end gradient calculations handled one or two coincident
control points, but not three. The test for coincidence of P2 and P3
was not quite correct. The case of 3 coincident control points is not
addressed in the spec, and an issue has been filed.

https://github.com/w3c/csswg-drafts/issues/4046

The algorithm as proposed in the spec is also missing a constraint for
detection of coincident control points.

Bug: 827560
Change-Id: I8b216c8ab32881d6e100d338f4601c30e658bfb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1670150
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Ian Vollick <vollick@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671411}