AW: Set bg color to black when fullscreen in EvictFrameIfNecessary
In https://chromium-review.googlesource.com/c/chromium/src/+/1614798 we
added logic to set the background color to black for fullscreen webview
in OnRenderFrameMetadataChangedBeforeActivation. It turns out that on
pre-N Android, this is too late and misses the view being re-painted,
leaving a white background up until the next invalidation is forced.

This change in OnRenderFrameMetadataChangedBeforeActivation paralleled
the EvictFrameIfNecessary from OnFrameMetadataUpdated, but missed
other cases of EvictFrameIfNecessary.

In order to more exactly match the pre-surface-sync behavior, this CL
moves this workaround to a helper function and calls this from the
appropriate places.

Bug: 961223, 973365, 974692
Change-Id: I6067e4687fef918c97804634c32ad70be163182d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1670359
Commit-Queue: Eric Karl <ericrk@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Auto-Submit: Eric Karl <ericrk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671438}