WebXR layout test for XRInputSource re-creation
This test verifies that XRInputSource objects are re-created when
required by the SameObject attributes: when a gamepad is connected or
disconnected from the input source, when handedness changes, and when
target ray mode changes. The test also makes sure the input change
events are sent correctly.

This can't be an external WPT test because it uses MockXRInputSource
and a cross-browser XR input test API is not available yet.

Bug: 976978
Change-Id: I5fd5e90daf78c40fc8c57f70f0a389274222eb95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1672108
Commit-Queue: Jacob DeWitt <jacde@chromium.org>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671491}