Check if updating |CachedLayoutResult| failed for unexpected cause
r671460 (crrev.com/c/1670146) called |SetNeedsLayout| when
updating |CachedLayoutResult| failed for any reasons. It was
pointed out that we'd like to notice if it failed for other
reasons than |ComputeMinMax|, so this patch changes it to
|NOTREACHED()|.

Still calls |SetNeedsLayout()| because not doing so can lead
to bad results.

Change-Id: I39b77f748628efda354c64258dffbd92b8d59b1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1673845
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671678}