[iOS] Don't fallback to google servers when fetching favicons in omnibox
When favicons are loaded, if there's no favicon, we currently fall back
to fetching it from a google server. This creates a visual flash when
the user first sees a synchronously retrieved default image, and then,
moments later, a favicon.
This CL disables this fallback. This means favicons are not shown for
unvisited sites now.

Bug: 964918
Change-Id: If8d07cda3400b74b2a438c95cc2ccf14785465a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1631698
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671680}