Some optimization on download shelf warning dialog creation
Currently the logic is to try to append a link break after
each word, and see if the warning dialog length inceases.
This is very inefficient as GetPreferredSize() call is very
expensive.
This CL improves the logic by starting at the middle of the
text, and start searching forward or backward.

BUG=972090

Change-Id: I5c96e52d44af9d8e82690814e5856381b08ab554
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1670368
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Min Qin <qinmin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671714}