Make ContentShellBrowserTestActivity startup asynchronous.
This moves content_browsertests and component_browsertests on Android
off the initChromiumBrowserProcessForTests() and onto production code
paths for initialization. We use the standard BrowserStartupController
startBrowserProcessesAsync() method and wait for it to complete. The
BrowserTestBase ui_task waits for the async initialization before
running the test, by checking with the NativeBrowserTest.java class.

R=avi@chromium.org, boliu@chromium.org, dtrainor@chromium.org, jbudorick@chromium.org

Bug: 961849
Change-Id: I6b6ea2225a342e79e1f906d618fdc21c8b97568b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1669981
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#671716}