Key feature metrics for autofill
This CL introduces a sequence of metrics representing the autofill funnel as
well as a set of key feature metrics to judge the success of autofill. These
together are intended to serve as a long term set of metrics to guide our
feature development.

This CL has a major effect on the existing Autofill.FormEvents.Address.*
metrics: In the past, everything that was not a credit card, was considered
an address form (i.e. also password forms). This changes with this CL.
Only forms that contain an address field are considered an address form.

Bug: 966475
Change-Id: Ifb3f64a16988ed7939e76cad2dd3b58aad64d9dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1628607
Commit-Queue: Dominic Battré <battre@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Fabio Tirelo <ftirelo@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Maxim Kolosovskiy <kolos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672016}