Patch Windows 10.0.18362.0 SDK (1903) toolchain
This updates the previous package to use the 10.0.17763 version of the
Debuggers directory. This is necessary because the 10.0.18362 version of
dbghelp.dll doesn't load on Windows 7 and we rely on it. This package
was created by replacing the Debuggers directory in the previous
package. Therefore it is:

Mostly 10.0.18362
Debuggers from 10.0.17763
d3dcompiler_47.dll from 10.0.17134

The instructions for the previous package are included below:

Packaging was done on a Windows Server 2019 VM, cleanly created for this
purpose.

The package was created by downloading VS 2017 Update 9.3, from
https://visualstudio.microsoft.com/vs/older-downloads and downloading
the 15.9 Visual Studio Professional 2017, and then running the installer
like this:

$ PATH_TO_INSTALLER.EXE ^
    --add Microsoft.VisualStudio.Workload.NativeDesktop ^
    --add Microsoft.VisualStudio.Component.VC.ATLMFC ^
    --add Microsoft.VisualStudio.Component.VC.Tools.ARM64 ^
    --add Microsoft.VisualStudio.Component.VC.MFC.ARM64 ^
    --includeRecommended --passive

Then the 10.0.18362.0 SDK was installed from:
https://developer.microsoft.com/en-US/windows/downloads/windows-10-sdk

Then the packaging script was run like this:

  python depot_tools\win_toolchain\package_from_installed.py 2017 -w 10.0.18362.0

Since the new d3dcompiler_47.dll uses the UCRT and we want to avoid shipping
that the final packaging step was to unzip the package, copy over the two copies
of that DLL from win_sdk\Redist, and then repackage the toolchain with:
  > python package_from_installed.py --repackage=<full-path-to-toolchain-dir>

This is needed to make progress towards two-phase lookup.

Bug: 970064, 976161, 919163
Change-Id: I1043a01c91b040fd7e47a160f953fd5b6502ed0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1674628
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672194}