Expand multiple-acitvity PIP when the shelf icon is pressed
Currenlty, if you click on the shelf icon when Netflix shows PIP,
the home activity is activated, the menu is shown above the shelf
and PIP gets killed.

To be consistent with other PIP, we just expand PIP in this case
as well.

the PIP got expanded properly.

Bug: b/129367828
Test: Got Netflix into PIP, pressed the shelf icon and confirmed
Change-Id: Id677e8224ba55cd9863cd450b7551af5e2d35d82
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1675484
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>
Commit-Queue: Kazuki Takise <takise@chromium.org>
Auto-Submit: Kazuki Takise <takise@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672395}