media/gpu/test: Fix decoder not being destroyed in new video decoder tests.
The new video_decode_accelerator_tests built on the video player test framework
are designed to test VideoDecoder-based video decoders. To allow the tests to
work against our VDA-based video decoders the TestVDAVideoDecoder wrapper is
used. The VideoDecoder interface expects a decoder to destroy itself in the
Destroy() function. This was not properly done in the wrapper causing resource
exhaustion and crashes in some tests.

TEST=ran new VD tests on grunt

BUG=978714

Change-Id: I20f2aa1ed229fc3583892d7f990c722c01ad56ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1677770
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672400}