viz/gpu/media: Plumb VaapiJpegDecodeAcceleratorWorker to GpuServiceImpl.
This CL modifies GpuServiceImpl to create and attempt to initialize a
VaapiJpegDecodeAcceleratorWorker object to be used for accelerated JPEG
decodes when the USE_VAAPI build flag is used. If the creation succeeds,
the client is informed of the decoder's capabilities through the GPUInfo
structure.

The VaapiImageDecoder and the VaapiJpegDecodeAcceleratorWorker must be
modified to compute the decoder capabilities and convey them through a
new method in the ImageDecodeAcceleratorWorker interface,
GetSupportedProfiles().

Additionally, we remove the obsolete check in ImageDecodeAcceleratorStub
that restricts decode requests that need mipmap chain generation. This
is supported already.

Bug: 868400
Test: None.
Change-Id: I127013f7aa3ff4f02387ed1ceebfbed2b371d20a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1518234
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Antoine Labour <piman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672403}