ime: Add error handling for SetCompositionRange.
We need to reject invalid and empty ranges.

For invalid ranges, we can reject them in InputMethodChromeOS, which
would handle it for all TextInputClients.

For empty ranges however, only the individual TextInputClients can
handle them, as they are a valid call. Currently setting an empty
composition range in views::Textfield crashes, as Textfield uses the
InvalidRange to represent no composition, rather than an empty range.

Hence, we change views::Textfield::SetCompositionFromExistingText to
clear the composition underline when the provided range is empty.

Note that InputMethodChromeOS also has a ClearComposition, but it
is different to the one in Textfield: the former will actually delete
any composing text, whereas the latter only deletes the underline.
This is why we can't clear the composition underline from
InputMethodChromeOS.

Bug: 976585
Change-Id: I42c6f26a1a8aa40560017ccdf9834096fd8de6b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1668711
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Auto-Submit: Darren Shen <shend@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672418}