Show a message if sync is paused due to clear cookies on exit.
This CL implements a mockup that shows a message in the profile chooser view
explaining that sync is paused due to cookie settings set to clear cookies on
exit. On click on 'your settings' within the message, the user gets directed to
the cookie settings page. This view is only visible if sync is paused and
cookie settings is set to 'clear cookies on exit'.

Bug: 942456
Change-Id: Ia353fe6e5750b56a61d8771c09bea651a267810f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1664081
Reviewed-by: Thomas Tangl <tangltom@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Dan Beam <dbeam@chromium.org>
Commit-Queue: Monica Basta <msalama@google.com>
Cr-Commit-Position: refs/heads/master@{#672456}