Replace WebMediaStreamTrack with MediaStreamComponent in Audio/VideoTrackRecorder
This is a part of moving media recorder implementation to blink and
a follow-up CL of https://crrev.com/c/1673406.
This CL replaces WebMediaStreamTrack with MediaStreamComponent in
AudioTrackRecorder and VideoTrackRecorder and calls directly
AddSink and RemoveSink of track instead of passing through
WebMediaStreamAudioSink or WebMediaStreamVideoSink.

Bug: 960665
Change-Id: I3df392a26311a4f10388c63ba88c6c8397b74677
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1676204
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Miyoung Shin <myid.shin@igalia.com>
Cr-Commit-Position: refs/heads/master@{#672509}