Fall back to main thread for non-compositable paint worklet
The worklet thread currently does not support all data types for the
native and custom properties in CSSPaintValue::GetImage.  To resolve
this, we will fall back to using the main thread for any CSSPaintValue
that contains an unsupported property type.  This fall back feature
makes use of the extra set of global scopes created on the main thread
when CSSPaint is running off main thread.

A unit test was also added for this functionality.

Bug: 948761
Change-Id: I65dea64b1d6c4f6e8a5ea3319953bcf8ebe68fcd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1667751
Commit-Queue: Adam Raine <asraine@google.com>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672554}