[CrOS Network] Fix "view password" icon dimensions.
Before this CL, the icon had the correct height but incorrect width.
This was due to the fact that the <cr-input> to its left was taking up
too much room.

The fix was to change "width: 100%" to "flex: 1" on the input element.

Bug: 977117
Change-Id: I27063c67b039fc4547615d2e9c9382e5e795249a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1677552
Auto-Submit: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672560}