DL: Ensure SimplifiedLayout handles DL elements without processing children
This patch ensures that when we're doing a simplified layout, we
consider that children might be dirty due to being blocked by a display
lock. We skip layout out the children, and also loosen a DCHECK to allow
for display locked elements to enter the simplified layout code path.

R=cbiesinger@chromium.org, chrishtr@chromium.org, ikilpatrick@chromium.org

Bug: 978166
Change-Id: I4c79c49c0dd470e06b4cbd5091b5d8125d1cf4a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1675228
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672568}