[CrOS Network] Fix crash when opening Internet detail dialog.
A previous CL [1] used a newly-added string as part of the
<network-proxy-exclusions> element. This element is shared between both
the settings page and the Internet detail dialog.

However, the newly-added string was only added to the WebUIDataSource
belonging to settings and was not added to the dialog's. This CL fixes
this crash by adding the string to the shared
chromeos::network_element::AddDetailsLocalizedStrings() function
instead.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1669183

Bug: 978762
Change-Id: Ibc5f06d6f3db71b06e55c8c96e717fb08c4d42ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1679099
Auto-Submit: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672701}