Use long for Android APK version codes.
Android supports 64-bit version codes for APKs in P and later. Use a
long instead of an int when retrieving the version code of the host app
and of Chromium itself in BuildInfo. On older OS versions, we just
promote the existing int versionCode to be a long which is the same
semantics that the platform uses for existing APKs.

All the places that currently consume these versionCodes just convert it
to a string anyway, so will not be impacted.

Bug: 967775
Change-Id: I00d8616d6a6974244823a4ed59122a96607dfea8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1678800
Commit-Queue: Richard Coles <torne@chromium.org>
Reviewed-by: Changwan Ryu <changwan@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Auto-Submit: Richard Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672948}