Unify accuracy of cubic bezier calculations in cc and blink.
Currently, there is an inconsistency between blink and cc for the
accuracy of bezier curve calculations. Short animations could use a
coarser resolution than gfx::kBezierEpsilon with Blink.  In a recent
patch (https://chromium-review.googlesource.com/c/chromium/src/+/1643973),
the accuracy of solver was improved to resolve WPT failures. The
coarser accuracy for Blink is only used in the event of falling back
to the bisect method. With the improvement to the initial guess
fed into the Newton-Raphson solver, the fallback is limited to
rare cases. Removing the accuracy parameter from the Evaluate calls
brings the API in line with the spec, and makes composited and main thread
animations consistent.

Bug: 577016
Change-Id: Ibd01f41b82da189d0fd1b722c5eba39a9a28e7d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1674446
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672983}