Avoid caching outdated resolution state in LocalFontFaceSource
When there are chained local('unresolvable font'), local('Roboto')
sources listed for a @font-face, and font scanning completes, so that
the first LocalFontFaceSource receives the
NotifyFontUniqueNameLookupReady() callback, CSSFontFace::Load() is
triggered, which will try to load the next source, since the first one
is unresolvable.

Previously, then the second LocalFontFaceSource would have its
was_resolved_ initialize to false,because when it was created, the font
table was not available.  But because now the font unique name lookup
table has become available, the second LocalFontFaceSource is no longer
in IsLoading() state, but it is in IsLoaded() state when
CSSFontFace::Load() checks for that. However, its IsValid() would have
returned false, since was_resolved_ was still false, and there was no
way it would have been turned to true.

It turns out we can work without a temporary was_resolved_ state for
determining IsValid(), as the LocalFontFaceSource is either in
IsLoading() state, or we can cheaply ask the FontCache for whether it is
available.

I was unable to come up with a unit test for this, as it would require
controlling when the browser side service signals completion of
scanning, while at the same time verifying which fonts get used on the
page.

This fixes accidentally attempting to load a web font for Roboto Regular
in the browse_media_googleplaystore_2019 perf test, which triggers
console messages which were recorded as perf regression for the
v8.browsing_mobile benchmark.

Bug: 976766
Change-Id: I5b401539e1765f0e87f1412451b8b488d77e807b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1680412
Reviewed-by: Emil A Eklund <eae@chromium.org>
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/master@{#672986}