WebUI: Replace all cr-input mixins with part selectors or variables
This CL removes the rest of the remaining mixins that were used to style
<cr-input>'s inner elements, and moves the styles to exposed CSS parts
or into a CSS variable if enough users of <cr-input> are modifying the
same property. This CL also removes some unused CSS properties that
seemed to have no effect on the component.

Bug: 973674
Change-Id: Id733292ade29eb40a52854e60b019c29b3848964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1671351
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673019}