[AW NS] Allow loads when io_thread_client is null
This is to address recent issues with the network service roll-out.

Align the loading behavior with the old (non-network service) code path,
see go/wv-ns-behavior-differences item 4. When IoThreadClient is null we
currently abort loading with net::ERR_ABORTED, while in the old code
path this is allowed to go through.

This patch makes sure to not issue net::ERR_ABORTED when GetIoThreadClient()
returns null.

BUG=977784,841556

Change-Id: If47486587e2b39ff25c57db797cbb4372187b5db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1675358
Reviewed-by: Tobias Sargeant <tobiasjs@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Commit-Queue: Tim Volodine <timvolodine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673025}