Fix crash during ToolbarActionsBar destruction
Prevents observing changes to the ToolbarActionsModel due to destruction
of members of ToolbarActionsBar that are declared after the
ScopedObserver.

Observing the destruction of |pending_bubble_controller_| is theorized
as the root cause for a crash that calls back into
BrowserActionsContainer while it's being destructed (which is what
triggers ToolbarActionsBar destruction).

Bug: chromium:978367
Change-Id: Ib4372c46776dca1ddb085408b7f80ac027a610f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1680377
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Commit-Queue: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673102}