[Dolphin] Implement a first-order Markov chain predictor.
This is a simple predictor that ranks the next target based on the
previous one. The returned rankings for a given previous target are a
distribution over all targets that have been used after that previous
target.

This is implemented with the ConditionalFrequencyPredictor, which maps
targets to frequencies given some condition. In this case, the
condition is the previous target.

Bug: 921444
Change-Id: Ib4a20857e55103cb4cd7d1878b8d5c22e3de6d6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1677864
Reviewed-by: Jia Meng <jiameng@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673201}