Reset SpatNav binding only on main frame detach
We detach the spatial navigation Mojo binding when a FrameView is
removed so that, for example, navigation to a new page causes us to
create a new binding.

The bug here is that we do this for any FrameView being detached. We
shouldn't do this if an iframe is detached.

Bug: 976892
Change-Id: I52ca672b7868bc6e309c5d0967c882fad53fca5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1672016
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673253}