IDL compiler: Define IdentifierIRMap
In compiling Web IDL, it's a common task to look for an IDL definition
with its identifier.  And in compiling IDLs, the target definition may
change depending on the compilation phase.

IdentifierIRMap is a support tool to make it easy to find the target
definition(s) in complation.


Bug: 839389
Change-Id: I42c4a0cc1730b920eba63418ade7a0bbb185d52f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1680262
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Commit-Queue: Hitoshi Yoshida <peria@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673273}