android: Align the reached code profiler field trials
The ReachedCodeProfiler feature is controlled by the finch experiment but the
feature is in practice enabled with a lag. To track the sessions in which the
profiler is actually enabled, we register a synthetic field trial.

The one disadvantage of this approach is that the synthetic trial was recorded
for 100% of users, making hard the comparison between control and enabled
groups.

This CL modifies recording of the synthetic trial so that it records the
exact finch trial group that was selected in a previous session.

Bug: 916263
Change-Id: I7a7668667861175cdc0b2068a99b1fe619dd7c61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1543398
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673300}