Change all ReusedPasswordType to PasswordType.
Previously, Chrome Safe Browsing code converted PasswordType
from the password manager code to ReusedPasswordType which
we define in the protos. However, the only difference is that
SYNC_PASSWORD=>SIGN_IN_PASSWORD. SIGN_IN_PASSWORD is confusing.
Thus, this CL is changing to use the original PasswordType
from the PasswordManager as it will be a clearer source of truth.
In addition, the SYNC_PASSWORD originally meant a password for
an account that is considered "blessed" for sync. This is
equivalent to the primary account, which is the a currently
Chrome signed in profile that the user has that is
closest to being set to syncing if the user chooses to set it up.

This is only part 1 of the refactoring. The next part is to update
SyncAccountType with the new field from the new proto.

Bug: 914410
Change-Id: I35948bc745e568cb93e5f852cf4ff9a2b4bc1393
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1639717
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Reviewed-by: Nathan Parker <nparker@chromium.org>
Commit-Queue: Bettina Dea <bdea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673400}