Align the texture format between frame buffer and skia wrapped texture
Previous to this CL, it was possible on some platforms for the frame
buffer copy texture (used by CopyTexImage2D) to have one format, and
the skia wrapped texture (used by SkImage::MakeFromTexture) to have
a different format. This lead to intermittent errors in
GLUtils::ValidateCopyTexFormatHelper() for mismatched formats. And
that lead to the backdrop-filter feature flickering on/off when a
re-layout occured. This is now fixed.

Bug: 978858,977264
Change-Id: Ie6bb6d41293092eb860429dccf274502c0b1faef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1681086
Commit-Queue: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Eric Karl <ericrk@chromium.org>
Auto-Submit: Mason Freed <masonfreed@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673442}