Set WebappManifestScope from WebappActivity
Chrome determines which keys to make cancellable or not based on the
webapp manifest scope. This is currently set from the WebApkActivity, a
subclass of WebappActivity. However, a simple page with a manifest added
with "Add to Homescreen" doesn't launch with a WebApkactivity. In this
CL we set the WebappManifestScope from the WebappActivity as well.

This was fixed in https://crrev.com/c/1621415 in M76 by an unrelated
refactoring. However, that CL is too large to merge to M75. In this CL
we add a very simple, very safe, touchless-only change directly on the
branch.

Bug: 978624
Change-Id: Ife330daecba72e1cf8329559e412824fc088bc75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1682565
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Cr-Commit-Position: refs/branch-heads/3770@{#1120}
Cr-Branched-From: a9eee1c7c727ef42a15d86e9fa7b77ff0e63840a-refs/heads/master@{#652427}