viz: Defer BeginFrame throttling until one vsync after max swap limit.
Currently we mark the Gpu busy which throttles BeginFrames immediately
after the swap which hits this limit. This delays the start of the next
vsync until the previous swap is acked in turn delaying the draw for
the next vsync. This unnecessarily reduces pipelining for the next frame
if the swap will be acked before the next frame needs to be drawn.

Avoid the above by allowing one vsync after the max swap limit is hit.

R=sadrul@chromium.org, sunnyps@chromium.org

Bug: 973703
Change-Id: Iec513cd01c4bde98536bc9094ad92e92d95ce7d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1675464
Commit-Queue: Khushal <khushalsagar@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673558}