Puts navigation score normalization behind a finch param.
A navigation score is normalized by the sum of metrics weights,
and the set of navigation scores for a page are normalized from 0 to 100.
The removal of the normalization is temporarily behind a finch param,
until we ship weights and a threshold that do not rely on the normalization.

Bug: 979050
Change-Id: I36afef487a3870476ecc1004bd83a8947de08d42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1682561
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Commit-Queue: Sofiya Semenova <sofiyase@google.com>
Cr-Commit-Position: refs/heads/master@{#673773}