Fix DBus hanging when initializing status icons or notifications
The initial calls to
  service = org.kde.StatusNotifierWatcher
  path=service = org/kde/StatusNotifierWatcher
  interface = org.kde.StatusNotifierWatcher
  method = GetCapabilities
and
  service = org.kde.Notifications
  path=service = org/freedesktop/Notifications
  interface = org.freedesktop.DBus.Properties
  method = Get
may fail when the service is unowned.  In this case, a timeout will have
to expire which blocks the DBus thread.

On my gLinux machine, the timeout is very short (< 1s) and there's no
user-visible hang.  But for some systems (like current Debian Sid), the
timeout is on the order of 10s.  If DBus status icons were unavailable
but X11 status icons were, this causes an issue where the icon wouldn't
be shown for about 20s.  This is because
NotificationPlatformBridgeLinuxImpl gets initialized before the status
icons, so the DBus thread spends 10s timing out if notifications were
not available.  Then it times out for another 10s when making the
initial call to org.kde.StatusNotifierWatcher.

This CL avoids both timeouts by checking with the org.freedesktop.DBus
service (which will always be available) that the services in question
are owned before making method calls to them.

R=thestig
BUG=419673

Change-Id: Ia9c284b8f7e1a2fdb0f8314ef974e4dca6c31063
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1682953
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673873}