media/gpu/vaapi: Remove usage of VaapiPicture in VaapiVideoDecoder.
The buffer creation flow in the new VaapiVideoDecoder is currently based closely
on the VaapiVideoDecodeAccelerator flow. This flow makes use of the VaapiPicture
object, which contains a lot of GL-specific code. As only import mode is
supported in the VaVD and most of the code in VaapiPicture is not used when
importing buffers, we can remove the use of VaapiPicture altogether.

TEST=ran new VD tests on eve

BUG=977935

Change-Id: I296b140c6c1e807633956caf6880d06a794788a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1672081
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673916}