Fix tab hover card bug causing flickering.
Previously a timing buffer was created to suppress the delay before a tab hover card was shown when a user had moused out of the tab for a short amount of time. This was being set whenever a hover card was hidden. This change instead suppresses that time delay only after the hover card was closed from the mouse hover exiting the tab strip.

This change also hides the hover card when tabs are animating.

Bug: 973382, 910739
Change-Id: I6a7b5b3432d01cb3d45e78cd069cd45d2b05895b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1680948
Commit-Queue: Caroline Rising <corising@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#673926}