Add deviceName update to CastSettingsManager
This change includes functions in
CastSettingsManager to update the native
device name, including the JNI interface
to allow the methods to be called from C++
layer.

This is a pre-requisite to enabling
the ability of the ApiSetEurekaInfo
API to accept changes to the
deviceName.

Bug: internal b/133863638
Test:cast_base_junit_tests and manual
on device testing

Change-Id: I7983bc5b63f8d55aa82647c808c7074e4226d0d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1678230
Commit-Queue: Brandon Mousseau <bmoose@google.com>
Reviewed-by: Luke Halliwell <halliwell@chromium.org>
Reviewed-by: Simeon Anfinrud <sanfin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674256}