Add feature flag for use of DISPLAY priority for browser UI & IO
This adds a feature flag BrowserUseDisplayThreadPriority to control
raising the priority of the UI & IO threads in the browser.

Together with the recently added flags GpuUseDisplayThreadPriority and
BlinkCompositorUseDisplayThreadPriority, this allows Linux based builds
to set the same thread priority Chrome OS uses for display threads
(which is not the same as on Android; that boost is more modest).

The Browser UI thread requires special care on several platforms as it
starts with a higher priority by default.

Bug: 937462

Change-Id: I807f9d88db83c63f3c10ab38467db5a86b932659
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1682802
Commit-Queue: Michael Spang <spang@chromium.org>
Reviewed-by: Antoine Labour <piman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674260}