[LCP] Text: name-change from node to object
Since the landing of crrev.com/c/1649467, TextPaintTimingDetector
started to use LayoutObject instead of Node to keep track of a text
object. However, the CL didn't change the variable names and
function names to reflect this change. This CL is to fix this issue.

Bug: 967837

Change-Id: Iae0b0c0ac228b594bf9f61f47c707a060ab6c689
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1681046
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674311}