Add app icon and name to lock screen media controls
The app icon and name of the active media should be displayed in the
upper left corner of the lock screen media controls view. If there is
no provided name or icon, then the defaults should be used. See the
corresponding bug for visual examples.

Bug: 965570, 965569
Change-Id: Ic22bab77f51575f090f7ff1296771888a0def591
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1681043
Commit-Queue: Mia Bergeron <miaber@google.com>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674322}