Reland "Migrate WebDatabaseHostImpl to using the new Mojo types"
This is a reland of 2d150b082938299f6aeb95cc29c91fb5c6d7fecd

Using registry->AddInterface() and a function to do the binding with an
unretained pointer to the RenderProcessHostImpl object is not safe because
such object might be destroyed before the function gets called if its
invocation doesn't happen in the main thread.

To prevent this, use AddUIThreadInterface() to make sure that the binding
function runs on the main thread and never beyond the life time of the
RenderProcessHostImpl object, and then post a task to the sequenced
TaskRunner of the DatabaseTracker from there to create the actual instance
of WebDatabaseHostImpl from there from the correct thread as expected.

Additionally, stop passing a reference to the DatabaseTracker to the
binder function, since we can access to it anyway from the function's
scope when running it.

Original change's description:
> Migrate WebDatabaseHostImpl to using the new Mojo types
>
> This converts the part that interacts with the remote implementation
> of mojom::WebDatabase, by moving to using a mojo::Remote.
>
> Likewise, this patch migrates the part that implements the interface
> mojo::WebDatabaseHost, which relied on mojo::MakeStrongBinding, by
> using mojo::MakeSelfOwnedReceiver instead.
>
> As for the client side that uses the remove mojo::WebDatabaseHost
> interface from Blink, this CL does not migrate those bits since they
> will be taken care of as part of CLs 1669941 and 1675356.
>
> Bug: 955171
> Change-Id: I4da1e5b3b97e0034872f7b748818b15ffe5a55ad
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1678317
> Reviewed-by: Oksana Zhuravlova <oksamyt@chromium.org>
> Reviewed-by: Ken Buchanan <kenrb@chromium.org>
> Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
> Cr-Commit-Position: refs/heads/master@{#673242}

Bug: 955171, 980087
Change-Id: Ied6cdcb8da92e163d00afc44415a8ddf1b5704a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1683720
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Oksana Zhuravlova <oksamyt@chromium.org>
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Cr-Commit-Position: refs/heads/master@{#674362}