[Mfill Android] Fix key icon showing too late
This CL removes a redundant check: since the native rework, we check
early whether the accessory has contents worth showing.
That makes the check in the UI redundant and even wrong: the very first
time a field was focused on a page, the accessory would be allowed to
show because it would be populated a splitsecond later but the check for
whether it's empty would suppress it.

Bug: 978757
Change-Id: I01d6e9706160a2b1d292c05f9d7a59915d5de8a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1685553
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674383}