Does not allow casting to a table related IA2 interface if the object is not table related
This is in order to address a high priority bug affecting the NVDA screen reader.

In many cases, our implementation of QueryService and QueryInterface casts to requested interfaces without checking
if the operation is allowed on that object.
This is being addressed in the following patch:
https://chromium-review.googlesource.com/c/chromium/src/+/1584824
However, QueryService and QueryInterface need to return the same result if called on the same object multiple times.
This is not guaranteed in our current implementation because accessible objects can sometimes switch their roles without being destroyed and re-created.
Ideally, if an object is destroyed, any QueryService or QueryInterface calls on that object should continue to function as before,
but any attempts to use the APIs exposed through any returned interfaces should fail via an HRESULT failure code.

Until the above issue concerning object roles is resolved, this patch introduces a partial implementation
of InternalQueryInterface which prevents casting to table related interfaces on non table related objects.
If a role change happens on a table related object, it will be destroyed and re-created in Blink
so the above issue should not affect table-like objects.

Change-Id: I599de935075793768f73d2a9ef60732a76de2dc5
Bug: 974062
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1685497
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Auto-Submit: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674550}