Close the new tab after launching it to an external protocol handler
If a new tab is opened to a non-web protocol URL, close that tab if the
protocol is launched to avoid leaving an orphaned tab behind. Matching
the behavior for tabs spawned to file downloads, this cleanup avoids
leaving orphaned tabs around.

Original credit for fix to @ericlaw, this CL simply adds a test.

Bug: 223334
Change-Id: I96b40068f0f01a52f0219fab0be01911c904d9ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1683507
Commit-Queue: Callum May <camay@microsoft.com>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674585}