[LargestContentfulPaint] Allow cached_image to be null
|cached_image| can be null on OnLargestContentfulPaintUpdated in a
scenario like this:
1. OnLargestImageUpdated() is called due to an image content.
2. OnLargestTextUpdated() is called to update on some text content.
3. Some image and text content is removed, so now the remaining text is
smaller than the removed largest image.
4. Garbage Collection occurs.
5. OnLargestTextUpdated() is called. We update the largest text, but we
have not updated the image removal, so we |cached_image| is null. Since
the new largest text is smaller than the removed largest image, we try
to use the null variable.

I think the better solution would be to sync image and text updates so
that the LCP calculator does not receive them out of sync, which can
cause problems like this. For now, we fix this by removing the DCHECK.

Bug: 980662
Change-Id: I037b4f92a4517e05ffe4c57292f48f2de089052a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1686213
Reviewed-by: Liquan (Max) Gu <maxlg@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674596}