vk: Rename keyboard::KeyboardController to KeyboardUIController.
We currently have four "keyboard controllers":

- keyboard::KeyboardController: Controls the keyboard.
- ash::KeyboardController: An interface for controlling the keyboard.
- ash::KeyboardControllerImpl: The ash impl of the interface, wraps
    keyboard::KeyboardController.
- ash::VirtualKeyboardController: Monitors hardware changes and decides
    when to enable / disable the keyboard.

Obviously this is very confusing, so we want to do some refactoring so
that we end up with something where KeyboardControllerImpl has several
sub-controllers that is only responsible for a well-defined area.

This patch changes keyboard::KeyboardController to KeyboardUIController.
There's still non-UI related logic in KeyboardUIController, so we'll
need to move that elsewhere (either to KeyboardControllerImpl or to
a sub-controller).

Bug: 964191
Change-Id: Ia73a16622ed204d65c5575530c1d57d9e9f1904a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1686638
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674625}