fido/bio: use more specific bio callbacks
Bio-enrollment callbacks currently return a BioEnrollmentCallback,
which is unnecessary. This CL switches it out, meaning the caller
doesn't have to see/understand a BioEnrollmentResponse. This also
involves adding a new status callback for the enrollment process,
which is tested in the BioEnrollmentHandlerTest class.

Note GetModality/SensorInfo are left out because they aren't used
(and may be deleted in the future).

Bug: 974046
Change-Id: I45b4bfe0b5ab168415595b4b4073b02b0f0e56af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1680370
Commit-Queue: Matthew Webb <noviv@google.com>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#674684}