Fix crash in VotesUploader::SetKnownValueFlag.
SetKnownValueFlag has a very strong assumption that best_matches has
known_username as a key. As a result it crashes. It should be more
defensive and do nothing when a key is missing (nothing bad would happen
if this function would do nothing, since it adds additional info into
votes, and it's ok when it would add skip some votes).


Bug: 974900
Change-Id: I3ac35f64d06540bc1c6d9e872935a376260c4bb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1687811
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Vadym Doroshenko <dvadym@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674755}