Move org name into certificate button subtitle
This change moves the org name into the certificate button subtitle. To
do this and fix layout issues, this also adds a new PageInfoHoverButton
implementation that uses the hover behavior from HoverButton but sets
up its own GridLayout and contents to avoid layout issues stemming
from the HoverButton inheritance chain. This updates the certificate
button, cookie button, and site settings link button to use the new
PageInfoHoverButton view.

The change to the new PageInfoHoverButton also restores the tooltips
specified for each hover button in Page Info. These had previously been
broken during the transition to Harmony styling such that they never
displayed for the Cookies and Site Settings buttons.

This change also makes it harder to test for the presence of the org name
sub-view, so this removes the expectation from the browsertest case
(unfortunately), but it is still exercised by the ShowUI() code.

Screenshots:
- Single line org name https://screenshot.googleplex.com/ynwcgisiXgO
- Multi-line org name https://screenshot.googleplex.com/ncrxSJhQxOo

Bug: 969734
Change-Id: I73e3716c5bc49b3d724f5ed00aea39fcd1a93435
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1686172
Commit-Queue: Christopher Thompson <cthomp@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#674898}