Enable overscroll for the file grid view
The feedback panels can overlap content in the main FilesApp
grid view.

Modifying the minimum height of the spacer at the end of the
grid view allows the user to scroll the view past the end of the
files grid to allow view of the contents while a feedback panel
is being displayed.

Bug: 932274
Change-Id: Ib90207ed024b291c4381b2dcd0e380702a1af564
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1690403
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675168}