Allow DeviceAccountsProvider to return hosted_domain
Before signin-in Chrome on iOS forces a fetch of the soon-to-be
primary account's hosted domain from the ChromeIdentityService.

According to the design [1], the sign-in workflow will be changed
to reload all accounts from the system instead of seeding the
soon-to-be primary account. In order to avoid breaking anything,
this needs to be able to seed the hosted domain for this account.

In order to perform this, the DeviceAccountsProvider is extended
to allow returning this information via its AccountInfo struct,
and ProfileOAuth2TokenServiceIOSDelegate is changed to use the
value if present (via the override of SeedAccountInfo -- if the
value is not set it is ignored internally).

Extend ChromeIdentity class to also expose the information via
a property (the value may be nil if not available yet).

[1]: https://docs.google.com/document/d/1_NknywZB7UCQdCiCxEPQU85lXouid_aMxdzujEJSiEM/preview

Bug: 957887
Change-Id: I567b9c029b6123469e312df11c884da91fe00636
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1687254
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675191}