Clamp queries to FontCache::GetFontPlatformData to a font size limit
When querying the font cache with a font size large enough so that
multiplied by a 100 it would reach above UINT_MAX, this leads to
computing HashTable entries that are equal to the sentinel values for
empty and deleted items in the HashTable. This in turns leads to
segmentation faults when writing a hash table entry to a nullptr
address.

As a mitigation, when we have incoming FontDescriptions with too large
sizes, as far as FontCache lookups go treat them as if their size was
smaller. Plus, for FontCacheKey, used for the FontPlatformDataCache
HashMap, use more space for the sentinel value.

Bug: 969402
Change-Id: I3a5fcd4ad003a24256f21e0f325856c8578af42e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1688938
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675240}