arc: Add power.mojom method to signal power_supply update.
On older kernels, the ARC implementation relied on udev events
to be fired for this type of event. Newer kernels made the udev
events namespace-aware, and the containerized ARC framework
cannot anymore listen them. We need to this explicit event to
recover the feature to update power state in ARC.

BUG=b:130842405
TEST=ArcPowerBridgeTest.PowerSupplyInfoChanged

Change-Id: Iafa2d9e59079fbec455b453019cdf2a27d44d520
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1682677
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Kazuhiro Inaba <kinaba@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675281}