Bump expiry Enterprise.ExistingInstallAttributesLock
Revive the UMA histogram Enterprise.ExistingInstallAttributesLock
to continue collecting the stats about the enterprise enrollment
flow when the existing install attributes are used (which means
that the previous enrollment was lost, but the device was not
powerwashed before the new enrollment was performed).

It was assumed before that this flow is likely to be broken and
not officially supported, but later additional investigation was
done, showing that it still makes sense to support this flow, and
the test coverage was improved (in https://crrev.com/c/1628726).

Bug: 871508
Change-Id: I58d602ea4fbd4d4c46a3e08cf267a52f18fc4a80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1684342
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675339}