Refactors copying of data to child processes
The same sequence of calls was used in two places and is now
provided by a function in win_utils.

This may change reporting of the histogram referenced by
https://cs.chromium.org/chromium/src/sandbox/win/src/sandbox_types.h?l=29
as SBOX_ERROR_NO_SPACE can no longer be reported by
interceptions.cc.

sbox_unittests do not exercise the previous or changed code
but I have manually verified that it is working.

Also updates a comment in target_process.h which triggered this.

The function is sometimes used to copy quite large variables
into child processes and is not limited to 32bits.

I've checked all uses of this function and they expect
it to work as indicated by the updated comment.

Change-Id: Ia158b89054f8d2e1a05e5400cddf08108755e949
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1684504
Commit-Queue: Alex Gough <ajgo@chromium.org>
Auto-Submit: Alex Gough <ajgo@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675340}