Add histogram metric of tabs 'scrubbed' by mouse or key press per given interval of time.
This change considers 'scrubbed' to mean a tab that is activated for <= 1.5 seconds before the user navigates away. We are making the assumption that when a tab is 'scrubbed' it is likely that the user did not navigate to that tab to use it. The goal is to see this metric go down with use of the tab hover cards feature. This would show that tab hover cards help users find the tab they are looking for without having to activate as many tabs as they would have without the feature.

(cherry picked from commit 46890485c259c623e465e9c53e87a80b0586da5d)

Bug: 910739, 981911
Change-Id: Ia00b0e5ec4ce0969bd8807ac1564fceed0682651
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1636346
Commit-Queue: Caroline Rising <corising@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Reviewed-by: Nik Bhagat <nikunjb@chromium.org>
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#666499}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1690682
Reviewed-by: Caroline Rising <corising@chromium.org>
Cr-Commit-Position: refs/branch-heads/3809@{#764}
Cr-Branched-From: d82dec1a818f378c464ba307ddd9c92133eac355-refs/heads/master@{#665002}