Merge local and server sharing prefs.
We need to properly resolve local and server updates to the sharing
prefs. The vapid key with the oldest timestamp value wins. For devices
we assume that only one device will ever update its own entry in the
dictionary and pick the values with the highest timestamp.

Bug: 968195
Change-Id: I67971357b69cde6e77da621458dde213d3ce9fa5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1687232
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Michael van Ouwerkerk <mvanouwerkerk@chromium.org>
Commit-Queue: Richard Knoll <knollr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675356}