[FractionalScroll] Use ScrollableArea offset for commit
After a commit, Blink sets the scroll offset in the main thread property
trees. When we push the main thread scroll offsets to the compositor
thread, its the values in the property tree that are used.

Currently, Blink uses LayoutBox::ScrolledContentOffset, which explicitly
uses the floored scroll delta. This means that when fractional scroll
offsets are enabled, we continue to use a floored offset here. The
result is that the fractional parts of the delta sent from the
compositor thread aren't reflected back. This causes the compositor to
send more delta (since it normally subtracts the reflected delta from
the current active delta) causing inaccurate scrolling.

This CL makes the relevant Blink code use the ScrollableArea's
GetScrollOffset method which floors the offset only if the
FractionalScrollOffsets feature is disabled. It fixes several web tests
broken when the REF is enabled.

TEST=virtual/threaded/synthetic_gestures/animated-wheel-tiny-delta.html,
     virtual/threaded/fast/scroll-behavior/main-frame-scroll.html

Bug: 414283
Change-Id: I98160d759799ff7ad0adaa97ecf56ae3b0a47edc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1690292
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675360}