Ensure RunningChromeVersion symlink be created
This symlink is used to indicate to app shims which version of the
Chrome framework should be loaded. In the event that mojo-on-mach was
used, the function to create the symlink would not be created due to
an early-out.

Move the symlink to above mojo-on-mach initialization.

Bug: 980717
Change-Id: I40b7af08a164ea5706820cf144e1f20bc42e5b14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1686618
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/branch-heads/3809@{#772}
Cr-Branched-From: d82dec1a818f378c464ba307ddd9c92133eac355-refs/heads/master@{#665002}