Print Preview: Do not focus search box until signin succeeds
The search box should be focused only when the user account is set. If
the tab is just closed, the "sign in" link should remain focused.

Also adding a test, to validate this behavior and the previous change
to focus the search box after sign in.

Bug: 979603
Change-Id: I7bd3ffe8218c5feb310d740f028523df2929929f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1682331
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675486}