Always guard BuildBaseTextCodecMaps by mutex
Before this CL, BuildBaseTextCodecMaps() assumed that it is called
on the main thread. This assumption isn't always correct because a
service worker thread could exist without Document and a service
worker may construct a TextEncoding when it fetches the toplevel
script on the worker thread [1]. To make BuildBaseTextCodecMaps()
thread-safe, guard the function by a mutex.

[1]
https://cs.chromium.org/chromium/src/third_party/blink/renderer/core/workers/worker_classic_script_loader.cc?l=229&rcl=b9eca7bc28c7a78ef63e449f72fa91a2b82f8876

Bug: 967272
Change-Id: I327871257182ab90b3b23da085d0633026cbb347
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1691601
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675501}