Reduces layering in the implementation of BadgeService.
This merges the implementation of the BadgeService
into the BadgeManager. BadgeManager is a KeyedService,
of which there is only ever one instance. This is
modelled on NativeFileSystemManagerImpl. The renderer
requesting that a badge be set is tracked through a
mojo::BindingSet, which allows the requesting
RenderFrameHost (and thus, app) to be looked up.

Bug: 981447
Change-Id: I97fbc66a5c7ca8cac203ff860ac8d73302e3fb31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1690570
Commit-Queue: Jay Harris <harrisjay@chromium.org>
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675505}