Fix wrench menu disappearing under certain conditions
Changing the call from end -> cancel causes this animation to complete.
Note that this was how it was implemented previously, but I canged it
in: https://chromium-review.googlesource.com/c/chromium/src/+/1631947.
It seems that the accessibility issue that the change above fixed is
still fixed when "reverting" my change.

Bug: 968956
Change-Id: I47951bb793b1eebf7f6086eacdc1e151f9cbcd37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1691703
Reviewed-by: Theresa <twellington@chromium.org>
Commit-Queue: Brandon Wylie <wylieb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#675906}