[quarantine] Sanitize persisted quarantine URLs
The quarantine component attaches metadata to files downloaded on
Windows and macOS indicating their source. This is done so that the OS
and AV products can factor the source of the file into evaluating
whether the file is safe.

Previously the source URL was presisted in its entirety. This change
introduces a sanitization step that removes unnecessary information from
the URLs being persisted. See bug for details on what's being removed
and why.

R: wfh@chromium.org
Bug: 917842
Change-Id: I7831d1486b73a7e2858004e719c9680ea7a43d36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1407439
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Commit-Queue: Asanka Herath <asanka@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676150}