[LCP] Image: remove the record cap
Largest Image Paint used to have a cap on the map & set, containers for
the recorded nodes/objects/images. The cap was placed in order to
prevent the containers from growing uncontrollably. But actually, if
it grows in proportion to the size of the DOM tree, its size should
be deemed under controlled. Based on this idea, we should remove the
cap of containers in LIP.

This CL also fixes the issue that the |invisible_node_ids_| set didn't
remove node_id. Otherwise, its size would be grow uncontrollably.

Related CL:
crrev.com/c/1685325: the text counterpart

Bug: 966476,967837
Change-Id: I29462450311b337bf635bc3cfd68379286422e6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1692199
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676154}