Add the ability to dump platform accessibility trees for PDF
For HTML accessibility testing, we've found it valuable to dump platform
accessibility trees and check them against expectation files. This CL
brings that capability to PDF by integrating the tree formatter into
PdfExtensionTest. Further tuning of the property filters for PDF may be
required in the future, but this gives us a starting point.

This CL also adds functions to browser_test_utils to find a node in
the accessibility tree by name and/or role. For PDF this is used to
scope the dumps to the PDF itself, so that changes to the host HTML do
not impact the tests. There are other examples of these find functions
in content_browsertests today, but none are suitable for export.
http://crbug.com/980391 is for future work to consolidate these
functions.

Bug: 974397
Change-Id: I0cfd2a7b456550a94a3587f36c0ee6f5d8dc4821
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1674498
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676179}