Add Viz timestamps to FrameTimingDetails
This CL starts adding more timestamps to the new FrameTimingDetails that
is sent back to Viz clients after a frame is presented. Eventually these
timings will be used to determine abnormally long stages of the
rendering pipeline in the case of missed frames.  This CL simply adds
the timestamp of when the display compositor receives a submitted
compositor frame and when it begins the DrawAndSwap.

Bug: 963780
Change-Id: I3ca4715689401d0544c7efb17a9a096e2534131d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1655738
Commit-Queue: Nathan Zabriskie <nazabris@microsoft.com>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676207}