[DeviceSync v2] Support devices without decrypted metadata
In v2 DeviceSync, CryptAuth will not return the encrypted group private
key in a SyncMetadata response if other devices in the group have not
uploaded the encrypted key yet, for instance if they are offline when a
new device joins the group. And, devices without the group private key
cannot decrypt device metadata sent in the SyncMetadata response. This
metadata contains a public key used for multidevice (BetterTogether)
features and a PII-free device name at the moment.

In this CL, we allow devices to be stored without decrypted metadata.
This is necessary to preserve the existing multidevice setup flow after
migrating to v2 DeviceSync; only device names and supported features are
needed to preserve the flow. See go/cros-devicesync-v2-multidevice.

Bug: 951969
Change-Id: Id2986f866be5a86ea0b20afe5fffc1b9693c070c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1695963
Commit-Queue: Josh Nohle <nohle@chromium.org>
Auto-Submit: Josh Nohle <nohle@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676220}