[filelist] Block invalid FileList thumbnail loader requests early
ImageLoaderClient is throwing DOM exceptions while processing requests
{ImageLoaderRequest} that have no request URL. Odd, the request URL is
required (it keys the ImageLoader caches). Tracked the source of these
“no-url” requests down to the FileList thumbnail loader.

The FileList thumbnail loader sends ImageLoaderClient requests for the
file list entries while the user is scrolling the FileList. So it is a
good idea to avoid doing unnecessary work therein (it’s hot code).

Change the FileList thumbnail loader to setup the URL first and reject
early with an ImageLoaderResponse error if there is no URL to request.
Enjoy the faster scrolling, CPU and power savings, that result.

Bug: 965370
Change-Id: I5dd9a1390bc1f92010aea1b4ca2e8a50b67177df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1695661
Reviewed-by: Alex Danilo <adanilo@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676225}