Add a margin to the "Open with" context menu item
The "Open with" context menu item can contain an icon displayed on the
right. This icon is implemented as a background image, so the text can
overlap the icon if it is too long. Fix this by adding a margin to the
end of the element when we show an icon.

Bug: 958719
Change-Id: Ic8108e8300cdd1c584d483ce442e2a6362814bae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1697001
Reviewed-by: Noel Gordon <noel@chromium.org>
Reviewed-by: Alex Danilo <adanilo@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676320}