[tools/perf] Remove regex to ignore stats in memory metrics
These were added long time ago as uploading statistics for all memory
metrics in chart json format overwhelmed the dashboard.

But these should be a no-op now, as we upload histograms instead.
The code shows that as long as any of the stats is uploaded, then
the entire histogram is added [1].

[1]: https://cs.chromium.org/chromium/src/third_party/catapult/telemetry/telemetry/internal/results/page_test_results.py?rcl=d2a5fee6e9409c644ea111f2a7d89e03c0ee1c05&l=438

TBR=etienneb@chromium.org

Bug: 921000
Change-Id: I8c3c2e1e972966cce3a0f660f87710db1b7d64f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1688935
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Ben Hayden <benjhayden@chromium.org>
Reviewed-by: Mythri Alle <mythria@chromium.org>
Reviewed-by: Ulan Degenbaev <ulan@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676349}