Extract remaining platform-dependent parameters from PO2TSBuilder
This is an intermediate step in the process of moving all of
ProfileOAuth2TokenServiceBuilder code into IdentityManagerBuilder. To do
this, the building of PO2TS cannot use Profile not any code from
//chrome/browser.

In this CL, ProfileOAuth2TokenServiceBuilder is modified to access
cookie settings, the web data service and the reauth callback from
parameters provided by IdentityManagerFactory.

Bug: 974198
Change-Id: I9fe95b2f63859e5bcdd8128b362846ed5b212a50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1690843
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Henrique Ferreiro <hferreiro@igalia.com>
Cr-Commit-Position: refs/heads/master@{#676370}