Refactor AutofillAssistantModuleEntry, getting rid of the factory.
This change is meant to be a pure refactoring. It is not meant to change
behavior.

The current design of AutofillAssistantModuleEntry makes them specific
to a WebContents. This assumes that everything needs a tab and
WebContents, and that we should wait if there isn't one, which won't be
true for the DirectActionHandler to be returned by
AutofillAssistantFacade.

This changes makes AutofillAssistantModuleEntry global and stateless,
getting rid of the intermediate factory instance.

This change also extracts the AssistantOnboardingCoordinator that was in
AutofillAssistantModuleEntryImpl, to make it reusable in the
implementation of direct actions and adds a test for the new use
case (onboarding without tabs.)

Bug: b/134741524
Change-Id: Iadf0831cdbf252910de841996b99943c7336036f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1695305
Commit-Queue: Stephane Zermatten <szermatt@chromium.org>
Reviewed-by: Jordan Demeulenaere <jdemeulenaere@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676438}