Fix NPE while accessing AutocompleteCoordinator during deferred startup.
This change resolves a NPE happening during deferred startup on slower devices.
It looks like these devices do not have AutocompleteCoordinator ready in time
to invoke the call to prefetch ZPS suggestions.

The downside of this change is that in these cases prefetch would not happen,
but there is a possibility we do not need to trigger it either; the prefetch
is executed currently whenever user accesses a new tab page, too.

Bug: 982343
Change-Id: Ida84fedcb8f0f057c95979d8571cb9033f2884c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1693337
Commit-Queue: Ender <ender@google.com>
Reviewed-by: Theresa <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676458}