Display mount notification at most once per disk
Certain operations on disks (e.g. format or rename) require a remount of
the device. These cause spurious notifications when the disk is mounted
again. As the disk object in disk mount manager only gets destroyed when
the disk is unplugged, store a flag there to indicate whether or not it
is currently the first time this disk is being mounted, and check that
before displaying the mount notification.

Bug: 521054
Change-Id: I3f28437ca79a9599b5bb2992bb1ac10877f68083
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1687649
Reviewed-by: Anand Mistry <amistry@chromium.org>
Commit-Queue: Austin Tankiang <austinct@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676696}