grit: fix node unittests
These unittests have been broken since they were added, but because
of the way the verification was written, all the checks were skipped
so no one noticed.

Add some explicit checks so we don't rely on the for loop which would
be skipped in the case of empty results.

Bug: 983071
Test: `./grit/test_suite_all.py` passes
Change-Id: I1db23c87c541fcfc27062415b67c5536363604ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1699161
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676707}