media/gpu/test: Skip video decoder thumbnail test when no checksums are present.
On platforms where the frame validator is not supported, the thumbnails tests is
used as a fallback. This test doesn't work on resolution switching streams
however.

Currently no thumbnail checksums are present in the json files accompanying
resolution switching test videos, so we can simply skip the test if no checksums
are present.

TEST=./video_decode_accelerator_tests on nyan_big and

BUG=981715

Change-Id: If9c95f7cd0228c9e1f27edf594265aa45ecb744c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1696459
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676720}