Make completed panels auto-delete
Adds a setTimeout that when fired, deletes the file operation completed
panel. If the user cancels before it fires, it doesn't matter since the
logic inside the display panel checks if it exists anyway.

Timeout is set to 7 seconds, might revisit if it's too long.

Bug: 947388
Change-Id: I561bb910b3e74ecb48fe30e7dcfb551d2639058e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1698366
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676770}