Simplify CWVSyncController unit tests
CWVSyncController wants to test the interaction with the production
workflow of the sign-in which IdentityTestEnvironment cannot properly
model. This is because IdentityTestEnvironment is generic code that put
the IdentityManager in a known state for code that do not care how the
list of accounts is updated.

The only method from IdentityTestEnvironment that is used by the test
is UpdatePersistentErrorOfRefreshTokenForAccount() and it does not
depends on having fake backing the IdentityManager. So change the test
to create a real instance and to invoke the test util method directly.

This removes one usage of IdentityTestEnvironment::ExtraParams (which
will eventually be removed and will allow to ensure IdentityManager is
always backed by fake when using IdentityTestEnvironment).

Also do some cleanup of the test:

-   use getter to fetch the different KeyedServices instead of storing
    pointer to the services as CWVSyncControllerTest member field

-   move initialisation and cleanup to constructor and destructor as
    this is preferred to overloading SetUp() and TearDown()

-   introduce a separate class TestIdentityManagerObserver to count
    the number of times OnRefreshTokenUpdatedForAccount() is called

-   remove unnecessary member field and duplicate initialisation

This also required changing IOSWebViewSigninClient constructor to take
a WebViewBrowserState* instead of the shared url loader factory and
the cookie manager as the getter are creating the objects lazily which
require properly initializing the network stack which is unnecessary
for tests.

Bug: 982259
Change-Id: Ia99bef652386d89f683db82a09c6d23cfe037ce6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1695285
Reviewed-by: John Wu <jzw@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676806}