Remove unused IdentityTestEnvironment::ExtraParams
All the tests that were making use of that feature have been converted
to use the real production workflow as IdentityTestEnvironment is not
designed to allow testing sign-in or sign-out workflow since they are
different on each platform.

Remove the feature as it has no use, thus all the platform behave the
same way when using IdentityTestEnvironment.

Bug: 982259
Change-Id: I3809280d769863b80f12dcea5ac04f83ce855ee8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1695344
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676808}