BackForwardCache: Track presence of WebGL for bfcache metrics
WebGL uses GPU resources. If a page in the back forward cache uses
WebGL, the page would keep the resources on GPU and this might be
problematic for other pages as GPU resources are limited.

Another option would be to cause context lost, but this is still
controversial. Most websites don't prepare for context lost and
increasing possibility of context lost might break websites.

Bug: 981740
Change-Id: Ib04680d9be76258c220d2f0e385eb10a17de567d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1697189
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Hajime Hoshi <hajimehoshi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676829}