[tools/perf] Remove unused "action_name" arg in _OilpanGCTimesBase
None of the subclasses use the argument, so it always gets the default
value of ''. Confusingly this is only passed to the LegacyPageTest
constructor, whose first and only argument is about cache clearing [1]
and has nothing to do with names of actions.

There was possibly an API change at some point and this only works
because of a happy accident.

[1]: https://cs.chromium.org/chromium/src/third_party/catapult/telemetry/telemetry/page/legacy_page_test.py?rcl=1abe66f1c25324dc3e5c2bf84f3d5c5aed105196&l=48

Bug: 921000
Change-Id: I6fd512cf410d639f7fbca2d51db0ce801e6f2bef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1699664
Auto-Submit: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>
Cr-Commit-Position: refs/heads/master@{#676929}