Remove "POINTER_FLAG_INRANGE" flag when filtering hovering touch moves
When we filter the touch move messages which are not in contact with
the digitizer surface with a flag "POINTER_FLAG_INCONTACT", we should
not check for "POINTER_FLAG_INRANGE", because they are not in the
detection range of the screen of the digitizer. This is based on the
messages we get from the smart board touch device.

TBR=lanwei@chromium.org

(cherry picked from commit e522d2c476e59c08dde531f91b4a428928f1b895)

Bug: 874948
Change-Id: I06bfd9a4048aa9800c59d974548bc8d82375b8ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1691150
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#675632}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1700287
Reviewed-by: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/branch-heads/3809@{#830}
Cr-Branched-From: d82dec1a818f378c464ba307ddd9c92133eac355-refs/heads/master@{#665002}