[Touchless] End NavigationRecorder on external navigations.
At some point the implementation of launchNtp() that is called from the
context menu changed. It stopped replacing the tab, and thus no longer
called onDestroy(). It seems to be coming in as an intent, which we
can pick up easily by adding FROM_API to our mask.

Bug: 983628
Change-Id: I686f4b6ce48cf25c6dec8bc1a4db83708e7e3ed2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1701956
Reviewed-by: Theresa <twellington@chromium.org>
Commit-Queue: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677444}