[mojo] Introduce connection groups
Introduces a new "connection group" concept to C++ bindings. A
ConnectionGroup object is a lightweight bit of ref-counted state to
which Receivers and PendingReceivers can hold an opaque reference.

If a PendingReceiver holds a reference, any receiver which binds it
retains the reference. If a Receiver holds a reference, any new
PendingReceiver it dispatches in an inbound message will automatically
get its own copied reference.

The creator of a ConnectionGroup can observe whenever the group's
reference count is decremented to zero.

ConnectionGroup thus provides a sort of viral reference count of
actively bound Receivers stemming from some root Receiver. This is
intended to serve as a generic and completely passive replacement for
the ServiceKeepalive mechanism provided by Service Manager client
library, in preparation for the removal of Service Manager and the
general simplification of service definitions.

Support for connection groups covers both serialized and unserialized
(i.e. lazy s11n) messages, as well as new and old bindings endpoint
types.

Motivation for this feature is covered by
https://docs.google.com/document/d/1wFKCupPftQUH4OAbLHHYcDJkjWX5a2fi9Kk7c3lnCs8/edit

Bug: 977637
Change-Id: I66704dd0e0d08e2652c4a49e8ad7fab64286eafa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1699212
Reviewed-by: Oksana Zhuravlova <oksamyt@chromium.org>
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#677515}