Worker: Set "mode" for browser-initiated worker script requests
This CL sets ResourceRequest::mode to kSameOrigin for browser-initiated worker
script requests (also known as Plz{Dedicated,Shared}Worker) as the HTML spec
requires (see spec comments in the patch).

This changes test expectation of worker-redirect.html. Before this change,
cross-origin redirect is denied in a worker's renderer process after redirect
chains are handled in the browser process (see
WorkerScriptLoader::FollowRedirect()), and BaseFetchContext prints the console
message. After this change, cross-origin redirect is denied while redirect
chains are handled in the browser process, so the console message is not
printed in the renderer process.

Bug: 983389
Change-Id: Ia22a1619b0661fe8ad778dc00b80df030f2fbd55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1699881
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677628}