Switch WebMediaPlayerMSCompositor to RefCountedThreadSafe + Traits
... fromRefCountedDeleteOnSequence.

As a preparation step to Onion soup'ing webmediaplayer_ms_compositor.cc
et al, this CL switches WebMediaPlayerMSCompositor ref counting model
from deriving from RefCountedDeleteOnSequence to RefCountedThreadSafe + Traits.

Among the reasons, it is the fact that base::RefCountedThreadSafe itself
is allowed** in Blink, and it does support the requirement of objects that
need to be created and destructed on different threads, through the use of Traits.

** It is its Blink counterpart, WTF::ThreadSafeRefCounted, that is allowed.

Also quoting haraken@ from [1]:

"
I think that a more important point is that this is introducing a concept of
sequence to Blink. Sequences are not yet widely used in Blink (this needs
some architectural discussion), so I'd prefer avoiding it if possible.
"

[1] https://crrev.com/c/1698346/3#message-485babfda60c5e60c025a3c18934ab1789459304

BUG=704136
R=guidou@chromium.org, haraken@chromium.org

Change-Id: Ic3bd79fce2756bca2c621801cf5a2c76564a8e0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1698128
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Auto-Submit: Antonio Gomes <tonikitoo@igalia.com>
Commit-Queue: Antonio Gomes <tonikitoo@igalia.com>
Cr-Commit-Position: refs/heads/master@{#677635}