DL: Check for lock in RebuildLayoutTree
It's possible to get into RebuildLayoutTree with an element in a
locked subtree, because it's possible in some cases to style recalc
elements in a locked subtree (because style dirtiness propagates
through the DOM tree for now) + mark those elements for reattachments.

If we ever encounter a locked element (that doesn't need reattach but
has a descendant that needs reattachment) we stop the rebuild there
and propagate the dirtiness later.

Bug: 971901
Change-Id: Ie64a8d2a897b5d99995fe559095a0f35a1ae958f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1687210
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677639}