service worker: Refine the UI thread interceptor.
Before this CL we posted a task to the IO thread to make the inner
interceptor, hopped back to the UI thread to save it, then posted a task to
the IO thread to start it.

This CL posts a task to the IO thread to create and start it, and keeps
it in ServiceWorkerNavigationHandleCore. This eliminates a thread hop.

Bug: 824858
Change-Id: If97fdeb6ef655f4a7203502a1323b2fdc8faec60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1699352
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677672}