Flattened SMILTimeContainer::AnimationsMap
The HashMap that was SMILTimeContainer::AnimationsMap was
changed to only be one level deep. This should save some memory
but potentially increase the number of collisions. The code is
also easier to follow and a lot of lines have been simplified out
thanks to this change.

This is a refactor change so nothing should have changed.

Bug: 981522
Change-Id: If4d89ce423d6149c3e0ddc86e5818490bec37187
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1700059
Commit-Queue: Edvard Thörnros <edvardt@opera.com>
Auto-Submit: Edvard Thörnros <edvardt@opera.com>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#677757}