SplitSettings: Side navigation style to spec
Use an internal <div> to represent the rounded blue or grey background,
such that the "margin" around the div is still the clickable <a> tag.

See bug for spec and screenshots.

The "Advanced" button spec isn't finalized, so keep it the same as now.

Bug: 967923
Test: manual
Change-Id: I73d863cd774c38ca0c528d0a3748d3487b62fd8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1702788
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Michael Giuffrida <michaelpg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677852}