Fix bug in Widget::Show() so that widgets cannot activate the window when CanActivate() is false.
CanActivate() was already being checked for cases when showing a widget and non_client_view_ did not exist but this check was missing otherwise. This caused widgets to activate the window regardless of CanActivate() when Show() was called after the first time the window was shown. This change preserves CanActivate state for multiple widget Show() calls.

Bug: 910739
Change-Id: I95defcf00ad2c2c17ef387c4de31d3c19aa3e02b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1499953
Reviewed-by: Michael Wasserman <msw@chromium.org>
Commit-Queue: Caroline Rising <corising@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677888}