Add Pinpoint as trybot scheduler
Adds the Pinpoint service account as a scheduler for the tryserver ACL
set. This is so that it can trigger builds for bisection. This is
currently only expected to actually schedule builds on
gpu-fyi-try-win-xr-builder, but other bisection builders are likely to
be added in the future.

Bug: 977735
Change-Id: If63bd750e931d9002f06a43080ef9aad34fb0e45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1704322
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677896}