Fix an issue that download will fail after restarting on Q
if a download is restarted during resumption, the content URI will
be preserved in target path. However, because of the restart, the
DownloadItemImpl::ReleaseDownloadFile() has been called and the
content URI is already deleted.
As a result, writing to the content URI will always fail.
This CL fixes the issue by regenerating a new content URI if the
previous one has been destroyed.

BUG=984169

Change-Id: I2a0611f7061a5a373492cfc18a3a09935544e70e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1702180
Commit-Queue: Min Qin <qinmin@chromium.org>
Reviewed-by: Xing Liu <xingliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677913}