Fix HoverButton to active on mouse release rather than mouse press.
Currently, the code path does not get to
HoverButton::IsTriggerableEventType since HoverButton has a
MenuButtonController which has an implementation of
IsTriggerableEventType.

This fixes the bug by adding an additional condition for
IsTriggerableEventType in the controller.

This CL also moves IsTriggerableEventType into the Controller and
non-virtual.

Bug: 981335
Change-Id: Id8a8cf07d393de31be6ff991be2e95b858194f73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1695649
Commit-Queue: Charlene Yan <cyan@chromium.org>
Reviewed-by: Trent Apted <tapted@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#677943}