webauthn: get dialog manager from top-level WebContents
In |ShowAuthenticatorRequestDialog|, the top-level WebContents is used
to find the dialog manager but, later, the specific WebContents that's
using WebAuthn is used. In the context of a Chrome OS local login flow,
only the top-level WebContents has a dialog manager attached.

Thus, for Chrome OS local logins, the WebAuthn dialog was never resized.
This caused the dialog to permanently take the size of the first sheet.
If a PIN was needed, the extra PIN text-field would not fit in the given
size and would be invisible---it could be tabbed to but not seen.

This shouldn't matter because Google login doesn't use PINs. However,
due to a server-side issue, that WebAuthn setting is not being expressed
and thus users with PINs configured on their security keys would be
unable to use them for local Chrome OS logins.

BUG=983576

(cherry picked from commit 442acf5ae32238cc45d977e1a611f4b36a1c8929)

Change-Id: Ia80fdd0c910948fbef89b38527deaff670cd53df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1700376
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#677139}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1703389
Reviewed-by: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/branch-heads/3809@{#882}
Cr-Branched-From: d82dec1a818f378c464ba307ddd9c92133eac355-refs/heads/master@{#665002}