Optionally block downloads of password protected archives
This CL adds a DownloadCheckResult of BLOCKED_PASSWORD_PROTECTED and
propagates that into the download item's danger type for downloads that
are detected to be password protected when the user preferences indicate
password protected downloads should be blocked. We currently only detect
ZIP/RAR as password protected, but that may be extended in the future.

Future work will add the enterprise policy controlling this feature, and
the UX to reflect a blocked download.

Bug: 982492
Change-Id: I32af438385a6d3d6eca772e7fb048453d71194ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1694146
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678076}