Move font fallback test data, add language information and CJK tests
The font fallback tests previously did not contain information about the
language for which fallback was being requested. Language tags added as
much as possible, with a TODO left for figuring out good language tests
for some historic languages, which however, still receive correct
fallback.

Added tests for covering CJK fallback to ensure that depending on
language, a correct Simplified or Traditional Chinese, Korean or
Japanese font is selected respectively.

Moving the data to a separate file in preparation for using it in
DWriteFontProxyImpl unit tests which also deal with similar fallback
requests.

Bug: 976737
Change-Id: I5ac49403bdd4a8c68309bb1f41e82e979c7119bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1704436
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678205}