Expose editable combobox as a group of controls for accessibility.
- Editable combobox is a combobox group. Inside there is a text field and a button.
Windows correctly allows to focus only the textfield and it's properly announced as a combobox.
Mac is tricky. User selects a group of controls and then can step into. Inside both the text field
and the button are accessible and announced properly.
- The dropdown isn't shown by default on focus anymore. This is not how a standard OS combobox
works. Also it confused the screen reader on Win as the menu was immediately announced skipping the
actual combobox.
- The dropdown can be opened with Arrow Up/Down.

Bug: 972069
Change-Id: I1089f335ccf0870699946f87617fd4533577b149
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1683265
Commit-Queue: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678245}