ui: Introduce GetNumberOfPlanes
We can't determine the number of planes for a buffer based solely on
the format.  Some hardware formats (compressed buffers, for one)
require a plane of metadata to track the state of the buffer
contents. This patch introduces a new NativePixmap method that returns
the the number of planes.

BUG=979736

Change-Id: I61a874eb4c7800266c0e08d46bfaa693ad56b3d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1678541
Commit-Queue: Mark Yacoub <markyacoub@google.com>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Antoine Labour <piman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678322}