Add PictureLayerImpl histograms for non-LCD text
This change introduces the following histogram.

Compositing.Renderer.PercentPictureLayersWithTextButLCDTextDisabled.<suffix>

The percentage of picture layers that include a DrawTextBlobOp but do
not allow LCD-text to be used.  The suffix is LessThan10, 10To30, or
MoreThan30.

This will be used to understand how composition decisions affect text
quality.

Change-Id: I8f1c2a76705d07b833631a0c8717a089eb71bb65
Bug: 984346
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1693804
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Commit-Queue: Sam Fortiner <samfort@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#678452}