Use async targeting to find drag and drop targets
Add callback supports to the RenderWidgetTargeter
and use the async targeting in drag and drop code
and register a callback to send drag events using
the callback. Note that since drag and drop events
aren't WebInputEvent it is not possible to use
FindTargetAndDispatch function of RenderWidgetTargeter
at this time.

This is exactly the same as the CL before:
https://chromium-review.googlesource.com/c/chromium/src/+/1641383
but rebased on top of another change that moves the
drop operations to take ownership of the OSExchangeData:
https://chromium-review.googlesource.com/c/chromium/src/+/1666910

TBR=sadrul@chromium.org

Bug: 804633, 967801
Change-Id: If4f53e880bb805c87d13869348a57cd857adee5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1693442
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Commit-Queue: Navid Zolghadr <nzolghadr@chromium.org>
Auto-Submit: Navid Zolghadr <nzolghadr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678475}