Add CodecImage::NowUnusedCB.
Previously, CodecImageGroup forwarded image destruction to any
interested parties.  MaybeRenderEarlyManager used this to know when
to try early rendering.

However, with image pooling, early rendering isn't always caused by
image destruction; the same image can be reused.

This CL adds a NowUnusedCB to CodecImage that is called when the
CodecImage won't be used for display anymore.  Early rendering uses
this signal, rather than destruction, to check for rendering.  Also,
it removes the image from the CodecImageGroup once it's unused,
since the image no longer depends on the output surface.

Note that the CodecImageGroup still uses CodecImage destruction to
keep track of images; it's not necessary to remove them from the
group like MaybeRenderEarlyManager does.  This guarantees that the
CodecImageGroup will track images properly, even if the
MaybeRenderEarlyManager is destroyed first.  This can happen if, for
example, the decoder is destroyed while VideoFrames are outstanding.

Also note that the NowUnusedCB is only called during image
destruction right now.  A future CL will allow it to be triggered at
other times.

Change-Id: Id26e6eb11e178586fac17ead0a585ed6bbfef854
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1707190
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678549}