Remove minimum size field from FaviconServerFetcherParams
No caller of CreateForMobile sets this field above 16px, which is the
minimal value enforced by LargeIconService, so in fact all mobile UIs
that download from the favicon server use this default value. Behavior
is not changed for them.

For desktop UIs, this CL decouples the minimum_size from desired_size,
keeping the desired size of 16 DIP while decreasing the minimum size to
16px. We argue that this is beneficial for the UI because it improves
availability of icons. If any desktop UI in the future needs to enforce
the minimum size to display, it can do so when loading the icon from
local cache.

Bug: 982810
Change-Id: I1160316e93fd721cc977833aa9635d40971a2da6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1695366
Reviewed-by: Nicolas Zea <zea@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Kristi Park <kristipark@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Jan Krcal <jkrcal@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#678736}