[PM] Create FramePriority interfaces and helper classes.
Declares the "VoteConsumer" interface.

Also creates Vote, AcceptedVote, VoteReceipt, VotingChannel and
VotingChannelFactory helper classes. A voter registers with a
VoteConsumer, and receives a VotingChannel built by a
VotingChannelFactory. A Vote is created by a voter, submitted to a
VoteConsumer over its VotingChannel, and if accepted becomes an
AcceptedVote to be stored by the VoteConsumer. An AcceptedVote is
entangled with exactly one VoteReceipt, which is handed back to the
voter.

The VoteReceipt is effectively a handle to an AcceptedVote, and allows
a Voter to withdraw their vote explicitly after initially submitted it.
Both VoteReceipt and AcceptedVote are move-only types to ensure the
relationship is strictly maintained.

The next CL will introduce a voter/consumer that aggregates votes.

BUG=971272

Change-Id: I77ac705a614d7ecc89083f65aa80d3a3684f6cb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1697138
Commit-Queue: Chris Hamilton <chrisha@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678856}