Fix Non-BrowserAccessibility GetNextSibling/GetPreviousSibling
All delegates that inherit from AXPlatformNodeDelegateBase do not
implement the GetIndexInParent method. This makes generic
GetNextSibling/GetPreviousSibling impossible on the delegate. To
mitigate, I have implemented a generic GetIndexInParent on
AXPlatformNodeDelegateBase. It has a complexity of O(n), but this
is not different that the various platform implementations of
AXPlatformNode::GetIndexInParent.

Change in browser_accessibility_mac was necessary to avoid using a
const cast as GetIndexInParent is not const. (Relies on non-const
method: GetParent)
The other option would be to make GetParent const, which is
trivial to do, but requires changes to an additional 20 or so files.

Bug: 984995
Change-Id: If061c212eee974534aad588b9d225f7cff9b00df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1707132
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Commit-Queue: Jacques Newman <janewman@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#678863}