Improve size of trie encoding for HSTS preload list
The trie used for the HSTS preload list is a series of dispatch tables,
each starting with a prefix of zero or more characters. The length of
that prefix was previously encoded in unary, so the number of bits
required to encode a length was one larger than the value of that
length. This changes the format to use a slightly more complicated
prefix code (https://en.wikipedia.org/wiki/Prefix_code) which based on
the current distribution of lengths encoded in the trie reduce the size
of the encoded trie by about 18 KB.

Change-Id: I3ff03c613bd266010b3ff1d8c69f9eae73474e1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1708248
Commit-Queue: Nick Harper <nharper@chromium.org>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678889}