Don't reuse a task runner between system_perfetto_unittests.
The issue was that PerfettoTracedProcess only constructs a task runner
once and stores the pointer. If the same process was reused to
run a new test we still have this pointer, but the ScopedTaskEnvironment
it belonged to has been destroyed.

While investigating this I discovered a couple of minor issues that I
also fixed in this patch because I was there.

1) Technically there is a potential race if you construct a TestDataSource
while a trace is active. The race is between the constructor finishing and
the DataSource being registered on a different thread and being started.
Fixed by making a static create method.

2) ResetTaskRunnerForTesting has to be called before the
PerfettoTracedProcess has been used, which consumer_host_unittests
violated.

3) PerfettoTracedProcess::ClearDataSourcesForTesting allowed access to
|data_source_| off sequence, which could cause a data race if you aren't
aware of it. Fix that by making the class PostTask to clear it.

4) Ensure that we delete the AndroidSystemProducer when destroying
the test to make sure it gets destroyed on the same threadpool sequence.

Bug: 983509
Change-Id: Ib1025697c182f746e26a9a75307804cb65886386
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1702305
Commit-Queue: Stephen Nusko <nuskos@chromium.org>
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Auto-Submit: Stephen Nusko <nuskos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679043}