Separate buckets for AnimateMotion and Discard tags
Previously in the animation code the AnimateMotion and
Discard tags were grouped together in the same sandwich
when animated, but they require different paths further
down in the call tree. This made it hard to refactor the
SMILAnimationSandwich to hold the value being animated,
and results in messy code. This is one step to clean it
up.

This is a refactor commit and no functionality should be
changed with it.

Bug: 641437, 981522
Change-Id: Ibef821300606c407a96a1bb4917103ff2b54fcd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1709578
Commit-Queue: Edvard Thörnros <edvardt@opera.com>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Auto-Submit: Edvard Thörnros <edvardt@opera.com>
Cr-Commit-Position: refs/heads/master@{#679062}