Use IdentityManagerBuilder in //ios/chrome/browser
IdentityManagerBuilder was introduced to abstract the building of the
Identity Manager from the embedders, initially supporting all platforms
but iOS, which lacked a way to build a ProfileOAuth2TokenService.

In this CL, profile_oauth2_token_service_builder.cc gets support for
creating a ProfileOAuth2TokenServiceDelegateIOS and
//ios/chrome/browser/signin/identity_manager_factory.cc is modified to
delegate the building of IdentityManager to IdentityManagerBuilder.

Bug: 974198
Change-Id: I54746e2f75e60afa330999a369f1e0f9788bd70a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1699811
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Henrique Ferreiro <hferreiro@igalia.com>
Cr-Commit-Position: refs/heads/master@{#679071}