Fixed crash when slider-vertical appearance is used for input without shadow root.
Since webkit-appearance: slider-vertical / slider-horizontal can be used
for other elements (including inputs), we need to add a check to treat
the case when the appearance is used for an input without a shadow root.

Added a test to validate the fix and prevent further regressions.

Bug: 985642, 985643, 985647
Change-Id: I97e8e5ad9492f7688f145c5c791f4bb9603f3cef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1710069
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Ionel Popescu <iopopesc@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#679087}