Implement GetFallbackFont on Linux
This CL implements the GetFontFallback on Linux by using the FontConfig
API.

This code is used in ShapeRuns when a fallback font is required.
Without the patch, the ShapeRuns only rely on the slow GetFallbackFonts
which returns every fallback fonts (no matter of the characters or locale).
The GetFallbackFont is a fast path with provide a font for most of the
cases.

The code is based on SkFontMgr_fontconfig::onMatchFamilyStyleCharacter.
Unfortunately, the new SkFontMgr is SkFontMgr_FCI and doesn't provide that
function.

see: https://cs.chromium.org/chromium/src/third_party/skia/src/ports/SkFontMgr_fontconfig.cpp?l=910


R=robliao@chromium.org,asvitkine@chromium.org

Bug: 904994
Change-Id: I7156a68f14d0183a0564396e0902910de829ba5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1707834
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679209}