[omnibox] Fix Paste and Go context menu item for URL-like strings
The omnibox classifies URL-like strings as URL matches, even if it's
not a valid GURL. For example, "test.com" is treated as a URL match
even though it's not a valid GURL (because it has no scheme).

The current code for rendering the Paste and Go context menu item
naively converts the input text to a GURL, which doesn't work in some
cases, so for "test.com", it would show: "Paste and go to".

This is not correct. We fix this by using the destination URL that's
derived during string classification, rather than just blindly putting
the input text into the GURL constructor.

This CL also adds a test for this case.

This CL also eliminates the OmniboxEditModel::ClassifiesAsSearch
method, because there's only one caller after this CL, and instead
makes OmniboxEditModel::ClassifyString public.

Bug: 980002
Change-Id: Id411f4cc6e23902206e813302e315523a7f553af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1710780
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679248}