touchless: Enable pinning by default
Browser restarts usually happen when device is booting, and requests to
pin are made during the startup sequence, hence ramping this up from
DISABLED to ENABLED via Finch would be slow, and will hold users from
benefiting the optimization for quite long.

In order to solve that we enable the feature by default.

There is still a chance to make the Finch experiment by going from
ENABLED to DISABLED in the experiment arm.

Safety:

* On devices where pinner service does not support pinning Chrome, the
  attempt would fail on a background thread after a reflection call and
  would send no binder IPCs.

* Pinning won't work on platforms that do not support "code ordering"
  because there is no region to pin. All of the arm/arm64/x86 do support
  code ordering now.

Bug: 929926
Change-Id: I28804b99fcf9960fe200da276995c07e5e2567c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1710545
Reviewed-by: Benoit L <lizeb@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679387}