Add a 'held' concept to InProgressTouchEvdev, and implement handling.
We add a concept of "held" to in progress touches to allow for
delaying events sent to clients. If an InProgressTouchEvdev is marked
as held, it's added to a queue for that slot until the event is no
longer marked as held, is marked as cancelled or the stroke is ended.

When no longer held or the stroke is ended, all elements in the queue
are reported with their original timestamps.

When a stroke is marked as cancelled, the elements in the queue are
deleted and cancellation proceeds as normal.

Bug: 982118
Change-Id: Ia84ef6477cb42babdd6bda0a7335684a23c2210f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1697182
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Rob Schonberger <robsc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679468}