[iOS] Add dark mode colors to omnibox in BVC toolbar
This mostly adds new colors and uses them. One other small change is
related to highlighting the fake omnibox on the NTP. Before, it was
colored with alpha, so the alpha could just be changed on highlight.
Now, the color is opaque, so I added another view right over the
fakebox itself that changes from alpha 0 -> highlighted.

Bug: 976643
Change-Id: Ie681a109f6e69846036f932cfe7dba3820aaea51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1705793
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#679521}