Remove the PerfettoTracedProcess WeakPtr which isn't needed in real code.
The PerfettoTracedProcess is only destroyed in tests, so we don't need
to protect it, but there was a bug since we didn't check the weak_ptr
which means when we in place reconstructed the PerfettoTracedProcess the
WeakPtrs would be invalidated. This fixes this by ensuring the the task
can't run before we've constructed the object and also drops the WeakPtr
which wasn't needed.

Bug: 983509
Change-Id: I87d3ab16b49358742a1a27fd7d6d179b43d90e76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1712576
Commit-Queue: Stephen Nusko <nuskos@chromium.org>
Commit-Queue: Eric Seckler <eseckler@chromium.org>
Auto-Submit: Stephen Nusko <nuskos@chromium.org>
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679587}