Add a method in PDFiumPage to calculate the text range for an input rect
This CL introduces a new method GetUnderlyingTextRangeForRect() in
PDFiumPage which given an input rect, returns the text range found
inside the rect in the form of start index and count of characters.
This method has zero tolerance and relies on the fact that the mid-point
of the char rectangle should lie inside the input rectangle.

This new method will be used to create a relation between text and
annotations (Links) in a subsequent change to provide accessibility
support for links.

This CL includes unit tests to verify the text range of the underlying
content for various input test rects in pdf file.

Bug: 981448
Change-Id: Ibf051cff31e59e63d8fafb480adb0c6e8bb1f588
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1692608
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#679620}