Support running signin::URLLoaderThrottle on the UI thread.
This CL is part of a larger effort (crbug.com/824840) to move
NavigationURLLoader and its dependencies from the IO to UI thread now
that the Network Service has launched. The URLLoaderThrottle will only
be run on the UI thread when the NavigationLoaderOnUI feature is
enabled, and will never run on both threads at the same time. This CL
adds some complexity to handle determining which thread to run on,
which will need to be cleaned up once we've finished the migration and
the feature is launched.

Design doc: doc/1bjW2JlY_IS8TAC9x2wML8vu1ufrbYOAzAOoR9P0xvuM

Bug: 978617
Change-Id: Ic97e66ed770eb8e6c4e55796e0aa208cca0d507a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1680993
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679738}