Make sure to paint the frame opaquely to the bottom.
The header is painted with anti aliasing and it can blend
pixel with transparent pixel below, which can leave the alpha
value at the bottom. (removing the antialias fixed the issue)
Instead of painting twice (once for each to and bottom) this CL
simply extend one dp to bottom to make sure it's fully opaque.

Bug: 985903
Test: manual.
Change-Id: Iaf3b7113d996ae39c3fd49420e21e09b05cec0d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1711890
Reviewed-by: Malay Keshav <malaykeshav@chromium.org>
Commit-Queue: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679749}