[gaia auth] Remove default constructor
We are reading at least one of the bools in ClientOAuthResult without
initialization. This was due to the default constructor missing it,
but this constructor is only used in one place, undesired anyways,
so this CL removes it and calls the regular constructor in its place.

Change-Id: I3e9478c35be5ebafd7422a80b5c7d274e7ffbdd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1708018
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Commit-Queue: Kevin Bailey <krb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679766}