[WebAPK] chrome://webapks disable updates in server cache window
The WebAPK server only updates its cache every minute or so. If
an update is attempted and re-attempted within this window then
the update will fail even if a change to the manifest occurs.
To prevent this we will disable the update button for 1 minute
after an update attempt is completed.

Bug: 984586
Change-Id: I1dece1f0cb7236159b105ebe5e4caaa354645573
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1704499
Commit-Queue: Hector Carmona <hcarmona@chromium.org>
Reviewed-by: Hector Carmona <hcarmona@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Peter Kotwicz <pkotwicz@chromium.org>
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679779}