assistant oobe: fix the incorrect clickable zone.
This change fixes the incorrect clickable zone of |zippy-additional| on
the assistant third party oobe page. In this way, a modal dialog only
shows up when user clicks the link (the <a> element). Clicks on any
other parts besides <a> in |zippy-additional| will have no effect on
the page.

Bug: none.
Test: manually.
Change-Id: Id5f2f136afcb2fdc677f79cae597c0909d44f141
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1713945
Commit-Queue: Meilin Wang <meilinw@chromium.org>
Reviewed-by: Yue Li <updowndota@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679818}