Avoid always hopping to IO for service worker for NavigationLoaderOnUI
I noticed the service worker interceptor blocks navigation on a
UI -> IO -> UI thread hop while looking up registrations in
MaybeCreateLoader. This CL keeps a map of the origins with
registrations on the UI thread, so we can skip this hop for origins
that have no service workers registered.

This will no longer be needed if service workers are migrated fully to
UI, so depending on what the timeline of that is this may not be
needed.

Bug: 824858
Change-Id: I2ce383ce4b8a370617c823d954376b9dd9f84a6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1710848
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#679875}