Make window capture perf & behavior consistent between new/old APIs
This change makes window capture via the getDisplayMedia API on Windows
perform & behave consistently with window capture via the
chrome.desktopCapture API from an extension. (Also, with the previous
experimental window capture support via the getUserMedia API with non-
standard constraints.)

This change allows window capture via getDisplayMedia to use the
capture-screen-and-crop approach when the window is occluded only by the
'sharing' notification bar and some other conditions are met. It
broadens the scope of existing logic to plumb the notification window ID
to the capturer, so that it may be special-cased there (ignored when
checking for occluding windows). This plumbing is now enabled for a
stream type of DISPLAY_VIDEO_CAPTURE (screen/window capture via
getDisplayMedia) in addition to a stream type of
GUM_DESKTOP_VIDEO_CAPTURE (screen/window capture via older APIs).

Benefits of this change:
+ Allows use of the significantly faster capture-screen-and-crop path in
more cases, which can be up to ~36ms/frame (~160x) faster than the
capture-window-contents path in my testing (more details are in the
bug). This difference increased with the recent fix for
https://crbug.com/webrtc/10734 .
+ Allows capture of menus & tooltips in more cases (plus dialogs if
https://crbug.com/webrtc/10767 is fixed), partially mitigating
https://crbug.com/980864 .

Downsides of this change:
- Includes the 'sharing' notification bar in the captured frames in more
cases, when it overlaps the window. This is similar to the behavior of
capturing the main screen, and identical to the behavior of capturing a
window via the chrome.desktopCapture API from an extension.
- May decrease test coverage / issue discovery around the capture-
window-contents path, by reducing the likelihood it's used via Chromium.

Bug: 973245
Change-Id: I224b20137ba701150ed01c6a1fb5dba134478c8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1711183
Reviewed-by: Weiyong Yao <braveyao@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680152}