[Autofill Assistant] Fixed flaky test.
Before this patch, we would wait for AA to be shown before invoking the espresso check. This would sometimes fail on slower devices, because asynchronous events on the native side might not have finished processing at that point.

This CL instead continuously polls the UI thread until the specified condition is matched or the default timeout is reached.

Bug: b/986026
Change-Id: I311f94dd8fd4383dfcce35656299bddce366b7bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1712564
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Stephane Zermatten <szermatt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680362}