[Autofill Assistant] scroll element to center before clicking it
Before this change, an element was scrolled into view before it was
clicked. On website that use top or bottom bar, this means that the
element could be under the bar, so the click would never succeed.

This change attempts to increase the probability of an element to be
fully uncovered, even if it's inside of a container, by moving it first
to the center of the screen just before clicking. scrollIntoViewIfNeeded
is still used afterwards, while waiting for the button position to
become stable to avoid bad interactions with event listeners.

Bug: b/135909926
Change-Id: Idea978b2c3407f41b53fee02feb92c86015122f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1708080
Commit-Queue: Stephane Zermatten <szermatt@chromium.org>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Cr-Commit-Position: refs/heads/master@{#680425}