Buffer layout shifts after pointerdown.
We treat tap but not scroll as an excluding input for the purposes of
cumulative layout shift calculation and the hadRecentInput bit on the
performance entry.

NotifyInput looks at WebInputEvent types to determine exclusion status.
If kPointerDown is followed by kPointerUp, the pending shifts are
excluded.  But if followed by kPointerCancel or kPointerCausedUaAction,
the pending shifts are not excluded.

Bug: 978027
Change-Id: I43f48b0001df168a46e2b44955742144f22c691b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1715232
Commit-Queue: Steve Kobes <skobes@chromium.org>
Reviewed-by: Bryan McQuade <bmcquade@chromium.org>
Reviewed-by: Nicolás Peña Moreno <npm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680598}