Fix UAF of URLLoaderRequestController by using WeakPtr.
(This CL was written by falken@. bashi@ will submit on behalf of him)

URLLoaderRequestController was giving callbacks to interceptors using
base::Unretained(this). For |fallback_callback|, this is problematic as
the interceptor might invoke it after the controller/interceptor might
have already been destroyed. For example, the service worker interceptor
makes a ServiceWorkerNavigationLoader object, which outlives the
interceptor and calls |fallback_callback| after getting a response from
the service worker in the renderer.

For |loader_callback|, it probably doesn't need to be a WeakPtr, but
just change it for consistency and so we don't need to make assumptions
about when interceptors invoke the callbacks.

Bug: 984460
Change-Id: Id529f55412ef7ed1be83d6607d27590b8859a26d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1717885
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#680773}