Add close button to CrOS lock screen media controls
This CL adds a close button to the upper right corner of the CrOS lock
screen media controls. This button appears if the mouse enters the view
and disappears if the mouse exits the view. Upon clicking this button,
the media controls are dismissed and media playback is paused.

Bug: 965574
Change-Id: Ic3ac8345c25ed200db924ff337610aaac95d89a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1713038
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Commit-Queue: Mia Bergeron <miaber@google.com>
Cr-Commit-Position: refs/heads/master@{#680968}