Add MediaSessionPositionChanged to MediaSessionObserver
A MediaSessionPositionChanged event has been added to
MediaSessionObserver. This is passed along to MediaControllerObserver
as well. This event notifies MediaSessionObservers when the position of
the current media session has changed. It will also provide the new
position as a media_session::MediaPosition object.

To start, this method will only pass on a fixed, mock MediaPosition
object (this is done in MediaSessionImpl).

Bug: 985077
Change-Id: I189031a7e380da27a1c0d7d446f2e325a6624875
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1707311
Reviewed-by: Sean Topping <seantopping@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Chrome Cunningham <chcunningham@chromium.org>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Commit-Queue: Mia Bergeron <miaber@google.com>
Cr-Commit-Position: refs/heads/master@{#680970}