kerberos: Prevent UI calls when disabled
Prevents UI code from calling into KerberosCredentialsManager when
the KerberosEnabled policy is off. Usually, the UI is hidden when the
policy is off, but it is possible that the user opened the UI while
Kerberos was still enabled and kept it open. This CL prevents calls
when the user clicks buttons in this still-open UI.

BUG=chromium:961246
TEST=Turn on KerberosEnabled policy, log in, go to
     chrome:settings/kerberosAccounts, verify that adding an account
     works (e.g. using chromeadm-lab.com account), keep UI open and turn
     off KerberosEnabled policy, reload policy, verify that account
     vanished and that no new accounts can be added.

Change-Id: Id473cf3229864a9be140f7abfd250a027eaf48ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1704760
Commit-Queue: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681000}