Use floating new tab button for non-compact layout in TabGrid
TabGrid uses UIToolbar for both compact layout (Regular(V)xCompact(H))
and non-compact layout. For non-compact layout, the new tab button is
wrapped in to a container UIView with insets, then inserted into a
transparent UIToolbar so that it looks like a floating button but still
has the built-in a11y modal panel for long-press under a11y font-size.

The non-compact layout is no longer working in iOS13 and the floating
new tab button is compressed probably because the calculation of
UIToolbar's height is changed in UIKit. After a discussion with UX, we
decided to remove the transparent UIToolbar for non-compact layout and
make the new tab button a real floating button.

Bug: 971786
Change-Id: I7efddd5e044ce19a469c01ca75535a22b1b5da6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1715188
Commit-Queue: Yi Su <mrsuyi@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681036}