Do not strip resource names in debug builds
Resource name stripping was enabled in both debug and release when it
was new so that issues are more likely to be found early. Now that the
optimization has been on for some time without issue, it makes sense to
disable it in debug mode so as to not annoy developers that are looking
into resource related issues.

Change-Id: I68ac00396dc8e2b5d37d276c22c58a1d020e1392
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1719451
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Auto-Submit: Mohamed Heikal <mheikal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681143}