Extend lifetime of some SBDownloadFeedback metrics
I'm working on a more reliable method to upload files for another project,
and was looking into using that in the DownloadFeedbackService as well.
I expect these metrics to be important if that pans out, so extend the
lifetime of these metrics by a little bit.

Bug: 969929
Change-Id: I1d26b502ad76cf4aa753344a19b0bd2e96c83a6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1719653
Auto-Submit: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: Robert Kaplow <rkaplow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681277}