[sms] Add UMA tracking for SMSReceiver API for measuring user friction
This change adds UMA tracking for SMSReceiver API to give us data on friction
and performance of UI decisions made for this feature. Because there was some
disagreement between adding the user interaction after a successful call, this
will give us a better idea of what kind of impact the UI has or doesn't have.
Specifically it will track the following.

1. Duration between the time a SMS is successfully received and when the user
   selects Continue.

2. Duration between the time a SMS is successfully received and when the user
   selects Cancel.

More details on metrics and its motivation here in the API Design Doc [1].

[1] https://docs.google.com/document/d/1dB5UM9x8Ap2-bs6Xn0KnbC_B1KNLIUv4W05MunuXYh0/edit#heading=h.bbporvvwzoo7


Bug: 976332
Change-Id: Id689c9937c17a9dc930cce89ba0965e8f20a96d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1716961
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Jun Cai <juncai@chromium.org>
Commit-Queue: Ayu Ishii <ayui@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681306}