[AA Direct Actions] Distinguish RPCs coming from direct actions.
This change marks the following RPCs with enough information to figure
out whether they've been triggered by direct actions:

When SupportsSite is called from a Chrome CCT created for AA, this
change sets ClientContext.is_cct.

When a script is run by using a direct action, instead of the UI,
this change sets ClientContext.direct_action

The two are necessary, as it is possible to use direct actions on a AA
started by CCT. It is possible to use the UI to answer prompts triggered
by a script run by a direct action.

Additionally, when a prompt choice is made by using a direct action,
instead of the UI, this change sets ProcessedActionResut.direct_action

Bug: b/138292695
Change-Id: I196d9d7492cc27fce164e9a671cf794732bf1d21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1715379
Commit-Queue: Stephane Zermatten <szermatt@chromium.org>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Cr-Commit-Position: refs/heads/master@{#681315}