Speculative fix for object element crash
This CL replaced a bunch of element_->GetLayoutObject() calls with
GetLayoutEmbeddedContent():

https://chromium-review.googlesource.com/c/chromium/src/+/1597534

However, those two calls are not equivalent if the LayoutObject is not
a LayoutEmbeddedContent. This patch restores the calls to
element_->GetLayoutObject().

I was unable to reproduce the crash, even using the clusterfuzz tools,
so I can't write a test.

BUG=982214
R=bokan@chromium.org

Change-Id: If2dee39a235ee961d7f27258e7a04e8165d2ab3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1719911
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Stefan Zager <szager@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681407}