Settings: Remove basic_page references to non-existent paths
3 of the pageVisibility paths (search, androidApps, system) that are
read by the settings-basic-page [1] are never defined and are not part
of the PathVisibility typedef [2]. I think the dom-if they control were
simply never removed when the paths ceased to be used.

This CL cleans up settings-basic-page by removing the references to
non-existent paths.

[1] https://screenshot.googleplex.com/CPO8xHQ9WSx
[2] https://screenshot.googleplex.com/nxGJbNy8vbx

Change-Id: I25a99720a6f4df25a6e5b1ac64acd419bba15eaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1716338
Reviewed-by: Hector Carmona <hcarmona@chromium.org>
Commit-Queue: Jordy Greenblatt <jordynass@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681437}