Send GesturePinchUpdate events as non-blocking
GesturePinchUpdates are currently blocking even though they are always
handled on the compositor. The only case where we inspect the ack of a
GesturePinchUpdate event is for touchpad pinch, but in this case, the
browser creates a synthetic wheel event and sends that. So it's only the
wheel event that needs to block.

Bug: 787924
Change-Id: I1f5e25289eb8bde2264cec851df880baf31dcab1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1717693
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681462}