jni_generator: Fix jcaller usage.
This CL changes static methods to not pass a jcaller object to
the native side. This isn't necessary because there is no Java object
instance for static methods and it was actually just passing a jclass
as a jobject which was always unused.

The @JCaller annotation is also removed. It isn't strictly necessary
anymore and is often used incorrectly (i.e. using @JCaller and adding
an unused param on the native side).

All Java and native definitions and uses are also updated to reflect
the changed logic.

Tbr: agrieve@chromium.org  # Large mechanical refactor.
Bug: 986930
Change-Id: I2ef62d189dad484779524d0d5edadeaf51204395
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1719427
Commit-Queue: Eric Stevenson <estevenson@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681466}