Add a mechanism to skip the built in expectations in run_web_tests.
We are adding a variant of run_web_tests that the GPU team will
use to run a subset of the WPT tests that will only ever work
on bots with actual GPU cards. In order for this to work, we
will always skip (WontFix) the tests in the main run_web_tests runs,
but we will run the tests separately.

In order to do that, we need a way to bypass the default expectations
(i.e., skip the lines that will be in NeverFixTests).

This CL adds a --no-default-expectations arg to run_web_tests to
do so.

R=rmhasan@google.com, kainino@chromium.org
BUG=984815

Change-Id: Ia855f70c0f8cdf3b272f700e7e08349b8c5b1694
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1710888
Reviewed-by: Rakib Hasan <rmhasan@google.com>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681468}