win: Remove win_is_xtree_patched machinery.
Back when we still used MSVC, this checked if our hermetic toolchain had
a local change to make MSVC's unreachable code warning work with MSVC's
headers.

We haven't used MSVC in a while and even if someone were to use it, /wd4702
is currently disabled globally unconditionally.

Eventually we should turn on clang's -Wunreachable-code, but for now
this removes an exec_script() invocation that infinitesimally slows down
gn and that doesn't have any effect in practice.

Bug: 346399
Change-Id: I0d68086ce0bb859ce58c579723e92f64c1c3c256
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1721457
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681529}