[TextFragmentAnchor] Add UMA metrics
This CL adds two metrics. The first is an UMA histogram tracking how
often we try to use an element-id based fragment and succeed vs fail.
Note that if the TextFragmentAnchor feature is enabled, we won't count
this metric if the fragment successfully parses to a |targetText|
anchor.

The second metric measures how often we see a hash ("#") character in
the fragment. This is technically an invalid URL but we'd like to
evaluate how often it happens in practice. Similarly to above, if
targetText is enabled, we won't count cases of ##targetText as being
part of this metric.

TBR=bokan@chromium.org

(cherry picked from commit 427b2ca2ff003fed78074d8f54b6548ed2757f48)

Bug: 987027
Change-Id: Ic3b6cf1a85e590449af12c131edc60b13bad0dfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1719651
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#681420}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1723298
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/branch-heads/3865@{#18}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}