Modules: Enable partitioned native libraries for VR
This change flips on the switch for putting the VR UI's native code in
the feature DFM.

In doing so, it enables a partitioned Chrome library for Android, which
changes how libchrome.so (and libmonochrome.so) are generated.  The
link stage now generates libchrome__combined.so, a special library that
uses lld linker capabilities to put feature code into separate
sections.  The main and feature libraries are then extracted in a
follow-on step, producing discrete libchrome.so and feature libs.

As a result, this change has the potential to break scripts that (for
example) induce re-linking of libraries by deleting libchrome.so.

Bug: 874564
Change-Id: I708e990b3aa6b1abe71b28a67add51662c268276
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1663486
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: Bill Orr <billorr@chromium.org>
Commit-Queue: Christopher Grant <cjgrant@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681883}