Stop overriding defaults for most UI Automation window properties
With UI Automation enabled, the top-level node for the Chrome window
would show most properties as empty. This was happening because our
fragment root was mapping values from an empty AXNodeData.

UI Automation has a built-in provider for HWNDs. In our implementation,
fragment roots do not contain any useful accessibility information
beyond what the built-in provider would expose, so we simply return
VT_EMPTY for most properties, so that we don't override values returned
by the built-in provider.

The exceptions are IsControlElement and IsContentElement. We implement
these properties to exclude the legacy HWND from the control and content
views, because we don't want screen readers to stop on the legacy HWND
and announce "Chrome Legacy Window." Other HWNDs are allowed in the
control and content views so that, for example, an assistive technology
can manipulate the window using WindowPattern.

R=iapres@microsoft.com

Bug: 928811
Change-Id: If606f399c386109adc3bd56bf081b2e5c2362070
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1716010
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Ian Prest <iapres@microsoft.com>
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#681894}