[IndexedDB] LevelDB Scopes integration with simple Iterator.
This patch integrates IndexedDB on top of LevelDB Scopes. It tries to do
this as simply as possible, and doesn't have sophisticated iterator
invalidation logic.

The task for this change is here:
https://docs.google.com/document/d/1jX6OOQJ0aLrWAeqYLtARXPFawUlJPbHNIKAK-0vmFHU/edit#heading=h.80ylb2v9t1ff

refactoring, as well as an increase needed to implement a feature, and I've
already spent a non-trivial amount of time trying to reduce its size.

Binary-Size: Increase is temporary due to a "new way" / "old way"
Bug: 862456
Change-Id: I03bc400fcc23d4dcfe98c3d8722c2cff2895dd0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1676828
Commit-Queue: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Reviewed-by: Chase Phillips <cmp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681924}