[iOS] Update the navigation state before finishing history navigation.
With the non-modal dialog rollout, the back/forward buttons on the
bottom toolbar are now visible when HTTP authentication dialogs are
displayed.  When performing one of these navigations, the navigation
state is still REQUESTED, which causes a DCHECK in CRWWebController's
|-didFinishWithURL:loadSuccess:context:|.  This CL updates the
navigation state to FINISHED before finishing back/forward navigations
that occur while in this state.

Bug: 988548
Change-Id: I11fdcc03a90aa50896069354c4399cafb896c74c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1723806
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#681994}