Android: Configure proxy_resolver for in-process
proxy_resolver is supposed to run in-process on Android, but r663141
unintentionally broke this. This fixes it.

Note that this is a direct merge to M76 branch, as the bug was
coincidentally fixed by a series of much more complex changes on ToT
since branch.

Bug: 980743
Change-Id: Ic40289651aea0bfd5df5fbfa020b336466ec2ba5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1725123
Reviewed-by: Eric Roman <eroman@chromium.org>
Cr-Commit-Position: refs/branch-heads/3809@{#971}
Cr-Branched-From: d82dec1a818f378c464ba307ddd9c92133eac355-refs/heads/master@{#665002}