DNR: Implement schema and indexing of url transformation rules.
This CL adds support for a new Redirect dictionary to the "action" field of a
Rule. This allows specifying rules which redirect to extension paths or
performing various url transformations. Eventually this is meant to replace the
"redirectUrl" field.

This CL adds support for the schema and indexing of these rules. In a subsequent
CL, these rules will be evaluated.

This also fixes a bug with the current handling of redirectUrl key when the
redirect url is relative. The query and fragment paths within the relative url
weren't being handled correctly.

DOC=https://docs.google.com/document/d/1foL79VwByaVeUwYPBfZeqyC2SYcAawhjjH05eqHqW_E/edit
BUG=983685

No-Presubmit: True
Change-Id: Ief61eed0a640f5fcadbc17e58fe611691fa64e8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1705019
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#682081}