Add draw_utils::GetBottomGapBetweenRects().
Currently draw_utils::GetBottomFillRect() returns the gap on the
bottom side of a page created by insetting the page. However,
in two-up view, the bottom gap for a page has height equal to the
difference in height between the current page and its adjacent one.

draw_utils::GetBottomGapBetweenRects() takes the bottom component of
a page and a rectangle, returning the portion of the rectangle that
is below the page. This will be used in PDFiumEngine to compute the
bottom empty space for a page in two-up view as mentioned above.

Bug: 51472
Change-Id: I436600264106340ca6699799359149f29c797844
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1721318
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Jeremy Chinsen <chinsenj@google.com>
Cr-Commit-Position: refs/heads/master@{#682097}