Remove IndexedDBClient
IndexedDBClient was a WorkerClient and it provided AllowIndexedDB().
IDBFactory extracted an IndexedDBClient from execution context to
call AllowIndexedDB(). However this was an unnecessary abstraction
and we can directly implement AllowIndexedDB() in IDBFactory.
This CL removes IndexedDBClient to remove the abstraction.
This eliminates creating a GarbageCollection object.

This CL doesn't have behavioral changes.

Bug: 988335
Change-Id: Ia6948cf195abf53335609d2ddb2a02bd7c7faa84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1722276
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Joshua Bell <jsbell@chromium.org>
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#682146}