Detach the SSLVersionMax policy from prefs
The SSLVersionMax policy was introduced in 2017 to allow administrators
to disable TLS 1.3, as a temporary workaround for buggy middleboxes. It
was always documented in policy_templates.json to be temporary and stop
working in M75, with notice in both M74 and M75 release notes.

Remove the implementation of the policy.

Bug: none
Change-Id: Idec7c14101d2199cd5fad23897a5e76cd26025c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1716059
Auto-Submit: David Benjamin <davidben@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Commit-Queue: Pavol Marko <pmarko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#682284}