Update suggestion chips on reset for show
AppsContainerView stops changing its state on closing, which
causes to remain the suggestion chips on inconsistent state with
its view.

I believe it is still good to keep the suppressing the state
change, but we need to reset the suggestion chip state on re-show.

Bug: 988566
Test: manually
Change-Id: Iddb2a4a69c17f601a66e80419ce58f28df4ca18e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1725236
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#682338}