Evaluate the delay mapping with a score value
Evaluate the mapping between the count of in-flight requests to the peak
queueing delay by computing a score for samples.

The score is the percentage of samples that conform to the rule: the
more in-flight requests, the higher peak queueing delay. The score is
the metrics that evaluates whether the queueing delay prediction
algorithm is effective.

Change-Id: I1d6fdb9cccc9b52ff5b8d7849ae358d465dce73a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1717565
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Commit-Queue: Jianfeng Wang <jfwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#682342}