Remove CrosNetworkConfigTestHelper from ShellBrowserMainParts
CrosNetworkConfigTestHelper was intended as a convenience class for
tests that require specific network configuration setup. It relies
on running the message loop to ensure complete setup before tests run.

For ShellBrowserMainParts, we just need a CorsNetworkConfig instance.
The easiest way to provide this is to initialize the fake shill clients
and NetworkHandler, then create an owned CrosNetworkConfig instance,
without the additional complexity in the test helper.

Bug: 980951
Change-Id: I4b4a72ec7258f5010e0e24750e2f37f6bd884138
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1726932
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#682458}