Instantiate and use PalmDetectionFilter within TouchEventConverterEvdev.
Do so by creating a Factory for PalmDetectionFilter, controlled by
flags and feature parameters, called
PalmDetectionFilterFactory. Thoroughly unit test this factory.

Add an instance of SharedPalmDetectionFilterState to
InputDeviceFactoryEvdev, and pipe it through to instances of
TouchEventConverterEvdev so that all PalmDetectionFilter get the same
shared instance, as intended.


Bug: 982118
Change-Id: I5878bc126a9e4bf809b9d5869520bef814aace8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1716000
Commit-Queue: Rob Schonberger <robsc@chromium.org>
Reviewed-by: Michael Spang <spang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#682567}