Clarify namespace for "features" in performance_manager
Inside the namespace performance_manager the string features
can refer either to ::performance_manager::features or to
::features depending on what the compiler knows about.

In jumbo builds the compiler has more information and will always
resolve features to ::performance_manager::features. This broke
the windows jumbo fyi builder, but the fix is trivial, just
write ::features instead of features.

(Having a sub namespace with the same name as a common top-level
namespace violates the code style guide but this is common
for features and switches and need a large cleanup.)

TBR=treib@chromium.org

Bug: 973963
Change-Id: Ida06c223f595130b48a8e4442a8e17f5476de8ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1728551
Reviewed-by: Daniel Bratell <bratell@opera.com>
Commit-Queue: Daniel Bratell <bratell@opera.com>
Cr-Commit-Position: refs/heads/master@{#682657}