Navigation: Move commit_timeout from NavigationHandleImpl.
This includes moving all related functions and members.

This has a potential of reproducing this bug: https://crbug.com/957212.

The bug was initially produced after a big CL was landed that included
some changes from this CL. But due to the size of the changes, it was
hard to point out what the actual problem was.

Bug: 916537
Change-Id: I2c8b11f82e5b5197976b3cdb15de40f3032e68ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1720775
Commit-Queue: Mohamed Abdelhalim <zetamoo@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Lowell Manners <lowell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#682680}