flags: expire M77 flag set
This change implements step 2 of the process outlined in
//docs/flag_expiry.md.  This change will be merged to the M77 branch
after it lands so that these flags are actually expired in M77 :)

This change: 1) Adds a new bit to the supported_platforms mask in
flag entries, named
   'kExpireM77', which is set if a flag expired in M77;
2) Adds a new feature "TemporaryUnexpireFlagsM77" and a matching flag
   "temporary-unexpire-flags-m77";
3) Adds code to the flags UI to hide flags with kExpireM77 in their
   supported_platforms mask unless TemporaryUnexpireFlagsM77 is
   enabled.
4) Removes already-gone flags from //c/b/flag-metadata.json
5) Removes already-gone flags from //docs/flag_expiry.md

Future iterations of this process will use two more bits in the
supported_platforms mask, since the flag expiry process maintains a
window of two milestones between flag expiration and final removal,
so (e.g.) at M78 kExpireM77 and kExpireM78 will be live, and at M79
kExpireM77 will be removed and kExpireM79 added, such that kExpireM78
and kExpireM79 will be live.

After this change, all the flags annotated with kExpireM77 show
as "unsupported platform" in the flags UI and are greyed out. A
follow-up change will add a specific annotation to these flags in
that UI marking them as expired.

Bug: 953690
Change-Id: I2d7361b6f3f070ecf469bf4eccb9b42b5a19d704
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1729449
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#682945}