Reset title of "New Tab" button when page change
When a11y font size is used, long press on the "New Tab" button in tab
grid will show a built-in a11y modal panel which contains an image and
a title. Currently the title is missing, and this CL fixes it.

BTW, the image is also changed to the larger image for incognito page
instead of the small one.

Bug: 989016
Change-Id: I706570f1002b82eaf33908006d09cc3498cb6e66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1726090
Reviewed-by: edchin <edchin@chromium.org>
Commit-Queue: Yi Su <mrsuyi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683095}