Fix failing ContextMenu tests
Tests testCopyImageLinkCopiesLinkURL and
testContextMenuRetrievesImageLinkOptions were failing when LayoutNG
is enabled. The issue seems to be that the context menu is
populated for the HTML element directly below the one the test is
trying to long press.

This CL works around the issue by adding an extra line break after the
element in the test HTML.

Bug: 959265
Change-Id: Ia5dd780c2b5edd4152412c6e61c8e6fb8749b0b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1721809
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Cr-Commit-Position: refs/heads/master@{#683256}