Add PDFiumEngine::LoadPagesInSingleView().
Currently PDFiumEngine::LoadPageInfo() loads pages into PDFiumEngine
only in a single-view layout. However, in the future LoadPageInfo() will
be able to load pages in a two-up view layout as well.

This CL extracts the logic for formatting pages in single-view and
appending them to |pages_| to a helper function
PDFiumEngine::LoadPagesInSingleView(). This helps manage the complexity
of LoadPageInfo().

Bug: 51472
Change-Id: I087f171308e4425dd3bae8b8abe5f0f70e0c0eed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1728059
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Jeremy Chinsen <chinsenj@google.com>
Cr-Commit-Position: refs/heads/master@{#683296}