[CrOS Celluar] Implement ActivationDelegate in setup flow UI.
ActivationDelegate is the primary interface from cellular_setup.mojom to
be implemented in JavaScript. This class deals with sending and
receiving status update messages to the service.

This CL implements this delegate and deals with routing between the
first page (SIM detect) and either the second page (provisioning) or the
third page (final) depending on the response received.

Note that the second page (provisioning) still isn't implemented as part
of this CL; this functionality is part of a follow-up CL.

Bug: 968858
Change-Id: I67d86647237dc47a1c4660bf7160e5e993f45f13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1721042
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Ryan Hansberry <hansberry@chromium.org>
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Auto-Submit: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683384}