media/gpu/v4l2svda: query output planes count during format setup
So far we were forcibly setting the number of output planes to 1, but
the Kukui decoder uses a multiplanar format.

The V4L2 driver will update the number of planes if it is incorrect, so
use this information when attempting to set the format to get the
correct number of planes. This is what the v4l2vda already does.

Bug: b:132589320
Test: vdaunittest and vdatests pass on Minnie.

Change-Id: Idb4170875adf1df2fdbd1cbe65c233937aed0702
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1730732
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683465}