Attempt to make test more stable
The animation test "correct-events-for-short-animations-with-syncbases.html"
used to be flaky. This is an attempt to fix it and also convert it to a
web platform tests.

Verified to run in FF and Chromium.

Colors were also changed reflect their usage.

Bug: 981522
Change-Id: Ieead9ca1c5ccf050ac2e886ecf630d9a9f61b6e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1732097
Auto-Submit: Edvard Thörnros <edvardt@opera.com>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683542}