HttpServerPropertiesImpl:  Refactor test-only methods.
With https://chromium-review.googlesource.com/c/chromium/src/+/1710779,
a number of methods in HttpServerPropertiesImpl are only used by
HttpServerPropertiesImpl itself when loading prefs and tests. This CL
makes these methods private, with public ForTesting methods for tests
to call them. It also renames them, making it clear that they're for
dealing with loading data, and not general purpose setters.

This CL also renames spdy_servers_map() and GetCanonicalSuffix, which
are only used by tests, removes broken_alternative_service_list() and
recently_broken_alternative_services(), which are not used, removes
GetFlattenedSpdyServer(), which is not actually defined, and reorders
methods defined in http_server_properties_impl.cc, which did not match
the order of the header file.

This is done in preparation to merge HttpServerPropertiesImpl with
HttpServerProperties.

Bug: 984996
Change-Id: I1122ea2d2576696877634d3a4e9485edfa72847f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1724497
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Zhongyi Shi <zhongyi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683627}