Tab screenshots for tabs that were never visible.
This is a collaboration of fdoray@ (Catan) and dfried@ (Top Chrome) to
improve how preview images are captured for tab hover cards and for
the new touch-enhanced tablet tab-switcher being developed by the
WebUI team in collaboration with the Views and Top Chrome teams.

Tab preview thumbnails are an integral part of our UX plan for tab-
switching improvements in the next few quarters. Previous
infrastructure was aimed at saving images for the NTP, which was a very
different set of capture situations than what we want for tab previews
(specifically, background-loaded tabs were never captured). This logic
allows the capture of tabs loaded in the background or restored from a
previous session by using a different capture mechanism.

This functionality is entirely behind a flag and is known to be slightly
buggy. We plan to use volunteers to teamfood the functionality and send
feedback/issues so it can be made ready for production.

Change-Id: Ie356b95dc42b75729860c772e9a01a19028d0fee
Bug: 981103
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1689397
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683732}