Settings: When saved password check closes, focus on export dialog
The bug in refers to the fact that before this change, pressing 'ENTER'
with the exportPasswords dialog up would cause a password check
initially, however if the user cancelled the password check prompt,
they will again see the exportPasswords but this time 'ENTER' does
nothing.

hcarmona@ pointed out that this is part of a broader accessibility
issue on this page, namely that the password prompts' closing doesn't
put focus back on the element that triggered it to open. This can lead
to a confusing and difficult UX for users who are using tab to navigate
through the interactive elements on the page.

I ran through a wide range of use cases affected by this change and
recorded them, mostly captured in [1]. Note that the transition at
0:29-0:30 in [1] resulted from pressing enter at the
passwords-export-dialog.

I recorded successful password entries ([2], [3], [4]) separately
because if they are done in close succession, the auth token is
reused and the password prompt is skipped, which negates defeats the
purpose of the recording.

[1] https://drive.google.com/file/d/13naHaZSf7hG3pC09XaO7WW4bmDmmhYW6/view
[2] Successful password prompt from passwords-export-dialog:
https://drive.google.com/file/d/1LLN0dJ8NY54zSTOoZL6F9AcPSFHF0EFy/view
[3] Successful password prompt from password-list-item:
https://drive.google.com/file/d/1hIDOFFrEMRClFlT0VfVTXtyw-YweIfsp/view
[4] Successful password entry from password-edit-dialog
https://drive.google.com/file/d/1XJaGTrmDm4dOnc6a1tQyKnNxaHWU_NzR/view

Bug: 989449
Change-Id: Ic5a363d217cb70e76aa6acb5104ec2fea3a00b4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1731733
Commit-Queue: Jordy Greenblatt <jordynass@chromium.org>
Reviewed-by: Hector Carmona <hcarmona@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683788}