Reland "[Cros SR] Reflect user history deletion actions in the model."
The original CL had a potential thread safety issue, which is fixed
in this CL. This also adds a sequence checker to RecurrenceRanker to
ensure this is fixed.

PS1 is the original CL. PS2 is the fix and sequence checker.

Original CL: crrev.com/c/1725032

Bug: 931149
Change-Id: Iaa5d87d38fa6a9bd5a5605b9feb9a5fd9154d3f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1732269
Reviewed-by: Jia Meng <jiameng@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683903}