Convert AdvancedProtectionStatusManagerTest to IdentityTestEnv
Inject dependencies into AdvancedProtectionStateManager instead of
passing a Profile* and letting the class access to its dependencies
internally (this allow for easier injection in tests).

Change AdvancedProtectionStatusManagerFactory to return the same
instance for off-the-record Profiles as for regular Profile (thus
eliminate static methods on AdvancedProtectionStateManager) and
fix the getter to take a Profile* as parameter.

Add unit test for AdvancedProtectionStatusManagerFactory to ensure
the correct instance is returned for off-the-record Profiles (and
remove uses of TestingProfile in AdvancedProtectionStatusManagerTest).

Change RequestsAdvancedProtectionVerdicts() method to not be static
(as the code to get the correct Profile is now in the factory) and
remove IsUnderAdvancedProtection() static method as the getter can
be used instead.

Since the AdvancedProtectionStateManager instances are only created
by the factory in production code, there is no need to check if the
profile is off-the-record or not (the factory does the check) and
it is possible to use a SequenceChecker instead of checking for the
UI thread (since the factory will create the instance on the UI
thread/sequence anyway).

Fix client code to uses the factory to call the method that were
previously static methods.

Bug: 984487
Change-Id: I8f6a57afeee438419b6e4db59147164d5471fb9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1707353
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Nathan Parker <nparker@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683945}