Navigation: Clarify loading assumptions in unit tests.
The current behavior is to have the NavigationSimulator handle only the
navigation itself and leave the loading to unit tests. In practice a
lot of code works by "accident", never stopping load between different
navigations, etc.

To make the overall approach about loading more manageable, we switch
the paradigm to finish loads by default. This CL is mostly about
maintaining tests that rely on the load being ongoing after a navigation.
For these, we introduce explicit parameters in the NavigationSimulator.

Bug: 979200
Change-Id: I50bf458c487023843d01f753bc5a5c309dc094b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1709582
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Commit-Queue: Arthur Hemery <ahemery@chromium.org>
Cr-Commit-Position: refs/heads/master@{#683953}