[Autofill Assistant] Enable support for requiring a billing postcode.
This CL enables support for requiring a billing postcode in the
Payment Request, based on a flag in the service proto.
If the flag is set to false (default), then a postcode is not required
to complete a Payment Request. If the flag is set to true, and the
billing address doesn't contain a postcode (based on regional settings),
an error message is shown and the flow cannot be continued.

(cherry picked from commit 391f98ff40e178ecd03f1f0f71e9c3024f88ef73)

Bug: 990298
Bug: b/133731582
Change-Id: Ie6f0199b9dc558d642f1bc4c1dd4faf301d6d512
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1722940
Commit-Queue: Marian Fechete <marianfe@google.com>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#682218}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1735453
Reviewed-by: Jordan Demeulenaere <jdemeulenaere@chromium.org>
Cr-Commit-Position: refs/branch-heads/3865@{#179}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}