[Autofill Assistant] Add WaitForDocument action to check document ready state.
This change adds an experimental action WaitForDocument for waiting for
the main document or a document in a frame to reach a certain
ready-state. This is similar to what happens implicitly in the click
action, but gives us more control and better understanding of what's
going on.

This is experimental, because it's unknown whether and how much it'll
improve the scripts. The action has been designed to allow extra control
over its operation and report the start and end ready state, for
monitoring and debugging.

This is a cherry-pick of http://crrev/c/1713519 into the M77 branch, approved on http://crbug/990361

(cherry picked from commit 8910940e340843af591eb603a692e22bee42dde6)

Bug: b/137176709
Bug: 990361
Change-Id: Ief95a5f7644e243625f698385010dfdc08b1aa94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1713519
Commit-Queue: Stephane Zermatten <szermatt@chromium.org>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#682334}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1735462
Reviewed-by: Jordan Demeulenaere <jdemeulenaere@chromium.org>
Cr-Commit-Position: refs/branch-heads/3865@{#184}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}