[merge to 77] desks: Fix overview highlight related issues.
Fix highlights not being positioned correctly if the shelf is not bottom
aligned. This is because we use the mini-views layer target bounds,
which are in local bounds, so we need to account for the parents target
bounds, in this case the desk bar view.

Fix highlights not repositioning after adding an item.

Fix being able to highlight the new desk button even if it is disabled.

Also fixes a crash when the highlight is on mini view, and a desk other
than the highlighted one is deleted.

Also the rounded corner of the highlight matches the new desk button
when it is highlighting it.

(cherry picked from commit bf681b5e21a23a70b02529c9e85d4521cf356ec1)

Test: added tests
Bug: 988183, 988134, 988126, 989795
Change-Id: I6a44ded06bce1089946dcfdc793ac4b35be4ca4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1722193
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#683695}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1737374
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/branch-heads/3865@{#209}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}