[iOS] -ChromeIdentity.hostedDomain cleanup
When sign-in with a new gmail account, the Chrome identity was
initialized with no hosted domain since the SSO profile was not fetched
yet.
The property was never updated later. Even when the AuthenticationFlow
was fetching hosted domain using the AuthenticationFlowPerformer.
During the sign-in process

To fix the issue, -ChromeIdentity.hostedDomain is remove. ChromeIdentity
stays just a proxy for SSOIdentity.
ChromeIdentityService::GetCachedHostedDomainForIdentity() is added to
get the hosted domain.

Bug introduced with:
  + crrev.com/c/1687254
  + crrev.com/i/1434627
  + crrev.com/c/1687254

Related to:
  + crrev.com/c/1732084 (adding new API)
  + crrev.com/i/1560989 (adding new implementation)
  + crrev.com/c/1732085 (switching implementation)
  + crrev.com/i/1560990 (old implementation cleanup)
  => crrev.com/c/1732101 (cleanup API) <=

Bug: 987380
Change-Id: I77b4a1428ce43aca81ae93a0875b3052be97bea5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1732101
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684258}