Add boolean indicating whether cross browsing instance frame lookup is allowed.
WebView and extensions allow frame lookup across browsing instances. We wish
to enforce Same Agent restrictions in bindings but for these cases we
cannot. This change adds the ability to mark frames as having
CrossBrowsingInstanceFrameLookup. A folloup change will add a SECURITY_CHECK
to the bindings interface for this.

BUG=970360

Change-Id: I76334a82644161f7b44b07a32847514bf20acecf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1708752
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684460}