Do not limit concurrent readbacks in TabContentManager
The maximum number of concurrent readbacks was capped at 1. When a
readback cannot be returned because the tab is no longer attached to
the view, all the subsequent readback would be abandoned, causing
blank or stale thumbnails in Grid Tab Switcher. When this happens,
the symptom persists until the tab is reattached to the view or a
restart. The tab can be reattached to the view by switching to it,
but there is no way for a user to know which tab to switch to.

This CL removes the limitation. The maximum number of concurrent
readbacks is still limited by the number of tabs, since a tab can only
have one readback. Removing the limitation should not increase the
system loading because the extra concurrent readbacks would be in
pending state waiting for their tabs to be reattached.

(cherry picked from commit f8ba2a26188020c963f68c8d2d9140ebe2341764)

Bug: 986047
Change-Id: I741159e208757839bc25a9ead386a464e4759736
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1713847
Commit-Queue: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#683848}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1740268
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/branch-heads/3865@{#249}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}