[CrOS Cellular] Fix flash of "missing SIM" in Cellular settings.
This UI jank was introduced by [1], which migrated from using
networkingPrivate to the NetworkConfig Mojo service.

The issue was that when deviceStateProperties_ is null, hasSim_()
returns false, so the "missing SIM" section is not hidden. This CL
updates the logic to ensure that this is not the case.

I also added an OWNERS file for this directory, since I forgot to do so
when adding the rest of them in [2]

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1662636
[2] https://chromium-review.googlesource.com/c/chromium/src/+/1532905

Bug: 985678
Change-Id: Ifbf07702869e854862a3296b83475def0c534ad2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1732429
Auto-Submit: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684560}