DL: Copy existing WPTs to rendersubtree directory
Copy tests in lock-after-append/ and activation/ to the
rendersubtree directory (and for some tests, further organized
to subdirectories) to later be changed to use the rendersubtree
attribute. This CL only moves the tests, no changes
to the contents (except for relative paths etc).

This CL doesn't move tests under lock-before-append/ (some are
the same as tests in lock-after-append/, so need a further look),
sizing/, and paint/ (will be moved just before content-size CSS
property probably).

Bug: 991095
Change-Id: I61356a84af0bae9f250badf766ad283833463846
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1738413
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684614}