Prevent potential float cast overflow in SVGSMILElement
There was a possibility that the RepeatingDuration of an element
would be infinite, which is a perfectly acceptable value, would
try to do a division which made a lot less sense.

SVGSMILElement::CalculateAnimationRepeat now has a check to make
sure the number is actually something worth dividing with.

Bug: 990257
Change-Id: I43032e350bb199375feeae0f6e12b179b77b15f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1735368
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Edvard Thörnros <edvardt@opera.com>
Cr-Commit-Position: refs/heads/master@{#684664}