Web Share Target for TWA 1/2
This CL refactors CustomTabActivityInitialPageLoader to suit the needs
of Web Share Target implementation. The issues addressed are:

1. The first intent is handled in InitialPageLoader while the subsequent
intents (received in onNewIntent) are handled elsewhere. Thus the code
that picks up a sharing intent would be also scattered. This CL adds
onNewIntent() method to InitialPageLoader and renames it to
CustomTabActivityIntentHandler.

2. In TWA mode we'll have to override intent handling. There is
currently no way for that, so this CL introduces one: a Strategy interface
is extracted out of IntentHandler and injected into constructor.
IntentHandler does work that presumably should never depend on the mode,
like checking session validity and waiting for initial tab, then calls
the strategy.

3. For initial intents we parse the extras in CustomTabIntentDataProvider,
for new intents we parse them directly. With sharing this duplication would
get worse. This CL reuses CustomTabIntentDataProvider for new intents.

How this helps the share target implementation can be seen in the next CL:
https://chromium-review.googlesource.com/c/chromium/src/+/1724686


Bug: 985331
Change-Id: If92f755bfc302aea31af730c241ac652221a82fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1715374
Commit-Queue: Pavel Shmakov <pshmakov@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684720}