mediastreamaudiosourcenode-routing.html is not flaky
This test should not be flaky anymore (because we sort the tracks by id
as expected by this test).  And it looks like the WPT bots have
updated to a version of chrome that has the fix.

This test also runs correctly on the blink/chromium bots now too.

Let's try to remove this expectation again.

Bug: 989493
Change-Id: I74cb4239cda9f8f6e411c2db1a636c0d1841699c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1737820
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684759}