Force download of USS sync datatypes upon configuration
As opposed to directory datatypes, non-blocking (USS) datatypes rely on
DataTypeController::RegisterWithBackend() upon configuration to
determine whether synced data is locally available, and use that to
decide whether a download should occur.

For the case where a datatype has been "paused" (usually via a
precondition failure kMustStopAndKeepData), reenabling the datatype
should incur in a immediate download.

Note that this aligns with the historic behavior for Directory
datatypes, which read the is-initial-sync-done state only during
startup and injected in the constructor to DataTypeManagerImpl, but
later reconfigurations don't honor such state.

Bug: 991519
Change-Id: I581ee4b42f5a6e98a672120d0d78763f6aaaeb33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1741685
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684798}