[Signin]: Persist the cached cookie accounts in GaiaCookieManagerService.
The Butter account is not known immediately on startup. Instead, we need
to wait for refresh tokens to be loaded from disk and for a
|ListAccounts| network call to determine the cookie account.

To avoid the need to wait for a network call, we persist the cached
cookie accounts in |GaiaCookieManagerService| in a pref. On startup, we
initilaze the accounts using the pref but consider them as stale, so
that they will get refreshed as soon as possible if there is a network
connection. This way the butter account is immediately available on
startup, even if the user is offline.

Bug: 990286
Change-Id: Ifc0406aff726385367a9f6bdc20dd9e636725fe5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1731940
Commit-Queue: Monica Basta <msalama@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684807}