[webauthn] Add getCredential to DevTools Domain
This patch adds the `getCredential` command to the WebAuthn DevTools
Domain. This allows retrieving a single credential by ID.

This is one in a series of patches intended to create a Testing API for
WebAuthn, for use in Web Platform Tests and by external webauthn tests.

For an overview of overall design, please see
https://docs.google.com/document/d/1bp2cMgjm2HSpvL9-WsJoIQMsBi1oKGQY6CvWD-9WmIQ/edit?usp=sharing

Bug: 922572
Change-Id: Id18716cfab8f63501797a730818ee505ddcab067
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1742294
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Nina Satragno <nsatragno@chromium.org>
Cr-Commit-Position: refs/heads/master@{#684968}