[Payments Autofill] Refactoring OnDidGetRealPan() to take UnmaskResponseDetails struct.
Instead of passing a std::string for the real pan response, this CL is replacing it with a struct. The reason for this is that future projects will have additional data in the response (e.g. tokenized cvc and fido enrollment challenge). The struct will make it easier to append to.

Bug: 949269
Change-Id: I3724ca885871e8f4dbeb9d85b3c01995136b9103
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1719969
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Jared Saul <jsaul@google.com>
Commit-Queue: Manas Verma <manasverma@google.com>
Cr-Commit-Position: refs/heads/master@{#684987}