[SmartLock] Migrate GetRemoteStatus metric entirely to UnlockManager.
This fixes the success bucket under-reporting described in
crbug.com/987315, and also simplifies the code.

This is a first step to fix crbug.com/987315 before a subsequent CL
will split the metric into one that measures the success of finding
and connecting to the and one that measures the success of fetching
the remote status connecting.

Bug: 987315
Change-Id: Ia702978b07a47958e8457cef5632b4bd3086efbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1717538
Commit-Queue: Ryan Hansberry <hansberry@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685013}