Send recently-set SameSite-Lax-by-default cookies with "unsafe" requests
This implements an (ideally temporary) intervention which allows cookies
that don't specify SameSite and get defaulted into SameSite=Lax mode
(under SameSiteByDefaultCookies) to be sent on requests with "unsafe"
http methods (e.g. POST) in addition to the usual Lax-mode restrictions,
if they are recently-created (within some threshold time delta).

This adds a CookieSameSite enum value LAX_MODE_ALLOW_UNSAFE which is
applied to default-Lax cookies that are recently-creatd. If a cookie has
effective SameSite of LAX_MODE_ALLOW_UNSAFE, it is sent with a request
if it meets all the Lax-mode requirements except for "safe" http method.

Bug: 990439
Change-Id: I5a650383699562a71ddf01447d3e23c6836c6362
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1734031
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685187}