Allow length interpolation to fail when comparisons are involved
Current interpolation code assumes that a specified length can always
be converted into a |CSSLengthArray|, which is no longer true if the
length calculation involves comparison functions min/max().

Hence, this patch allows |CSSPrimitiveValue::AccumulateLengthArray| to
to fail, so that interpolation can fail gracefully.

This patch is preparation for min/max implementation. Note that CSS
comparison functions are not implemented yet, so there is no real code
path that fails such interpolation at this moment.

See the all-in-one implementation of min/max for reference:
https://crrev.com/c/1737874

Bug: 991672
Change-Id: I1b5fe2d324f0836b1e21f6ef12abe21e1d7307c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1743135
Reviewed-by: Emil A Eklund <eae@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685302}