[Switch Access] Create preference and slider for keyboard auto-scan time
Added a preference for the auto-scan speed specific to navigating
within the on-screen keyboard. Added a slider to the Switch Access
settings page to adjust this preference. The new preference and slider
are not yet connected to the behavior of Switch Access (so the auto-scan
time in the keyboard will not yet change to reflect the value of the
preference).

Test: With emulated Chrome OS on Linux and both the Switch Access and
text input command line flags enabled
(--enable-experimental-accessibility-switch-access and
--enable-experimental-accessibility-switch-access-text), manually
checked the slider for keyboard scanning speed in Switch Access
settings.

Screenshot of the change in chrome://settings:
https://screenshot.googleplex.com/yoxXvHa67F1.png
This change adds the slider for 'Keyboard Scanning Speed'
in the 'Auto-Scan' section.

Bug: 958081
Change-Id: I24550276bd6bfe360e9703ce677ab62c3a83b033
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1728428
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Commit-Queue: Sophie Yang <sophyang@google.com>
Cr-Commit-Position: refs/heads/master@{#685306}