Add prefetch redirect histogram
This CL adds a histogram to count the number of prefetch requests that
get redirected vs the number of prefetch requests in general. The goal
is to collect some data to see how often prefetch redirects happen in
the wild, since for privacy reasons, we're interested in changing the
redirect mode of prefetch requests to ether kFollow or kManual.

Bug: 988956
Change-Id: I30ca86dad106fbd35351f3e609a9d60ee01aad31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1743237
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685426}