WebApp: Create ExternalWebAppManager and WebAppPolicyManager subsystems for BMO
This CL instantiates the ExternalWebAppManager and WebAppPolicyManager
subsystems when BMO is enabled. These subsystems are generalised to work
with or without BMO and are needed for general web app operation.

This CL also reorders the subsystems to be alphabetical in the header and
ordered by their interdependencies in the initialisation code.

Bug: 916381
Change-Id: Ie0f146c41bde8355243e0b380dee5e1c642596e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1745672
Commit-Queue: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685459}