[Direct Actions] Add simple metrics to track usage.
This change adds metric to track calls to:

ChromeActivity.onGetDirectActions, which is made by the system when the
user triggers the assist app while Chrome is the foreground app.

This is mapped to a user action.

ChromeActivity.onPerformDirectAction which is made by the system when the
user triggers the assist app while Chrome is the foreground app. This is
mapped to a enum histogram, to be able to track the
different outcomes of such a call:
- the action was not recognized
- the action was recognized and performed and which action that was

Since the set of direct actions is open, there's a
catch-all "other", which shouldn't be triggered, as all actions
allowed by the DirectAction feature are part of the histogram
value.

This is the cherry pick of http://crrev/c/1736354 into M-77, which was approved in http://crbug/991530 

(cherry picked from commit df18ba918fc1a0418434540b369c01c757a4d5b0)

Bug: 959841
Bug: 991530
Bug: b/138918753
Change-Id: I833bfbeb03e832e2c770e104685b34afb56bd55f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1736354
Commit-Queue: Stephane Zermatten <szermatt@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#684290}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1745480
Reviewed-by: Stephane Zermatten <szermatt@chromium.org>
Cr-Commit-Position: refs/branch-heads/3865@{#297}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}