Handle disconnected ABSNs that have been stopped
Normallly UpdateSchedulingInfo makes sure the right things happen when
a node is stopped.  However, if the node is also disconnected,
UpdateSchedulingInfo may never get a chance to do anything because the
node is no longer in the graph.

Hence HandleStoppableSourceNodes needs to handle this case.  If the
node has been scheduled to stop and the current time is past the end
time, run all the code to finish the node.  This doesn't guarantee
the node will be collected right away, but at least we no longer
wait until the end of the buffer to enable cleaning up of the node.

Previously, we'd wait to do the cleanup at a time after the entire
duration of the buffer.  This allows nodes to be collected a bit sooner.

Bug: 991491
Change-Id: I4ef86c302bce41be81b0b0e9e69cb33f3c1ddce9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1745486
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685610}