Add delayloads for full browser test libraries
Add delayloads for browser_tests, interactive_ui_tests, and
content_browsertests. This change also creates a common config to
define the delayloads so that all the different binaries (test and
product) are kept in sync. Two lists are maintained: one for all
binaries and one that cannot be included in chrome_child.dll or other
test binaries which are used within sandboxed processes.

This change also cleans up the delayloads unittests a bit.

Bug: 990943
Change-Id: I7d0e96bc1b8af8cf08b84e368218e27c53ce4d12
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1735901
Reviewed-by: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Commit-Queue: Cliff Smolinsky <cliffsmo@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#685630}