[VizHitTesting] Add new bucket to VizHitTestResultsMatch
When a frame is loaded it may change the layout which causes the hit
test data on the viz side being outdated. As a result, we may see a
mismatch between viz and blink hit test results. This could be very
common on popular sites. See [1] for details.

[1] https://docs.google.com/document/d/1AlDyVvKtZ_SZZey_76srFqaTM-QPor0pKc6VXE7lEMU/edit?usp=sharing

This patch adds a new bucket to catch it.

Change-Id: I4d9d41c7a6be9d82ab616d2092c2c599e72b6e6e
Bug: 992064
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1743257
Commit-Queue: Yi Gu <yigu@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685752}