arc_net_host: Allow for Shill to expose the kStateDisconnect state
Shill is going to be exposing the kStateDisconnect Service state when
a Service is in the process of disconnecting. This change prevents us
from hitting the NOTREACHED() when that state is exposed. Note that
system_api will have to be uprev'd after CL:1745568 is merged, as
kStateDisconnect is being changed from "disconnect" to
"disconnecting".

Bug: 989361
Change-Id: Ic3ce54663973dabaad38375528e587df75985e95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1746831
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685864}