[NGFragmentItem] Add NGInlineCursor
This patch adds |NGInlineCursor| to traverse fragments within
an inline formatting context.

The plan is to add necessary functions for the current code
base, supporting both the current |NGPaintFragment| and the
new |NGFragmentItem| to ease the transition and also to hide
the implmentation details, based on the idea we discussed at
crrev.com/c/1583036. Unlike the |TextCursor| API, this class
can traverse fragments but not text.

The structure of |FragmentItems| is not final yet, and there
are several cases known to be wrong yet, but it's good enough
for a minimal test as in the unit test.

Bug: 982194
Change-Id: Ibf43ae01c9c5d55242a66d436857f817cb5fc04a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1745118
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685872}