Portals: Flush tap timeout when cancelling gesture detection
In a situation where we're waiting for a timeout after a single tap and
we reset gesture detection, we want to ensure that the ongoing tap
gesture is cancelled, so that the gesture stream to make sense.
Otherwise, we could have a case where we have a GESTURE_TAP_DOWN
followed by another GESTURE_TAP_DOWN, if a down event is sent after
resetting gesture detection.

Bug: 990555
Change-Id: I07410907a8c2737785f3b83a6964f777b6220fcd
Tbr: nasko@chromium.org
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1739982
Commit-Queue: Adithya Srinivasan <adithyas@chromium.org>
Reviewed-by: Timothy Dresser <tdresser@chromium.org>
Cr-Commit-Position: refs/heads/master@{#685994}