Add text presentation arrows to fallback test
The CSS tests use a lot of horizontal and vertical arrows to point at
particular parts of the text. In order to document that the arrow
fallback changes when we move Windows fallback to OOP (from Cambria Math
to Segoe UI Emoji), add the set of text presentation arrows to the
languages and rare glyphs font identification test.

Bug: 976737
Change-Id: If80ee6965b73130aceba41a47174b2ed4a880d04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1748752
Commit-Queue: Emil A Eklund <eae@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686023}