M77: service worker: Check for null context_ in provider host for UpdateUrls.
The crash bug was introduced in https://crrev.com/c/1705478 because
ServiceWorkerControlleeRequestHandler::MaybeCreateLoader changed
such that the |context_| check below was removed:

  if (!context_ || !provider_host_) {
    ... return;
  }
  provider_host_->UpdateUrls(...);

The desired behavior is a bit subtle. The only reason |context_| could
be destroyed while the provider host is still alive is if the provider
host transferred contexts using the context copy mechanism we want to
remove per https://crbug.com/877356. In this state, we end up with a
provider host that's fairly useless: it can't serve container host
messages etc., because there are early returns for !IsContextAlive().
However, it seems scary to have an inaccurate URL in the case of a
cross-origin redirect.  The previous behavior before the CL mentioned
earlier would keep the old URL in the normal case, but in the
|skip_service_worker| case it would attempt to update the URL via
NavigationUrlTracker.

This CL makes the decision to update the URL as far as possible, and
just avoid touching |context_| if it's null.

The real solution would be to remove the context copy mechanism.

(cherry picked from commit c8bb5163bbc278feb2ecbf038a6c3a05784ae7fc)

Bug: 988551, 877356
Change-Id: I692b0c73dd10cf4af24275eb709382b6ebbf73e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1743230
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#685079}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1749483
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/branch-heads/3865@{#323}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}