Android Photo Picker: Fix DecoderServiceHostTest.
Before we switched to multiple queues, it was enough to
check if an item was being added to an empty queue to
figure out whether to start the decoding process. Now we have
to look at the processing queue instead, because things will
move from the high and low priority queues over to the
processing queue when the processing starts, and we don't
want to start another batch when there already something in
progress.

This explains the test error: we were getting multiple batches
started, which confused testing.

Once I fixed that error, the test started erroring out because
we're no longer decoding 20 frames but 10 (to avoid slowdown).
Fixed that also.

Bug: 990128, 656015
Change-Id: I60f7643fe3b2b1d3aebfdceedc8571bac87a651e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1746474
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Finnur Thorarinsson <finnur@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686027}