Add dialog for password leak detection
Implements dialog for desktop that should pop up when user used
leaked credential. The picture used is a placeholder and the dialog
itself is not called anywhere. At the moment it only covers the case
for user who is not saving passwords in Chrome or Google Account.

Bug: 986317
Change-Id: Ia725c1282110b4b6657001d3c58943569b77cc15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1738459
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Rafał Godlewski <rgod@google.com>
Cr-Commit-Position: refs/heads/master@{#686062}