Create PaintLayer for inline elements with backdrop filter
Prior to this CL, inline elements with backdrop-filter did not
get a PaintLayer, because BackdropFilter wasn't in the list
of attributes in CreatesGroup. That is now fixed. This was not
noticed until recently because the backdrop-filter element did
not create a containing block prior to [1].

[1] https://chromium.googlesource.com/chromium/src/+/4be7920fee20fd21f5c4dc432dc18d2579a487fd

Bug: 992549
Change-Id: I79d8777a178eca488bb76b0909ca815467e34fdb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1745600
Auto-Submit: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Commit-Queue: Mason Freed <masonfreed@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686106}