[serial] Make options parameter to requestPort() optional
Unlike Web Bluetooth or WebUSB where developers are expected to filter
devices based on model or capability it is common to not be able to
determine the device connected to a serial port and so many applications
will want to match all ports. This change makes the requestPort()
options parameter optional.

Bug: 884928
Change-Id: If3017c7331020fdb4a62c9b195bef8ba0ef0169e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1749460
Reviewed-by: Ovidio de Jesús Ruiz-Henríquez <odejesush@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686112}