HttpCache::Transaction: Read buffer length to be correctly set
It is possible in case of partial requests for transactions to go from
reading the response body back to reading headers for the next range
and then back to reading the response body.
In some code paths the |io_buf_len_| is set to the headers length and
then used as the length of the read buffer provided by the calling site
which is incorrect. More details in
https://bugs.chromium.org/p/chromium/issues/detail?id=986211#c66

This CL fixes the issue by having a separate member field read_buf_len_
which is only set to length of the read buffer provided by the calling
site in Read().

Manually verified the asan issue reproduced while loading the site
http://mahokubota.com doesn't come after this patch.

Bug: 986211
Change-Id: I223656e7e1eb8906d1e0415fac184faea1042b6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1747389
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Commit-Queue: Shivani Sharma <shivanisha@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686118}