[omnibox] Enable preserve default from fakebox if the input is a URL.
Previously, if the user was starting from the fakebox, type demotion
could result in a query suggestion being the default even when the
user's input was classified as a URL. With this change, we use the
preserve-default mechanism to prevent type demotion in this case. This
almost guarantees that a URL suggestion will be default because the top
local and server suggestions for a URL input are always and likely will
always be URL suggestions.

This change also makes it so that we no longer use type demotion of
relevance when deduping but instead always choose the natural highest
relevance candidate for each suggestion. This ensures that we actually
get the highest relevance suggestion when applying preserve-default.

TBR=jdonnelly@chromium.org

(cherry picked from commit ff56e15623b53b112e860986f6ef874f5437926e)

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1723807
Commit-Queue: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#684984}
Change-Id: If4d53112a50a9884654a6a26c5fb52c494849040
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1749715
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/branch-heads/3865@{#336}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}