Add WebView for WebUI-based tab strip
This is gated by both the buildflag |enable_webui_tab_strip| and a new
kWebUITabStrip feature flag.

The change adds toggle button and new-tab button next to the toolbar as
well. This is still rough (has no animation, shows scroll bars), but
checking this in should make it easier to iterate on. There are also
placeholder icons and hard-coded colors in use for now.

Bug: chromium:989131

Change-Id: I5bc2a20e85c8b98ca882d536f2f7937233304de1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1742934
Commit-Queue: Peter Boström <pbos@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Collin Baker <collinbaker@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686207}