Reland isUVPAA implementation on clank
Currently, clank is using the isUVPAA that gmscore supports, and
the behavior of isUVPAA is intentionally narrowed to return true
only if fingerprint is available. Since gmscore is going to change
the behavior of isUVPAA to return true if fingerprint or screen
lock is available, clank should reland its own isUVPAA implementation
to return true only if fingerprint is available.

This CL is the first of 3 CLs to add isUVPAA implemenration on clank
and remove the flow to call isUVPAA that gmscore supports.

Bug: n/a
Change-Id: I7f0a025b49597fe73e070562bf96921e64c24622
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1739973
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Suzy Li <suzyli@google.com>
Cr-Commit-Position: refs/heads/master@{#686226}