Reland "Move implementation of Mojo method GetTextSurroundingSelection() to Blink"
We can't construct a WTF::String using the default constructor when running
the GetTextSurroundingSelectionCallback directly for those cases where the
SurroundingText is empty, since that will construct a null string instead
of an empty string, causing a crash at the Mojo level due to trying to send
a null message over the wire, which is not permitted.

To avoid this crash in Mojo, make sure we create a WTF::String out of an
empty string, instead of using the default constructor.

Last, this CL also incorporates the required changes to address Daniel's
comments after the original patch landed in CL 1696968 [1], to rename the
newly added Mojo interface to become blink::mojom::Frame.

[1] https://crrev.com/c/1696968/11/third_party/blink/public/mojom/editing/editing.mojom#9

Original change's description:
> Move this method from the content.mojom.Frame Mojo interface into Blink,
> as part of a new blink.mojom.SurroundingText Mojo interface, and add a new
> class there to implement such interface (i.e. blink::SurroundingTextImpl),
> to replace usages of content.mojom.Frame's GetTextSurroundingSelection()
> from the browser process.
>
> Note that this SurroundingTextImpl class still relies on WebSurroundingText
> and WebLocalFrame to keep the change smaller, but this is a temporary step
> that will be corrected once WebSurroundingText has been moved out of the
> public API and into renderer/core, which will happen on a follow-up CL.
>
> Bug: 980151
> Change-Id: I62fb2d5b24a98773c1acddbe1c58217c1a2d53e4
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1696968
> Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
> Reviewed-by: Mike West <mkwst@chromium.org>
> Reviewed-by: Ken Buchanan <kenrb@chromium.org>
> Reviewed-by: Kentaro Hara <haraken@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#679981}

Bug: 980151, 987214, 987191, 989438
Change-Id: Ia32be36c1a7524930ddcecc79e7fac5bcb514a13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1735459
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Cr-Commit-Position: refs/heads/master@{#686348}