Use tree positions for CreatePositionAt{Start,End}OfDocument
Narrator frequently calls ExpandToEnclosingUnit() for page or document
units when moving by words. This call was accounting for a sizable
fraction of each movement step, and about 56% of it was spent creating
the start and end positions. Switching to tree positions to find them
eliminates almost all of that cost.

Bug: 928948
Change-Id: I447aae5f089a27d041c2b173b16bf40629cb64ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1745547
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686426}