AW: add CHECK to understand lack of UMA records
WebView currently has no UMA records. One possible explanation is if the
client_id is always a consistent value. It's hard to check if it's
always a consistent value, but we can check for a few classes of invalid
values (incorrect length, invalid characters).

Tentatively add CHECK()s so we can look for crash data. We'll remove
these CHECK()s when we know the root cause.

Bug: 992250
Test: None
Change-Id: I920061f973a8320bb5b6df14cf2a3a193751e82f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1750346
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Changwan Ryu <changwan@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686519}