Tidy up JavaScriptDialog dependencies and factories.
1) JavascriptDialogViews is moved from //c/b/ui/ to //c/b/ui/views/, so
that it will be able to depend on other code in //c/b/ui/views/.

2) JavaScriptDialog::CreateNewDialog is introduced as the new unified
abstraction for instantiating a tab-modal Javascript dialog. The factory
function is declared in //c/b/ui/, but defined in //c/b/ui/views/ and
//c/b/ui/android/; so callers don't have to directly depend on either.
Existing factory methods are simplified and merged into this.

3) The global function InstallChromeJavaScriptNativeDialogFactory() is
renamed to InstallChromeJavaScriptNativeAppModalDialogFactory() to make
it clearer it is distinct from the tab-modal dialogs.

Change-Id: Ia4c89a70167175de73613ecda0220e389294d741
Bug: 987465
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1743630
Commit-Queue: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686751}