Use TimeDelta in DocumentTimeline::WakeAfter
This is another step in the process of moving the animation engine to
base::Time* types. We're replacing doubles by base::TimeDelta which is
the perfect fit to represent durations.

We are not using AnimationTimeDelta, the wrapper
class we're using to ease the transition, because the animation timers
require a base::TimeDelta anyway so, the conversion from double to
TimeDelta needs to be done anyway.

Bug: 737867
Change-Id: I028fec11be2120548f829a1090e26ee4430a8a8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1746253
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Commit-Queue: Sergio Villar <svillar@igalia.com>
Cr-Commit-Position: refs/heads/master@{#686802}