Ozone/DRM: Send initial bounds with window creation message
Since the bounds affect which device the window is associated with,
it's better to send it in the same message that creates the window.
That way we will not run the blocked tasks with the wrong device. This
also saves one IPC message.

Bug: 620927
Change-Id: Idce8664f35484d14c7a36104b616ed7d8028358e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1752289
Commit-Queue: Saman Sami <samans@chromium.org>
Auto-Submit: Saman Sami <samans@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686891}