[css-layout-api] Convert to LayoutNG.
I begun this as an incremental patch, but ended up being easier to
reason about as a large conversion instead of temporarily duplicating
logic and methods.

This changes the LayoutCustom object to LayoutNGCustom. This inherits
from LayoutNGBlockFlow to pick up the fallback logic.

A new NGCustomLayoutAlgorithm is created, which handles most of what
LayoutCustom::Perform layout did previously. This invokes the
NGBlockLayoutAlgorithm when the fallback logic is required which is a
nicer design.

All tests pass, except for passing additional data being parent<->child
custom layouts which I'll fix later.

Bug: 726125
Change-Id: I9463cacc44a0c2160db0fcea75944f475ea6b547
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1745396
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686959}