Removing cast_framework resources without breaking cast
We had always been removing cast_framework resources before, which worked
because we had a proguard rule that ensured we didn't hit the path which
loaded these resources. Recently, we switched to R8, which subtly broke
the proguard rule which prevents this path. As a hotfix, we stopped
removing these resources
(https://chromium-review.googlesource.com/c/chromium/src/+/1745441), but
this cost us 240kb of binary size.

This change continues to fix crbug.com/985139, while also not costing us
the binary size of the hotfix.


Bug: 992461, 985139
Change-Id: Ie3437289d531c3a9d8ccf8b320443173762239c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1746795
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Sam Maier <smaier@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686972}