Fix sending scrollend on main thread snap
This CL sends scrollend event if GSE is being
handled on the main thread after the snap (if
the snap has started after GSE). It also takes
care of sending scrollend event for snap after
fling.

This is the second try to solve the lsan bot failure.
This CL is exactly the same as previous CL that got reverted:
https://chromium-review.googlesource.com/c/chromium/src/+/1743253

with some modifications to the test wait condition and
input injection instrumentation to solve the leak bots.

TBR=bokan@chromium.org

Bug: 907601
Change-Id: If877310b966040606062bcc72e8d12ffd19d013f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1753679
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Commit-Queue: Navid Zolghadr <nzolghadr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#686993}