Call the SecurityEventRecorder for non-sync reuses.
A new security event service was created to be able
to record all gaia account security events. This change
depends on cr/1693188. This CL calls the SecurityEventRecorder
from the ChromePasswordProtectionService on PasswordReuseLookup
with verdict. This CL also logs when a user clicks on "Protect
Account" button on the modal warning dialog.

Bug: 914410
Change-Id: If4d13bc3f2fa5d159d11817eead2cdc3ddcf7b39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1711201
Reviewed-by: Hector Carmona <hcarmona@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Markus Heintz <markusheintz@chromium.org>
Reviewed-by: Nathan Parker <nparker@chromium.org>
Commit-Queue: Bettina Dea <bdea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687066}