Add initial impl for CanApplyOptimization in OptGuideKeyedService
Note that this currently just checks the hints and filters for whether
an opt type can be applied on the page load and the target is just
passed through.

We add an initial interface in //components/optimization_guide since we
expect that there will be callers coming from //components (i.e. previews)

Bug: 969558
Change-Id: I465feed24dbedb90a793c3b2151896f5923b4d09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1747972
Commit-Queue: Sophie Chang <sophiechang@chromium.org>
Reviewed-by: Doug Arnett <dougarnett@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687121}