[threadpool] Query OS version info at USER_VISIBLE priority for chrome://version.
A BEST_EFFORT task can be preempted when resources are needed for higher
priority tasks, for example during startup. That means that it is not
an appropriate priority for a task that produces a user visible result.

The appropriate priority when the result of a task is visible to the
user is USER_VISIBLE.

Bug: 990313
Change-Id: Id5de1feac5db052f5069480baf17576c68bd3f08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1752631
Commit-Queue: François Doray <fdoray@chromium.org>
Reviewed-by: Hector Carmona <hcarmona@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687252}