grit: improve unhandled node error messages
Currently if a node parsing logic fails in an unexpected way, we
just show the exception name.  This isn't super useful by itself
as it lacks specific context, so include the exception value too.
In the case of missing files, this will be the filename that the
code was unable to locate.

Bug: 983071
Test: `./grit/test_suite_all.py` passes
Change-Id: Id9b9fd3370c648099ded6bafa816323bdf0de96d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1755392
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687256}