Foreground service: Fix a crash in OnePlus devices.
OnePlus devices uses modified Android OS and track analytics when foreground
service is stopped. This introduces a crash in the process that runs
ActivityManagerService, and send a signal to app process and triggers
NullPointerException.

This CL tries catch this exception, and make call sites to use the new
utility function.

TBR=mlamouri@chromium.org

Bug: 992347
Change-Id: I8ec0e6b1ccb1a23469369811f3f74bf83dbb4612
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1753675
Commit-Queue: Xing Liu <xingliu@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687393}