[ios] Switches a callsite back to PostTaskWithTraits().
This code landed on trunk as a call to PostTask(), but that method
signature did not exist in M76 and M77.  Switching back to the legacy
API will let us cherrypick this CL to the branch and fix the build.
Afterwards, this CL can be reverted on trunk.

BUG=994313

Change-Id: Id691ea0f171f2ac3095a5a446a5dc2fed1364d01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1756158
Reviewed-by: Kariah Davis <kariahda@chromium.org>
Reviewed-by: Peter Lee <pkl@chromium.org>
Cr-Commit-Position: refs/branch-heads/3809@{#1018}
Cr-Branched-From: d82dec1a818f378c464ba307ddd9c92133eac355-refs/heads/master@{#665002}