service worker: Add ServiceWorkerOnUI feature flag.
Flag to control whether the service worker code in the browser process
(the "service worker core thread") runs on UI thread.

This is not in any shape that works yet, but it will help make code more
readable as the feature is built out.

Bug: 824858
Change-Id: Icb5ffb0dbfbf18bc0a9a03c845fb28dbd284da7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1757248
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687538}