Fix ARC++ screen capture.
ARC++ screen capture is returning an empty buffer with OOP-D enabled.
This is caused by a synchronization problem in
ArcScreenCaptureSession::CopyDesktopTextureToGpuBuffer() where the
CopyOutputResult release callback is run immediately after GL commands
for texture copy are issued.

The release callback would return resources on the same GPU channel as
ArcScreenCaptureSession before OOP-D so no explicit synchronization was
needed. The display compositor is moved out of the browser GPU so the
resources are no longer returned on the same GPU channel. Include a
SyncToken from the browser main thread context with the release callback
to avoid this.

Bug: b/138331095, 994336
Change-Id: I79f417775a259dba0d1f1c586600856c14c2940f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1756704
Commit-Queue: kylechar <kylechar@chromium.org>
Reviewed-by: J Kardatzke <jkardatzke@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Yoshiki Iguchi <yoshiki@chromium.org>
Reviewed-by: Yury Khmel <khmel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#687645}