Detects/breaks a client redirect loop with DeferAllScript for a specific site
Navigating to britishairways.com with DeferAllScript enabled for that site
causes a client redirect loop. We don't plan to enable it for that site but
we would like to detect and break this kind of loop if it does happen.
The simplest place I found to detect and break the loop is in
NavigatorImpl::OnBeginNavigation() which has access to the Referrer URL
of the previous navigation along with the new URL and if they are the
same, it is a loop.


(cherry picked from commit b3b99bc47bf062c3e622776cfd265591e1df822a)

Bug: 987062
Change-Id: I24280584775f24f0b1a7e6db36c07e7135f31a5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1739493
Commit-Queue: Doug Arnett <dougarnett@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#685605}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1758924
Reviewed-by: Doug Arnett <dougarnett@chromium.org>
Cr-Commit-Position: refs/branch-heads/3865@{#435}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}