Service worker update checking works when OnComplete arrives later
ServiceWorkerSingleScriptUpdateChecker used to be confused that the script is
identical when the new script is truncated and OnComplete() is received after
the end of the body. This fixes it by calling CompareData(nullptr, 0) instead of
Succeed(kIdentical), which is the same with a fix in the past
(https://crrev.com/c/1714584).

Bug: 995146
Change-Id: I06b324ed0b999addb49d232901beeab67c93e2b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1760629
Commit-Queue: Makoto Shimazu <shimazu@chromium.org>
Auto-Submit: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#688041}