[PWD Editing Android] Add password unmasking icon
Add an icon to PasswordEntryEditor layout so passwords could be unmasked
and masked, like in PasswordEntryViewer. The unmasking/masking logic
is going to be implemented in a future CL. The positioning of the icon might
slightly change in the future, but it's going to stay in this fragment.

Bug: 377410
Change-Id: I9fbd6295d23901b1c825d2758c56d990656d2052
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1748894
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Ivana Zuzic <izuzic@google.com>
Cr-Commit-Position: refs/heads/master@{#688062}