Add font_size and script to the hashing of ShapeRunWithFontInput
This CL replacing the usage of the address of the SkTypeface by it's
uniqueID.

It's adding the script and font to the hashing function since text is
often rendered with fonts of different size. We also observed many
collisions while rendering the same text but with different text
range.

The previous code was working correctly since the HashingMRUCache is
looking for every member with the same hash. It is reducing the amount
of collisions. Under the hood, HashingMRUCache is using an
std::unordered_map.

R=msw@chromium.org

Change-Id: I916798ccd218be2f959aff272768c05633138140
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1710131
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Michael Wasserman <msw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#688105}