[LayoutNG] Implement basic float painting in NG block painter
This patch implements basic float painting in
|NGBoxFragmentPainter| when children are not inline.
1. Change |PaintFloats| to work without |NGPaintFragment|.
2. Apply the |kFloat| paint phase logic in |PaintObject| to
   when |ChildrenInline| is false.

The block logic is very similar to the inline logic, but this
patch does little work to share because the inline logic will
need to revise by the NGFragmentItem work.

Fixes ~200 tests, failures is down to ~500. There are still
some floats failures and they seem complex, such as when
floats are under the engine boundaries. They need to be
investigated in future patches.

There should be no behavior changes unless NG painter is
enabled for block boxes.

Bug: 988015
Change-Id: I90a152ee31a6db5d69d611a097e28af0823a47a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1757988
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Cr-Commit-Position: refs/heads/master@{#688123}