[M77 Merge] cros: Stop launcher from being blank on secondary monitor
Previously, the contents of the launcher would not appear if the bottom of the
secondary display was not aligned with the bottom of the primary display as a
condition for when to fade in the launcher. This change will always attempt to
fade in the launcher on a state change to a non-closed state. The fade-in
animation will not be performed unless the opacity of the launcher contents
is < 1.

TBR=newcomer@chromium.org

(cherry picked from commit 4803c8a3f2782b367b7894c822e0cdd31255e9f6)

Bug: 991297, 990888
Change-Id: Ia745be8180a6a2531fbdbe32b96bd9f5b0de4b2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1752867
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#687087}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1761293
Reviewed-by: Matthew Mourgos <mmourgos@chromium.org>
Cr-Commit-Position: refs/branch-heads/3865@{#480}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}