[Android] Generalize the PWM onboarding dialog
The leak detection dialog follows the same UI pattern, thus the
onboarding dialog code can be reused.

This is mostly a rename CL, the only difference being that the dialog
now receives the drawable id from the bridge, which is meant to
allow for drawable customization for different dialogs.


Change-Id: I3178bb1040e9d2495464063f642a1357117ca479
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1751252
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Commit-Queue: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/master@{#688440}