[PWD Editing Android] Change the place where the observer is removed
Observer for the password list in PasswordEntryViewer should be removed while the view still
exists, to avoid calling findViewById on an empty View.

Previously, the observer was removed in onDestroy() which happens when the View is already
gone and passwordListAvailable could call setRowText which uses findViewById.

Bug: 994010
Change-Id: I10fbea31dd6faad777b24ec8783c7572e91484ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1761589
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Commit-Queue: Ivana Zuzic <izuzic@google.com>
Cr-Commit-Position: refs/heads/master@{#688490}