Fix an ARC test flake when the App Service is enabled.
ArcAppsPrivateApiTest.OnInstalled creates a FakeAppInstance, but the instance
is torn down without clearing the underlying connection. This causes test
flakes where the App Service is still trying to resolve callbacks against the
FakeAppInstance that is now destroyed at the end of the test.

This CL fixes the test by creating the FakeAppInstance as a member variable on
the test, and adding a TearDown step to properly clean up the connection.

BUG=992839

Change-Id: I748fcc6c64d66392848ad0c49ec7916e7e8686b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1757950
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Yury Khmel <khmel@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#688574}