Integrate always on top into commit logic in exo
Along with the changes to move always on top to ARC attributes in
Android, this CL integrates always on top into commit logic in exo.

For PIP, as we want to ensure that always on top is toggled after
window state and bounds changes have been done, we update always
on top in onPostWidgetCommit.

BUG=b:138768141
BUG=b:138741400
TEST=Manually confirmed always on top works properly with ArcCompanionLibDemo
TEST=Manually confirmed YouTube doesn't flicker when entering PIP

Change-Id: I22dbad22985bd6e0249b207509e1b0563486edbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1732703
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Eliot Courtney <edcourtney@chromium.org>
Commit-Queue: Kazuki Takise <takise@chromium.org>
Auto-Submit: Kazuki Takise <takise@chromium.org>
Cr-Commit-Position: refs/heads/master@{#688670}