Update GpuIntegrationTest minidump handling
GpuIntegrationTest will catch errors from test runs, then nicely format the
exception. As well as symbolize minidumps.

However it is possible for this to lead to a second nested exception. If startup
failed, there won't be a browser object to attempt to find minidumps from.

This updates the handling to only attempt to symbolize the minidump if the
browser exists, and therefore succeeded at startup.

TEST=pixel_tests

Bug: 993379
Change-Id: I79f575556c5785b4aa6e6c6c031067cb1c5f993a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1761773
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/master@{#688777}