Reporting compositor latency of each interactions to UMA
The latency of each stage for missed and non-missed cases is reported to UMA for each of the
existing interactions:

CompositorAnimation
MainThreadAnimation
PinchZoom
RAF
TouchScroll
WheelScroll

For example, the histogram for CompositorLatency.MissedFrame.EndCommitToActivation shows the
duration between commit and activation for the frames which have been missed. This CL in
addition to what was reported will report histograms below:

CompositorLatency.MissedFrame.CompositorAnimation.EndCommitToActivation
CompositorLatency.MissedFrame.MainThreadAnimation.EndCommitToActivation
CompositorLatency.MissedFrame.PinchZoom.EndCommitToActivation
...

Which reports the duration of commit to activation for the frames which are missed while
CompositorAnimation or MainThreadAnimation or PinchZoom interactions were active.

The intention of adding these breakdowns to compositor latency is to create the possibility
of comparing different interactions in terms of latency. In order to prioritize various
possible optimization of each interaction, we would need to know how each of them impact the
performance. This would also help with figuring out which of the interactions has the most
missed frames or latency increase.


Bug: chromium:976757
Change-Id: Icdd5efcaf25e91d161ccfe3ab4dcfdb9bdc29932
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1751902
Reviewed-by: Behdad Bakhshinategh <behdadb@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Commit-Queue: Behdad Bakhshinategh <behdadb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#688779}