Show blank pages underneath HTTP auth prompts
Originally, we designed HTTP auth committed interstitials to show the
error page body from the server underneath the login prompt. This
proved to be problematic for some servers and confusing for users, who
expect to see a blank page. This CL introduces a
LoginNavigationThrottle to rewrite the error page content to a blank
page. The login prompt is shown on top of this blank page.

Trickiness arises when handling prompt cancellation, when we do want
to show the error page content from the server to the user. The
throttle has to differentiate the refresh navigation from other
navigations with auth challenges, and avoid rewriting the response on
the refresh navigation. We also have to avoid re-showing a login
prompt when the refresh navigation finishes. LoginTabHelper therefore
tracks the navigation entry ID and navigation handle ID when
refreshing on prompt cancellation to avoid rewriting the response on
this navigation and to avoid re-showing the prompt.

Bug: 992639
Change-Id: Ife6cd6c3b1e9dba58a7dfee8033c31331133cad3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1761392
Commit-Queue: Emily Stark <estark@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/master@{#688871}