PDF Accessibility: Passing links and images to PdfAccessibilityTree
This CL passes the vector of links and images received from the plugin
process to the PdfAccessibilityTree. The links and images are validated
to check whether any invalid data is received or not.

Tests for all scenarios of invalid data have been added using mocked
data from plugin layer to validate the scenarios.

Bug: 981448
Change-Id: I2587188255f7db4fa6d824064a46095acbfba61b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1753366
Commit-Queue: Ankit Kumar 🌪️ <ankk@microsoft.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#688893}