[Extensions] CHECK that loginScreenUi usage is allowed
This CL adss a CHECK to verify that the calling extension is running in
the sign-in profile, has the "loginScreenUi" permission and the device
is managed by policy. This should already be enforced by the feature
definition for the "loginScreenUi" permission. However, since showing
login screen UI has further security implications, we want to be extra
secure and prevent any side-channels from calling this API.

Bug: 957573
Change-Id: Ia8fcdf083b1e77aa5458c66da12b5616938c04f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1762217
Commit-Queue: Alexander Hendrich <hendrich@chromium.org>
Reviewed-by: Zakhar Voit <voit@google.com>
Cr-Commit-Position: refs/heads/master@{#688936}