[M77]Fix LCP UseCounter and add one for buffered flag
This CL fixes the UseCounter for LargestContentfulPaint by moving it to
the PerformanceObserver::observe() method. Before, it would be triggered
in Performance::getEntriesByTypeInternal(), which would be triggered
only when requesting LCP entries with the buffered flag. While at it, we
move ElementTiming UseCounter to this method too (no important change).
We do not move the EventTiming UseCounter because if 'event' is observed
then we will add to the UseCounter even though it's not shipped yet.
Once issue 994332 is fixed, we could move the UseCounter logic.

This CL also adds a UseCounter for calling observe() with the buffered
flag. We trigger it with any call to observe() where the buffered flag
is set to true.

Bug: 995455

TBR: skobes@chromium.org
Change-Id: I088f90b861fe3d2a9d4071421ebd97e78cf62316
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1763256
Reviewed-by: Nicolás Peña Moreno <npm@chromium.org>
Cr-Commit-Position: refs/branch-heads/3865@{#530}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}