Add linear interpolation to resample input/scroll events.
This CL adds the linear interpolation that resample events to 5ms before
frame time. This is trying to match the resampling behavior in Android.

This CL also removes the "Disable prediction when dt < 0" check. The
check is not valid when we resample to frame_time - 5ms. And according
to UMA Event.Latency.ScrollBegin.Touch.EventTimeToRAFTime, we only have
around 0.05% data that frame_time - last_event_time < 1ms.

Bug: 969697
Change-Id: I5e08b4e8f158e7bae60ef3869c2d56ebadfe4fca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1761310
Commit-Queue: Ella Ge <eirage@chromium.org>
Reviewed-by: Navid Zolghadr <nzolghadr@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689057}