Reland "Implement ProtoChromePromptIPC."
This is a reland of e5f1b06d49463065691dfe49b82f1498dae7bd6f

It was reverted in https://crrev.com/c/1760644 for causing test failures on CI
builders that couldn't be reproduced on other machines.

This patch disables the failing tests. A followup will reenable the tests with
extra debugging to diagnose the failures on the CI builders.

TBR=wfh
TBR_REASON=Reland of test-only proto change that was already reviewed.

Original change's description:
> Implement ProtoChromePromptIPC.
>
> Bug: 969139
> Change-Id: I35d551b6bdb41b4e6bd7689120c21d9d683ad72f
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1733747
> Commit-Queue: Oliver Li <olivierli@chromium.org>
> Reviewed-by: Will Harris <wfh@chromium.org>
> Reviewed-by: Joe Mason <joenotcharles@google.com>
> Cr-Commit-Position: refs/heads/master@{#688357}

Bug: 969139
Change-Id: Iafdc22cc06fbedd5de73bf8bb7881fc46e01f350
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1759154
Reviewed-by: Joe Mason <joenotcharles@google.com>
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Joe Mason <joenotcharles@google.com>
Cr-Commit-Position: refs/heads/master@{#689074}