Ignore inline background images in forced colors mode
Inline background images are painted after backplates are drawn in
forced colors mode. This breaks the ensured level of contrast for such
elements. To account for this, background images are ignored for inline
elements when forced colors mode is enabled and forced-color-adjust is
auto.

Bug: 970285
Change-Id: I609eb1779ed3a9d964cb6165e5452b9d2a9666e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1757685
Commit-Queue: Alison Maher <almaher@microsoft.com>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#689095}