Correct comments for the |request_initiator| parameter in WillCreateURLLoaderFactory.
This CL corrects comments for the |request_initiator| parameter in
WillCreateURLLoaderFactory. The current comments say that navigation requests
are not associated with an initiator which is incorrect.

Also, add a check for a null |render_frame_host| to
ProxyingURLLoaderFactory::MaybeProxyRequest to prevent a null pointer dereference
and fix associated crashes.

BUG=996041

Change-Id: I4a468ded6006c19f0a321d685fa40aa7f6ad1331
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1761607
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689202}