[sms] Use an unblocking InfoBar rather than a blocking modal dialog
Sample:

https://storage.cloud.google.com/chromium-translation-screenshots/d2da91cc316334522197c0b6cb4eb0c8dddc58f5

After doing one round of outreach, we learned quickly that the modal dialog
was too disruptive for developers and an imposition on their flow.

https://github.com/samuelgoto/sms-receiver#opt-in-ux

The blocking nature of the modal assumes there is nothing else to be done,
which quickly proved to be too strict for partners (e.g. some forms have
first/last names input fields that can be entered while one waits for
an OTP).

In this CL, we move our implementation from a blocking UX to an unblocking UX,
 reusing the InfoBar visual language:

https://github.com/samuelgoto/sms-receiver#unblocking-ux

We debated for a while keeping both implementations in the code base, but
decided against it due to the (a) maintenance cost involved, (b) memory
/ strings consequences involved, (c) code complexity introduced and (d)
our intuition that the infobar is a strictly better UX than modal dialogs
for this feature.

This CL:

- deletes the old modal dialog implementation (and its resources)
- creates an info bar dialog implementation
- removes the content/public/SmsDialog interface

Bug: 994874
Change-Id: I1360de9fa0bb10f99df71684e6f617b7c0c4f5cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1751966
Commit-Queue: Sam Goto <goto@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689335}