[Native File System] Quarantine files after save
After an atomic write finalizes, this applies Mark-Of-The-Web on the
resulting file. This is done via the Quarantine component, which on Mac
OS X, simply applies some file metadata.

On Windows, the Quarantine component may call the AttachmentServices
which could start an Anti-Virus check and that may cause the file to be
deleted if malware is found.

Bug: 986588
Change-Id: Iebefe25abbc32501b55e29bc8196f8a049dee2c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1737163
Auto-Submit: Olivier Yiptong <oyiptong@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Olivier Yiptong <oyiptong@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689606}