NativeFS: Extend the restricted directory dialog for file selections
Add strings for the dialog to use when the user is selecting files.
(selection of files within sensitive directories is restricted)
Plumb a flag for whether the user is selecting a directory or file(s).
Upload screenshots and add sha1 files to the repo.

TBR=nasko@chromium.org

Bug: 992627
Change-Id: Ifd5f0a1c048fe2a7a50d3cd8eb5db0e01e0a8d26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1766435
Reviewed-by: Michael Wasserman <msw@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Commit-Queue: Michael Wasserman <msw@chromium.org>
Auto-Submit: Michael Wasserman <msw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689717}