background sync: Prepare for service worker UI thread core.
The thread ServiceWorkerContextCore lives on (the "core thread") will
move from the IO thread to the UI thread when the ServiceWorkerOnUI
feature is enabled.

This makes background sync code aware of the core thread rather
than assuming the IO thread. This enables content_browsertests
for "*BackgroundSync*" to pass when ServiceWorkerOnUI is enabled.

Bug: 824858
Change-Id: I2cbba7fbf9ee2bfabacf82a05b6398260c636ba3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1763273
Reviewed-by: Mugdha Lakhani <nator@chromium.org>
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689906}