Undo some changes to MainIntentBehaviorMetricsIntegrationTest.java
We had problems going on with test flakiness caused when
ChromeTabbedActivity.setupCompositorContent runs earlier than it used to,
typically before NewTabPage is constructed.  This is problematic because the
LayoutManager tries to hide the empty page when the NTP is created resulting in
MainIntentActionType.SWITCH_TABS getting recorded.

With the exception of MainIntentBehaviorMetricsIntegrationTest#testCreateNtp
these seem to have gone away so I'm reverting the changes.

Bug: 982081, 980587
Change-Id: I64809492ed8bfda932f9a93ff8e3329e61e57e51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1749189
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Alex Clarke <alexclarke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#689952}