PDF a11y: Added new heuristic for text runs delimitation
The previously merged text runs heuristic change to support vertical
text runs (https://crrev.com/687072) introduced a bug described in the
linked bug.

This change fixes it. It adds a new heuristic that verifies if the
difference between the angle formed by the first and last character of
a run (the |text_run_angle|) and the angle formed by the current and
previous character (the |current_angle|) is greater than 90 degrees.
If so, the text run breaks.

The file directional-text-runs.pdf has been updated to detect any
future regression.

Bug: 994958
Change-Id: I21fbe389fe5527c6dba9f05a23fd865a5b3a221e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1758833
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Ian Prest <iapres@microsoft.com>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#690011}