[DevTools] Rework the console-worker.js test for flatten=true.
In my rework, I think there could be some race conditions
in the order of the output. I'll do the dry-run for now
and see whether anything manifests. Perhaps it's worth
doing some sorting of the logs, e.g. we could make the
test insensitive to the order of some of the events in
the golden file, by keeping two or three separate logging
streams for each section.

Bug: 991325
Change-Id: Ib9b8fb0732f5efbf9ad619ff12d5a36bccb02a16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1762823
Commit-Queue: Johannes Henkel <johannes@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#690125}