Android: Fix NPE in gesture navigation.
The reported crash shows that BottomSheetController might not be ready
at the beginning when Tab starts getting first overscroll events. This
CL adds nul checks in the interface implementation of NavigationSheet
against the controller to not perform the action until the controller
is ready for use.

Bug: 996184
Change-Id: I6624282caaf79c8d2bd24ce8128aa24f16848cd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1765055
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#690130}