cc: Fix query for RasterTaskSchedulingDelay
This CL fixes the order in which the command buffer queries are created
for the RasterTaskTotalDuration and RasterTaskSchedulingDelay UMAs. The
CL that introduced the latter assumed that the query for it was issued
before the query for the RasterTaskTotalDuration UMA [1]. However, this
was not the case.

[1] https://cs.chromium.org/chromium/src/cc/raster/gpu_raster_buffer_provider.cc?l=630-632&rcl=1eab052d8f52e823b71fd7bdd5fea70a8f95324f

Bug: 995155
Test: None.
Change-Id: Ie7f216c7900c7e99377622c3c8ba11001c2cae84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1769078
Commit-Queue: Khushal <khushalsagar@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#690185}