Use a sibling invalidation set for :nth-selectors.
We used to have a descendant invalidation set for :nth-selectors which
was scheduled on the parent node of those elements potentially affected
by DOM mutations. The problem was that such sets would invalidate a
bunch of descendants of affected siblings in the case where the
invalidation set features are common like tag names. E.g.:

div:nth-of-type(2n) {}

Would have an invalidation set invalidating all div siblings, but also
all their div descendants.

Instead we make the nth invalidation set a sibling invalidation set. It
is still scheduled on the parent node, but pushed as a sibling
invalidation set on SiblingData before we start invalidating the
children. The new NthSiblingInvalidationSet initially sets the
max_direct_adjacent_selectors_ count to the max value to make sure the
sibling invalidation reaches all siblings.

This fixes a performance issue for salesforce.com.

Bug: 992160
Change-Id: Ibe2d28da95286185cdebc7894b029e8741a6216c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1763761
Reviewed-by: Eric Willigers <ericwilligers@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#690296}