CreateNavigation test helper should use a real content::NavigationEntry.
In the near future (see https://crrev.com/c/1737488) we want to avoid
creating sessions::SerializedNavigationEntry objects with a bogus
PageState.  This requires creating a SerializedNavigationEntry based on
a real content::NavigationEntry object.  This CL helps with this goal.

This CL moves the test helper method for creating
SerializedNavigationEntry objects from
//components/sessions/core/serialized_navigation_entry_test_helper.h
to a new
//components/sessions/content/content_test_helper.h

The implementation of the test helper is modified to use a real
content::NavigationEntry object.

Bug: 976055
Change-Id: I0c8400f4a340e3cb03ecdc43fb82cb40d8c29f67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1769657
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#690561}