gpu: Flush context after inserting presentation feedback event query
Flush context after inserting presentation feedback event query so that
we don't keep waiting indefinitely for the query to retire.  While the
query is active we keep vsync on, so also ensure that we don't call the
client vsync callback if it doesn't ask for it.

This is a speculative fix for some perf regressions when presentation
feedback was enabled on the waterfall.

Bug: 892874, 994277
Change-Id: I748b4c2427b0c78f681d73f967c6d6756f7228ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1769730
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/master@{#690594}