Reland "Enable the jankiness profiler on ChromeOS"
This is a reland of [1], originally revert in [2] for causing [3].
The flakes result from not checking for base::TimeTicks::is_null() in
ProfileProvider::OnJankStarted(). The random initial now ticks in
MockTimeDomain resulted in the flake in the test. Fix by adding the
base::TimeTicks::is_null() check.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1692209
[2] https://chromium-review.googlesource.com/c/chromium/src/+/1765134
[3] https://crbug.com/996647

Original commit message:

Enable the jankiness profiler on ChromeOS.

This implements the jankiness profiler on ChromeOS. The ProfilePrivder
observes jankiness through JankMonitor and then sends notification to
each registered MetricCollector implementation class. The profiles are
collected with the JANKY_TASK trigger event.

Bug: 904785
Test: unit_tests.
Change-Id: I36a4bd4f95a35482156253b9d031693d02371440
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1768836
Reviewed-by: Gabriel Marin <gmx@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Commit-Queue: Chinglin Yu <chinglinyu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#690657}