Increase the storage limit of App Cache Memdisk.
There has been a fixed limit on the size of the app cache memdisk
which resulted in failing to load scripts of over 10 MB in
off-the-record mode.
Apart from the unexpected user experience, this could be abused to
detect private browsing.
To reduce this issue, the upper limit is changed to cache storage max
limit.

Bug: 818045
Change-Id: I5901d4a8cdafec2d5afbc8c77db56c9d13399f6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1771809
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#690696}