[Traffic Annotations] Fix --no-filtering with extractor.py
Running the auditor with these arguments would fail:
    --no-filtering --extractor-backend=python_script

This is because the auditor passes a single argument, "./", to
extractor.py, which doesn't know what to do with it. Rather than change
the behavior of extractor.py, this CL makes the auditor pass arguments
that extractor.py knows how to handle (i.e. a list of .cc files).

Bug: 966883
Change-Id: If5d6877ba116ee701157d38b9bc89405f2322180
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1769104
Commit-Queue: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/master@{#690722}