Don't quantize build timestamps on official builds
Official builds - the PDBs and the PE files - often get published to a
symbol server. For PE files the path name in the symbol server is:

    "%s\%08X%s\%s" % (peName, timeStamp, imageSize, peName)

Since the peName for a particular DLL/EXE never changes and since the
size often doesn't change this means that the timeStamp is the only
differentiator between nearby builds. With the strategy of setting the
build timestamp to 5 am of the last commit time it is easy to get
a build that overwrites the previous build on the symbol server. This
happened when build 75.0.3770.143 overwrote all of the PE files from
build 75.0.3770.142, which complicated the investigation of (restricted
view, sorry) crbug.com/964273. This probably happened many other times.
The PDB files were never overwritten which is why this was not noticed
earlier.

When the 5 am quantization was added we were using the current time for
the build timestamps. Now that we are using the last commit time it is
less important to quantize to 5 am so this change removes that
quantization, for official builds only.

An increased number of days where we do multiple builds of one channel
means that this issue is hit more frequently than when the
quantization was initially added.

TBR=brucedawson@chromium.org

(cherry picked from commit 02f0edc3fb60b55db652fce1c298beda75e7d1d9)

Bug: 993509
Change-Id: Ibfac95569b713ede056d3ff070db0c05b4a38c77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1754527
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#687349}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1772551
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/branch-heads/3865@{#597}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}