Hide invisible Tab in GTS layout better
After switching to another tab from GTS, a thumbnail of the original
tab is taken. However, the thumbnail taking would be blocked if the
web contents of the original tab is detached. Therefore, when showing
the GTS-to-Tab transition animation to another tab, we cannot detach
the original tab.

The original tab is supposed to be invisible, but the tab decoration
was not properly hidden.

This CL hides the decoration of that invisible tab.

(cherry picked from commit f9355db345457fdab987846d5474c7e076100bf6)

Bug: 995992
Change-Id: Ia55cc21ddcbbc59674fa8c8268faaddc458cd4a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1764660
Reviewed-by: Mei Liang <meiliang@chromium.org>
Commit-Queue: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#689621}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1773525
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/branch-heads/3865@{#629}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}