Remove a few related uses of NOTIFICATION_SESSION_STARTED
In these files, it was being used to clear notification registrations
to fix http://crbug.com/125276

In NetworkStateInformer, it was doing nothing except removing its own
registration.

In SignInScreenController, the referenced UI object is owned, and
therefore can't have been freed, so the precaution was no longer
necessary.

In ExistingUserController, we can replace with a check to
IsSessionStarted before handling NOTIFICATION_AUTH_SUPPLIED. I don't
understand why this code lives in ExistingUserController to begin with,
and there's a TODO/bug to rework the bodge, but this should preserve
behavior.

Bug: 268984
Change-Id: I219b69997e4873ed60695442a757bd8d549dc00a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1771614
Commit-Queue: Evan Stade <estade@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#690917}