End overview when opening app list (or home launcher in tablet mode).
Previously it was by design that we don't end overview when opening
app list while splitview + overview are both active. But after the
discussion in Issue 991390, we decided to follow the current existing
behavior and let app list end splitview+overview in this case.

Bug: 996869
Change-Id: I15689b5cf61872110254446bdecce3bc3103aca3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1766090
Commit-Queue: Xiaoqian Dai <xdai@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#690973}