[Passwords] Introduce PasswordForm.date_last_used
This CL introduces the field PasswordForm.date_last_used and make sure
it's gets populated when a login is added/updated from the UI or via
Sync. In addition, it migrates the data stored in login database
such that any preferred login has a more recent
PasswordForm.date_last_used than the non-preferred counter-part.

This is in preparation to replace the "PasswordForm.preferred"
field.

The plan is to land this CL for M78, and in M79 we deprecate the
"PasswordForm.preferred" field and use the PasswordForm.date_last_used
field instead to pick the login to show to the user in case
they have multiple logins with different usernames.

This change will simplify the Password Manager logic especially
with the upcoming Butter4Passwords.

Change-Id: Id9af62764052f556f2405c4cb3123b54ba5161ab
Big: 991510
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1768320
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691128}