[FCM laziness] Query persisted messages by prefix.
The infrastructure for persisting messages saves the messages using
appId + senderId as an id to identify the sender entity.
When the message is queried only the appId is available.
Currently the code checks if the messages for appId are saved, which
is always false.
The fix checks if the messages for the sender with |appId| prefix are persisted.

Bug: 989441
Change-Id: Ibb28b49d2da5f899b86365f7b33d82b00d61f34c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1771415
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Richard Knoll <knollr@chromium.org>
Commit-Queue: Tatiana Gornak <melandory@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691159}