[Mfill Android] Fetch favicons by origin directly in filling controller
Before this CL, we would issue favicon requests as soon as we sent the
suggestions to the UI. In practice, the caching overhead we introduced
isn't really worth the code since the favicon service relies on the
history database (and it's caches) anyway.

Therefore, this CL drops the caches and moves the favicon fetching one
layer higher (into the manual filling controller). This simplifies the
code and unblocks the passwords controller from being frame-based.

Additionally, the origin is now part of the request (if it is different
from the frame-origin). This ensures that favicons are origin-exact and
if a PSL-matched suggestion (e.g. from a mobile site) uses a different
favicon, it will use that instead.

Bug: 945300
Change-Id: Iddbc074b0fd463089a9d5ce2e88f0c5490ca362f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1768526
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691172}