[LayoutNG] Fix (one-type) of baseline cache misses.
We don't hit our cache when our baseline requests don't match. This
typically isn't an issue, however whenever we start layout from a
layout-root, we have a high likely-hood of missing the cache due to
this.

This patch (when calling NGConstraintSpace::CreateFromLayoutObject)
passes in the previous baseline request lists in order to increase
the cache-hit rate.

Change-Id: I43190eaaa86a3584581bb5abff937d20e41ce3e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1772524
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691174}