[ios] Refactor BadgeMediator
This change moves the ownership of the state of displayed badges to
BadgeMediator. This will centralize the state in preparation of
the use of PopupMenuTableViewController for badges. Because of this,
BadgeConsumer's API is modified to just notify the receiver about
any updates to the displayed and fullscreen badges.

Bug: 976901
Change-Id: I4d4cf7c5209312c231b46a228810f04fdd9ae1e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1769083
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691220}