[Mfill Android] Update accessory bar when filtering autofill chips
Whenever a user enters characters, the autofill chips are filtered to
only show chips with matching substrings. When scrolled to the end, the
automatic spacing would push off the left-most chips since it received
no signal that new chips might want to take this space.

With this CL, the automatic spacing is recalculated whenever the any
number of chips changes.

Additionally, we scroll to the left-most position so newly filtered
chips are always fully visible.

Binary-Size: Seems to be a quirk, unrelated to the introduced changes (https://crbug.com/998206).
Bug: 997609
Change-Id: Id250a8b74d35689bb24eb03619c262ccce3a059e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1771401
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691223}