[sensors] Remove enabled-by-default flags and bump milestones
This change removes the kGenericSensor because the Generic Sensor API
has shipped. The Blink runtime enabled feature flag is also removed.

The kSensorContentSetting flag is removed as it has been enabled for a
few milestones and there is no longer a need to support disabling it.

The expiration milestone for the kGenericSensorExtraClasses flag is
bumped because the work to ship that feature is still ongoing.

Bug: 750018,796904
Change-Id: I2704ef824e73f7dc95d42466b605bb461931a162
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1761255
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Dan Beam <dbeam@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691252}