Update content change behavior of bottom sheet
The bottom sheet now completely closes before changing content due to
higher priority content showing. This patch is in response to the
changing use case for the bottom sheet: there are many users with
different behavior patterns that don't allow for content changes while
the sheet is open.

In general, content shows via showNextContent(...) after the sheet is
hidden is response to a particular event. The next content in the
queue will be shown at that point.

This patch also updates a few names and fixes a long-standing priority
bug where priority was reversed. This happened because the previous
logic (and test) never resulted in more than one item in the queue at
a time.

Bug: 986310
Change-Id: Ie51baa3617e256277c815a9c1ff7239b58a99187
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1735049
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691262}