Remove old and misleading TODO.
Removing a TODO that is no longer needed now that conversion of empty
SiteInstances to default SiteInstance has landed in
http://crrev.com/c/1693778 . The change actually suggested in the
comment is wrong and I believe should have actually applied to a
site URL check and not a SiteInstance check. The SiteInstances
are expected to be different even if default SiteInstances are enabled
because they come from 2 different BrowsingInstances.

Bug: 958060
Change-Id: I5ae816f96ac097ce361af0ffcd236a43cd93241d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1774103
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Aaron Colwell <acolwell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691273}