[PWD Editing Android] Change the title displayed for an exception site
When a site is saved as one of those for which we want to stop
showing the InfoBar that offers password saving, that site appears
in Never saved list in Settings->Passwords. When a click is
performed on an entry in the list, that entry is shown in a new
Fragment. For this purpose we are using PasswordEntryViewer, which
is also used for displaying passwords. Currently, in both cases
the title says "Saved password", which doesn't make sense for both
cases, so in this CL another title ("Never saved") is used for
displaying a password exception site.

Bug: None
Change-Id: I2ee9533a098b913eccc687912e97fb059b5c26da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1773154
Commit-Queue: Ivana Zuzic <izuzic@google.com>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691279}