[DevUI DFM] Add comments related to Native Resource Splitting for DFM, with minor cleanup.
Native Resource Splitting for DFM (on Android) changes the dynamics of
PAK file usage. For context, key changes are:
* Creating split resource PAK for DFMs
* Downloading split resources on on-demand DFM install.
* Loading split resource PAK *on use*.
* Serving resource data bytes from split resources only in code that's
  deemed safe to do so (e.g., by "gating" code so its reachable only
  when the associated DFM is installed).

Also important are what's *not* done for split resource PAKs (but are
still done for the main resources.pak, which is installed in the base
module).
* Loading split resource PAKs on startup.
* Sharing {file descriptor, memory mapped file region} of split
  resource PAKs with child processes (i.e., split resource PACK are
  accessible from UI thread only).
* Retrieving {file descriptor, memory mapped file region} of split
  resource PAKs from child processes (complements the above).

This CL adds comments to code sites related to loading (at startup),
sharing, and retrieving operations for resources.pak. The comments
state that in general, these are not done for DFM split resources.

The CL also performs some minor cleanup.

Bug: 927131
Change-Id: Ib087b02087355043e6a0983c85345da14e64b4d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1774661
Commit-Queue: Samuel Huang <huangs@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691286}