VideoEncodeAccelerator: Each VEA selects H264 level if it is unspecified or invalid
A VideoEncodeAccelerator (VEA) client needs to specify a proper h264 level in
the original VEA API. If the h264 level is unspecified, Level 4.1 is used by
default. This design causes a failure of recording 4k video with media recorder
because most VEA clients don't specify the level.

This CL modifies the VEA API design. A VEA client is still able to specify h264
level. However, if it is invalid (or unspecified), VideoEncodeAccelerator s
selects one of proper h264 levels for the specified resolution, bitrate and
framerate.

Bug: b:139788862
Test: video_encode_accelerator_unittest on hana and eve
Test: Record 4k video
Change-Id: I807a6fce10929739fe6ab161d2faddfd9e39d3a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1765131
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691468}