media/gpu/v4l2svda: use VideoFrame to wrap output DMAs
A VideoFrame is a better container for the output DMAs, especially if we
add support for ImageProcessor which requires it as a frame abstraction.

Prepare for ImageProcessor support by wrapping the VideoFrame in a
single place instead of passing DMAbufs around.

Bug: 132589320
Test: VDA unittest passes on Minnie in regular and import mode.

Change-Id: I10d0ee8fd749778ba5aba06612a7c4d9c46d4786
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1768478
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691531}