Simplify prefetch + cross-origin prefetch logic
This CL centralizes the cross-origin-prefetch-specific logic to
PrefetchURLLoaderService, whereas previously it was split between
PrefetchURLLoaderService and PrefetchURLLoader. This does not come with
any observable change, however simplifies the implementation in a way
that will make the preload-header-on-prefetch code cleaner.

This CL also eliminates the
PrefetchURLLoaderService::CreateLoaderAndStart method that was called
from the network::mojom::URLLoaderFactory override, leaving only the
URLLoaderFactory override method left. This is much cleaner, as both
were not necessary now that the network service has been enabled by
default.

Bug: 939317
Change-Id: I06b57222ea02a4ac9594eec1f89700a1011655b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1775646
Commit-Queue: Dominic Farolino <dom@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691583}