Prolong the expiry milestone for enable-webrtc-pipewire-capturer
The main reason for the flag is so called dialog hell, when you try to
share the screen on Wayland through PipeWire. To address it we are
waiting for work that needs to be done in the upstream[0] and Chromium
needs to be adapted to it together with other changes that will remove
the UX problems. Once that's done the flag could be removed.

[0] - https://github.com/flatpak/xdg-desktop-portal-gtk/issues/204

Bug: 682122
Change-Id: Iff3cad65dcce40de18968f6b7e3dfe4ecb8e55bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1768354
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Auto-Submit: Tomáš Popela <tomas.popela@gmail.com>
Cr-Commit-Position: refs/heads/master@{#691651}