devtools: Prepare for service worker UI thread core.
The thread ServiceWorkerContextCore lives on (the "core thread") will move from
the IO thread to the UI thread when the ServiceWorkerOnUI feature is enabled.

This CL makes DevTools aware of the core thread instead of assuming the IO
thread. This makes https/tests/devtools web_tests pass when the feature is
enabled.

Bug: 824858
Change-Id: I20e38f50998f4b0cd9d8b8cc43526665d92dbf94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1775655
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691689}