[sequencemanager] Set PendingTask::delayed_run_time as before.
Prior to
https://chromium-review.googlesource.com/c/chromium/src/+/1706829,
PendingTask::delayed_run_time was set by SequenceManager for
non-delayed tasks when "set queue time for tasks" was enabled. The
jank histograms depend on that behavior, and have regressed since it
changed.

To ensure that we can track of the evolution of jank over time
and detect regressions, this CL restores the previous behavior of
setting PendingTask::delayed_run_time for non-delayed tasks when
"set queue time for tasks" is enabled. We could explore changing
that behavior in a separate CL.

Test: Launch Chrome on a fast machine. After 30 seconds, navigate
to chrome://histograms. The value recorded for
Browser.Responsiveness.JankyIntervalsPerThirtySeconds.Startup should
be equal or very close to zero.

Bug: 997203
Change-Id: I60649e9415579e62a2fe2206c8307ed7a4e66668
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1775660
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691792}