Follow power saving mode in custom tabs before Q
When a Custom Tab is started with COLOR_SCHEME_SYSTEM on Android versions
before Q, it will now follow the system power saving mode. Behavior on
Q is unchanged.

Monitoring power saving mode is part of GlobalNightModeStateController.
We can't reuse it as a whole, because it also looks at Chrome's theme
setting. So tracking power saving is extracted into PowerSavingModeMonitor
to be reused in Custom Tabs.

Bug: 994702
Change-Id: Ib7550eddf2c70ca68c25f654fffdd2922f62e2f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1762219
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Pavel Shmakov <pshmakov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691865}