Save a ReusedPasswordAccountType instance type.
Previously, we were saving |username_| but still
having to create a new ReusedPasswordAccountType in
multiple calls from different classes. This could
provide inconsistency as there's a possibility that each
time we create a new instance, it might refer to a different
account. Now, we have one source of truth for an
instance of ReusedPasswordAccountType where a warning was shown
last for it.

Note that we do still have to keep |username_| because
GetOrganizationName needs it to grab the AccountInfo which
gives us the email to grab the domain.

Bug: 914410
Change-Id: I39c0c09baaa03158727a52ac631cbaed7e57d30f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1754793
Reviewed-by: Hector Carmona <hcarmona@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Reviewed-by: Nathan Parker <nparker@chromium.org>
Commit-Queue: Bettina Dea <bdea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691873}