Registers schedulers to the EventLoop
When CooperativeScheduling runs a nested loop, it should not run any tasks from
the yielded EventLoop, as this will change the order of task execution.
To stop all task queues belonging to a EventLoop, this CL registers the frame
schedulers to the EventLoop.
Documents with ContextDocument should not register their schedulers as they
share the context document's scheduler, and would cause duplicated
registration.

Bug: 804661
Change-Id: Iec2d772090c39740337400d7ff547465aeb582c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1743290
Commit-Queue: Keishi Hattori <keishi@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Yuta Kitamura <yutak@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691903}