Forced colors mode backplate (non-LayoutNG)
This change adds logic for painting a readability backplate behind
text in forced colors mode for non-LayoutNG. Backplates are split by
paragraph, as defined by two consecutive line breaks. A backplate is
not drawn if forced-color-adjust is none.

Tests are added for this in the LayoutNG change:
https://chromium-review.googlesource.com/c/chromium/src/+/1749299

Bug: 970285
Change-Id: I51aec229d905de5fe7b233792c41555037c5b172
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1713333
Commit-Queue: Alison Maher <almaher@microsoft.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#691937}