media/gpu: Cancel timer when simulating drain
If a driver doesn't support the drain operation, the drain is
simulated but the buffer ready timer is not cancelled.
If the buffer ready timer is not cancelled, the unittest will return
an error if the unittest doesn't end before 3s since processing the last
frame.
This patch simulate the drain operation by calling FlushEncoderDone
function and cancelling the buffer ready timer in the same time.

BUG=999013
TEST=tast with tast video.EncodeAccel* on veyron_minnie-kernelnext

Signed-off-by: Francois Buergisser <fbuergisser@chromium.org>
Change-Id: Id2a8bcd71a8bf82e84970bebf6eb18f3b68ee636
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1772853
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691947}