Remove NullValue() from CalculateIterationTime
NullValue() is represented as a NaN in a double. We could better return
a base::Optional<double> and use the undefined base::Optional as null
value. Also there is no need for the IsNull() call as we could just use
the returned value as a boolean.

Bug: 994811
Change-Id: Ic2a84e8f329c2b790dcd0b996439942974255b4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1760745
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Sergio Villar <svillar@igalia.com>
Cr-Commit-Position: refs/heads/master@{#691994}