service worker: Remove "doom installing worker" mechanism.
The "doom installing worker" was introduced at a time when we didn't
have timeouts for starting a worker or for lifecycle events. The purpose
was to prevent the job queue from being stuck forever on a bad worker.
Now that we have these timeouts, we can remove the mechanism. This
also aligns with the spec, as seen in the passing WPT test.

A previous attempt was made at
https://chromium-review.googlesource.com/c/chromium/src/+/560633/.
That attempt added a timeout timer for the job queue, but was reverted
due to crashes with it. I considered adding a timeout timer here too,
but it seems to have little utility considering the job should only be
stalled by a worker stuck in starting or install. If there is another
failure, it's a bug and timing out might not actually help.

Bug: 723037, 999027
Change-Id: I0c6248db5a2e834f986218e61cec82e4b8e0f4a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1776156
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#691996}