Ansible package management playbook generation
Implement Ansible playbook generation for APT package management
in the Crostini container. The code is not wired up with preferences
or policies at this point and accepts strings of JSON input.

This code will be used in a handler for an experimental enterprise
policy on Chrome OS.

Bug: 993321
Change-Id: Ifb07f787a1daf6b2829a8cce73e30b990e25116e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1751253
Commit-Queue: Ilya Bizyaev <ilyabiz@google.com>
Reviewed-by: Olya Kalitova <okalitova@chromium.org>
Reviewed-by: Timothy Loh <timloh@chromium.org>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Cr-Commit-Position: refs/heads/master@{#692047}