[scroll-snap] Treat wheel scroll as directional scrolling for snapping
Previously wheel scrolling was being treated similar to touchpad.
This lead to behavior where it was very hard to escape snap points
since at the end of every wheel tick which scrolls a small amount
Blink would snap back. The only escape was to wheel continuously
to get closer to the next snap point.

This is not ideal. It is much more natural for wheel ticks which
are discrete to be treated as "direction scroll" for snapping
purpose. So single wheel tick would take user to the next snap points.

See [1] for a before/after video.

[1] https://bugs.chromium.org/p/chromium/issues/detail?id=963609#c5


Bug: 963609
Change-Id: I72d1f23ab25e3d1ce29afa47d2d591497e8f1c6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1758498
Commit-Queue: Majid Valipour <majidvp@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692075}