[BackgroundTaskscheduler] AlarmManager delegate
The BackgroundTaskScheduler is used to schedule one-off or periodic
tasks for Chromium, by delegating the actual scheduling to the
corresponding system API. GcmNetworkManager is used in L and lower
Android versions, while JobScheduler is used in the other newer versions
of Android.

Since a new data type is added in order to support scheduling of
tasks at exact times, a new API will be used to schedule this type
of tasks: AlarmManager. The reason for using this API is that it offers
more guarantees regarding the scheduling of the tasks: even if the
device is in idle mode (added in the newer versions of Android), the
task will still be scheduled at the wanted time, with a maximum delay
of 9 minutes.

This CL adds a new BackgroundTaskSchedulerDelegate implementation,
responsible for the delegation of tasks to AlarmManager. The necessary
updates in the existing classes and new tests are added with this
change.

Bug: 970160
Change-Id: I87d090c64854dab209a444a7cae65c2dcb6c2430
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1755989
Commit-Queue: Ioana Stefan <ioanastefan@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Richard Knoll <knollr@chromium.org>
Reviewed-by: Mugdha Lakhani <nator@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692079}