Fix the GPU watchdog restart on Linux
After the GPU watchdog starts in the GPU initialization, the watchdog
thread is stopped before sandbox initialization and then restarted after
the sandbox. This CL fixes the bug that causes the watchdog restart failure.

Because of this bug, the GPU watchdog did not run on Linux. This CL will
enable the GPU watchdog on Linux again.


Bug: 992572
Change-Id: I9dc1da1ef23bda772f2d47d9f4a733b624dce517
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1772516
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Maggie Chen <magchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692113}