New user pod UI for chall-resp login begin
Update the Ash UI for the challenge-response authentication (a.k.a.
smart card login) according to the newest UI mocks (see
https://crbug.com/993788 ) :

* The smart card icon and the "Sign in with smart card" label are NOT
  clickable anymore;
* A separate "->" arrow button is added. Tapping it starts the
  authentication.

Implementation-wise, additionally the existing user password input field
was wrapped into another wrapper, to allow hiding it WITHOUT continuing
to consume the screen space. This was done, as the UI mocks require the
challenge-response "->" button to be drawn exactly in the place where
the password input field is normally shown.

Bug: 983103
Test: enroll a Chromebook, configure device policies to force-install smart card middleware onto Login Screen, log in using a SAML account that uses smart card authentication, lock the Chromebook, check that an arrow button, a smart card icon and "Sign in with smart card" appears, click the arrow button, check that the PIN keyboard appears, enter the smart card PIN, check that the user session gets unlocked
Change-Id: Ie210d47409e9f8a1fdd44a70ebb2efd37ca1311b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1776540
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692134}