Onion soup content/renderer/media/webrtc/webrtc_audio_sink.cc|h
This is part of phase 1.9 on the design document [1].

[1] https://docs.google.com/document/d/1AJKVA5U4nDkyDB9p4ROrggWXadCxyy-grKaE9KS5vOU/

Since there are clients in //content directly referencing to this class,
it was temporarily placed in blink/public/platform.

Also, once its clients get moved to Blink, and so this API, it can
switch over from STL types (std::string and vector) to its WTF
counterparts. TODO is left.

BUG=787254
R=guidou@chromium.org, haraken@chromium.org

Change-Id: I3519318d5ae6f5b3641442f78d5e6116801dbbe9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1778005
Commit-Queue: Antonio Gomes <tonikitoo@igalia.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692180}