[CrOS Cellular] Fix click on Cellular network in mobile subpage.
Before this CL, clicking on a non-connectable Cellular network in the
mobile subpage would result in no action except for some UI jank. This
was due to the fact that the network cannot be configured and thus it is
not possible to attempt a connection.

This CL fixes this situation by navigating to the network's detail page
in this situation.

Bug: 947351
Change-Id: If390de1f8fd5476fbae193a9597f64dc31fa9143
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1777067
Auto-Submit: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692212}