Add tablet-mode case of LauncherDragTest
This CL adds the home-launcher state changes in ShellTestApi, and
so LauncherDragTest can conduct within tablet mode.

As far as I tested, the input latency for dragging in tablet mode is
not quite bad. It is 50-60msecs in scarlet; that is comparable to
the latency of clamshell mode in nocturne.

Bug: 995991
Test: interactive_ui_tests
Change-Id: Ifd7d2694bdcf5a60eb295d11773df1eb73ec181b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1775598
Commit-Queue: Jun Mukai <mukai@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692271}