Bindings: Early exit in value->IDBKey conversion after exception
In cases where an array is being processed to generate a key (either a
JS array, or an array key path), if an exception occurs the conversion
needs to terminate immediately so that the exception can be delivered
to script.

Add early exits in a couple missing places, and add tests. Note that
these cases are not reachable from script, because the conversion code
is defined to run on "structured clones" of input data, so any getters
that throw (for example) will have been filtered out by the previous
serialize/deserialize process producing the clone.

Change-Id: I53576eb1d47d23b0be4b712bc12e9de47078a419
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1777060
Commit-Queue: Joshua Bell <jsbell@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692282}