[PWD Editing Android] Start the password editing activity from the native side
The password editing activity is now started from the native side,
so it can be started the same way from the info bar in the future.
It is still shown when editing icon in password viewing activity is
clicked, but now the displaying of the password doesn't depend on
its id in the password list.  The saving needs to be redone to be
independent of the id, so it doesn't work now. It will be done in
another CL.

Bug: 377410
Change-Id: Ia10556d69332f3131ae0d257c39a951faab9a08a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1776359
Commit-Queue: Ivana Zuzic <izuzic@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692494}