android: Don't transition Surfaces at app startup.
There seems to be a race condition in the framework which results in a
crash if we transition between SurfaceViews too soon at app startup.
This change removes that to verify the theory. It was initially added
to avoid additional color layers when we use video overlays with
SurfaceControl, but its unclear what the cost of this is. For now, just
remove the transition altogether and if it turns out that this is an
issue we can try delaying the transition after startup.

R=​dtrainor@chromium.org

(cherry picked from commit e1830003682069460f18ea7d7ae4551aadd755f2)

Change-Id: I422a396aa3616773629933657dd1a5baf1341ce7
Bug: 998311
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1776906
Commit-Queue: Khushal <khushalsagar@chromium.org>
Commit-Queue: David Trainor <dtrainor@chromium.org>
Auto-Submit: Khushal <khushalsagar@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#691863}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1781817
Reviewed-by: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/branch-heads/3865@{#703}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}