LayoutVideo should not mark NeedsLayout in layout
This patch fixes |LayoutVideo| not to mark |NeedsLayout|
even when intrinsic size was changed, if it is during the
layout. Layout clears |NeedsLayout| of its inline ancestors
that doing this can leave them |NeedsLayout| set after the
layout was complete.

This is probably an old issue, but there were no code path
to hit it, unitl r684192 (crrev.com/c/1703540) changed the
intrinsic size computations.

yosin@ pointed out that |intrinsic_size_| is not used at all
when |ShouldApplySizeContainment()| since r684192, so we can
probably skip |LayoutVideo::UpdateIntrinsicSize()|. This
should fix the problem and also is supposed to improve the
performance a bit. This patch takes more conservative
approach to ensure we never |SetNeedsLayout| of inline
ancestors during the layout.

Bug: 998395
Change-Id: I735780bbc3b65321796d217934b57c6be2d2adbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1780986
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692541}