Clear the account-scoped PasswordStore on signout
This adds an "is_account_store" flag to LoginDatabase, which specifies
whether this DB stores profile-scoped or Gaia-account-scoped data.
Sync uses this flag to clear the account-scoped data on signout.
In the future, the flag will also be used by the DB itself to "tag"
credentials it returns as coming from the account or from the profile.

Bug: 998455
Change-Id: Ibcdc1082fab5085ef354be7c4055958165135f39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1748755
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692606}