Fix TabSharingUI crash on killing a renderer process.
The TabSharingUI maintains a map of pointers to tab sharing infobars on
all current tabs, that gets updated OnTabStripModelChanged. However,
killing a renderer process does not trigger a tab strip model change
since the tab still exists, but its WebContents is replaced by the sad tab.

Make TabSharingUI observe InfoBarManager for OnInfoBarRemoved events to
remove the infobars from the map, and for TabChangedAt events to
recreate the infobar for tabs that don't display it, i.e. tab that
crashed and got reloaded.
Don't display the infobar on the Sad Tab, because its content cannot be
shared.

Bug: 999760
Change-Id: If33943ee86d409d2c18b93b3e04ef0d3e2cfb559
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1781593
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Marina Ciocea <marinaciocea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692645}