[Sync:USS] Implement reading NigoriLocalData
We should restore state of the bridge and the processor during
construction of the bridge. We currently don't verify the data, since
it's unlikely to successfully decrypt and deserialize corrupted data
(see NigoriStorageImpl), but have plan to add some verifications in the
future.

Additionally, we expose encryption state through notifications inside
Init(), because if we have some data restored from the file, we may
end up in ApplySyncChanges() not being called.

Bug: 922900
Change-Id: Ib70af41b85c16a22ef30d1508d1da05e5e918162
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1781432
Commit-Queue: Maksim Moskvitin <mmoskvitin@google.com>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692699}