Configure streams according to the capture intent.
Chrome OS VCD configures the still capture stream lazily when taking the
first photo to optimize the performance for video conferencing scenario.
But the underlying camera pipeline might have different tuning for
video/photo mode based on the streams configured, so the preview in CCA
might change significantly after taking the first photo, which confuses
the users and image quality engineers.

This CL adds SetCaptureIntent() in the CameraAppDevice interface, so CCA
could set the capture intent to help the camera device decide optimal
configurations, which is similar to CONTROL_CAPTURE_INTENT [1] in
Android.

[1] https://developer.android.com/reference/android/hardware/camera2/CaptureRequest#CONTROL_CAPTURE_INTENT

Bug: b:139648548, b:135715807
Test: Check the streams configured properly in CCA for each mode, and
the color of preview does not change after the first capture.

Change-Id: Ib15e608299cf4a90ff061fb69c2d95a79f0addab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1780990
Commit-Queue: Shik Chen <shik@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Wei Lee <wtlee@chromium.org>
Reviewed-by: Kuo Jen Wei <inker@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692723}