Fix ARC_PACKAGE sync not waiting after reenabled
The patches fixes an issue where, if the user goes to Chrome OS settings
and disables ARC++, the later reenables it, sync won't start until the
browser restarts.

The underlying issue is that the SyncableService conceptually
transitions from being ready to sync to being unready, which is not
supported by sync. This means SyncableService::WaitUntilReadyToSync() is
not called for the second time, and things fall appart.

Extending sync for this exceptional case would incur in quite some
complexity (along the lines of reverting
https://chromium-review.googlesource.com/c/1477896 and plumbing
analogous APIs to SyncableServices).

Instead, we address this problem in the controller, where the datatype
becomes "unready" by failing to meet preconditions, but without clearing
sync metadata.

Bug: 978837
Change-Id: If7a7f00e022f67af73162a0da874b97e4f067245
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1642632
Auto-Submit: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Long Cheng <lgcheng@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692751}