Don't ask for validation if download request will start from beginning
When strong validators are missing, download could request for some
downloaded bytes to validate the content on disk. However, if the
file on disk is smaller than the data required for validation, there is
no need to validate the content as download can just overwrite the whole
file. This also avoids that if content validation fails, download will
have to restart again, thus increasing the interruption count.

BUG=999746

Change-Id: Ic2e3e5ffb98579acbde9ae5b4dffb7f8f5302ae4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1774501
Reviewed-by: Xing Liu <xingliu@chromium.org>
Commit-Queue: Min Qin <qinmin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692788}