Fix beta build
When https://crrev.com/81c02ad landed in master, it was between a revert
and a re-land of a patch that required changes in
validation_message_overlay_delegate_test.cc. The revert never happened
in the beta branch, so the test does not compile in that world.

To fix that, bring the test fix from the re-land[0] to the beta branch.

[0]: https://chromium.googlesource.com/chromium/src/+/ee668df54e5715835c809d33c6698bd9535da667%5E%21/#F12

Bug: None
Change-Id: I91e35101efa3651cd7bccdca4ccd14ebd61d2c49

TBR=flackr@chromium.org,pdr@chromium.org

Change-Id: I91e35101efa3651cd7bccdca4ccd14ebd61d2c49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1782752
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/branch-heads/3865@{#715}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}