Add explicit flag for compositor scrollbar injected gestures
The original change to enable scrollbar latency for the composited
scrollbars incorrectly used an existing member to try and determine
whether a GestureScrollUpdate was the first one in an injected sequence
or not. is_first_gesture_scroll_update_ was incorrect because it is only
updated when input is actually dispatched to InputHandlerProxy, and the
flag is cleared for all GSUs before the location where it was being
read.

This bug was missed because of incorrect tests. The
VerifyRecordedSamplesForHistogram method doesn't actually assert or
expect anything - the return value must be inspected.

As part of fixing up the tests, I made a few other changes to get them
passing consistently across all platforms:
- turn on main thread scrollbar injection feature (in case it's ever
  turned off we don't want the tests to start failing)
- enable mock scrollbars
- disable smooth scrolling
- don't run scrollbar tests on Android

The composited scrollbar button test is disabled due to a bug in how
the mock theme reports its button sizes, which throws off the region
detection in ScrollbarLayerImplBase::IdentifyScrollbarPart (filed
crbug.com/974063 for this issue).

This previously landed as crrev.com/c/1652741 but was reverted for
flakiness on ChromeOS. I'm unable to repro (ran 100s of times both
locally and on swarming bots) the flakiness at this point,
but have expanded out the histogram names to not use variables
in order to better determine where the flakiness might be coming from.
(Also crrev.com/c/1725224 added a RunUntilInputProcessed that could
have fixed this flakiness).

Bug: 954007
Change-Id: Ic102aebbcbd627ecb5edd0a0633ac2e8b1d17db5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1774911
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692814}