[DevUI DFM] Add handler to rewrite URL of hosts in the DevUI DFM (behind dfmify_dev_ui build flag).
This CL is the last piece for minimally viable DevUI DFM: Detect
whether the host of a chrome:// URL request belongs in the DevUI DFM.
If so, and if the DevUI DFM has not been installed and loaded, then
rewrite the URL to chrome://dev-ui-loader, which is the interstitial to
install / load the DFM, and on success, redirect to the original
chrome:// page. Details:
* URL rewrite uses existing BrowserURLHandler framework, which also has
  the advantage showing the original URL in the Omnibox.
* Currently the only DevUI DFM host is chrome://bluetooth-internals.
  More will be added in the upcoming weeks, but these require case-by-
  case refactoring and BUILD file changes.
* Add dev_ui_url_handler.* to implement detection logic and rendering
  of chrome://dev-ui-loader URL.
  * Currently redirection only keeps original URL's host, but not
    /path?query#fragment. This will be fixed in a follow-up.
  * The current chrome:// restrictions should suffice in preventing
    abuse of chrome://dev-ui-loader's ability to redirect.

Meanwhile, the DevUI DFM feature is still behind the |dfmify_dev_ui|
build flag.

Bug: 927131
Change-Id: I2d108e648a9e8ac8f67a2b697ea615b01a1cf826
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1782016
Reviewed-by: Samuel Huang <huangs@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Commit-Queue: Samuel Huang <huangs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692818}