Fix race condition during kiosk launch error check
When a kiosk app is set to be auto-launched and the user cancels its
launch, we should disable auto-launch on next boot. This was
achieved by setting "KioskAppLaunch" variable into the local state,
and subsequently resetting it every time. This variable was reused
by both plain kiosk and arc kiosk modes.
The variable was reset in only one of the code paths, leading to a race
condition when second one tries to read from already reset variable.

There are many possible solutions for this, one of the simplest ones
is to persist the error state during the whole boot.

Bug: 984740
Change-Id: I161daae80a4fedb51d47ebff93b84e34bfc475b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1781431
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Anatoliy Potapchuk <apotapchuk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692819}