Fix JS errors in internet-detail-dialog
internet-detail-dialog is a top level element that may be loaded before
InternetDetailDialogUI::BindCrosNetworkConfig() is called, causing
'chromeos.networkConfig.mojom' to be undefined when the HTML is loaded.

Because of this, defining 'const mojom = chromeos.networkConfig.mojom'
at the top level triggers an error. To address this, avoid using the
alias outside of individual JS methods.

This also fixes an error where cr_network_listener_behavior.html was
not being included.

Bug: 1000207
Change-Id: Id1c87c08a1c019d4c4dafbbc668dcce63c1c9c4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1783064
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692831}