Use the correct process for about:srcdoc
It turns out the process used to load about:srcdoc was not always
correct. For instance, when starting from A(B), if B navigates to
about:srcdoc, it was sometimes loading in B's process instead of
A's process. This is now fixed and enforced.
This is not a regression. The previous code about srcdoc in
RenderFrameHostManager::CanSubframeSwapProcess() isn't really effective
when the about:srcdoc navigation happens AFTER a cross-process one.

Bug: 996044, 996725
Change-Id: I670ee6dc917cfe4cb02721a58a79a1a415b45e61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1773264
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692853}