Fix spacing around subtitle in SAML PIN dialog
Decrease the spacing between the title and the subtitle in the security
token PIN dialog shown during SAML sign-in.

To achieve this, remove the extra <p> element that was adding extra
margin/padding.

Also adjust accordingly the position of the PIN keyboard, so that the
spacing between the subtitle and the PIN keyboard stays the same after
this CL.

Bug: 964069
Test: verify in DevTools that (1) distance between bottom edge of title and subtitle is 32dp, and (2) distance between bottom edge of subtitle and top edge of the PIN input field is 78dp
Change-Id: I6fc9e073e100870755830a8011087ecd90d4a49a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1777865
Reviewed-by: Alexander Alekseev <alemate@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692921}