[Files Ranking] Prefetch QuickAccess results.
The latency of the call to the Item Suggest API is too long, causing
UI issues with results popping into the list.

This CL moves the call to immediately after the launcher is opened into
peeking state, giving a longer window for the call to complete.

Bug: 959679
Change-Id: Idc10c56f3807ed9079ac8924fc49417b1d427d8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1772788
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Reviewed-by: Jia Meng <jiameng@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692923}