Fix TrustStoreMacTest.SystemCerts on macOS 10.15.
The output of 'security find-certificate -a -p -Z' in 10.15 includes both a
SHA-256 and SHA-1 hash for each cert. The old parser split on the SHA-1 hash,
confusing the rest of the parsing. Now it splits on the
"-----END CERTIFICATE-----" PEM marker, which should be a consistent place to
split regardless of any future changes.
The parsing of the hash value looks for the first hash mentioned in the text
regardless of algorithm. (It's only used for SCOPED_TRACE and LOG messages,
so the particular algorithm isn't important.)

Bug: 998845
Change-Id: Iaea008a253d0dc4d25314010b2803b841206b7ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1783661
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Commit-Queue: Matt Mueller <mattm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692931}