Settings/languages: Prevent focus on input method list items.
The input method list item element is essentially an abuse of the
default list item implementation in that selecting the main area of the
input item sets the selected input method as the enabled input method.
After selection, this part of the list item element should not be
focusable or respond to activation.

This change provides a few things:
* Override the cursor for an actionable input list item element so it
does not appear as an activatable button.
* Only set the tabindex on the non-enabled input list item elements.
* Prevent focusing the input list item element such that a focus ring
does not flash then disappear on click.

R=jamescook@chromium.org, khorimoto@chromium.org

Bug: 987607
Test: none
Change-Id: I60c79b50a4e8691ab378d43721116fbbc7202ace
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1783239
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: James Hawkins <jhawkins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692933}