CrOS Shelf: Fix shelf centering on non-primary displays
The issue was that we were calculating centering coordinates according
to the whole screen rather than each individual display.

Add a test for this.

This change also fixes a bug in ash/screen_util where the shelf bounds
were not returned in terms of the whole screen's coordinates. A few
unit tests in ash/wm also needed fixing in that sense.

Bug: 997768
Change-Id: I91ab6ef2f85601edc115743dd0ab0ae83076be94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1773891
Commit-Queue: Manu Cornet <manucornet@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#692973}