[Dice] Migrate tokens more aggressively
As measured by the histogram Signin.TokenServiceDiceCompatible, we
see that ~1% of the users on Canary are not migrating their tokens
correctly for Dice.

This CL removes the edge cases in ShouldMigrateToDice(), so that all
users should now migrate their token service for Dice.

The edge cases were:
1) account tracker service migration not completed
2) hosted domain information missing

The migration will now have the following behavior for these cases:
1) Dice migration will happen even if the account tracker migration
is not done
2) if the hosted domain information is missing, we use
IsNonEnterpriseUser(). This parses the email of the user and matches
the domain with a list of known non-enterprise domains. If the domain
is known to be non-enterprise, the token won't be revoked. Otherwise,
the token is likely to be enterprise and should be revoked.

Change-Id: I1d7fb12874e274edd8aea8af27fb5f77fa9f8b24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1782617
Reviewed-by: Monica Basta <msalama@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#693075}