Make color-scheme affect used system color.
We currently compute system colors as their rgba values based on
color-scheme. However, color-scheme was not a high priority property, so
changing color-scheme on the same element as the system color was set
did not always yield the correct result as the color-scheme might have
been applied too late.

There is an open issue[1] for having system colors computing to
themselves and resolve to an rgba value for the used value. That would
mean we will have to change the internal storage for colors in
ComputedStyle and possibly store a StyleColor with system color keywords
in addition to currentcolor which currently computes to the keyword.

For now, instead of introducing an even higher priority level, make
color-scheme a high priority property, and delay application of the
cascaded color property value to the ComputedStyle until we have applied
other high priority properties.

[1] https://github.com/w3c/csswg-drafts/issues/3847

Bug: 939811
Change-Id: I91ce04a55886507ee20aace9b2470a5ac5b43fbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1773073
Commit-Queue: Emil A Eklund <eae@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Cr-Commit-Position: refs/heads/master@{#693151}