Fix the device disabled case at FrE check.
When the device is disabled, we receive from the server
RESTORE_MODE_DISABLED mode. The client has no information if the policy
forces to re-enroll or not. In the initial implementation it was treated
as device not being forced to re-enroll. This change fixes that case, so
that the flags from VPD don't get deleted for disabled device.

Also with a test image, the install_attributes file is deleted while
the checksum remain on disk. This has yet to be investigated, but to
avoid any problems when install_attributes file is corrupted, we set
the device as being in DEVICE_MODE_ENTERPRISE if auto_enrollment
detects the device is disabled.

BUG=chromium:996059
TEST=Manual tests and unit tests

(cherry picked from commit cbe2bc9f3d31b695c7dcf87ed066e1521bbb241d)

Change-Id: I8bc30028af043fdb563e6374d7e5db4b30830218
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1771966
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Bartosz Fabianowski <bartfab@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Igor <igorcov@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#692052}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1782906
Cr-Commit-Position: refs/branch-heads/3865@{#734}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}