[NTP] Fix white flash before the page is initialized in dark mode
The NTP body is hidden before initialization is complete, so the theme
background color must be set on the HTML element instead. This prevents
a white flash before the NTP fully initializes.

The background color will be removed after initialization in order to
dynamically update the color when the theme is changed.

Bug: 1000186
Change-Id: I6b8c4b2b76599af41df088214f372506eb48bc8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1783407
Reviewed-by: Dan Beam <dbeam@chromium.org>
Commit-Queue: Kristi Park <kristipark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#693255}