[BrowserSwitcher] Break BHO's dependency on CRT
When compiled with use_mini_bho=true, the Legacy Browser Support addon
won't depend on the CRT, which reduces binary size by a lot
(~800kB => ~20kB).

This is done by essentially rewriting the BHO to use built-in Windows
function instead of the stdlib, and implementing some of the functions
we need.

Before:
    browser_switcher_bho.dll    368640
    browser_switcher_bho_64.dll 443392

After:
    browser_switcher_bho.dll     12288
    browser_switcher_bho_64.dll  15872

Bug: 950039
Change-Id: Icf8f73b8d4e4b029761f04d5877f7fc773697a2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1699279
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Commit-Queue: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Cr-Commit-Position: refs/heads/master@{#693266}