[PM] Aggregate per-frame freeze origin policy using a decorator.
With this CL, the freeze origin policies of frames of a page that are
"current" are aggregated in a decorator. In a future CL, the
aggregation will be forwarded to TabManager to be taken into account
when deciding whether a page can be frozen.

Bug: 999594
Change-Id: I5729caaf646017af611e3e77a30e576cb4525baf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1779062
Commit-Queue: François Doray <fdoray@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Sigurður Ásgeirsson <siggi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#693325}