ozone: Fix compilation of views_unittests
After https://crrev.com/c/1666910, views_unittests compilation for ozone
was broken. This CL fixes this issue.

The views_unittests will be enabled in the follow up CL. However, there
are some tests that fail and have to be disabled for ozone/linux. Once
they are fixed, they will be enabled back. There ~20 failing tests in total,
which mostly fail because of not implemented methods in the
DesktopWindowTreeHostPlatform.

Bug: 700309
Change-Id: I14faa0bac9d937731d774bd039e3dfb793e0ea66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1782845
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#693559}