[iOS][Dark Mode] Extend default appearance customization to all views
The major changes are in appearance_customization. Before, we were in
the middle of migration, so I didn't want to set the tintColor by
default on all views. Now, migration is complete, so this is safer
moving forward.

The rest of the changes are deleting places where tintColor is set
unnecessarily.

Bug: 974255
Change-Id: I2700cdca98c30602e0025481a3da6800bd1f8388
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1787283
Reviewed-by: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#693697}