DL: Undo the frame throttling update during will begin main frame.
This complicates the code a little bit. I don't think this is strictly
necessary, since the savings it would provide for the single frame are
defeated by the fact that we need to paint that frame, so we process
the lifecycle anyway. When intersection observer runs, we'll still
update the throttling status.

I've re-opened the referenced bug with hopes of coming up with (and
documenting) a good solution that would save us as much lifecycle
processing time as possible.

R=chrishtr@chromium.org, szager@chromium.org, erikchen@chromium.org

Bug: 976391
Change-Id: I209ce44e4a17042c937de31d538e85548b583ac7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1783876
Reviewed-by: Stefan Zager <szager@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#693765}