[M77] Revert "Move Video Capture service off Service Manager"
This reverts commits 2f8c1dd283a265c587267232577cbe8b5e6043aa.
 and b35ff4b6e01e4b53f416feb1795056869ca242e9.

2f8c1dd283a265c587267232577cbe8b5e6043aa was originally reviewed at
https://chromium-review.googlesource.com/c/chromium/src/+/1710051

b35ff4b6e01e4b53f416feb1795056869ca242e9 was originally reviewed for M77 at
https://chromium-review.googlesource.com/c/chromium/src/+/1725324 and
is a minor bug fix on top of 2f8c1dd283a265c587267232577cbe8b5e6043aa
that was made in M78 and was merged to M77.

The reason for this revert is that it makes Webcams not work on Windows 7.

Bug: 992328
Change-Id: I02f9a7e4f29eb4810233c72eb14ceb565a0a613b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1786450
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/branch-heads/3865@{#748}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}