Add trace event to debug RenderText Itemize
This CL is adding a trace event to get details on the way the text is
split into runs before text shaping.

The split text is passed to ShapeRuns which is trying to find
appropriate fonts for the best glyphs coverage. Unfortunately, some
unexplained cases are really slow (see crbug/995789).

We determined that some cases were related to Emoji (see crbug/873423).
This code is intended to determine other block of unicode are causing an
expensive font fallbacks.

This code can safely be reverted when the bug (crbug/995789) is fixed.

R=robliao@chromium.org, asvitkine@chromium.org
CC=​drott@chromium.org

Bug: 995789
Change-Id: I7406c80b6acf9faaf7dd46184e6d04fdd7d4c5c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1785099
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#693777}