[LayoutNG] Only propagate break-before of the first child.
The break-before value should only be propagated if we're not at a valid
class A breakpoint. This is only the case at the first in-flow child.
Propagating it in other situations could cause spurious forced breaks in
the ancestry, as the new test would demonstrate.

The actual fix here is to simply check that we're at the first child,
but I rearranged the code a little while I was at it, to keep the block
fragmentation code more closely together. We'll also no longer update
the has_processed_first_child_ member unless we're block-fragmenting.

Change-Id: I831e162adae0d4b56231c118274ba1083da1fc97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1787320
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#693788}