Restore AnimationClock to its old behavior when outside of animation frames
This is essentially a hack to fix crbug.com/995806. The HTML spec says that
the clock for animations is only updated once per rendering lifecycle, at the
start. However the spec also assumes that the user agent runs rendering
lifecycles constantly, back-to-back. In Blink we attempt to *not* run rendering
as much as possible, to avoid unnecessary CPU usage. Unfortunately that then
means that when we are outside of the lifecycle, e.g. in a setInterval, the
AnimationClock time no longer updated after https://crrev.com/629180f !

This is a short-term fix for that problem, by re-introducing the old dynamic
self-update behavior - but only when outside of rendering lifecycles. Note
that it may still have an effect on the overall time, as the logic sometimes
guesses too far and the next real frame may have an older time than we updated
the clock to. In that case there would be a discrepancy between the animation
clock time and things like the rAF timestamp, but it isn't the worst bug in
the world.


(cherry picked from commit 44f1bd10c558e5f0f420255916ee364ffda42352)

Bug: 995806
Change-Id: I66642c28026c9370c73b5ce4d7db2fa2981ec7e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1778598
Reviewed-by: Robert Flack <flackr@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Majid Valipour <majidvp@chromium.org>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#692841}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1787698
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/branch-heads/3865@{#755}
Cr-Branched-From: 0cdcc6158160790658d1f033d3db873603250124-refs/heads/master@{#681094}