Web UI: Re-enable CrElementsInputV3Test on non-Mac
This test is flaky only on Mac, just like CrElementsInputTest. Re-enable
it on other platforms and add a comment referencing the bug for
CrElementsInputTest. The flakes are likely due to the same cause, since
these tests are identical other than the version of Polymer.

Bug: 997944
Change-Id: Ife5e77c066496f702a3661457aa68d0f2af808c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1785624
Reviewed-by: Hector Carmona <hcarmona@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#693869}