[PM] Cleanup: move win WorkingSetTrimmer feature with others.
The only thing in resource_coordinator_features is a flag for the
performance_manager Windows WorkingSetTrimmer. That feature flag
should really live in performance_manager features with the other
similar flags. This CL moves that flag and then cleans up the files
that are no longer needed after that move.

BUG=chromium:988602

Change-Id: I0531ead60d5d1117046607a38d8165936f008ef0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1783299
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Alexei Filippov <alph@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Sigurður Ásgeirsson <siggi@chromium.org>
Commit-Queue: Brian Geffon <bgeffon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#693878}