Introduce FROM_STARTUP as a TabLaunchType.
This allows the code to distinguish a tab created on startup from
tab created FROM_CHROME_UI.  We animate tabs created that use
FROM_CHROME_UI, but we do not want to do that for tabs created on
startup.

The aim is that this will fix the flakiness of the
MainIntentBehaviorMetrics.

BUG=980587

Change-Id: Id00df9944c30a2b7b120ecb29b268af8ea3618c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1786672
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#693919}