DevTools: aXe core linting for DOM breakpoints in Elements.
Added DOM breakpoint accessibility testing.  Slightly
modified the structure of the DOM Breakpoints list DOM (but
didn't alter its display) for more specificity in the CSS.

The test here is a little bit more involved than most of the
tests in the Elements pane, primarily because the process of
launching the DOM breakpoints list and creating them is a
bit more involved than other components of the tool.

Bug: 963183
Change-Id: I2543331f81d1ba96a7a0ddb6ca31dba214a1761a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1683826
Commit-Queue: Junyi Xiao <juxiao@microsoft.com>
Reviewed-by: Erik Luo <luoe@chromium.org>
Cr-Commit-Position: refs/heads/master@{#693957}