Forced colors background color at the root
As a result of the following change, background-color should compute to
the Window system color for all values except for the alpha channel
in forced colors mode: https://chromium-review.googlesource.com/c/chromium/src/+/1764317

The change above removes |background-color: Window| from the forced
colors stylesheet. Being that the default background color for the body
is transparent and alpha channels are preserved, the default background
color for the body will remain transparent in forced colors mode as a
result of the above change.

Since this is not the expected behavior, the style was re-applied to
the forced colors stylesheet in the following change:
https://chromium-review.googlesource.com/c/chromium/src/+/1749299

The default background color for the body is now correct, but if an
author/user defines a background color with a specific alpha channel,
the alpha channel gets overridden for the body when the UA
stylesheet is re-applied. This does not match the spec.

As a result, this change adds logic to apply the appropriate alpha
channel after the UA stylesheet is reapplied. This ensures that the
default background color for body is still Window and that any alpha
channel defined by the author/user is in fact preserved as expected in
forced colors mode. This also ensures that default styles, such as
a yellow background for marked text, are not overridden in forced
colors mode.

Bug: 970285
Change-Id: Ia294764dd6f4373bec0a94adf1c8f042053f2e77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1779070
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#693988}