New Tab Page: continue adding a "real" search box ("realbox")
* Implements a UI identical to the "fakebox" (but with real <input>)
* Does some embeddedSearch plumbing
* Queries the autocomplete system (powers omnibox)
* Renders results based on autocomplete classifications
* Updates externs

It may make sense to launch this earlier on fullscreen, where the
"fakebox" + omnibox experience is pretty bad.  If so, we'll either
implement that aspect in JS (fullscreen detection) and let it flow
out embedded in the client or add a separate flag.

Still to do:
* favicons & bookmark star for suggestions
* favicons, bookmark star, & URL icons for realbox
* shift+delete to remove suggestions(?)
* more bug fixes :)

Bug: 243926, 778489, 996516
Change-Id: I87d4c695c19149ea771229f7ae026aa8727f61ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1761688
Commit-Queue: Dan Beam <dbeam@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Auto-Submit: Dan Beam <dbeam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#694051}