Adds UMA histograms for Assistant proactive suggestions.
This CL adds the following UMA histograms:
* Assistant.ProactiveSuggestions.RequestResult
* Assistant.ProactiveSuggestions.RequestResult.Error
* Assistant.ProactiveSuggestions.RequestResult.SuccessWithContent
* Assistant.ProactiveSuggestions.RequestResult.SuccessWithoutContent

* Assistant.ProactiveSuggestions.ShowAttempt
* Assistant.ProactiveSuggestions.ShowAttempt.Success
* Assistant.ProactiveSuggestions.ShowAttempt
    .AbortedByDuplicateSuppression

* Assistant.ProactiveSuggestions.ShowResult
* Assistant.ProactiveSuggestions.ShowResult.Click
* Assistant.ProactiveSuggestions.ShowResult.CloseByContextChange
* Assistant.ProactiveSuggestions.ShowResult.CloseByTimeout
* Assistant.ProactiveSuggestions.ShowResult.CloseByUser

Top level metrics (for aggregate level reporting):
.RequestResult
.ShowAttempt
.ShowResult

Metrics segmented by category (for deeper slicing and dicing):
.RequestResult.*
.ShowAttempt.*
.ShowResult.*

Note that this CL reserves 3 categories on the client for future use by
the server to loosen the coupling between the client and server
rollout for support of new content categories. As new categories are
rolled out on the server, the client will be updated as well but in a
way that is not concerned with being strictly synchronized.

Bug: b:140199998
Change-Id: I9f945bf2be4889eaf324ed59343454a114a610ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1775368
Reviewed-by: Tao Wu <wutao@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: David Black <dmblack@google.com>
Cr-Commit-Position: refs/heads/master@{#694053}