Make sure ContentsView resets everything synchronously
ContentsView::ResetForShow invokes SetActiveState with default
animate parameter, which means it can animate when the previously
state (before closing) was not kStateApps. This can confuses the
all further state transition which leads to weird visual effect
as reported. But it doesn't have to animate at all -- the intention
was to reset to a state just before showing.

This suppression of animation in ResetForShow omits some processing,
so this CL changes a few other parts:
- SearchBoxView needs to reset before SetActiveState; the reset
  might call ShowSearchResults() which is contradicting to
  SetActiveState().
- non-apps pages should become invisible. This is done on
  ContentsView::Init() but has not been done in ResetForShow.
- Layout should reset some intermediate states of the search box;
  its layout, background, and the transform.

Bug: 1000920
Test: manually
Change-Id: I53031ad699dae9bd5d6ea7787bfdeea74a986572
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1788309
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#694061}