tab_ranker: add FrecencyScorer.
This is a first try of using RecurrenceRanker for tabdiscarder with a
much simpler version.

(1) Each tab is recorded with a FrecencyScore.

(2) The score is updated on every reactivation with UpdateFrecencyScore

(3) The score is populated to TabFeatures.

(4) TabScorePredictor uses this score as the reactivation score if
    kFrecencyScorer is enabled.

Bug: 999908
Change-Id: Id7aa5ff971f7a6f7d22d5fa4e9329755894a3d1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1781899
Commit-Queue: Charles . <charleszhao@chromium.org>
Auto-Submit: Charles . <charleszhao@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#694130}