Remove NullValue() from CalculateCurrentIteration
NullValue() is represented as a NaN in a double. We could better return
a base::Optional<double> and use the undefined base::Optional as null
value. Also there is no need for the IsNull() call as we could just use
the returned value as a boolean.

Bug: 994811
Change-Id: I60818193d785cb574b0a1fef284decf2fdfaf624
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1763693
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Commit-Queue: Sergio Villar <svillar@igalia.com>
Cr-Commit-Position: refs/heads/master@{#694175}