[iOS] Switch omnibox answers to use different AutocompleteMatch field
This follows a similar approach to https://crrev.com/c/1070506. Instead
of always using match.answer->first_line() to get the text of the
main line of the question (the part corresponding to what the user has
typed), we now use match.contents to get that part, and then append on
any additional_text or status_text from the first line.

One small visual change is that now, the "user text" part of answers
have bolding (from the match contents and classification). Before, they
didn't because we ignored the server-provided HTML bold tags and didn't
provide anything in their place. My asusmption is that this is fine.

Bug: 763894
Change-Id: I4a9ec9da56454433dfe33aa3db2c0ee5b555c2a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1787508
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#694203}