[Video Capture Manager] Convert pointers from Unretained to WeakPtr.
This CL replaces the usage of unretained pointers with weak pointers
in VideoCaptureManager.

This conversion is safe because all places where the pointers are saved
are on the IO thread as well as the place were the callbacks are then
executed (see line 326 and 348).

BUG=998548

Change-Id: I47bda798fa7bcbd66bf23682ee6c6dd26b5642c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1789223
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Armando Miraglia <armax@chromium.org>
Cr-Commit-Position: refs/heads/master@{#694214}