Send app URL from app shim to browser
This serves to identify app shims that should be treated as
multi-profile (only shortcut apps and PWAs are to be multi-profile),
so that, in a future patch, they can ignore their profile path.
Also (in the future), in the event that an app is not installed on any
profile, but the shim still exists (e.g, a user wipes their prefs),
then this URL can be used to open that shortcut URL in a browser tab
(rather than just fail).

Rename the function that is called when a shim connects to the browser
from LaunchApp to OnShimConnected. Change the function from taking a
half-dozen arguments to taking an AppShimInfo structure, and pass that
structure directly to AppShimHostBootstrap.

Bug: 982024
Change-Id: Iae45c54751fd66e3dbdb3b711e0bf07b7f901dd9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1786560
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#694312}