Correct hit testing local offset in text control inner editor
The previous calculation was:
 local_offset = hit_test_location - accumulated_offset
                - inner_editor->PhysicalLocation();
which was correct when the inner editor was the top element in the
shadow dom or its PhysicalLocation included all offset from the text
control.

In a search input box, there are extra levels of layout objects
between the text control and the inner editor. We need to exclude
all offsets from the text control to the inner editor when calculating
the local offset.

Bug: 1000657
Change-Id: If1053ceffe556c2ca5be4639b21e669465c93b0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1789983
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#694502}