Added a callback to notify before a shell surface tries to close.
We would like to intercept these messages so that clients can monitor
attempts to close shell surfaces. The intention is to bring up UI
for forcibly closing windows that do not close themselves quickly
enough.

In discussion with sky@ it is undesirable to have this kind of
functionality put into views::Widget, so it was added as a callback
to ShellSurfaceBase.

Bug: 914322
Change-Id: Ieed675c937176283356bbccdf99055138b5166ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1776140
Commit-Queue: Nic Hollingum <hollingum@google.com>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#694628}