[iOS] Save current session on page commit
Before this CL, the current session was saved when the page was loaded.
The page is loaded when the main frame has finished loading its content.
When there is a push state event, this callback isn't called.
This CL changes it to store the current session when the navigation is
finished, which happens when the item is committed.

Bug: 998563
Change-Id: Ibc77d38cf4e130477471161d28a4d6cb6c03365f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1778483
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#694732}