Add TestNavigationObserver constructor that takes a desired net::Error
Add a constructor to TestNavigationObserver that takes a target
net:Error, and will wait until a navigation with that error finishes.
This simplifies the testing of error pages in the browser process.
Prevents racing with other navigations when registering an observer for
an error.

This is a precursor to crrev.com/c/1783300 which introduces tests that
have many subframe navigations which eventually cause an error page to
load.

Change-Id: I7ad301c1d4999e79b6590cd4d5d2d4dccaf3aeb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1789427
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#694785}