[Prefetch+SplitCache]: Prefetch's preload headers are cached correctly
This CL introduces a new |preload_header_token| member to
ResourceResponseInfo; it optionally holds an UnguessableToken set by
the PrefetchURLLoader on cross-origin-main-resource prefetch responses.
The response and token are sent to the renderer, and passed to
PreloadHelper. PreloadHelper associates this token with link header
preload ResourceRequests, via a new |preload_header_token| member on
ResourceRequest.

The preload headers are then marked as prefetches, and sent to the
PrefetchURLLoaderService. The service re-associates the token with the
parent prefetch's NetworkIsolationKey, so that the preloads are fetched
with the correct NetworkIsolationKey, and can be reused by the destination
origin.

Bug: 939317
Change-Id: I6a589846b804fcb9b90f7b7bb381808f860cca4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1772896
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/master@{#694800}