[DevUI DFM] Enable DevUI DFM (Android Chrome only).
This CL sets the default value for |dfmify_dev_ui_dfm| to true, to put
the DevUI DFM on track for launch: Canary -> Dev -> Beta -> Stable.
Depending on approval status, we can easily disable this. For Canary,
Play Store is also expected to gracefully handle disappearance /
reappearance of the DevUI DFM, if we ever need to disable / re-enable
it.

Currently only chrome://bluetooth-internals is in the DevUI DFM, but
more will be moved in the upcoming weeks. Also, only native resources
split is supported (native code split is future work).

Bug: 927131
Change-Id: Ie5c8b821ca067a19f4c69dbd7af863816a89c460
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1793233
Commit-Queue: Samuel Huang <huangs@chromium.org>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#694885}