cc: Cap raster scale ratio between horizontal and vertical scaling.
Since we do uniform scaling horizontally and vertically when rastering
content and choose the higher scale between the 2, this can lead to
excessive memory use if the ratio between the 2 scales and layer bounds
is extremely high. Cap this ratio to some limit (arbitrarily chosen).

Change-Id: If90b101d5a096c03dfb22fb633f3b33e2600eb74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1790958
Reviewed-by: enne <enne@chromium.org>
Reviewed-by: Eric Karl <ericrk@chromium.org>
Commit-Queue: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#694889}