Remove unneeded SurfaceSync CHECK
We added a CHECK to RenderWidgetHostImpl::SynchronizeVisualProperties. This was
in order to help track down the route cause of a rare crash on Mac. Adding a
CHECK in the browser process in order to crash earlier than in the renderer.

The original bug has long since been fixed, but we kept this around in case we
detected crashes in the wild. However aside from ClusterFuzz, and one project
embedding Chromium, none arose.

We do not feel it is valid to keep this CHECK around anymore. Our worry about
this affecting end users is gone.

Bug: 993274
Change-Id: I6de4c737f575d5d0fb183d93f1a5d3a092fd87a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1792459
Reviewed-by: kylechar <kylechar@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/master@{#694900}