Start untying SVGSMILElement::Progress
To be able to separate the resolution of intervals from the update of
the "position" information (percent, repeat) we need to keep the "full"
previous interval. Extend the existing |previous_interval_begin_| to a
full interval, add a new GetActiveInterval(...) helper and use that
in CalculateAnimation{Repeat,Percent}. Then move the calls to the latter
two functions in Progress() after the interval check/resolution.

Bug: 998526
Change-Id: Ia7154e6612252d21b04ab6302c8ef75bc7664b45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1790449
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#694909}