Recompute touch action regions on interest rect changes
The interest rect is a window within a scroller which has display items.
The typical invalidation mechanisms will handle changes to drawing
display items. Hit test display items are a little different because we
need to ensure ScrollingCoordinator::UpdateTouchEventTargetRectsIfNeeded
is called post-paint to map them into the space of the cc::Layer. The
bug was that this mapping update did not occur on interest rect changes.

This patch ensures the scrolling coordinator updates anytime display
items are painted.

Bug: 995715
Change-Id: I1428e956a83c4c07a18e40a8e36559f694eabf97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1793546
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#694967}