Handle VIEW_SELECTED event from list item in ArcA11yHelperBridge
Previously, Android's AccessibilityEvent.TYPE_VIEW_SELECTED is handled
as an event of changing value in ProgressBar or Slider in ARC++.
However, this event is usually used when selecting an item in
AdapterView (like ListView or GridView).

This CL modifies to handle both patterns of the event.

      app that ListView items can be selected and read by ChromeVox

Bug: b:130187244
Test: unit_tests --gtest_filter="ArcAccessibilityHelperBridgeTest.*"
Test: unit_tests --gtest_filter="AXTreeSourceArcTest.*"
Test: manually test with ag/9191465 using PlayStore and TalkBack test
Change-Id: If3af1bdc28490d00399fa89bd18258ff89731d19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1746326
Commit-Queue: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: Alice Boxhall <aboxhall@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Sara Kato <sarakato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#695041}