Reland viz: Partial swap without readback
viz::BufferQueue currently implements partial swap reading back
from the previously swapped buffer contents that are outside the
current swapbuffer damage.

This introduced a dependency on a GPU API, that we want to get rid
of (crbug.com/958670).
This technique also assumes it is safe to read back from a buffer
that is used from another driver (crbug.com/457511).

This CL removes the readback and solves the same problem exposing
a rectangle that the compositor needs to redraw to make sure contents
damaged during all the swapbuffers before the current one are
recomposited.

Bug: 958670, 457511
Test: BufferQueueTest.PartialSwapCurrentBufferDamage
Change-Id: Icda39c42053c2d1134f739485357009271b80634
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1761708
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#695064}