Adds sandbox API for policy diagnostics
This will eventually support a chrome://sandbox WebUI for Windows.

Adds PolicyList and PolicyInfo types, the GetPolicyInfo() interface
to the sandbox (broker) and a basic implementation which simply
returns an empty list.

The intention is that GetPolicyInfo() will be called on a low
priority thread, and will then wait until the tracking thread
can complete the work of cloning policy information before
returning.

PolicyInfo is merely stubbed out at the moment and will be expanded
to describe more of the policy once this gets plugged into the
webui later.

Change-Id: Ie6ea8180e26ad71fa9e1628a2e9eef8f46a93fbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1756952
Commit-Queue: Alex Gough <ajgo@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#695409}