OOR-CORS: Add an experimental Feature to enforce extraHeaders option
We are planning to provide a temporary policy escape hatch for
the OOR-CORS enterprise-friendly rollout. With this policy, Chrome
handles any webRequest API users as specifying "extraHeaders" option
so that such Extension can work without any fix for a while.

Before adding such enterprise policy, we want to run a field trial
if such approach does not introduce any other problems.

This patch adds this escape hatch as a base::Feature trial. We will
run a field trials in non-stable channels for a while, and convert this
to an enterprise policy for m79.

Bug: 1000982
Change-Id: I27b96d4fbc70fb52eb2c878058d0fd22078fd339
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1792506
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Commit-Queue: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/master@{#695549}