Remove |SVGSMILElement::next_progress_time_|
This state is fairly high maintenance and used only for frame scheduling
and to check if updates need to be triggered for newly added timed
elements. So get rid of it. (For now - it may reappear in the future but
then more likely as part of the time container.)

Also make a pass over arguments which passes the current presentation
time, and rename it to that as well as using double rather than
SMILTime.

Bug: 998526
Change-Id: Icdb477ad86c349d0ddab5b9843b87dc186fd3b67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1795764
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#695558}