Reland "gpu: Load vk functions for querying image/buffer properties"
Original CL: https://chromium-review.googlesource.com/1794142

The original CL was reverted due to using VK_VERSION_1_1 instead of
VK_API_VERSION_1_1 which causes Vulkan on Chrome to break for users with
Vulkan 1.0 or lower.

This CL reapplies the changes in the original CL but updates both the
new code and existing code with the proper version number.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1796527

BUG=996470
TEST=run ozone_test on a machine using Vulkan 1.0

Change-Id: I4fb7de01c6e7c24f88c3736c129bc08f822ba9a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1797245
Commit-Queue: Brian Ho <hob@chromium.org>
Reviewed-by: Peng Huang <penghuang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#695627}