Add drop shadow support for popups on Windows.
This CL adds drop shadow support for popups on Windows when
the FormControlsRefresh feature is enabled.

In order to add the drop shadow, the Window used for popups (TYPE_MENU)
is decorated with the WS_THICKFRAME style. Since WS_THICKFRAME also
provides a frame, the remove_standard_frame flag is set as we are interested
in keeping just the drop shadow.

Since drop shadow is not supported on Windows versions without DWM composition,
then the drop shadow is replaced by a thin-line border by using the
WS_BORDER style.

Bug: 1000898
Change-Id: I04588d8d1703cb98f4b7a239f9cf46d742324a1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1786247
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Ionel Popescu <iopopesc@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#695655}