Ensure that a submenu is shown within screen bounds
On smaller screen it is possible that there is no room for a submenu
on either side of the parent menu. In this case instead of changing
the side just move the submenu the minimum necessary amount to fit the
screen (see screenshots in the bug).

Bug: 991227
Test: MenuControllerTest.TestSubmenuFitsOnScreen
Change-Id: I149849e9b2cf44de1b79f45854c49af3f5fe9eb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1790475
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Vladislav Kaznacheev <kaznacheev@chromium.org>
Cr-Commit-Position: refs/heads/master@{#695667}