Do less initialization in PdfCompositorImpl ctor.
Currently it initializes some DirectWrite code on Windows in unit tests
that is only intended to be called once per process, whereas
PdfCompositorImplTest constructs multiple PdfCompositorImpl objects and
ends up calling that repeatedly, which then fails. Avoid doing this, as
the DirectWrite code is not needed in the test.

Bug: 1002663
Change-Id: I741690466d3111f0095a070bd64f39ebfb835804
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1797225
Reviewed-by: Wei Li <weili@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#695669}