Make sure F6 focus cycle falls on focusable view
Fixes regression after refactoring of DownloadItemView where the new
view implementation is not focusable (but a subset of its children are).

DownloadShelfView returns a DownloadItemView as its first focusable
item (totally reasonable), but since the item itself is not focusable we
need to forward focus to the open button inside it.

I still think it's reasonable to focus an unfocusable item if its
children are focusable, so ideally a fix would change ::SetFocusedView
to focus the first focusable child (if any) or otherwise clear focus.
There's a TODO present inside AccessiblePaneView to address this.

Bug: chromium:1000998
Change-Id: I071d1c59401b35f251cd429ae3b5892462f91adf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1798086
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#695678}