Notification scheduler: Call TaskFinishCallback at correct time.
This CL wraps a inner class in NotificationSceduler class to track all
notification display flow, and invoke the TaskFinishCallback when
all display flows are finished.

Also, fixed an issue that InitHelper should be owned by
NotificationScheduler or we may use invalid pointer when the service
is destroyed.

Bug: 998980
Change-Id: I5d3dfa488eaeaf8b99649987c4982ddb279a4504
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1791274
Reviewed-by: Hesen Zhang <hesen@chromium.org>
Commit-Queue: Xing Liu <xingliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#695754}