[iOS] Don't reset fullscreen for new toolbar heights.
The reset occurred under the previous assumption that the toolbar
heights only change for navigation to/from the NTP, where the primary
toolbar height is 0.0.  This assumption is incorrect because the heights
can also be updated due to trait collection changes.  The navigation
case for NTP is already handled by FullscreenWebStateObserver::
DidFinishNavigation(), so removing the calls to ResetForNavigation()
does not break that case.

Bug: 896675
Change-Id: I2fd1b101f3057e6f3f3bb61802fc1683a1ab0862
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1328234
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Auto-Submit: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#695765}