Rely on UKM filtering instead of custom filtering for NavPredictor
This CL replaces the behavior of sampling the 10 largest links on the
page with beahve to filter entire pages via UKM event filtering. We set
reasonable limits on the repeated evets (10 for clicked, and 40/100 for
anchor elements). The 40/100 is inherited from the renderer, which only
sends that many metrics to the browser.

Bug: 1001593
Change-Id: If12e30e1a2cf50cb3975f5bc3749faac4534e241
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1797329
Commit-Queue: Ryan Sturm <ryansturm@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#695782}