Check that Widevine is a directory
Now that Widevine is in a nested directory, check for it's existence as
a directory (and not a file). Also update the permissions on the
actual Widevine library to be ${SHLIB_PERMS}.

Tested by building "installer" and then installing it. Widevine is
installed and Shaka player works with Widevine.

Bug: 1002648
Test: Installed Chrome and played Widevine content.
Change-Id: I531c0f7da58b50d36bb4e72fdde11f05cfa9f5a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1799306
Reviewed-by: Michael Moss <mmoss@chromium.org>
Commit-Queue: John Rummell <jrummell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#695817}