Initially send the correct display info for WMR
The XR process initially sends back dummy values (1024) to the renderer
for the display's width/height and later sends back the correct values
on the first frame.

However, in WebVR, most web pages do not look for the updated values
after entering the rendering loop. As a result, WebVR pages present
with a resolution of 1024 until the user exit presentation and then
re-enters presentation without refreshing the page.

In WebXR, Blink does not update the XRWebGLLayer's width/height when
the XR process sends back the correct values, so presentation also
occurs at 1024 until exiting and re-entering presentation.

This change updates the display info in the XR process and initially
sends back the correct value so that both WebVR and WebXR pages has the
correct width/height when starting presentation.

Bug: 1000897
Change-Id: I3491d43245bb53100212fd2d0ab38d8f1222e7c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1785883
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Patrick To <patrto@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#695856}