Fix weird/incorrect accessibility text when the dse icon is enabled
Accessibility text should provide the user with a description of what
interaction with the given UI element will involve. The dse icon
is purely cosmetric so it doesn't need an accessibility desicrption
unless the underlying button actually does something. This change
turns off the important for accessibility flag when there's no
description of the button available. There will be no description
available when no interaction with the button is possible.

Bug: 995059
Change-Id: Ifed72d4f63d8ad58a615a583221d1a63940b72e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1799683
Commit-Queue: Brandon Wylie <wylieb@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#695888}