Update rules for opening dialer directly.
For Q and above, we don't open the dialer in any case.
For N and below, we always open the dialer. If the device is locked then
a notification is shown which is dismissed on unlock.
For O and P, we open dialer if devices is unlocked. If the devices is
locked then we don't open the dialer and keep the notification.

(cherry picked from commit 19eb6004a87af1c104486181ab2220895efdaec6)

Bug: 1002137
Change-Id: I81f4386439efe2b60b6b35a8bb8cdc7c59fb591c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1795766
Commit-Queue: Himanshu Jaju <himanshujaju@chromium.org>
Reviewed-by: Alex Chau <alexchau@chromium.org>
Reviewed-by: Michael van Ouwerkerk <mvanouwerkerk@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#695560}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1800667
Cr-Commit-Position: refs/branch-heads/3904@{#95}
Cr-Branched-From: 675968a8c657a3bd9c1c2c20c5d2935577bbc5e6-refs/heads/master@{#693954}