Move NetworkDelegate::AuthRequiredResponse enum to web_request_api.h.
The enum is no longer being used in net/. However, it's still used by
the WebRequestAPI to indicate how an auth challenge received through
the NetworkService API should be responded to, so rather than remove the
enum, this CL moves it to ExtensionWebRequestEventRouter.

This CL also converts it to an enum class, for better type safety, at
the cost of additional verbosity.

NetworkDelegate::AuthCallback is similarly moved to
ExtensionWebRequestEventRouter.

Bug: 934009
Change-Id: I04dca2c6d66ca52b8cc4c8fa44e7d5af1429646f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1799367
Reviewed-by: Eric Roman <eroman@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#696119}