Post OnRemotePlayStateChange
The FlingingRendererClient receives calls to OnRemotePlayStateChange on
the media task runner. It then propagates these changes to WMPI, which
calls into Blink. This might result in some crashes, when Blink tries
to allocate some resources on the wrong sequence.

This CL makes sure we are on the main task runner before calling into
Blink.

Bug: 983723
Change-Id: Ib3765c7dfd67c674e31becfc83c4ae7d4aeda612
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1802028
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#696219}