Put errors from AXTreeSourceChecker in a crash key.
The goal is to figure out why we're occasionally unable
to serialize the Views accessibility tree.

To make this possible, I refactored AXTreeSourceChecker
to return errors as a string rather than logging them.
I also added a unit test so we can be more confident that
AXTreeSourceChecker is working correctly and that it's
outputting useful error strings when something goes wrong.

The crash keys are limited in size, so I added a bit more
info at the top of each error string about the relevant
nodes.

Bug: 990282

Change-Id: I7b84ed26b360e6be8cb8a5ac574a48aa4e1f5efa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1797129
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#696221}