Animations: Accept 'accumulate' as a composite operation
This CL does the first step in adding support for the accumulate
operation, which is to allow it in the JavaScript APIs. Because
accumulate is often spec'd to be identical to add, this actually
passes the tests for many properties. The properties that actually
need real accumulation implementations (to come in future CLs)
include transform and filter lists.

As with 'add', 'accumulate' is hidden behind CSSAdditiveAnimationsEnabled.

Bug: 788440
Change-Id: I0eb3374a9b76a4e6a90ae85602fddac9ab0d3a7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1797652
Reviewed-by: Yi Gu <yigu@chromium.org>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#696368}