[Sync] Sync BookmarkNode GUIDs
This CL does the following:

1. Syncs BookmarkNode GUIDs back to the server, committing local
creations/updates and processing remote creations.
2. Modifies the APIs responsible for the addition of folder and URL
bookmarks to take in a GUID in order to make (1) possible.
3. Upon local creation of a bookmarks, it sets the temp server id (later
to become the originator_client_item_id) to the bookmark GUID.
4. Equivalently, in the case where a bookmark created remotely is
received from the server without a GUID (legacy client commit), sets the
GUID field to the value stored within originator_client_item_id, as long
as it is of valid GUID format; otherwise, a new GUID is assigned.

(3) and (4) allow for the maintenance of a single GUID throughout the
lifetime of bookmarks created by modern clients, and also for the
assignment of a GUID to bookmarks created by legacy clients that is
consistent among all modern clients.

Bug: 978430
Change-Id: I97efb13f825c83cfa31d59b8bcdf85eb23f0390c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1702310
Commit-Queue: Pauline Leitao <psivieroleitao@google.com>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#696385}