[iOS] Move and rename tab model web state list delegate.
To prepare for the change in WebStateList ownership
(see crrev.com/c/1796358), this CL renames and moves
TabModelWebStateDelegate to be BrowserWebStateDelegate,
living in c/b/main.

(The ownership change CL will have the Browser object
create and own this delegate instead of the TabModel).

Change-Id: I92efa11b88dbe4ffb561be485fcd836bf9e7a7ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1803312
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#696390}