Add NetworkIsolationKey to HttpServerProperties::*QuicServerInfo methods
Also replace QuicServerId with SchemeHostPort, like all other methods.
This both removes the unnecessary privacy mode field, and allows a bit
more code sharing with the ServerInfo logic, which is also keyed on
SchemeHostPort+NetworkIsolationKey.  Done in one CL so that we only
lose data once, and because sharing an index with the ServerInfo map
avoids introducing a new QuicServerProperties+NetworkIsolationKey
struct only to remove it in my next CL.

This CL does not make any code pass in a non-empty NetworkIsolationKey
in production. I'll address that in a followup CL.

Bug: 969890
Change-Id: If2bbe2277c99bcba3cf7744abe358649c423f0e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1797386
Reviewed-by: Zhongyi Shi <zhongyi@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#696499}