Add comment for use of android/gyp script.
This adds a comment to the MacOS build's inclusion of
build/android/gyp/util/build_utils.py in "sources". The build works
fine without it, but the entry is needed to inform GN of the
dependency on that tool.

Change-Id: Ib17f6b76c3fd45a0bb5421496f3c85b1d756d8d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1802679
Auto-Submit: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Jamie Walch <jamiewalch@chromium.org>
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Cr-Commit-Position: refs/heads/master@{#696517}