PerfCollector: zero periodic interval disables periodic collection
A periodic interval of zero doesn't make sense. It would cause us to constantly
try to collect profiles.
Check for a zero value, and disable periodic collection in that case. We can
use this property to disable periodic collection via Finch in an emergency.

BUG: 1002754
TEST: unit_tests
Change-Id: Iec07d1c040e1a72cf86099fd13abd7af2e4e673d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1803908
Commit-Queue: Gabriel Marin <gmx@chromium.org>
Reviewed-by: Gabriel Marin <gmx@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Cr-Commit-Position: refs/heads/master@{#696532}