DevTools: Add ARIAUtils.setCheckboxAsIndeterminate
This adds a new API to UI.ARIAUtils to set an element's aria-checked
attribute to 'mixed', which is the equivalent of setting
intedeterminate=true on an actual checkbox element.
This is needed to support keyboard accessibility in the javascript and
event listener breakpoints views.

Equivalent to the Animation checkbox in this screenshot:
https://gyazo.com/99f2067f6810af3366943176393d8b2f

Bug: 963183
Change-Id: Ie09f8e993d1776774c410f6099356f3a1d912048
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1715815
Reviewed-by: Pavel Feldman <pfeldman@chromium.org>
Auto-Submit: Jack Lynch <jalyn@microsoft.com>
Commit-Queue: Jack Lynch <jalyn@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#696551}