Adds sandbox diagnostic api to service manager
This will provide information for the Windows chrome://sandbox webui.

The API is a little convoluted as it works like async-io and the
sandbox does not understand Chrome's threading model.

The service manager mediates this by taking in a callback to which a
completion is posted. This will be provided by
the sandbox webui which is intended as the caller of this API.

Bug: 997273
Change-Id: I6f8d56afa71376f6382913e87ee7ef86fe9c1068
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1766455
Commit-Queue: Alex Gough <ajgo@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: James Forshaw <forshaw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#696582}