gpu: Implement alpha:false support for webgl swap chain backing
This change adds support for alpha:false webgl contexts by overriding
the EGLImage internal format which was implemented in ANGLE recently:
https://chromium-review.googlesource.com/c/angle/angle/+/1777099

The client uses the RGBX8888 format when alpha: false is specified, and
the service maps this to internal format RGB with the same backing dxgi
format (BGRA8) as before which allows ANGLE to ignore the alpha channel
when doing blits, copies, read pixels, etc.

A pixel test for resolving a multisampled RGB8 renderbuffer to an
alpha:false drawing buffer is added.  The test fails before this change
because GLESv3 doesn't permit a blit from a multisampled renderbuffer to
a target with different format (RGB8 vs BGRA8).  The test passes after
this change.

This CL also adds unit tests for GLImageD3D that test storage format and
GL internal format combinations that we care about.

TEST=Pixel_CanvasLowLatencyWebGLSwapChainAlphaFalse

Bug: 939657
Change-Id: I2f4ad1add34386b298d12fc49867bbe4f88239e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1777119
Commit-Queue: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#696586}