Print Preview: Do not show N-up setting when it cannot be honored.
The internal printing code can only handle N-up for content from Blink
or the PDF plugin. For any other plugin, which is effectively just
Flash, the N-up setting does not work. So do not show the setting in the
Print Preview UI in this case.

To do this, take advantage of an earlier CL that exposes whether the
content being printed is PDF or not, and plumb that setting into the
Print Preview UI and show the pages-per-sheet control based on that.

Change-Id: Ic57f50a451aba9ba6918b59be572a371a1b4beca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1790256
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#696598}