[gamepad] Avoid leaking GamepadService in tests
GamepadService is implemented as a global singleton. Typically
there is only ever one GamepadService instance for the lifetime
of the browser, but in tests a new instance may be created for
each test case. Previously, when a new instance was created the
previous instance would leak. This CL makes sure the previous
instance is destroyed before a new one is created.

At startup, the GamepadService is passed a
service_manager::Connector to allow gamepad data fetchers to use
the device service to enumerate and communicate with gamepads
on the polling thread. The Connector must be destroyed on the
same thread it is bound to. Since the polling thread is owned by
GamepadProvider, the Connector has also been moved into
GamepadProvider so it can be destroyed along with the polling
thread. A pointer to the Connector is provided to the data
fetcher when it is added to the GamepadProvider.

BUG=961039

Change-Id: I66741b3ea5971cea3bbf2ba81bd1f734a7664ff7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1777024
Commit-Queue: Matt Reynolds <mattreynolds@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#696619}