[Files Ranking] Validate Drive files exist on-disk
There is a bug where a Drive file can be deleted via DriveFS, which
takes a few seconds to propagate back to the Drive QuickAccess results,
creating a window where we could surface a Drive file that can't be
launched.

We already deal with this for local files by checking they exist
on-disk before finalizing the results, so this CL does the same thing
for Drive files.

integration service right now.

Tested: manually, as we don't have a good way of mocking the drive
Bug: 1003588
Change-Id: I3ad839c18ad0dbb28fec13141e756993a4b65c74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1802775
Reviewed-by: Jia Meng <jiameng@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#696691}