Reduce the per type overhead of promises
Previously the executors where using OnceCallbacks. The problem with
that is we're going to redundantly invoke the callback destructor
when we run the callback. This is redundant because the executor will
overwrite itself (and the callback it contains) with the promise
result.

This patch takes advantage of the way Callbacks operate in chromium
and casts the callback to the repeating version and passes a const
ref to the executor.  This eliminates the overhead of passing a
scoped_refptr<Callback> by r-value.  We still respect the once-ness
of these callbacks despite these binary size shenanigans.

Note Promises are experimental, not used in production code at this
time. We are landing this only to make it easier to experiment with
binary size.

Bug: 906125
Change-Id: Ifd8f53e6dfdc6a65921b53222e05bdf384ac8df3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1803135
Commit-Queue: Alex Clarke <alexclarke@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#696750}