[Autofill Assistant] Ensure that UMA dropout is recorded only once.
The UMA histogram AutofillAssistant.DropOutReason is currently showing incorrect data, because the number of autobot drop-outs adds up to more than the number of autobot starts. This likely happens because the shutdown procedure for autobot is different between closing autobot from the user interface versus closing Chrome entirely. As a consequence, dropout reasons are logged in two different places in code. We might be able to refactor such that we can avoid this in the future.

For now and as a hotfix, this CL ensures that every instance of autobot can record at most one dropout reason.

Bug: b/140371674
Change-Id: I17b882e1327be6228867321b4ba6536b45a6822f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1780834
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Cr-Commit-Position: refs/heads/master@{#696751}