Attempt to de-flake ChildProcessTaskPortProviderTest.DeadTaskPort.
The ChildProcessTaskPortProvider does not provide any callback for when
a child process has exited. The test previously used the seqno on the
Mach port used for DEAD_NAME notifications to determine if the
process-died message had been delivered. But this is racy against
libdispatch delivering the callout to actual updating the bookkeeping.

Switch to to just looping the run loop until a timeout or exist
condition has been met.

Bug: 986288
Change-Id: Ic965b80878eb330caf6e83b307bed705dad4df1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1804521
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#696863}