UIA: Fix focus-event filter for event tests
Many of the UIA event tests are exhibiting a low-level of flakiness (too
low to be disabled, but high enough to be noisy).  These are mostly
related to spurious focus events.

It was determined that the logic that attempted to filter out focus
events outside the frame was faulty; it was testing the "automation id",
which is not necessarily unique.  This has been changed to the "runtime
id", which *is* unique.

Bug: 928949
Change-Id: I4e1c6f820f42c796f909513b9de5bb03a711eb4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1804382
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Commit-Queue: Ian Prest <iapres@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#696938}