Use base::TimeDelta as internal representation for SMILTime
This changes the internal representation of SMILTime to use
base::TimeDelta. This aligns better with handling of time in general,
and should hopefully avoid issues that arise from having a floating
point representation.

The special values "indefinite" and "unresolved" are stored as
TimeDelta::Max() and TimeDelta::Max() - 1us respectively to satisfy the
requirements for ordering.

Bug: 1003338
Change-Id: I26ef5d14b45f646dcfcfe672736aa7caf343a375
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1806753
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#697148}