[iOS] Add (and use) an explicit -disconnect for TabStripController.
TabStripController relies on -dalloc to stop its various WebStateList
observations. If the WebStateList is destroyed before the dealloc runs,
the non-empty observer DCHECKs are hit. The upcoming change to
WebStateList ownership (see crrev.com/c/1796358) will cause this to
happen, for example.

To fix this, an explicit -disconnect method is added to
TabStripController, and the tab strip coordinator calls it from -stop.

Change-Id: I22c2fc967c0b42c544ab93dce17518e93d272f43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1809298
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#697241}