[ios] Only Add BadgeView as AccessibilityElement if not already one
The BadgeView is added as an accessibilityElement when it is added
as a subview to LocationBarSteadyView. It is also removed and added
back in as a part of showing and hiding the SteadyView when the
Omnibox focuses/defocuses. At the beginning the SteadyView is shown,
so the BadgeView was being added twice. This change checks if the badgeView
is already an a part of the accessibilityElements before adding it.

Bug: 1004339
Change-Id: I59d6d7b148260ccbc61e68a1487e56c3bd2a72af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1807596
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/master@{#697274}