Remove RenderPassDrawQuad::mask_applies_to_backdrop
It didn't mean what the name suggests because
1. the mask didn't apply to backdrop, but backdrop filter;
2. in software renderer and gl renderer, it meant the mask applies to
   backdrop filter *only*; while in skia renderer, it was ignored.

The flag was never set in cc_unittests while mask on backdrop filter
tests passed without setting the flag.

Applying mask only on backdrop filter but not contents is not a case
in real world, so we don't need to support it.

With this CL alone, the mask will be applied on layer contents twice
but the pixel result is the same because the second application has
no effect (or the difference is below the fuzzy threshold of the
pixel tests).

https://chromium-review.googlesource.com/c/chromium/src/+/1799701
will fix the double mask issue by not issuing draw quads in
PictureLayerImpl::AppendQuads() for backdrop filter masks.

Test: All existing tests pass.

Bug: 1003414
Change-Id: I8c19017a95308c60904cb98ba879709e11a29d9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1804321
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Eric Karl <ericrk@chromium.org>
Reviewed-by: Martin Barbella <mbarbella@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#697303}