Let layer tree mode mask share most code path for layer list mode
In layer list mode, we treat a mask layer as a normal layer, except
that the masked layer is forced to create render surface, and the mask
layer blend on the masked layer in kDstIn blending mode.

Now we treat mask layer in layer tree mode in the same way.

Bug: 1003414
Change-Id: Ieb67009f4f2ae8faf9b7adf8195499649413238c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1799701
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Malay Keshav <malaykeshav@chromium.org>
Cr-Commit-Position: refs/heads/master@{#697315}