Fix settings search when matched text is outside a <settings-section>
There are settings promo banners that contain text that isn't inside
a <settings-section> element (e.g. the Chrome OS settings "If a setting
doesn't show on this page" banner).

If you type search text that matches this string you'll hit a JS
assertion because a "parent" is null while walking up the DOM tree.
The loop in question checks for null parent, but there's an if()
condition that checks parent.nodeName that needs to test for null
as well.

Bug: 1003970
Test: added to browser_tests
Change-Id: I3c5a7367464502461ab3a158c102f4ad246d9f99
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1807777
Reviewed-by: Michael Giuffrida <michaelpg@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#697347}