[App Management] Ignore app changes if the app is not in App Management.
App Serice updates are issued as delta of changes. This means that every
app update will not have the ShowInManagement field. This causes every
app update to be sent to the App Management front end. This CL ignores
any app updates to apps that are not already in App Management. Adding
new apps is unaffected.

Bug: 1002233
Change-Id: Ie2d674238a2d43d167c54f45c78bdb9060a770e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1807650
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Commit-Queue: Jeevan Shikaram <jshikaram@chromium.org>
Cr-Commit-Position: refs/heads/master@{#697426}