[CrOS Bluetooth] Add debug logs toggle to chrome://bluetooth-internals.
If supported, this toggle is displayed in a new tab named "Debug Logs",
and when the toggle's value is changed,
BluetoothInternalsHandler::ChangeDebugLogsState() is invoked.

In this context, the toggle is only supported if the associated flag is
enabled and if the primary account is a Googler's corp account.

Bug: 1004572
Change-Id: I2fd7a3a34d95cf5f642c6525a22d84245d81084b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1804683
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Auto-Submit: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#697434}