Include EthernetEAP properties with Ethernet in GetManagedProperties
Shill stores the Ethernet EAP properties in a separate service entry.

This separation creates unnecessary and awkward complexity for the UI
and serves little purpose at the UI layer.

Previously we merged the dictionaries in the JS layer. This change
does the merge in the mojo model layer, populating the managed EAP
properties for the Ethernet network when they exist. This is done
regardless of whether they are actually applied.

This CL also adds EthernetEAP properties to chrome://network for
debugging purposes.

Bug: 853953
For comment only changes to cros_network_config.mojom:
TBR=dcheng@chromium.org

Change-Id: I7c47c7f33c9df1515b1f8674b4e6178747fcd35f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1783880
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#697448}