Always plug "chrome://add-supervision" into FindInstance()
Currently, GetDialogId() starts out as the empty string, gets set to
kChromeUIAddSupervisionURL="chrome://add-supervision" when the Add
Supervision dialog is opened, and sometimes gets reset to empty string
when the dialog is closed depending on which button is pressed. This CL
ensures kChromeUIAddSupervisionURL is always passed to FindInstance()
instead of the empty string which can lead to unpredictable behavior
and can easily return something other than null.

Bug: 1000978
Change-Id: I0287aefdb3451db2e8f388dbff659623d4e7f982
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1787339
Commit-Queue: Toby Huang <tobyhuang@chromium.org>
Reviewed-by: Michael Giuffrida <michaelpg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#697485}