Handle display mode in GetToolbarActionViewForId
When the toolbar's display mode is not ::NORMAL the extensions container
is never initialized. Not checking for this (or container availability)
leads to crashes when the extension-uninstall dialog tries to find where
to anchor to in the window.

This intentionally checks the display mode instead of container
availability as the containers should always be created in ::NORMAL.

Bug: chromium:1005154
Change-Id: I5c51bf0f2a02c75896355ccee39077d6c55a0d3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1809974
Commit-Queue: Peter Boström <pbos@chromium.org>
Commit-Queue: Caroline Rising <corising@chromium.org>
Auto-Submit: Peter Boström <pbos@chromium.org>
Reviewed-by: Caroline Rising <corising@chromium.org>
Cr-Commit-Position: refs/heads/master@{#697669}