Further decouple emulation from OnSynchronizeVisualProperties
Emulation only changes a few VisualProperties, and then calls
SynchronizeVisualProperties(), which is a subset of
OnSynchronizeVisualProperties(). This moves code from
inside SynchronizeVisualProperties() to OnSynchronizeVisualProperties()
when the emulator does not change those values. This helps to clarify
which code needs to run from the emulator and which does not.

R=avi@chromium.org

Change-Id: Ia8253ab6a32d4fd2c5f0b69c1310536d75faa6b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1809604
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#697712}