Don't abandon ScriptPromiseResolver when rejecting via ExceptionState
ScriptPromiseResolver has logic in its destructor to ensure that you've
either resolved or rejected its promise (or not accessed the promise
at all), to catch abandoned promise issues. This technically wasn't
an issue for us because we were rejecting the promise via
ExceptionState, but since we access the ScriptPromiseResolver's promise
it thinks it was abandoned. 😢

Bug: 1003617
Change-Id: I276fde6ca0e2877cdc4793cb1674bca0821ca2bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1809598
Commit-Queue: Will Cassella <cassew@google.com>
Reviewed-by: Jacob DeWitt <jacde@chromium.org>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#697740}