In tab dragging, the window size should be larger than its minimum size.
In tablet mode, when doing a tab dragging, the dragged window's size
is 1/4 of the work area's size. It worked well for big screens, but on
smaller screens, the size might be smaller than the window's minimum
size. We should take the minimum size into consideration as well.

Bug: 1002728
Change-Id: I6a834150015890f7e9eb949b8d724c064474a84f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1799693
Commit-Queue: Xiaoqian Dai <xdai@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#697759}