Occlusion Tracker: Fix occlusion state of forced-visible descendants of hidden windows
This CL fixes a bug in which a forced-visible window that is a descendant
of a hidden window used to be considered occluded along with its descendants.
This causes problems with virtual desks, since a browser window may be
forced visible (while projecting), but the desk it resides on is switched,
hence its container is hidden.
In this case, the browser window, as well as its descendants should be
considered visible from an occlusion stand point.

BUG=998214
TEST=Added a new test

Change-Id: I7031f16f4fbfd6e86228670f5d58ef1ac135eab1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1810104
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#697826}