Abort chall-resp existing user login on PIN error
After the smart card PIN verification fails (e.g., when the user
entered the wrong PIN), immediately abort the user's sign-in attempt,
instead of trying to re-request the PIN.

This is a temporarily workaround until the proper UI gets implemented.
With the current UI, it's too confusing if we don't abort the sign-in
attempt, since the PIN keypad is shown again without displaying any
error.

This workaround will be removed after the UI is redesigned in
accordance to crbug.com/1001288.

Bug: 1002159
Test: start sign-in as an existing smart card user, enter wrong PIN, check that the PIN keypad disappears and the "Couldn't recognize your smart card" message appears
Change-Id: I12107dc3183c9428f3f5a17d3f33ddafea9573e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1793085
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#698010}