Make dark mode omnibox colors consistent with Clank.
Omnibox dark mode used to have a non-standard, lighter color.

On Clank it is 38% black over the toolbar color, which is exactly
GoogleGrey900. Since we do not use any math to derive this number in a
way that supports anything other than the default, we have just set the
value to GG900.

This also makes the omnibox consistent with the results pane, which
uses GG900 as its background, for a more seamless transition between
omnibox modes.

Bug: 1002249
Change-Id: I74cbff08150c5a0fd0beeb545b5ceb10cd7aab35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1814864
Auto-Submit: Dana Fried <dfried@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#698279}