[a11y] Shrink Cookies Dialog as much as possible while being usable.
Cookies dialog is our talled tab-modal dialog, so it often gets cut off
when working on a small screen or in a small browser window. This is
very bad for users who use high DPI due to low vision.

This change reduces vertical padding as much as is reasonably possible
and places the cookie info in a scrollable container while reducing its
height.

This is not perhaps the most ideal solution but it does not look bad.
Ultimately this dialog needs to be redesigned.

Bug: 872991
Change-Id: If768f5ecf4ef4cc1fc64a51ea8bd466622bf4f2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1815079
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Charlene Yan <cyan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#698355}