Remove URLRequest::set_delegate()
This method is no longer used. It was added so that
ResourceDispatcherHost could create a URLRequest itself, and then
pass it over to ResourceLoader, which took ownership and set itself as
the delegate. It also allowed the downloads code to create its own
URLRequests before handing them off to the RDH.

Also make URLRequest::get_upload() a test-only method. It had been used
to allow extensions to poke at upload data, which is no longer done in
the network stack. We could remove this method entirely, reworking the
tests that use it, but that's probably not worth the effort.

Bug: 934009
Change-Id: I9b2fc70da7b7d9867c396760b26d583067802b50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1814393
Reviewed-by: Eric Roman <eroman@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#698499}