[CryptAuth v2] Delay Instance ID retrieval in ClientAppMetadata
Wait to call InstanceIDDriver::GetInstanceID() until necessary. This is
a fix that allows enough time for the CryptAuthGCMManager to add itself
as an app handler for the app ID "com.google.chrome.cryptauth". If
GetInstanceID() is called before this happens, Instance ID data will not
be successfully retrieved from persisted storage and the data risks
being overwritten due to https://crbug.com/1006276.

In future work, we plan to delegate Instance ID tasks to the
CryptAuthGCMManager, but this quick fix works for now.

Bug: 1006280, 899080, 951969
Change-Id: I2ea092d97f2afc1bb995b1582cdc66366f530337
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1814660
Commit-Queue: Josh Nohle <nohle@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#698535}