Add decorated suggestion view class.
This change introduces DecoratedSuggestionView that is hosting the final suggestion
content view. This class features suggestion icon that is rendered on the left side
of the suggestion itself.
Since the ViewGroup serves as an isolated focus area, it must be a separate entity
from BaseSuggestionView. I chose to implement it in a separate file for clarity and
to ensure future changes do not increase entanglement between the classes.

Change-Id: I413c901b1e119355b27b1f8f970d0acc80cd1d9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1726900
Commit-Queue: Ender <ender@google.com>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#698641}