Move enums from CrostiniInstallerUIDelegate to a mojom file
Enums |InstallationState| and |Error| in class |CrostiniInstallerUIDelegate|
are moved to file crostini_installer_types.mojom and renamed to
`InstallerState` and `InstallerError`. All other changes to .h and .cc files
are just simple replacement to migrate to the new types.

The purpose of this CL is to allow the js code of the new Crostini WebUI
installer to use the same enums.

Bug: 929571
Change-Id: I185bb94b79187db98850dba4d23a7ba221e1d1e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1815797
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Timothy Loh <timloh@chromium.org>
Commit-Queue: Jason Lin <lxj@google.com>
Cr-Commit-Position: refs/heads/master@{#698836}