X11 and Ozone: Move setting transparency to DNWA
This functionality seems to be platform independent, so
the PlatformWindow interface got new method:
ShouldWindowContentsBeTransparent. It is now implemented
for both X11 and Windows platforms (the last one was taken
from DWTHWin).

Now, the DesktopNativeWidgetAura::UpdateWindowTransparency
sets the transparency based on the value returned from
ShouldWindowContentsBeTransparent.

In addition to it, the interface has default implementation
and returns false as DWTHPlatform used to be doing before.
The platforms that are interested in providing different
values must override and implement that method.

Also, setting native frame has been moved as long as
OnNativeWidgetCreate call is passed all the way down
from DWTHX11 to DWTHLinux to DWTHPlatform.

Bug: 990756
Change-Id: I7069609a2abe068b3553039d2e03ab98c3babd98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1789222
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#698843}