[Mfill Android] Fix crash on talkback selection
Speculative fix because I couldn't reproduce it.

Crashes occurring since M75 show seg faults in the method
AutofillKeyboardAccessoryAdapter::AcceptSuggestion. It uses exactly one
pointer (controller_) and this pointer is never set to null. If compared
with the old AutofillPopupViewAndroid, it shows that the adapter for the
accessory view doesn't reset the controller after |ViewDestroyed()|.

This methods does this exactly like the AutofillPopupViewAndroid which
is not perfect [1] but risk-free to merge to M77.


[1] An ideal solution would use unique_ptr instead of raw pointer and
have the controller delete the views instead of them self-destructing.
Besides the merge, it makes more sense to clean this up when we remove
the old popup code entirely.

Bug: 1006311
Change-Id: I6766504c1ed7fffd0befb872d1e4a745dfdb45ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1819441
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#698907}