viz: Always output full quad list on surfaceless platforms
in crrev.com/c/1761708 we changed how we do partial damage on surfaceless
output surfaces.
Instead of copying undamaged contents from the previous buffer, we started
expanding the damage rect and recompositing what changed onto the buffer
since the last time we composited to it.

This wrongly assumed that the damaged area could be arbitrarily expanded
and all the quads for the frame would be there.

On some platforms, this was not the case, since SurfaceAggregator would
prune quad that didn't intersect with the damage region.

This CL disables that optimization on platforms using surfaceless output
surfaces so that the new logic can expand the damage rect and there will
always be quads covering the expanded area.


Bug: 1004621
Test: OOBE on cave stopped flickering
Change-Id: Ifbafc294b96d183992483d833533668d9f2e0ea1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1818819
Reviewed-by: Yi Xu <yiyix@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Cr-Commit-Position: refs/heads/master@{#698926}