Pool SharedImageVideo for MCVD.
To reduce thread hops when generating VideoFrames, this CL adds
PooledSharedImageVideoProvider.  It builds on the direct provider by
maintaining a pool of SharedImageVideo objects, which it hands out
on its own thread to clients.

It also adds histograms to track VideoFrame creation time, so that
we can see which device(s) benefit from pooling.

All of this is turned off by default; it must be switched on by
enabling UsePooledSharedImageVideoProvider.

Change-Id: I4cfc03938dd4717252f3f527a3bcab0e7ae085a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1699138
Commit-Queue: Frank Liberato <liberato@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: vikas soni <vikassoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#698985}