[BackgroundSync] Check for connectivity before creating ready list.
When BackgroundSyncManager is notified of a change in network
connectivity, we immediately proceed to calculate two lists of one-shot
and periodic Background Sync registrations that we're ready to fire events for.

For each registration, we individually check whether there is connectivity.
This is wasteful. This CL adds a check before calling FireReadyEvents() and bails out
if there is no connectivity.

Bug: 574478
Change-Id: Iaf1bfe8552edbfbfbef8e6d8ffb0ab1acacaf341
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1816569
Auto-Submit: Mugdha Lakhani <nator@chromium.org>
Reviewed-by: Rayan Kanso <rayankans@chromium.org>
Commit-Queue: Mugdha Lakhani <nator@chromium.org>
Cr-Commit-Position: refs/heads/master@{#699298}