Restrict 'keySplines' x-values to the [0, 1] range
Per SMIL[1], all the values in a key spline should be within [0, 1], but
to be consistent with CSS' timing function, allow y-values to be outside
that range (at least for now).

While in ParseKeySplinesInternal(...), rename the |pos_*| local
variables to something that carries a little bit more semantic meaning.

[1] https://www.w3.org/TR/SMIL3/smil-animation.html#adef-keySplines

Bug: 1006544
Change-Id: I6baa4085943eca8144605fb0a5c442bc8486ad87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1821158
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#699373}