DevTools: Event Listeners a11y enhancement
In the Elements Event Listeners pane, one of the things that is
possible due to the "Ancestors" inclusion is that a parent element
with an event listener can be navigated-to via a mouse click on
the element name. This functionality is now extended to keyboard-
only users by handling the enter keypress for that same tree node.

Change-Id: I7603c43991079af24cbf749b10ff270928945d21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1815820
Reviewed-by: Yang Guo <yangguo@chromium.org>
Commit-Queue: Robert Paveza <Rob.Paveza@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#699407}