tablet: Remove set opacity of top window before screenshot.
It wasn't working as expected, if the screenshot was kept alive, you
can still see the top window. A patch [1] was added to make this not the
case, but the compositor still draws the frame where the top window
was hidden (seen as a flash) even with compositing locks applied. This
is OK for enter tablet animation as the entire animation will cover
the top window anyways. I will look into a way that works for exit
tablet animation.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1797482

Test: manual
Bug: 948706
Change-Id: Ic3b7843c70a5f71ac5c5ad0c2e7290af81d01551
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1804523
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#699519}