Use manifest.json to determine features supported by Widevine CDM on Linux
Use the manifest.json file bundled with Widevine CDM (on Linux) to
determine the features supported by Widevine rather than have them
explicitly specified in code. This is handled by the component updater
for Windows and Mac, so extending it to Linux now that the Widevine CDM
for Linux is in a separate directory.

Bug: 971433
Test: encrypted media browser_tests pass
Change-Id: I696ce5e77b675deb7f97b9b56a6da23814038521
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1808309
Commit-Queue: John Rummell <jrummell@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#699548}