Change the return type of GetOverflowStringName()
As a step to remove all uses of WebLocalizedString::kFooBar,
this CL changes the return type of GetOverflowStringName()
from WebLocalizedString::Name to int.

In addition, this CL renames GetOverflowStringName() with
GetOverflowStringId() according to the return type is changed.

As changing the return type, this CL replaces below ones with
matched IDS_FOO.

  - kAXAMPMFieldText
  - kOverflowMenuCaptions,
  - kOverflowMenuCast,
  - kOverflowMenuEnterFullscreen,
  - kOverflowMenuExitFullscreen,
  - kOverflowMenuMute,
  - kOverflowMenuPlay,
  - kOverflowMenuPause,
  - kOverflowMenuDownload,
  - kOverflowMenuEnterPictureInPicture,
  - kOverflowMenuExitPictureInPicture,

Bug: 995644
Change-Id: I0347751c8df31fb18e88f65e0024f1e4ec94b217
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1819285
Commit-Queue: Gyuyoung Kim <gyuyoung@igalia.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/master@{#699571}