Implement requesting WebGPU adapter with GPURequestAdapterOptions
This patch is the first one to support creating GPUDevice with extensions.
In this patch we implements requesting a WebGPU adapter with options.

Now we only support "powerPreference" as an option to request an adapter.
The option is treated as an enum and passed from Blink to the WebGPU
command buffer to create a Dawn adapter. The "high-performance" adapter
refers to the discrete GPU, and the "low-power" adapter refers to the
integrated GPU. The default adapter will be the one that refers to the
discrete GPU.

BUG=chromium:996713

Change-Id: Iffb98410cf2a2d6d4eb127bdd09a0886099db127
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1794033
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Austin Eng <enga@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>
Cr-Commit-Position: refs/heads/master@{#699598}