Load the default icon when Crostini app_id is invalid.
Crostini could generate a shelf id with the prefix 'crostini:', and use
this shelf id as the app_id, which couldn't match to an app in
AppService. Therefore AppService loads an empty icon. This CL checks
the prefix, and if it is a Crostini app, sets the resource_id as the
invalid resourceId, so that AppService can load the default penguin
icon as the expectation.

BUG=1006121

Change-Id: I08e4a21e4cf7119170e5c2eeb6d410020559af7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1820923
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#699619}