X11 and Ozone: Move IsVisible to DWTHPlatform
This moves IsVisible impl from DWTHX11 to DWTHPlatform
and adds IsVisible() method to PlatformWindow.

Scenic and Windows implementations are not implemented and
return true.

Drm's implementation is set to NOTREACHED as it is not expected
that the DrmWindowHost::IsVisible is ever called.

X11 and Wayland implementations return visibility for unminimized
windows and mapped windows (in case of Wayland, it means that
either xdg_surface or xdg_popup exist).

Bug: 990756
Change-Id: I09ff60a9064ea5bef674d202c5bee994e238c496
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1807239
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#699654}