[Device Info] Implement gc on startup
This CL implements a garbage collection routine for device info. This
routine is run on every startup of sync. This is okay from performance
point of view as all data is anyway in memory at this stage and number
of entities we iterate over is small.

Bug: 1007233
Change-Id: Icd6165a532a67218bcdf1d4d275b5b12cddbfaf9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1821785
Commit-Queue: Jan Krcal <jkrcal@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#699791}