Add profile menu for PWAs
Make ExtensionAppShimHandler be an AvatarMenuObserver (making it very
similar to the existing ProfileMenuControllerInternal::Observer).
Package up the profile list into mojo, and send it to
AppShimController, where it is un-packed and displayed.

Add a mojo method to communicate menu item selection back to the
AppShimHost in the browser process.

Note that this implementation is not complete yet.
- The UI design is that profiles for which the app is not installed
  should not be included in the menu (left as a TODO)
- Selection reaches ExtensionAppShimHandler::OnShimSelectedProfile,
  which just opens a new app instance -- it will need to load profiles
  or focus existing windows (and will need tests)
- The app shim will need a dock context menu as well

Bug: 982024
Change-Id: Ia72bc46b42b4ee3fa2866235c2e681aa690013a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1819853
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#699885}