JNI refactor: @NativeMethods conversion (UrlFormatter.java).
This CL was partially generated by
//base/android/jni_generator/jni_refactorer.py.

Some media unit tests needed to be refactored to use JNI mocks as they
depended on an UnsatisfiedLinkError to pass successfully.

In the process, converted MediaNotificationTestTabHolder to use
lambdas for stubbing to make the test code more readable.

Bug: 929661
Change-Id: I5937168ff0b20cd3e980fe08d877c558bfd70bad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1825583
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Eric Stevenson <estevenson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#699949}