Replace StatusView::Layout() with a LayoutManager.
The status view is the contents view of a widget, but for some reason we were
explicitly setting a |popup_size_| variable that it would base its size on, even
though that should have always stayed in sync with size().  In addition, layout
was mirroring the rect for the enclosed label, but the mirroring had no effect
since the view margins were the same on both ends.

By making the equal margins explicit (via collapsing to one constant) and
eliminating the |popup_size_| field, layout can be done with a simple
FillLayout.

This also reorders the StatusView functions so superclass, declaration, and
definition order all match.

Bug: 1005568
Change-Id: I3d21285d3f7885f61e5d781116562ebf4a822678
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1824350
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#699959}