Add early initialisation of FontConfig on Linux
This CL is proposing to lift the call to FcInit to ensures the
font-config is initialized.

Currently, the FcInit(...) is lazily called on first use. On windows,
the DWrite is early initialized to avoid the lazy initialisation. We
proposed to do the same on Linux.

The behavior should be the same on Windows. On Linux, FcInit should
be called on startup, as we need fonts for any rendering.


The main goal of this change is to help collapsing the bugs / crashes
stack to point to a single culprit. Currently, they are spread over
multiple stackframes in incorrectly reporter.

  Memory leak with FcInit(...) - crbug/1000881, crbug/977249
  Iterator corrupt, initialisation - crbug/1004254

More signatures are available here:
  https://bugs.chromium.org/p/chromium/issues/detail?id=1004254#c11

Change-Id: I1f6d80c790829b1d9416b773d5c9cdba16f90eef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1808447
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#699990}