FlexLayout: allow child views to float leading, trailing, and center
Provides ability to specify an alignment that kicks in after a child
view expands beyond its preferred size. The view can stretch to fill
the available flex space (kStretch), or float to the leading or
trailing edges or center of the container (kStart, kEnd, and kCenter,
respectively).

Change-Id: Iaeee58630ca4eab5a4f6a07554900cad0ad940f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1825717
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700013}