Add support for expanded/collapsed states in menu buttons
In UIA, buttons that set "HasPopup" to anything other than false are
not invokable. Instead the expectation is that buttons will implement
ExpandCollapse or Toggle patterns. Part of implementing those patterns
is ensuring that the correct state of the button (i.e. expanded or
collapsed) is also set appropriately. Otherwise, the default value will
be set to ExpandCollapseState_LeafNode which prevents screen reader
users from invoking the button.

ToolbarButton is a base class and calling
SetHasPopup(ax::mojom::HasPopup::kMenu) at this level seems misleading
(since they don't all have a menu popup), and does not allow us to
accurately set an expand/collapsed state.

This change sets HasPopup to the toolbar buttons that implement menu
model, and adds this property to the other subclasses that do have a
menu but do not implement menu model.

This fix also adds support for expanded/collapsed states in UIA if
HasPopup is set to kMenu.

Bug: 1004025, 1008052
Change-Id: I8a1fe0e8b13e91b469cf16cf5beecc24806bf3d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1803983
Commit-Queue: Maria Villarreal <mavill@microsoft.com>
Reviewed-by: Peter Boström <pbos@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700020}