grit: Removing option for compress = true
The plan was to migrate everything over to this, but it didn't happen.
Since there are no users of this, it's better to not support this anymore.
See https://chromium-review.googlesource.com/c/chromium/src/+/1803394/2
for an example of confusion with using compress = true.

Change-Id: Ib6d666c2a937c3e024ca5fde6b59eddbd560a13c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1825919
Auto-Submit: Sam Maier <smaier@chromium.org>
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700055}