ServiceWorker: Merge installed script evaluation into new script evaluation for cleanup
Before this CL, WebEmbeddedWorkerImpl on the initiator thread determined whether
this service worker is installed, and requested script loading and evaluation
based on the decision.

After this CL, the decision is made by ServiceWorkerGlobalScope on the worker
thread. This enables to remove thread-hop and script evaluation code per service
worker type (new or installed).

Bug: 938219
Change-Id: I55c49cf18998c80075e75df53c5c9983fddcaf78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1826301
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700116}