ServiceWorker: Pass BrowserInterfaceBroker etc as function arguments
This is a cleanup CL, and doesn't change existing behavior.

Before this CL, these values were kept as members of WebEmbeddedWorkerImpl, and
passed when a new service worker thread starts. After this CL, these are
directly passed through WebEmbeddedWorkerImpl as function arguments.

Keeping them as members was needed when asynchronous script fetch was done on
the initiator thread. Now, script fetch is done on the worker thread, and
operations on the initiator thread are synchronous. It's no longer necessary to
keep them as members.

Bug: n/a
Change-Id: Ib401370630e5593702ca05c5dc928baf53e7049d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1826304
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700125}