Push animation policy update into ImageResourceContent
It seem unnecessary that every client should need to handle this
(potentially forgetting to) - especially since what is needed is to call
back into ImageResourceContent which in turn checks with its observers.

Move this call so that it happens just after the ImageNotifyFinished()
notification, and then merge that whole sequence (which happens twice)
into a new method HandleObserverFinished(...). The existing
MarkObserverFinished(...) is merged into the new method since it is its
only caller.

Change-Id: I00fe6c4b31ee7499e7b89c4b2881beaeab95783d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1824893
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Hiroshige Hayashizaki <hiroshige@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#700142}