Do not recreate SkTypeface in fallback shaping stages
SetFontRenderParams() took only the family name of the returned fallback
font, and performed rematching based on the run's styling. This
rematching leads to an inaccurate recreation of the intended fallback
font. The intention with fallback fonts returned from Skia is to use the
exact SkTypeface that was returned by the fallback call, not to attempt
to extract family, weight, style parameters and recreate.

This CL split SetFontRenderParams() on TextRun::FontParams into two
cases: One that allows rematching, which is useful for styling when
doing the shaping runs with the primary fonts (as the Omnibox uses it
for styling substring matches, for example). Secondly, for fallback, the
new method SetRenderParamsOverrideSkiaFaceFromFont() is used which tries
to extract the SkTypeface from PlatformFontSkia if that is the
underlying PlatformFont implementation. Otherwise it performs a rematch
using CreateSkiaTypeface() as in SetFontRenderParams(). This gives us a
viable path forward towards moving to PlatformFontSkia exclusively, in
which case we can remove the last resort rematching step, as we will
always have an underlying SkTypeface.

This CL is also preparation for fixing issue 1003829 in which a specific
weight of Segoe UI is returned by the Skia fallback methods but
inaccurately recreated in SetFontRenderParams() as the weight
information is lost.

Bug: 1003829
Change-Id: I8d0ed08754b97ec0254325c4b09cd74aec9e07ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1824469
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700291}