Tab Strip WebUI: Add a custom element to display alert indicators
This CL copies over a SVG version of the tab alert indicator icons from
/src/chrome/app/vector_icons/ and creates a custom element to display
them. The icon dimensions have been customized for the WebUI version of
the tab stirp and may change. The animations have been made to act
as similar as possible with the existing implementations.

Although the back-end has not yet been provided, this CL assumes there
will be some mapping of TabAlertState to string, a string that will be
used to update the |alert-state| property of the custom
element.

Note: this CL does not add theme colors for all icons (for example, the
recording alert should display in red).

https://i.imgur.com/dq2SARY.png

Bug: 1004983
Change-Id: Ie0c20ae165bae935de1b21241815f7221351764a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1825793
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700357}