Log when we advance from one authentication method to another
This will be helpful e.g. when trying to match logs with behavior as
reported e.g. in crbug.com/1002617 (where logs don't match with the
customer story but we can't be 100% sure).

We log usin LOG(WARNING) as we expect the first auth to work, so this
is an exceptional condition.

Bug: chromium:1002617
Test: N/A
Change-Id: I4366445f9691e527218b907c5ce1debd19f37aa5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1827937
Commit-Queue: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Auto-Submit: Yves Arrouye <drcrash@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700370}