Add a device login policy for select to speak accessibility feature
A device policy which controls the select to speak accessibility feature on the
login screen.The policy can be applied as a recommended or mandatory policy.

Test:
1. Set the login_screen_select_to_speak_enabled policy to True, open System
Tray => "Accessibility", check that "Select-to-Speak" is displayed with a green
check in the login screen.
2. Set the policy to False, open System Tray, check that "Select-to-Speak" is
not displayed in the login screen.

Bug: 552439
Change-Id: I86bfea61ccf3cac74934f17181c260d5649e04c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1816558
Commit-Queue: Amr Aboelkher <amraboelkher@google.com>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700387}