Define guest view custom elements synchronously
Previously, we've had to wait for readystatechange before defining
the guest view custom elements because
1) there were circular dependencies that would cause errors if we
tried to perform the definition immediately, and
2) the Custom Elements V0 registration context was scoped to the
document rather than the window, so if we defined the elements
while the document was still on about:blank, the definition would
be available on about:blank, but not for the extension.

Now that the circular dependencies are fixed and guest view is
migrated to Custom Elements V1 (whose registration context is
scoped to the window), it is no longer necessary to wait for
readystatechange.

We now perform the definition immediately.

Bug: 810012
Change-Id: Ib5e8ba6f4e302eaae56c5566765521d64b1a0e15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1278287
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Reviewed-by: James MacLean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700425}