Fix the scrollable shelf's issue on relatively small screen
Now scrollable shelf ensures that when two arrow buttons show, the
right gradient zone always cuts the half of the rightmost icon. To
implement this feature, some available space is wasted. As a result,
there may be no sufficient space for scrollable shelf on
small-screen devices. In this CL, the width of the gradient zone
becomes a constant.

Bug: 1005989
Change-Id: Iaeb3c2c30c85746bfb21bbd82d3bb7f7a53783b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1815723
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700506}