LocalHistoryZeroSuggestProvider
Provides on-focus zero-prefix query suggestions from local history iff:
 1) input is empty.
 2) input is focused from the NTP (i.e., Address bar or NTP realbox).
 3) user is signed out and has set up Google as their default search engine.

Matches are created by performing a prefix search in the URLDatabase for
the Google search URL and extracting the search terms from the matching
URLs.

Once a match is deleted, all the URL records that could produce that
suggestion get deleted which prevents the suggestion from reappearing.
This is done by querying the HistoryService for a Google search URL
template that has its search terms replaced with those of the suggestion's.

Bug: 996516
Change-Id: I2cf767f0667a7b01652f49ccda627ae62cd97cc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1804374
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700511}