Add a device login policy for caret highlight accessibility feature
A device policy which controls the caret highlight accessibility feature on the
login screen.The policy can be applied as a recommended or mandatory policy.

Test:
1. Set the login_screen_caret_highlight_enabled policy to True, open System
Tray => "Accessibility", check that "Highlight text caret" is displayed with a green
check in the login screen.
2. Set the policy to False, open System Tray, check that "Highlight text caret" is
not displayed in the login screen.

Bug: 552439
Change-Id: Ibe294203977de06d8c5238e8db3d00a6bfa5fbf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1816506
Commit-Queue: Amr Aboelkher <amraboelkher@google.com>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700637}