Add constructor prototype.
This change adds a basic version of the library constructor that
decompresses all of the data at the start of the library.

This change adds patching of constructor magic bytes to
compress_section.py

This change updates tests from calculating the number of zeroes back
to the sum, since constructor keeps the zero-filled region under
control.

Bug: 998082
Change-Id: If34b66171965084f918b1393c21e7b7ef08e4bc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1808864
Commit-Queue: Denis Solonkov <solonkovda@google.com>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700639}