Add additional checks in tab_strip_model.
Checks for cases where methods might be called from the wrong thread,
triggering a race condition, and documents that tab_strip_model is not
thread-safe.

This does not close the associated bug; it simply provides us with
better diagnostic tools to figure out why it's happening and fix it.

Bug: 1006708
Change-Id: Ia722327b921e8183aebd750ba164ab92112c63b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1825918
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700789}