views: remove "splitter" role from Separator
At least on Mac, a "splitter" is an interactable thing that allows
the user to adjust the boundary between two groups of things. The
Views splitter is not like this - it is basically just an <hr> type
line. The Views control is more or less purely presentational and
should not have an AX role.

Bug: 961324
Change-Id: I8cfcd657c2ee9cae462a208e616dd8364cdb6428
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1827944
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700833}