AW docs: add "more docs" to navbar
This adds a "More docs" link to the end of the navbar, which links to
the /android_webview/docs/ directory listing. Not all docs are in the
navbar (not everything is relevant to everyone), but there should still
be a way to navigate there from the navbar interface.

This adds a comment to tell future contributors to leave this entry
last.

Bug: 918221
Test: Upload to gerrit > open file > click "gitiles"
Change-Id: Ibfd27d9d3724d31543532da0e5507dc97a6ba11f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1829241
Auto-Submit: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Chris Mumford <cmumford@google.com>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700834}