Refactor filter interpolation functions into an InterpolableFilter
This CL is part of the shift away from NonInterpolableValues to instead
having more InterpolableValue subclasses. It introduces a new subclass,
InterpolableFilter, which can represent a blink::FilterOperation. This
should enable us to more easily implement the correct additive behavior
for filter, see [0].

There should be no behavioral change from this CL, it is (intended to
be) a pure refactoring.

Bug: 1005828

[0]: https://docs.google.com/document/d/1aAvTNwxSSl0OWP3hQAV-Q7ru5-o5MzV4UEz7Uw9VHZk/edit#heading=h.9b2xtih585di

Change-Id: Ib60b5857167db8950f3eb457cfaeec98ee781ff1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1817284
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700843}