split view: Nix special apostrophe in comment in split_view_controller.h
While editing split_view_controller.h in emacs, I am always constantly
inconvenienced by error messages for reasons related to "it’s" in a
comment. The apostrophe is actually an unusual character. Compare:

' is on your keyboard.
’ is in the comment in split_view_controller.h.

There should not even be an apostrophe there, because the grammar of the
sentence requires the possessive "its" rather than the contraction.

Change-Id: I2b9e0397ad5019940769d224959a05bb073837a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1829651
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700858}