Merge JsonConfigConverter memory usage fix to M78 branch.
This cherry-pick contains fixes for a merge conflict with
crrev.com/c/1792395, which was landed on master after branch but
before the CL being cherry-picked.

The conflicting CL changed the JsonParser to use the new mojo types,
so I've converted them back to the old mojo types to fix the conflict.
I've tested that this compiles, passes unit tests, and has the correct
behaviour on-device.

Original description:

[Dolphin] Fix memory usage issue in JSON config converter.

We have a utility for converting JSON configurations for a
RecurrenceRanker into protos. To comply with the rule of two, this is
done in a separate process.

The implementation of this utility keeps a pipe open to that process
for its entire lifetime, which in practice is always the lifetime of
a user session. Keeping this process alive costs about 20MB of memory.

This CL refactors the JsonConfigConverter to have a more reasonable
lifetime story and prevent permanently allocating so much memory. The
two key changes here are:

 1. Explicitly close the parser Remote once parsing is complete.

 2. Allow use of the converter only through a factory method
    returning a unique ptr, encouraging the destruction of the
    converter itself once conversion is complete, and preventing
    an illegal double-use of one converter object.

(cherry picked from commit 83db999de8815a739b49e316b673912a9220f094)

Bug: 1003674
Change-Id: I5ca4cb520fb838cbda86cd41e62784a47aab8be1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1804946
Reviewed-by: Jia Meng <jiameng@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#698342}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1825102
Reviewed-by: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/branch-heads/3904@{#502}
Cr-Branched-From: 675968a8c657a3bd9c1c2c20c5d2935577bbc5e6-refs/heads/master@{#693954}