Local NTP: just send credentials with promo service requests
This was previously the case, at least for server-side opt-in to work.
Somewhere along the line this was changed (multiple times) and now
server-side opt-in to show "middle-slot" promos doesn't work any more.
This is a pain for developers but a bigger pain for less technical folks
to show demo promos.

Also, OGB and doodle requests already send creds as well.

R=mahmadi@chromium.org
BUG=1003508

Change-Id: I75869a818f7fcaeba08fa3babbfa30426fb06cce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1830365
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Dan Beam <dbeam@chromium.org>
Auto-Submit: Dan Beam <dbeam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#700936}