Fix deletion of error panels
The 'onclick' handler for panel items was being set up when the panel
was attached to the DOM. Side effect from that is creating a panel
type with buttons used 'null' for the click handler, causing the
buttons to fail to do anything after the panel is attached to
the DOM.

Set the 'onclick' handler inside the constructor to ensure it's
available for reference when setting the panel item type.

Needs test.

Bug: 1009318
Change-Id: I654626b59715d03e3ee14f6ac00eadb5bd9815ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1831618
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#701051}