[Autofill Assistant] Add UMA histogram for DFM installation.
This CL is adding a new UMA histogram for the Autofill Assistant DFM
installation. The metric contains the following values: installing of
the DFM is deferred, the installation is happening in the foreground and
it succeeded, the installation is foreground and failed, the DFM was
already installed. The metric cannot currently distinguish whether the
deferred installation succeeded or failed.

Bug: b/140369065
Change-Id: Ie250567996efeea809bd1be0c5fe5ef037b75be2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1825358
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Commit-Queue: Marian Fechete <marianfe@google.com>
Cr-Commit-Position: refs/heads/master@{#701063}