Add a device login policy for autoclick accessibility feature
A device policy which controls the autoclick accessibility feature on the
login screen.The policy can be applied as a recommended or mandatory policy.

Test:
1. Set the login_screen_autoclick_enabled policy to True, open System
Tray => "Accessibility", check that "Automatic clicks" is displayed with a green
check in the login screen.
2. Set the policy to False, open System Tray, check that "Automatic clicks" is
not displayed in the login screen.

Bug: 552439
Change-Id: I4c58b1e63ed32e2d861a062493a77d1dd09a7d5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1819439
Commit-Queue: Amr Aboelkher <amraboelkher@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#701115}