[Sync:USS] Factor out NigoriState
We factor out subset of field in NigoriSyncBridgeImpl to a separate
struct NigoriState. This struct directly maps to sync_pb::NigoriModel
proto and the key reason to do that is to move proto serialization/
deserialization away from the bridge.

This CL is pure refactoring.

Bug: 922900
Change-Id: I19b0425aeccce475f34dab6375f9929994f17ea9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1813311
Commit-Queue: Maksim Moskvitin <mmoskvitin@google.com>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#701161}