The ID used for an invalid AXNode ID should be 0 because -1 is used for the extra Mac nodes
This is a speculative fix for the busy problem on Mac
whereby some pages hit a DCHECK in AXPosition.
On Mac there are some extra AXNodes that are only created for table columns. These have negative
node IDs, and so using -1 as the invalid node ID might create a conflict with those nodes.
Table columns are not created on any other platform.

R=dmazzoni@chromium.org

Bug: 1006842
Change-Id: I38da62f41c570cebfde197895fb6bebb16649cc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1827999
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#701174}