Add seq DCHECK in ExternalPolicyDataFetcher destr
Add the DCHECK into the ExternalPolicyDataFetcher's destructor that it's
executed on the right sequence.

Before this CL, the destructor contained the comment that stated that
such DCHECK isn't technically possible due to unusual destruction order
in some tests, but we can't repro this problem with the current ToT
anymore.

Bug: 727645
Change-Id: I6490227656a876d546cd5b659f5d88b2d2ec0dfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1828820
Reviewed-by: Nikita Podguzov <nikitapodguzov@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#701264}