[WebGPU] Fixing ScriptValue::CreateNull crash in OnPopErrorScopeCallback
Recent changes in Blink caused this to start crashing. Directly using
v8::Null instead of ScriptValue::CreateNull fixes the issue. It looks like
is related to crbug.com/1008765 where it's no longer value to create a Null
ScriptValue when the value is no longer "in context".

This CL applies a small fix for the crash so we can roll the WebGPU CTS and
prevent further regressions.

Bug: 1009635
Change-Id: I4d4de6b499b05bb1833ec34128239ca135692378
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1830283
Commit-Queue: Austin Eng <enga@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Cr-Commit-Position: refs/heads/master@{#701278}