Add a device login policy for keyboard focus highlight accessibility feature
A device policy which controls the keyboard focus highlight accessibility
feature on the login screen.The policy can be applied as a recommended
or mandatory policy.

Test:
1. Set the login_screen_keyboard_focus_highlight_enabled policy to True, open System
Tray => "Accessibility", check that "Highlight object with keyboard focus" is
displayed with a green check in the login screen.
2. Set the policy to False, open System Tray, check that "Highlight object
with keyboard focus" is not displayed in the login screen.

Bug: 552439
Change-Id: Id640c6ba9aab7913c17e3ea43bc9c74d945c4a5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1826791
Commit-Queue: Amr Aboelkher <amraboelkher@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#701303}