Ash Tray: Add new overflow tray button for the in-app tablet shelf.
This button will be shown when the number of tray items exceeds what can fit in
the shelf normally. Click on it will show/hide the trays.

This CL just introduces the button behind a flag and it is not integrated with
the rest of the status area. That will come in later CLs.

Screenshot:
https://screenshot.googleplex.com/no5btxNAAOw

TEST=manual + stub unit tests
BUG=982511

Change-Id: I813b3f24e1c7172ca6ce0d058cd1376bf22a6bf9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1825921
Commit-Queue: Tim Song <tengs@chromium.org>
Reviewed-by: Tetsui Ohkubo <tetsui@chromium.org>
Reviewed-by: Ahmed Mehfooz <amehfooz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#701384}