Support navigation in bundled exchanges
This CL introduces BundledExchangesHandleTracker class to track navigations
within a bundled exchanges file.
A BundledExchangesHandleTracker is created by BundledExchangesHandle while
loading a page in the bundled exchanges file and passed to the next navigation
from the page. And MaybeCreateBundledExchangesHandle() creates a new
BundledExchangesHandle to load the page if the next page is in the bundled
exchanges file.

TBR=toyoshim@chromium.org

Bug: 995177
Change-Id: Ib4faaa70390d432f74424fd278f53d86d20f28d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1800811
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Kunihiko Sakamoto <ksakamoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#701496}