PDF Accessibility: Fix invoke on link
In CL:1730426 support for click action handling was added. In the struct
PP_PdfAccessibilityActionData two more members were added but the
members were not added in IPC_STRUCT in ppapi_messages.h file. Due to
this, the values were not being sent from the mimehandler process to the
plugin process. The default values for the members were being utilized
by the plugin process instead of the values in the mimehandler process.
Invoke on any link would always navigate to the first link of the page.
This CL adds the members in the IPC_STRUCT.

A test has been modified to invoke the second link instead of the first.

Bug: 981448, 1006729
Change-Id: I7b40753887916cd08a338c9d4c3c63f3dd432480
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1826743
Commit-Queue: Ankit Kumar 🌪️ <ankk@microsoft.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Ian Prest <iapres@microsoft.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#701529}