[iOS] Fixing FirstRunChromeSigninViewController leak
FirstRunChromeSigninViewController is being leaked when of all those
conditions are true:
 + it is the delegate of its own presentation controller.
 + it is shown inside a navigation view controller in second and last
position.
 + the navigation view controller is dismissed with the view controller
displayed.

Then a retain cycle is created with FirstRunChromeSigninViewController
and UIFormSheetPresentationController.
This is a bug in iOS, and should not happen.
Since FirstRunChromeSigninViewController cannot be swiped to be
dismissed, it doesn't need to be the delegate of its presentation
controller.

The fix is to not set the presentation controller delegate if the view
controller is not part of a navigation controller.
The fix is done in ChromeSigninViewController since
FirstRunChromeSigninViewController is a subclass of if.

Apple Feedback: https://feedbackassistant.apple.com/form-response/15480140

(cherry picked from commit 7d9b28d5acb289d889f9520b7109c7ed69536c5e)

Bug: 1004695
Change-Id: I85f7721c4bef1c95d6d30abfd06e4ce83954209f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1823198
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#699786}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1833614
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/branch-heads/3904@{#551}
Cr-Branched-From: 675968a8c657a3bd9c1c2c20c5d2935577bbc5e6-refs/heads/master@{#693954}