Remove DumpWithoutCrashing in ValidateAccessToCookiesAt.
This is intented to be merged into M78.

Bug 1003167 is intented to be fixed by:
https://chromium-review.googlesource.com/c/chromium/src/+/1829350/ but
we are not confident enough for merging it into M78 beta.

A lot of users are triggering this DumpWithoutCrashing(). To save their
resources, this patch remove it temporarily. It will be added back with
in M79 after a fix is landed.

Bug: 1003167
Change-Id: I17abcaebc3a6177c45a193ad47c4c731b340d12a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1832816
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#701602}