Add RenderFontParams to the PlatformFont constructor
This CL is adding the RenderFontParams explicit to the
font constructor.

On Linux, the call to GetFontRenderParams(..) is making
a call to the FcFontMatch to retrieve the rendering
parameters. It is quite common on Linux to iterate
through fonts returned by FcFontSort(...) and to create
gfx::Font with them. There is no need to make a second
call to GetFontRenderParams(..) [and to FcFontMatch] since
the rendering parameters are known
(see etFontRenderParamsFromFcPattern(...)).

On windows, the call to GetFontRenderParams(..) returns
the same parameters, no matter which font family is
provided.

R=robliao@chromium.org

Bug: 1008531
Change-Id: Ia368e3c797cf43b3bd726f871e01389b41092c61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1830061
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#701641}