X11 and Ozone: Move FlashFrame impl to DWTHPlatform.
FlashFrame method is meant for flashing a native window
to draw attention to it.

However, not all the platforms support that. Thus, there is
a default no-op impl of that method in the PlatformWindowBase.

As I know, currently X11 and Windows backends support that.

No functional changes are done in this patch. Just moving the code
to the "right" place.

Bug: 990756
Change-Id: Ib463ba75b0cdf5e49e705f0425241e242759adcc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1831897
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#701681}