Delete the main frame RenderWidget during detach if it's not undead
When RenderViewImpl::Destroy() happens, if there is no local main frame
it should Close the WebView and destroy the undead RenderWidget.

But if there is a local main frame, then Closing the WebView will
detach that main frame, and in the process we should destroy the
RenderWidget for the main frame in the same way that we would for a
child frame.

This makes RenderViewImpl::DetachWebFrameWidget destroy the
RenderWidget if the call happens as part of WebView closing, from
RenderViewImpl::Destroy().

Otherwise, it continues to only close the WebFrameWidget and remove it
from the undead RenderWidget.

R=ajwong@chromium.org

Bug: 419087
Change-Id: I642ccf0fe95547d5614af6b29c80a6176adaa231
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1825733
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#701727}