Extensions should appear force-disabled after adding supervision
Before, supervised users could still toggle extensions to "enabled" in
the UI. Doing this doesn't really enable the extension and has no
effect, which could be confusing for the user. Something that is WAI
could be perceived as a bug. This CL changes SupervisedUserService::
UserMayModifySettings() to force disable extension toggles for
supervised users.

Bug: 1003014
Change-Id: Ib00a6e3616d08c4c7791ef47854ebc1ae992ed5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1822797
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Reviewed-by: Michael Giuffrida <michaelpg@chromium.org>
Commit-Queue: Toby Huang <tobyhuang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#701772}