Plumb LocalFrame through ChromeClient::WindowToViewportScalar (part 1).
This CL introduces a new version of WindowToViewportScalar that takes a
LocalFrame parameter, marks the existing version as deprecated, and adds
one caller to the new version. The new caller is DataTimeChooserImpl,
and that change is what fixes bug 1006919.

What was happening before is that in the context of an out-of-process
subframe, WindowToViewportScalar was not finding a LocalFrame for the
top frame. It would then bail out early, not calling the appropriate
WebWidgetClient method. Hence, the error reported in the bug.

A follow-up CL will migrate the rest of the callers over to the version
of this method that takes a LocalFrame parameter and remove the other.

[Merged into M78 release branch.]

(cherry picked from commit b04bed18e551a19ac226a0ea1ded17a9357d2a3e)

Bug: 1006919
TBR: dcheng@chromium.org
Change-Id: I04bd2fa80abed139b90a42b81668bcc3aba8f967
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1831447
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Darin Fisher <darin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#701558}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1834269
Reviewed-by: Darin Fisher <darin@chromium.org>
Cr-Commit-Position: refs/branch-heads/3904@{#570}
Cr-Branched-From: 675968a8c657a3bd9c1c2c20c5d2935577bbc5e6-refs/heads/master@{#693954}