Downgrade security level on sites using legacy TLS versions
This adds a check to SecurityStateTabHelper to mark whether a page was
loaded over legacy TLS versions in the VisibleSecurityState struct. This
status is then used by the security_state code to downgrade the security
level to WARNING. The SecurityStateTabHelper code also checks the pref
set by the SSLVersionMin enterprise policy to determine whether to set
the flag in the VisibleSecurityState.

This version can be racy at browser startup, due to how the component
proto is loaded from disk (as BEST_EFFORT). A followup CL will change
that task to USER_VISIBLE to avoid this.

Another followup CL will change the Page Info security summary for pages
with the legacy TLS flag set in their VisibleSecurityState.

Bug: 989607
Change-Id: I510eafc0124f433ec4e4ce81e4ece4f1321bde6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1829462
Commit-Queue: Christopher Thompson <cthomp@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#701803}