Provide chrome://sandbox for Windows
Creates a webui to display a snapshot of sandboxed processes for
Windows. This CL includes only the basic plumbing and very basic
display of relevant information.

Process information must be obtained from the sandbox (broker) and
from the browser, as each knows different things about what is
running. For now we simply display this as json but in later CLs the
range of information available will be increased, and the lists of
processes from different sources will be joined together to form a
friendlier display of diagnostic information.

This will result in a larger |sandbox_internals_win.js| file which we
do not need on other platforms. Hence the file for Windows is split
from that for Linux or ChromeOS. We will display different information
to that provided by Linux or ChromeOS as the underlying sandbox
technology specific to the host OS so much of the data extraction code
is #ifdefed for OS_WIN or in _win.cc,h files.

Bug: 997273
TEST=navigate to chrome://sandbox on Windows, see a page

Change-Id: I91a4cb12a6a5b632d0da89f5ffccfcb30b177b62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1767660
Commit-Queue: Alex Gough <ajgo@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Michael Giuffrida <michaelpg@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#701897}