Add a user policy to configure the position of the shelf
Add a user policy to configure the position of the shelf within a session (Bottom, Left, Right).
The logic needed to change the shelf position already exists - right click on the tray and change its position -
The new policy just needs to be wired to the prefs which control this behavior.

Bug: 1003717
Change-Id: I574f69cf48f8a35d2ab76a9b128b781b1d47a7d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1829099
Reviewed-by: Manu Cornet <manucornet@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Alexander Hendrich <hendrich@chromium.org>
Commit-Queue: Aya Elsayed <ayaelattar@google.com>
Cr-Commit-Position: refs/heads/master@{#702085}