Fix preview thumbnail aspect ratio issues.
This CL fixes an error where the preview captured from a loading
webpage (used in hover cards and tablet tabstrip) could appear blurry
or stretched when the aspect ratio of the window did not match the
thumbnail.

It solves the problem by providing a max capture size based on the
ratio between the desired and current window size, then clipping the
image down later, just as we do with single-frame capture of active
windows on tab switch.

Still to do:
 - Trim scroll bars out of the final captured image.

Change-Id: I8cbeee5c9c5959163d2e7f0b6251802f40455f98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1818641
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Collin Baker <collinbaker@chromium.org>
Cr-Commit-Position: refs/heads/master@{#702127}