Don't stop asking for YCbCr data until it arrives.
Previously, VideoFrameFactoryImpl might be given a nullopt YCbCrInfo
in response to a request for it.  It would then clear the callback
and crash the next time it tried to return a frame; it didn't
check that the callback wasn't clear.

Now, it only clears the callback when it actually gets YCbCrInfo,
which is also when it quits asking for info on subsequent frames.

Also adds a test for it.

Bug: 1007196
Change-Id: Idfffe95b319e09285fdd4b6c8855d3e781fc606d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1829771
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#702196}