Apply UserGesture token to intents from incognito.
This doesn't fix any known (to me) bugs (and please revert if it breaks
anything).

However, it does two things:
1. It makes the code more consistent.
2. It fixes what might be a bad incognito detection method? The dialog
probably makes this too obnoxious to use in practice, but it might give
a site a way to know if the user is in incognito.

The user gesture code was added in
https://codereview.chromium.org/1243253004/, which doesn't mention
incognito at all, so I'm thinking it was probably just an oversight
that the gesture token doesn't get applied while in incognito?

Bug: 1006927
Change-Id: I74458549cea9f4aec65c6c41c0a73b4869d44115
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1832699
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#702309}