Changes ChromeOS external protocol dialog message for click to call.
When click to call feature is enabled on chrome os but no apps/devices
are available for the user to choose, the external dialog is fired.
We can use the space for educating the user on how to use the feature.
Currently we are only changing title of dialog as a hotfix.

Screenshot after change: https://drive.google.com/file/d/1WXrjcmvYhyjPXRx6p_Ucnp0RYO8fxuSP/view?usp=sharing

Bug: 1007995
Change-Id: Id97e84705f04f02d1e1716d628e560eea3e42473
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1829209
Commit-Queue: Himanshu Jaju <himanshujaju@chromium.org>
Reviewed-by: Alex Chau <alexchau@chromium.org>
Reviewed-by: Satoru Takabayashi <satorux@chromium.org>
Cr-Commit-Position: refs/heads/master@{#702368}