Check ComputedStyle of flat tree parent for skipping dirty marking.
We have an optimization where we skip marking an element style dirty if
both its own and the parent's ComputedStyle are null. The optimal check
is to look at the flat tree parent, and not go directly to the shadow
host from a shadow host's light tree child, since there may be a
display:none in the flat tree between the shadow host and the light
tree child.

For FlatTreeStyleRecalcEnabled, we allow walking the flat tree even if
it's dirty and we can use GetStyleRecalcParent() to check the flat tree
parent instead.

Bug: 972752
Change-Id: I955259bcbbbb4ab3dd91eca5f1dc3d631e26a921
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1836507
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#702417}