Fix H2+WebSockets when using an HTTPS proxy.
When using an HTTPS proxy, WebSockets would always try to use an
established H2 session for WebSockets, even if H2 over WebSockets was
disabled. Moreover, if H2 over WebSockets was disabled, it would even
use an H2 session that hadn't advertised WebSocket support.

These problems were a result of only checking the HTTPS proxy server
type, and not the scheme, when deciding if a request should try and
use H2, and using the wrong value when telling the H2 code that only
a session supporting WebSockets was needed.

This CL fixes these issues.

Bug: 1009612
Change-Id: Ic8ffdac454f3e079e47b8c35b1a5b4d8cac807e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1835253
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Reviewed-by: Bence Béky <bnc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#702441}