Remove use of deprecated function v8::V8::SetNativesDataBlob
With the recent port of the streams API to C++ [0], the natives blob is
now unused and will shortly be removed. V8 API related to the natives
blob is being deprecated [1] and will be removed in the next release.

This CL is the minimal change to keep up with these API deprecations.

In follow-up work, all other infrastructure related to the natives blob
should be removed.

[0] https://crrev.com/c/1746118
[1] https://crrev.com/c/1824944

Bug: v8:7624
Change-Id: I26e409660d177035cdb5d92c317fc2e0174be2f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1834352
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Ross McIlroy <rmcilroy@chromium.org>
Auto-Submit: Jakob Gruber <jgruber@chromium.org>
Commit-Queue: Jakob Gruber <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/master@{#702475}