Make ShouldOverrideUrlLoadingInternal more readable (9 of N)
This function is a large, and hard to follow. I'm hoping to break out a
bunch of functions that are clearer about what they're doing and make
ordering clearer.

In this change I do fix an exception handler to actually catch the
exception it was intending to catch (and you can easily repro the crash
by disabling the targeted Activity while the leaving Chrome alert dialog
is up).

I also make sure that when we fail to show the alert dialog for reasons
other than a BadTokenException, we also propagate this back so we don't
get into a state where we report an async action action is taking place
when it isn't.

Bug: 1006927
Change-Id: Ie09214f64ead85c31fb8e83e187878d984f2afc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1835144
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#702483}