Update git_cl to use buildbucket to fetch try job results.
Landing this will affect wpt importer on the next run, so if something breaks then
this CL should be reverted.

This deletes the old methods that used the `git cl try-results` command and
replaces them with the buildbucket rpc variants.

Also updated a lot of unit tests to specify try job results via MockWeb instead
of lambdas. This uncovered a couple minor bugs related to missing fields in
the response.

Bug: 1005801
Change-Id: Ia727635aeb1f2569707546454c52a0d1813683ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1837488
Commit-Queue: Luke Z <lpz@chromium.org>
Reviewed-by: Robert Ma <robertma@chromium.org>
Cr-Commit-Position: refs/heads/master@{#702603}