Add metric to detect support for OpenXPS
Windows printing path is being modernized from using the pre-WinXP
GDI print API to an XPS-based API.  This was first introduced
Windows 7 and was deprecated in favor of OpenXPS in Windows 8 and
beyond.  Some Windows 7 installations may be capable of OpenXPS if
they have been sufficiently patched [1].

Determining the number of Windows 7 machines which would still require
XPS instead of OpenXPS can provide guidance if it is worth the effort
to implement support for both.

[1] https://docs.microsoft.com/en-us/windows/win32/printdocs/app-support-for-openxps-printing#sending-xps-data-to-the-print-system

Bug: 1008222
Change-Id: Id528478fe4f90002aba09208cf34d74cfe787472
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1826278
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#702677}