[usb] Fix device info refresh on Android
On Android, after acquiring permission to access a device, the USB
descriptors are read directly from the device. This allows us to fill in
fields that are unavailable through the Android SDK. In r678566 the
UsbDeviceDescriptor object was updated to store properties in a
UsbDeviceInfo instance. In the process a bug was introduced into this
Android-specific logic which caused the device GUID to be reset in the
process of adding these fields.

This change updates the field updating logic to only copy the specific
fields that we expect to be updated by reading the USB descriptors. This
should be more future-proof.

(cherry picked from commit ed6b5331c28dcb37568e2f5b42883ef59006fb0d)

Bug: 1009362
Change-Id: Iafc9bad6855d8a6da5650d6179984cc40018d906
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1834033
Reviewed-by: Matt Reynolds <mattreynolds@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#701784}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1839637
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/branch-heads/3904@{#614}
Cr-Branched-From: 675968a8c657a3bd9c1c2c20c5d2935577bbc5e6-refs/heads/master@{#693954}