[ios] Disable animations on low battery level
Added disabling of animation when battery level is
below 20%. This is behind a flag that is false by default.
It will be enabled to 1% in next release to gather
information. Some animations not controlled by
[UIView setAnimationsEnabled:] are not handled for now.
Also some components imported from google3 do not
respect or restore the current animationsEnabled status:

MaterialOverlay, MaterialSnackbar: force animations enabled
but restore previous value.

NISnapshotRotation: is worse because it will leave it back to enabled
after doing an animation (not sure where we use it?)

Bug: 1006431
Change-Id: Ib2f32598575c0e63b579a2388dad914945714add
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1821917
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: David Jean <djean@chromium.org>
Cr-Commit-Position: refs/heads/master@{#702819}