Update AdPageLoad metrics to include all ad frames with non-zero CPU
Given that CPU usage is now reported to AdsPLMO, this updates the
criteria for when an ad frame is recorded. Specifically, it is recorded
iff it has either non-zero bytes or non-zero CPU usage. This change
also fixes an issue where metrics would still be collected for
resources blocked by the Subresource Filter.

PageLoad.FrameCounts.AnyParentFrame.AdFrames is versioned to
PageLoad.FrameCounts.AdFrames.Total. The other affected metrics have
not been versioned as the increase in the number of ad frames recorded
should be small (~5%).

Bug: 1000810
Change-Id: I7ec57326ca93f6e6686d84869bb7f4dfbf58bf15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1824124
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#702878}