desktop-pwas: Set display icon for Open With menu
This change enables PWAs registered as file handlers to appear in the
"Open With..." menu with a custom icon on Windows 8 and later.

On Windows 8 and later versions, when an application is registered as a
handler for a file type, the icon that is displayed in the "Open
With..." menu next to the application's name is controlled by registry
value HKCR/<application ProgID>/Application::ApplicationIcon. This value
can be a reference to an executable or .ICO file, and has the syntax
".../my_icon.ico,0".

This change extends ShellUtil::AddFileAssociations(), which will soon be
used to register Progressive Web Apps (PWAs) as file handlers on
Windows, to set ApplicationIcon.

Bug: 960245
Change-Id: If9ef55d21d71b00b47ce3246b333efa38c2965cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1832464
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Cr-Commit-Position: refs/heads/master@{#702949}