[ios] Fix flicker on first run on iOS 13
Prior to iOS 13, adding a subview to the key UIWindow would
place that view above the root view controller in that window.
In iOS 13, the UIWindow belongs to a UIWindowScene, and any view
added to that window is placed behind the root view controller.

Therefore, the hack to prevent the BVC from momentarily appearing
before the first run experience fails on iOS 13. To fix, this
CL adds the fake launch screen directly to the BVC container,
rather than to the window.

While this fixes the bug (and works correctly on iOS 12/13
and iPad/iPhone), it violates Law of Demeter or Least Knowledge
Principle. The violation is documented in the code and captured
in a TODO bug. But this violation is really just an extension of
the hack already in place.

Bug: 993742
Change-Id: I39eae9030ed77995d3711e5b6c3495490353446a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1838261
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Auto-Submit: edchin <edchin@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703053}