Block SMILTimeContainer updates when we're updating intervals
This adds a |is_updating_intervals_| flag to SMILTimeContainer, and sets
it in UpdateIntervals() and when calling Reset() on the sandwiches.
When this flag is true SMILTimeContainer::ScheduleIntervalUpdate() will
not schedule an update since UpdateIntervals() (in both cases) will
handle the update.
This should allow getting rid of the SVGImage::ScheduleTimelineRewind()
hack.

Bug: 998526
Change-Id: Iad31f2684db40dc71f254d7d19f57a4538272a5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1841332
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703245}