Using higher priority for actively exploited bugs.
This change notes that, for bugs being actively exploited in the wild,
we should use a higher priority such that they are fixed and merged as
rapidly as possible.

The change deliberately does not specify the priority to use in this
case, as this should be an extremely rare event and be subject to
extensive discussion. However, it does go so far as to say we 'should'
use a higher priority, which for a medium or high bug typically means it
would end up being Pri-0. That seems like it would typically, but not
always, be the right thing to do, so it feels like it gives the right
guidance.

Change-Id: Ied0ad4c9ab61425b748f4e8c78e7ea0a46bb060e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1841509
Commit-Queue: Adrian Taylor <adetaylor@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703315}