Set emulatedPosition for WMR controllers.
When a controller's PositionAccuracy is "approximate", emulatedPosition
should be true.

Also update a test page to only display the target ray/laser for a
controller when emulatedPosition = false (the controller mesh itself
remains visible). That gives an easy way to know the value of
emulatedPosition while still in VR.

Bug: 1011103
Change-Id: Ie90acc51afd914d1e082be7ddd12643257e9d9cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1842477
Commit-Queue: Jacob DeWitt <jacde@chromium.org>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703329}