Fix crash for Click to Call on pre M devices.
Android versions before M require a permission to read the call state
which we use to get metrics for Click to Call. As we do not want to add
permissions to the manifest, we won't try to log this metric on pre M
devices.

Bug: 1011767
Change-Id: I244295f57443c0831c488695b7ed2ea537e7c77c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1845007
Reviewed-by: Michael van Ouwerkerk <mvanouwerkerk@chromium.org>
Commit-Queue: Richard Knoll <knollr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703356}