[HaTS] Don't show survey when resource fails to load
There are cases that HaTS javascript library may not be loaded
successfully, for example, when the browser is offline. In this
case, we monitor the resource load complete call to make sure HaTS
library is loaded. Otherwise, we don't proceed with showing the
survey.

BUG=1006674

Change-Id: Ibcd725728978ce96fbbe263fb0c5d5c894d1ced9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1826291
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Wei Li <weili@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703407}