Allow moving accessibility-focused tab via keyboard.
SHIFT+[RIGHT|LEFT] now moves AX-selected tab. If a tab is moved into or
out of a group, its group will be updated appropriately.

Also added a handful of browser tests since the move right/left
functionality needs to be tested in integration with a live
TabStripModel.

Still to do: announce the change to the screen reader.

Bug: 996850
Change-Id: I0fe1b36c0bab0f7740cb0876bc94a04cfa092786
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1793547
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Charlene Yan <cyan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703540}