Implements the ShouldTargetNavigation interface for the PredictionManager.
This change builds the first piece of the PredictionManager to handle
the registration of OptimizationTargets and requests for whether a
navigation should be targeted for a particular optimization.

This change enables the PredictionManager to construct the feature
map needed to execute the PredictionModel to produce a  decision
for an optimization target.

The the next changes will handle loading PredictionModels from a
store and fetching model updates and host model features from
the remote Optimization Guide Service.

Bug: 1001194
Change-Id: Ib0af48cd2c81d0d4dfe92e4a99916b44eee92498
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1833783
Commit-Queue: Michael Crouse <mcrouse@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703596}