RecalcDescendantStyles for inactive insertion points.
For shadow dom V0, <content> elements which are not inside shadow trees
are not active insertion points and do not take part in shadow
distributions. They are part of the flat tree and otherwise behave as
any other elements. Call RecalcDescendantStyles to compute styles for
its descendants.

TEST=accessibility/table-row-role-crash.html
TEST=fast/dom/shadow/content-element-outside-shadow.html
TEST=fast/dom/shadow/shadow-dom-event-dispatching-child-of-inactive-content.html
TEST=fast/dom/shadow/shadow-insertion-point-rendering.html

Bug: 972752
Change-Id: I0037f5c3a1dc6d0f7746aa7e0eb9b118212e2047
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1846431
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703618}