Port GLContextCGL backpresure fence workarounds to EGL.
With the passthrough command decoder on Mac, backpressure fence and flush
workarounds need to be done in the GLContextEGL as well.  Move them to the
base GLContext class so they are usable by both context types.

Fixes flickering issues on UI elements on Mac when using the passthrough
command decoder.

BUG=1011017

Change-Id: Iba226caada24341abc58053e7d033e114d16df3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1838371
Commit-Queue: Geoff Lang <geofflang@chromium.org>
Reviewed-by: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703699}