[TouchToFill][Android] Fix wrong onDismissed calls
By setting the bottom sheet observer whenever we bring up the sheet, the
closing signal will be captured reliably. This wasn't the case before
which caused the bottom sheet to be permanently dismissed.

Note: I removed the NOTIMPLEMENTED() call from onDismiss() since it
crashes but has no real reason to do so — the cleanup happens in the
destructor anyway. It might be late but it's not wrong.

Side effect: for some reason, this seems to fix the intermediate peek
state (see screenshot 1 in the linked bug). This is unexpected and since
it's a timing-related issue probably not a reliable fix. It will be
addresses and tested separately, later.

Bug: 1011856
Change-Id: I8e2321ef920bc5adab116278521f95b449d10565
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1847214
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703707}