Replaces lower_term with normalized_term in keyword search terms table
Replaces the lower_term column in the keyword search terms table which
contains search term in lower case, with a new normalized_term column
containing the search term in lower case with the whitespaces collapsed
which will be used as a key to delete entries in addition to prefix search.

Depending on whether a search is initiated from the omnibox or the web,
the search term stored in this table may or may not have its whitespaces
collapsed. Since both of the following searches result in the same
zero-prefix suggestion, the new column can be used as a key to delete
both entries.

Omnibox search:
https://www.google.com/search?q=WEEKLY+NEWS&oq=WEEKLY+++NEWS+&sourceid=chrome&...

Web search:
https://www.google.com/search?q=WEEKLY+++NEWS+++&oq=WEEKLY+++NEWS+++&...

Bug: 996516
Change-Id: I7ae5bd6a6992480b3e394499a9b94ca11d1fc8ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1832903
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703869}