Clear autoscroll latched target when view is destroyed
This is a regression from a fix for middle-click auto-scroll for
OOPIFs where input events get locked to a single target when moving
over OOPIFs (https://crrev.com/c/1773540). When middle-click
auto-scroll is in progress and the view is destroyed, we need to
clear the state of latched target. However, the previous CL missed
this fact.

This CL fixes the issue by clearing the state of middle-click latched
target when it is destroyed.

Bug: 1007983
Change-Id: Ie4d98fdc97b7cf14e193857e61212e193cfd7f62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1825489
Commit-Queue: Sarthak Shah <sarsha@microsoft.com>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703918}