Overview: Use the spawn animation for desks removal only
https://crrev.com/c/1832477 previously added the spawn animation
and made it being used for all add-item-to-overview while overview
is active. This animation should not be used for dragging a window
to overview, or removing it from splitscreen, and should only be
limited to appending an item to overview as a result of desk removal.

BUG=1011981
TEST=Manual

Change-Id: I5a55b7bb7e49fb0fa9e5c7eb856ef7499675f698
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1847947
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703938}