[remoting][mobile] Remove TOAST notification appearance
Dialog notifications now support "Don't show again" checkbox, and it is
deemed nontrivial to implement toast with action button on Android, so
the usefulness and maintenance difficulty of having the TOAST appearance
is questionable. This CL removes the TOAST appearance and the appearance
enum altogether.

Bug: 1001291
Change-Id: I2fe082a80fd9f84dbda96faa453ef99d732e9dda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1848412
Reviewed-by: Yuwei Huang <yuweih@chromium.org>
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Cr-Commit-Position: refs/heads/master@{#703960}