Add UMA for whether the heavy ad intervention is blocked
This CL adds a boolean histogram that records true if an intervention
did not occur due to being over the per-origin limit of the heavy ad
intervention. This metric is important to monitor what percent of
interventions are lost due to the imposed limits.

Bug: 1012440
Change-Id: I28d5abaac8af3e80732478a5f2ca529ee7993520
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1846318
Commit-Queue: John Delaney <johnidel@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704034}