Mac: In the component build, make //chrome:chrome_framework a data_dep of :chrome_app.
The framework is bundled within the .app, but none of the dependent
shared libraries created by the component build are. Those need to be
expressed as runtime dependencies of the framework, and by extension the
app. But GN does not propagate that across the create_bundle rule so it
must be done manually.

Bug: 1010347
Change-Id: If4b08329c97e7ef5b0e5073fbb26252b922c50a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1846075
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704057}