DL: Update style & layout in IsFocusableStyle for elements in locked subtrees
Previously we used StyleForElement which is wrong. We should update
style to get the up-to-date style value with the correct inheritance etc
and also need to update layout because some things in accessibility
code requires the layout to be clean (just updating the style & layout
tree from IsFocusableStyle might dirty the layout).

Bug: 1002416
Change-Id: I7b891cde8861cdcb3c79f43b8fa48f13c163dda0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1846215
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704063}