Add Hangouts extension to the AppService.
For extensions, like Hangouts, since that is not app, it is not added
to the AppService, so the icon is not shown on the shelf. Hangouts is a
special extension, which shows an window, so it should be added to the
AppService to show the icon on the shelf, when launching the hangouts.

BUG=1009558

Change-Id: I7cfdbdcba68bb67aa7d70d10ab1a31ab6299bca4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1846213
Reviewed-by: Nigel Tao <nigeltao@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704069}