[omnibox] Remove answer from suggestion if keyword suggestion
When a suggestion has a keyword hint, we draw a lazy chevron to its
right, and offer the tab key hint. But if the suggestion is an
Answer, we replace this icon with the appropriate Answer icon.

When both of these occur within the same suggestion, it looks strange.
First, the suggestion looks like an answer, not a keyword suggestion.
Second, when the user hits tab, the Answer icon slides over as the
keyword chevron would do.

To fix this, we remove the answer from suggestions with a keyword
hint, reverting it to a normal search.

Bug: 1011876
Change-Id: I1f8cddd1fb1d3bd9e5c59d77edcaf92f776a94a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1847851
Commit-Queue: Kevin Bailey <krb@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704159}