Extract coalescing cert verification into a CoalescingCertVerifier
This moves the coalescing logic that is currently part of
MultiThreadedCertVerifier into a dedicated class, which
will coalesce multiple Verify() requests into a single
request to the underlying CertVerifier, if there is one
pending.

This allows for the simplification of the
TrialComparisonCertVerifier, which relied on an additional
signal from the MultiThreadedCertVerifier to know when to
trigger cert verification, allowing the TCCV to be composed
on the MTCV.

While this class can be composed on top of a
CachingCertVerifier, it will generally be more efficient if
it sits as the underlying CertVerifier for the cache, due to
additional bookkeeping involved with the coalescing.

Bug: None
Change-Id: I0c745cf52b83428549869a70aa397c1ef14de033
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1838657
Reviewed-by: Matt Mueller <mattm@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Ryan Sleevi <rsleevi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704176}