Reland "[iOS][MF] Fix search in all passwords"
The test requires more other commits to be on the branch. So I removed it from
the cherry pick.

This is a reland of b400d6c753b37e31373b024573d3373235a23465

The reason for the revert was that the test failing on the slimnav bot. This
is a legacy bot and runs with the feature off.

After investigating, it turns out it is a bug in Earl Grey where it is waiting
for a view controller to dismiss.

Since slim-navigation is not related to the failure, this CL disables the
test when slim-navigation is off, and relands the fix.

Original change's description:
> [iOS][MF] Fix search in all passwords
>
> If the password view controller was presenting something, it will stay
> around. This is fixed and a test is added.
>
> Bug: 1008421
> Change-Id: I8248cbbf28be94bc4b84c65ccda42129a192db0b
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1826895
> Commit-Queue: Javier Ernesto Flores Robles <javierrobles@chromium.org>
> Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#702358}

(cherry picked from commit f62ebc4f45e97e9d7669cac5123a6387870dd6e1)

Bug: 1008421
Change-Id: If336b1cc5b4d6427125921d01b636b0726b860f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1844979
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#704143}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1849684
Reviewed-by: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Cr-Commit-Position: refs/branch-heads/3904@{#700}
Cr-Branched-From: 675968a8c657a3bd9c1c2c20c5d2935577bbc5e6-refs/heads/master@{#693954}