Android: Ban System#exit with non-zero status
When System#exit is used, it bypasses our normal crash reporting
pipeline. Thus ban its usage in favour of simply crashing with an
exception. It used to be the case that crashes before native is
initialized are dropped. Since we now have PureJavaExceptionHandler this
is no longer a concern.

Explicit usages of System.exit(0) are allowed since in that case the
intent is for a normal exit. Specific instances of System.exit(non-zero)
can be allowed in a similar manner as the one allowed in this CL.

Bug: 1000651
Change-Id: Ie3042ec2782dd5577636fcaec65de43d8f0990e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1850352
Auto-Submit: Peter Wen <wnwen@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Peter Wen <wnwen@chromium.org>
Commit-Queue: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704308}