cc: Avoid Unnecessary Encoded Data Copies
In order to decide whether an image will go through hardware decode
acceleration, we currently obtain the encoded data, parse it, and
extract relevant information to compare it against the supported
profiles advertised by the GPU process through GPUInfo.

This is unfortunate because getting the encoded data may require a copy
if the data is not already contiguous in memory. Not only that, but the
copy would need to be done even if the image will ultimately not go
through hardware decode acceleration.

This CL solves this issue by abstracting and plumbing the necessary
information to make the decision up from the DeferredImageDecoder to
PaintImage, and then passing it to the ImageDecodeAcceleratorProxy for
final decision.

Finally, we only make a contiguous copy of the encoded data if hardware
accelerated decoding is supported.

Bug: 995149, 995887
Change-Id: I99f85e1324126b038a576149b0666245639dc03a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1778988
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Leon Scroggins <scroggo@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704355}