Remove net::LOAD_MAYBE_USER_GESTURE.
It was only being set by the AccountAvatarFetcher and the
AssistantImageDownloader, and only read by the ExtensionThrottleEntry
(which may well not even be on the path for the two classes that set
the flag).

As such, it's not serving any significant purpose, so can now be
removed.

Bug: 516495
Change-Id: Ib1b52ab6100a5810836a413e4b05f7a3d06ae7a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1827489
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Eric Roman <eroman@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Tao Wu <wutao@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704375}