ChromeVox: Don't populate tabs menu in the panel unless we're logged in
This change adds logic to ChromeVox to check the login state before
populating the tabs menu in the panel; if we are not logged in, then
we do not populate the menu. This logic was added due to privacy +
security concerns where it was possible to use ChromeVox to access a
user's open tabs without being logged in. More details can be found in
the associated bug.

Bug: 981100
Change-Id: I75afa213e8bf135540af5bbcdb408fcddb526a06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1832667
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704381}