Pull input initialization out of InitializeLayerTreeView
InitializeLayerTreeView has a weird connection with ScreenInfo which
causes us to require it to exist before doing
SynchronizeVisualProperties. We're going to pull it apart.

Here we pull the input stuff out to get out out of the way.

Use the WebMainThreadScheduler from CompositorDependencies instead of
going through a nullable RenderThreadImpl pointer in tests. Then make
the fake on in unit tests return a task runner for the input handler,
which can be the current (main) thread since these tests don't run a
compositor thread. Drop all the null checks around the input queue.

R=avi@chromium.org

Bug: 419087
Change-Id: I117b28eb157ab7b5dc61cef6b80f7f32243404e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1850416
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704389}