Enterprise Printers: Deleting BulkPrinterCalculator objects for User Policy
Since the CL:1838051, each BulkPrintersCalculator is created by the first
object that asks for it. BulkPrinterCalculator for User Policy is
deleted in NativePrintersExternalDataHandler or in the first called
destructor of corresponding EnterprisePrintersProvider or
CalculatorsPoliciesBinder objects. BulkPrinterCalculator for Device Policy
is deleted in DeviceNativePrintersExternalDataHandler. All instances of
BulkPrintersCalculator are accesses in the UI thread and by using WeakPtr.

BUG=chromium:1009062
TEST=on nautilus

Change-Id: I09d4b9b4deea4f0cc06847abbd18f4800a7c5e60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1836660
Reviewed-by: Sean Kau <skau@chromium.org>
Commit-Queue: Piotr Pawliczek <pawliczek@chromium.org>
Auto-Submit: Piotr Pawliczek <pawliczek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704424}