DevTools: add web audio context selector tests
This change adds basic tests that capture current behavior around the
Audio Context Selector.

It is meant as a start and not a complete list of tests,
more tests should be added in the future.

This also indirectly tests the UI.ListModel, UI.SoftDropDown, and
UI.ToolbarItem.

Change-Id: I7506c06e6cf8b865ab2645c9e9e4004df162e629
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1824060
Reviewed-by: Yang Guo <yangguo@chromium.org>
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Reviewed-by: Mandy Chen <mandy.chen@microsoft.com>
Commit-Queue: John Emau <John.Emau@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#704456}