[Upstream Feedback] Refactor how browser window shows autofill bubbles.
The idea is to move autofill bubble showing logic out of browser window
to keep it short and clean. Also this extracted class can be a observer
for autofill data saving and notify avatar so that avatar itself does
not need to be such observers.

Also add avatar button accessor in the ToolbarButtonProvider.

Bug: 964127
Change-Id: I46f85f01b1da9d8d85d9049a1acd95ebe9bb2c21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1845827
Commit-Queue: Siyu An <siyua@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Reviewed-by: Jared Saul <jsaul@google.com>
Cr-Commit-Position: refs/heads/master@{#704474}