[iOS] NTP: Allow Voiceover scroll back up after fakebox leaves screen
The root cause for the bug is that the fakebox fades out as it scrolls
up in landscape. After it gets all the way offscreen, its alpha is 0.
Once that happens, Voiceover doesn't recognize it as a view, so there's
no way to go back up to get it in view. To prevent this, this CL sets
the minimum alpha of the fakebox/header to 0.01, so Voiceover will
still see it. It still shouldn't affect the actual visibility because
the fakebox is behind the toolbar by the time it fades out, so it still
won't be visible.

Bug: 985726
Change-Id: Ibbb3947561e13a6312bcdc8973c7c71c25e728cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1849676
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#704532}