[Signin]: Show the concatenation of the GAIA name and profile name.
Use the concatenation of GAIA name and profile name as the identity of
the profile to be shown in the User Menu and the profile switcher. With
the efforts towards encouraging the use of multiple profiles, it becomes
more important to show a clear identity for the profile and clear
transitions on events like sign in, enable sync, sign out. If no primary
account or unconsented primary account exists for the profile, the local
profile name is used. Otherwise, we show the concatenation of GAIA name
and profile name in this format: |GAIA name (profile name)|. The only
exception where we only show the GAIA name is if the profile name is a
substring of the GAIA name. The implementation is behind a kill switch
|kConcatenateGaiaAndProfileName| enabled by default.

Bug: 1012182
Change-Id: I27626d9e9c77f38698db1012abf5e3fceccc266d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1845766
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Monica Basta <msalama@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704541}