[bfcache] Set all commit parameters on navigations to cached pages.
This CL addresses a bunch of TODOs to properly set
FrameHostMsg_DidCommitProvisionalLoad_Params on BackForwardCache
navigations.

With this CL, we store the DidCommitProvisionalLoad_Params used for
the navigation on RenderFrameHostImpl, so that it can be reused when
restoring from the BackForwardCache.

This approach should be somewhat more future proof, particlarly when
adding new fields to DidCommitProvisionalLoad_Params.

In order to prevent any change in memory usage, the parameters are only
stored on the RenderFrameHostImpl if BackForwardCache is enabled.

Change-Id: Idb72e7366876e59d1dfd046fa83284f6f280422c
Bug: 1005718
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1819251
Commit-Queue: Lowell Manners <lowell@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704612}