[LayoutNG] Support unknown fragmentainer block-size.
In the initial column balancing pass we don't know the block-size of the
columns (fragmentainers). We previously set the column size to
"infinity" (LayoutUnit::Max()) in this initial pass. Apart from being
slightly ugly, it is also problematic, as that would cause containers of
children with forced breaks to become "infinitely" tall, and we'd thus
end up with "infinite" column sizes.

Instead leave the fragmentainer block-size at unknown (kIndefiniteSize)
in the initial column balancing pass, and have the algorithms cope with
this.

At the same time, replace FragmentainerSpaceAtBfcStart in
NGConstraintSpace with an offset; i.e. instead of storing the remaining
*space* at the nearest BFC block-start, store the *offset* from the
fragmentainer block-start to the nearest BFC block-start. One good thing
about this is that we only need to deal with one kIndefiniteSize field,
instead of two, but this change will also come in handy when we add
support for calculating the tallest unbreakable piece of content in the
initial column balancing pass (in an upcoming CL).

Bug: 829028
Change-Id: Id66c44452601d499d3d22c9f231576d26e10c459
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1852290
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704802}