Don't receive any IPC messages in an undead/provisional RenderWidget*
* Except EnableDeviceEmulation.

Emulation on/off were the last messages. These messages should only be
sent in sequence with there being a RenderFrameHost present for the
RenderViewHost+RenderWidgetHost.

The EnableDeviceEmulation message is sent when navigation is ready to
commit, which is when the RenderWidget is still for a provisional
frame. At that point the widget is not undead, and the WebFrameWidget
is valid. So it's not too terrible.

Once the RenderWidget is deleted instead of made undead, we can just
accept messages while provisional without worry.

R=avi@chromium.org

Bug: 419087
Change-Id: I9ddfd81bc1ea0c448e86afe13e2b0a1b38dd6121
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1853675
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704922}