Restrict VoiceOver from navigating to programmatically hidden elements
This is an improved version of the original fix, which accidentally
restricted VoiceOver from navigating to elements that are partially
offscreen.

In this version of the fix, we still support that option, but change
the internal name of it from SetVisibleOnly() to SetOnscreenOnly(),
in order to avoid confusion.

In addition, whether the onscreen only option is used or not,
programmatically hidden elements (e.g. aria-hidden, visibility: hidden)
are now filtered out of the navigation.

Bug: 982408
Change-Id: I13441fe9f1e8243d4adbaf6fea9e169fc325f1ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1849044
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704923}