Cap the number of endpoints ReportingEndpointManager tracks.
The class tracks which endpoints uploads have succeeded/failed to.
With the introduction of NetworkIsolationKeys, seems like we should
bound its size, just in case.

This CL uses an MRUCache to do that. Ideally, we'd just evict URLs with
no entries in the ReportingCache, but that's a fair bit more complicaed,
and it's not clear if it's worth doing.

Bug: 993805, 1012931
Change-Id: I169882870e5ec7728c66653cea2134313905508d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1853860
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704925}