bfcache: Disable BFCache when WebSerial is used.
Similar to WebUSB, pages that uses the Web Serial APIs shouldn't be
frozen and enter the BFCache (at least for now).

This CL marks the RenderFrameHost that uses the SerialService as
disabled for BFCache.

See CL for WebUSB:
https://chromium-review.googlesource.com/c/chromium/src/+/1837933/

Some background on why we sometimes need to disable bfcache:
https://docs.google.com/document/d/1NjZeusdS1kyEkZyfLggndU1A6qVt0Y1sa-LRUxnMoK8

Bug: 992851
Change-Id: I26f48bb309c2c2c6f2737e52c1ce533f8e03f437
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1851785
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704928}