Replace WebLocalFrameClient::DidFailProvisionalLoad with AbortClientNavigation
These two mean the same: client navigation should be aborted, most likely
because a new one has started, or frame has been detached, or due to
window.stop() call.

This also adds missing RFO::DidFailProvisionalLoad call to AbortClientNavigation.
Previously, we notified about the navigation start, but if it was aborted we
did not notify RFOs.

This corresponds to the step 7.h from the doc linked to the bug.

Bug: 855189
Change-Id: I9a643870fd2f70645156bc6e5cc46da9c60a6c41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1802521
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Dmitry Gozman <dgozman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#704980}