Convert //components/web_resource away from base::Bind()
As part of the effort to replace base::Bind by
base::Bind{Once, Repeating}, this CL converts the one usage of
base::Bind() in //components/web_resource to instead be a usage of
base::BindRepeating(). The rational for using base::BindRepeating() is
that the callback is being passed in to a PrefChangeRegistrar, which
can invoke it multiple times in response to pref changes.

This CL also takes the opportunity to update the PrefChangeRegistrar
API in question to take in a base::RepeatingClosure rather than a
base::Closure.

TBR=jochen@chromium.org

Bug: 1007753
Change-Id: I7550f64eedb9d0da719f9ff8617e2c4d9d415eab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1847216
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705009}