views: remove DialogDelegate::ShouldSnapFrameWidth
This concept does not belong in DialogDelegate, since it is implemented only in
BubbleFrameView. This change moves the policy decision of when BubbleFrameView
should snap a bubble's width into BubbleFrameView itself. This change also
removes the only override outside Views of ShouldSnapFrameWidth, which was on
ExtensionsMenuView; this class was overriding it to return true while also
returning a static width from CalculatePreferredSize, meaning that it actually
has a fixed width anyway.

Bug: 1011446
Change-Id: I99369e83b7c467b8cbc2927727f0d268c44e6ea1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1854458
Reviewed-by: Peter Boström <pbos@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705054}