Adds pattern matching to DomainsToNotCheckComplianceOfUploadedContent, DomainsToCheckComplianceOfDownloadedContent and DomainsToCheckForMalwareOfUploadedContent policies.
This makes those policies more efficient by not needing to enumerate exact domains and also allows to match against more granular url properties, for instance by using patterns like 'http://*', 'https://server:8080/path' or '*'. The pattern matching used is the same one used by the URLBlacklist policy. To achieve this, this cl includes a refactor that puts that pattern matching code in a common util file.

It is worth noting that this changes what the policies actually do and so DomainsToCheck* is no longer an appropriate name, they will be renamed in a different cl.

Bug: 1008486
Change-Id: I71410849403f2fc54cefe34d94b0f64256f3e18a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1841772
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Owen Min <zmin@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Roger Tawa <rogerta@chromium.org>
Commit-Queue: Dominique Fauteux-Chapleau <domfc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705092}