Downgrade secure icon on pages that trigger a Safety Tip.
This CL downgrades the security icon of pages when a Safety Tip is
available for the page. The updated icon is the generic (i) NONE icon.
It now looks like this, even on secure sites:
https://drive.google.com/file/d/1J0oWA4yue1ydYTNT0R_bIj-cLm7LNUXV/view?usp=sharing

This CL also moves the Safety Tip UI feature flag to security_state's
feature set. This change was necessary to read the feature from within
components.

Bug: 1004849
Change-Id: I892b3d0c12d13e04bd486f33ed686b1f7398795a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1848413
Commit-Queue: Joe DeBlasio <jdeblasio@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Reviewed-by: Christopher Thompson <cthomp@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705145}