Make undead widget reviving act like new widget creation.
The order of operations for making a new RenderWidget for a frame is:
1. new RenderWidget.
   Now we have a pointer but it is not used because it is not
   initialized.
2. pass RenderWidget to WebFrameWidget as the WebFrameWidgetClient
   This gives the pointer out but it's just stored, which allows the
   WebFrameWidget to function correctly once it is used.
3. Initialize RenderWidget
   This includes calling into the WebWidget, which is why step 2 was
   important.

Reviving a widget from undead should look like this too, but currently
in step one, instead of just marking the widget as not undead, we also
jump ahead and start to do initialization. That means the
initialization code requires null checks around the WebFrameWidget
being present or not, and fails to do some of its normal initialization
steps in the revive-from-undead case.

This CL fixes revive from undead to look like creating a new
RenderWidget. Step one only marks the RenderWidget as no longer
undead and performs the equivalent steps if needed as the
RenderWidget's constructor. All initialization is deferred to the
InitForRevivedMainFrame() which will run as step 3.

The creation path is modified to be more explicit as well. When step 3
initialization happens, if the widget is undead, the LivingInit() steps
are skipped. They will be run when the widget is revived from undead.
And they will be run each time the widget is revived, if it happens
multiple times over its lifetime.

RenderWidget-specific logic is moved out of RenderViewImpl in the
process. The RenderWidget::SetIsUndead(true) method is the replacement
for Close() and it can do partial shutdown such as closing the
WebWidget. This removes the need for SetWebWidgetInternal(), making
the RenderWidget API for undead-ness closer to the normal APIs.

As compositor and WebWidget initialization in step 3 is now always
after the WebWidget is set, null pointer checks for the WebWidget can
be removed, as was the original intention in this CL.

R=avi@chromium.org

Bug: 419087
Change-Id: I024c852d00110015308640f272220ed7d560d5b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1854818
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705167}