[Fuchsia] Don't use protected memory on x64.
Protected memory and HW decoders are not supported on x64 devices,
so WebEngine wasn't working properly when WIDEVINE_CDM feature was
enabled. This CL adds IsFuchsiaCdmSupported() that indicates that
that FuchsiaCdm is supported on the device. For now it's just a stub
that returns true only in arm64 builds.

With this change WebEngine will not try to use protected memory on
x64 devices where it's not supported.

Bug: 1013412
Change-Id: Ib9c323c725886ea7397bc7f2633a44fad56ee13a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1854547
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Auto-Submit: Sergey Ulanov <sergeyu@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705189}