Fix accessibility issues with sliders on Android
First, re-lands the fix to bug 836554, which ensures TalkBack
announces when a slider changes. This time, adds a check to only fire
the TYPE_VIEW_SELECTED event if the slider has accessibility focus,
hoping that will address the problems with fullscreen video by suppressing
announcements when an unfocused slider changes.

Also, doesn't expose the "can scroll forwards / backwards" actions on a slider
that isn't an HTML INPUT control, because a pure ARIA slider won't respond
to those. That addresses bug 917845.

Finally, added code to ensure that a slider updates the accessibility tree
when its value changes - I found this was necessary in order to get
announcements after changing a slider with the volume keys.

Bug: 836554, 917845, 953652
Change-Id: I3f1f88321a4644bc5729c16c14e0bd9037a641ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1853857
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705199}