Show kite icon and parent tooltip in Site Settings for child users
Site Settings page currently shows the enterprise "office" icon
instead of the "kite" icon for supervised users. This CL fixes this
and ensures that the kite icon and "This setting is managed by a
parent." tooltip is shown in Site Settings when a supervised user
is logged in.

Bug: 1005547
Change-Id: I91d81337b3f1b5cd070163ecc032a1695592dc0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1838652
Commit-Queue: Toby Huang <tobyhuang@chromium.org>
Reviewed-by: Michael Giuffrida <michaelpg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705205}