Remove redundant null checks around input handling and context menu
The WidgetInputHandlerImpl always checks that the RenderWidget is not
undead before calling into it, so these methods always have a WebWidget
present. The context menu method is an IPC handler which is only
received when not undead or provisional, so it always has a WebWidget
present.

SimulateImeFinishComposingText() is never called.

ImeCommitTextOnMainThread() is called from RunOnMainThread() which
only executes the callback when the RenderWidget is not undead, so
a WebWidget is always present.

R=avi@chromium.org

Bug: 419087
Change-Id: I1b29030638f9e31f6c5646c5e6894f07a1e9b860
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1856718
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705240}