Keep all filter info as we pass through to adapters
Previously we were losing filter data be merging an array of filters
into one filter.  This patch restructured the BluetoothDiscoveryFilter
To have a list of DeviceInfoFilters which in turn keeps arrays of UUIDs
which more accurately represents the filter being passed in per the
web bluetooth specification.

Bug: 490401
Change-Id: Icdf8d102276db79eaff81e823d6b7654b65ba1c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1842559
Commit-Queue: James Hollyer <jameshollyer@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Ovidio de Jesús Ruiz-Henríquez <odejesush@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705349}