Unconstrain pointer when surface is destroyed
As per the wayland-protocol spec for locked pointer, the locked pointer
object is no longer valid if the surface to which it is constrained is
destroyed.

This CL makes sure to break the constraint and send an unlock event when
the surface associated with a locked pointer is destroyed.

Bug: None
Change-Id: I368af9a725da3c3b871c7574b2c76fe5fc113fec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1854879
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Prabir Pradhan <prabirmsp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705358}