Replace InternalAppName with BuiltInAppName
This is the follow up for crrev.com/c/1816495. InternalAppName has been
moved to BuiltInAppName in AppService, so replace InternalAppName with
BuiltInAppName, otherwise we have duplicate enum, which could cause
confusing for the future change.

Although the app_list/internal_app will be replaced by AppService for
the long term goal, in short term(one or two months), they are still
there, so as the temporary solution, update it to use the AppService's
enum BuiltInAppName.

BUG=1005640

Change-Id: I0992d549c25a40087212baba0b8d8305b1136d0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1852526
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705364}