Remove t/b/p/w/web_presentation_receiver_flags.h dependency from chrome/browser
In order to remove the dependency, the web_presentation_receiver_flags.h file
was moved to third_party/blink/public/common/presentation. Since the file is no
longer in the public/web/ directory we are also renaming it to
presentation_receiver_flags.h. The static keyword was also removed as constexpr
is static.

Bug: 1008303
Change-Id: I78c6641ac7c41fc7275b121ad5f7e68a9476f080
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1856079
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Sergio Villar <svillar@igalia.com>
Cr-Commit-Position: refs/heads/master@{#705402}