Use a range of dictionary sizes for hunspell_suggest_fuzzer
The dictionary used for fuzzing is ~500KB, which results in fuzzing at
~1 case/sec on clusterfuzz. Reducing the dictionary size allows the
fuzzer to run much faster, but may result in less coverage. Use a
FuzzedDataProvider to generate a range of dictionary sizes to get a mix
of coverage and make the fuzzer much faster.

BUG=1009087

Change-Id: Ia25feef99e48a1ed28ea6de636430d3509c69def
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1847603
Commit-Queue: Anand Mistry <amistry@chromium.org>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705459}