Group PDF page accessibility objects into new struct
This is a preparatory change for adding accessibility support for PDF
annotations. Currently, text, links and images within a PDF page are
supported in an accessibility context. These are passed as separate
parameters in Set/GetAccessibilityPageInfo methods exposed through the
Pepper API. When annotation support comes in, the number of parameters
would increase to 6 making those function calls unwieldy and requiring
additional extensions to the templated DispatchResourceCall function.
This change groups links and images into a new *PageObjects struct which
will also be the home for the upcoming representation of PDF
annotations.

Bug: 1008775
Change-Id: Icee1b9b17290614265be8c65c5b594d3a99fe767
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1831957
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Raymes Khoury <raymes@chromium.org>
Reviewed-by: Ian Prest <iapres@microsoft.com>
Commit-Queue: Kalpak Tapas <katapas@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#705481}