[cleanup][image_bitmap] Replace use of WTF::Uint8Array with WTF::Vector
This CL removes another use of WTF::ArrayBuffer that is unrelated with
JavaScript ArrayBuffers. We want to replace all uses of WTF::ArrayBuffer
which are unrelated to JavaScript ArrayBuffers to be able to refactor
the memory management of ArrayBuffer backing stores.

Even though the original code used ArrayBuffer::CreateOrNull, the
callers of CopyImageData did not handle the "Null" case. Therefore I
just replaced the WTF::ArrayBuffer with WTF::Vector.

R=jbroman@chromium.org, haraken@chromium.org

Bug: chromium:1008840
Change-Id: I03980eb402f943159465f13a5ea8b784c83f79ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1844956
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705502}