Remove t/b/p/w/web_fullscreen_options.h dependency from chrome/browser
In order to remove the dependency, the web_fullscreen_options.h file was moved
to third_party/blink/public/common/fullscreen. Since the file is no longer in
the public/web/ directory we are also renaming it to fullscreen_options.h (and
the struct defined in it to FullscreeOptions).

Bug: 1008303
Change-Id: If96a8263b3bb696931ce9af4846f3d6dcbc40698
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1852288
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Sergio Villar <svillar@igalia.com>
Cr-Commit-Position: refs/heads/master@{#705577}