AW UMA: persist install date
This persists the install date pref, so this will now log the install
date of the WebView-embedding application. For existing clients, this
will log the first startup in a milestone of WebView which persists this
pref.

This also includes logic to look up the install date if it's not already
logged, otherwise the install_date field would just log the date users
updated to a version of WebView which supports this field. This logic is
meant to be temporary.

I've added a histogram to track when we can safely remove this
workaround, and added a tests to verify the logic and that the histogram
is logged.

Bug: 1012025
Test: startup webview shell multiple times, log install_date inside
Test: AwMetricsLogUploader::UploadLog, verify it stays consistent
Test: run_android_webview_unittests --gtest_filter=AwMetricsServiceClientTest.*
Change-Id: I2d9238120b6bab2e08c116bf8f57e541b905534d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1846812
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705657}