[NativeFS] Decouple remaining IO thread bound work from the rest of the code.
This refactors the blob interactions to either be behind a PostTask to the IO
thread, or in the case of writes eliminates the direct interaction with the blob
system entirely by reading the blob through its mojo interface instead.

Also migrates one remaining FileSystemContext call to be behind a PostTask to
the IO thread.

This, combined with a previous CL, eliminates all code that required running on
the IO thread, paving the way for moving most of the native file system API
implementation to the UI thread.

Bug: 1011534
Change-Id: I8ef3ce97a77efb42bd12b020e98d7c1430ca3292
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1845914
Reviewed-by: Olivier Yiptong <oyiptong@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705698}