StartupObserver should detect downloads
The StartupObserver class monitors the startup of the browser and is
used to defer BEST_EFFORT tasks until after the browser has completed
startup and the active tab's page has finished loading.

The StartupObserver previously failed to detect the case where the
browser as launched to the url of a file download; it would then wait
for an additional navigation in the first tab or would wait until a
failsafe timeout was reached before indicating startup was complete.

This change signals completion in the scenario where a file download
takes place in the first tab. The failsafe timeout remains set at 3
minutes.

Bug: 1006954
Change-Id: Ia1e056f1a54ccc561d1dcf306161470beefe42e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1846099
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Eric Lawrence [MSFT] <ericlaw@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#705711}