[LayoutNG] Allow negative margins to stay on the line after overflow
This patch allows atomic inlines to stay on the line even
when it appears after the line overflowed, if its negative
margins can bring the position back to inside of the line.

This is not a well-defined behavior in the CSS spec, but
legacy/WebKit/Gecko do this at some levels, though not really
interoperable.

Bug: 1001000
Change-Id: If8cfc5d51de14c8de6a9caa78fa10abaf6f1b5b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1859424
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705848}