imf: Record key event latency in a more robust way.
Currently we record key event latency as measured as the time between
a pending key event was added to when the pending event was resolved.

However, this "pending key event" code path is only used by extensions
based input engines. We are planning to add a native engine that doesn't
use the pending key event dictionary approach. So those key events would
not be recorded right now.

A more robust way is to encode the latency measurements as part of the
key event callback. That way, as long as the key event is handled in
some way, we will record it correctly.

The metric was added recently and currently has no data, so this
doesn't invalidate our metrics.

Bug: 1009903
Change-Id: I747420bd7e143965f6f15c74fa81d55b6783b548
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1860197
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: Keith Lee <keithlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705850}