[iOS][MF] Consider iPad Pro Keyboard
On iPad Pro the keyboard layout is 3 points above the bottom, but not
one of its superview. The new checks goes up all the hierarchy until
the first view that is the size of the window. Then checks that it is
anchored to the bottom of the window.

Change-Id: I514910cd17253f9ad1f80ba11a2ba66aff6028a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1860033
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#705922}