views: default focus for tabstrip to active tab
When navigating to the tab as an accessible pane (via F6), focus currently gets
dropped at the first tab, which is not as useful as it might be. It is easy to
get to the start (via Home) or end (via End) of the strip but not to get to the
active tab. As such, focus should land on the active tab by default instead.

Bug: 910687
Change-Id: I0d70e171557d5591745c117bdca067f77c4feaec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1857303
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706048}