[omnibox] Clean up OmniboxEditModel::OnPopupDataChanged API
The API currently has an unused pointer argument that's used as a bool.
It just makes for a confusing read, and I'm trying to build on top of
it to fix some polish issues for on-focus suggestions (ZeroSuggest).

This CL turns that pointer-as-boolean argument into an explicit bool
argument that's properly labeled. No behavior changes.

Bug: 996516, 963173
Change-Id: I02767d26849dae035feeb54173ebee2fbc66ea03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1862256
Auto-Submit: Tommy Li <tommycli@chromium.org>
Reviewed-by: Kevin Bailey <krb@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706092}