[Files ranking] Add more DriveQuickAccess debug metrics.
We're unsure if there are errors stemming from one of the APIs that the
DriveQuickAccessProvider implicitly relies on. This CL adds debug
metrics for all operations that could error within the provider.

There are a lot of metrics here, and we plan to remove them once we're
confident the feature is working as intended.

Bug: 1011221
Change-Id: Iea9a3ee3237de88fe30bba6705adcebd962338b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1857800
Reviewed-by: Jia Meng <jiameng@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706244}