Send out Assistant ready signal when LibAssistant is started.
Currently the ash::mojom::AssistantState::READY signal is sent out when
LibAssistant is created, not when it finished starting.
This puts us in a limbo state where we're told everything is up and running, but
any query sent out will be rejected.

To fix that, a new signal ash::mojom::AssistantState::NEW_READY has been
introduced that will only be sent out when LibAssistant signals it is ready to
process queries.

This signal should eventually replace the old |READY| signal, but for now it is
added as a new signal to allow us to check it is sent out consistently on all
devices.

Internally this is achieved by introducing a new |AssistantStateObserver| which
is used to inform |Service| about all state changes to the
|AssistantManagerService::State|. These state changes are then forwarded to
|AssistantClient::OnStatusChanged|.

BUG: b/141264108
Change-Id: I38d9635001be5ef13e80090102646559e5c17158
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1835183
Commit-Queue: Jeroen Dhollander <jeroendh@google.com>
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Reviewed-by: Tao Wu <wutao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706268}