Move the continue reading to Built_in_app.
Built_in_apps is used to replace the internal apps. However,
InternalDataSource is still used for the continue reading and release notes,
which is confusing. So move the continue reading to Built_in_apps, and
AppSearchProvider should depend on AppServiceDataSource for all apps' search
and recommended. InternalDataSource and InternalAppResult are removed to
avoid duplicate code. All the related code from InternalDataSource and
InternalAppResult has been moved to AppServiceDataSource and
AppServiceAppResult.

Also GetInternalAppNameByAppId is removed, because no one uses it. It should
be removed by CL:1851890.

This CL is the preparation for the CL:1816256, which moves the internal app
list to Built_in_apps.

BUG=1013540

Change-Id: Ic3087807fe520ff5a1ffcb200cb1cf459b42e2cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1847698
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Tao Wu <wutao@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706351}