[iOS] Pass Browser into SettingsNavigationController initializers.
This is the first CL of several breaking crrev.com/c/1860025 into smaller
chunks.

This updates most of the SettingsNavigationController class inits to take
a Browser instead of a BrowserState. One of the inits (the import data
one) is called from AuthenticationFlowPerformer, which doesn't yet have
a Browser, so it's initialized with a BrowserState instead. Because of this,
SettingsNavigationController itself is initialized with a BrowserState.

There's also some cleanup of these init names. Prior to this change, none
of the class methods (for example: "newAccountsController:delegate:")
properly identified the first parameter. Some of the method signatures were
also impossible for clang-format to wrap to 80 characters. So:
 - The return type of all of these methods is now 'instancetype'.
 - The 'new' prefix is dropped from all of them, matching current style.
 - Methods that had 'WithBrowserState:' now have 'ForBrowser:', aside from
   -importDataControllerForBrowserState:delegate:importDataDelegate:
    fromEmail:toEmail:isSignedIn:.
 - Methods that didn't have a keyword for the BrowserState parameter also
   now have "ForBrowser:".
 - "syncEncryptionPassphrase" is shortened to "syncPassphrase", since
   "Encryption" doesn't add any useful information. The full method name is
   now +syncPassphraseControllerForBrowser:delegate:.
 - "autofillProfille" is corrected to "autofillProfile" (one "l" in
   "profile", The full method name is now
   +autofillProfileControllerForBrowser:delegate:
 - "settingsMainController" is reordered to "mainSettings". The full method
   name is now +mainSettingsControllerForBrowser:delegate:

BrowserInterface didn't have a 'browser' property, which it clearly needed.
This CL adds one.

Change-Id: I6b0a8c8583dc69485a0f65ac1d069de7b417d564
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1862915
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706481}