Content Verification: Make computed hashes aware of dot/space suffix treatment.
On windows, filename with (.| )+ suffix is ignored, i.e. "foo.html."
and "foo.html" are treated the same.

VerifiedContents is already aware of this and it stores a
canonicalized version of filename for filenames containing (.| )+
suffix.
This CL makes ComputedHashes aware of the change too, so that searching
for hashes will consider canonicalized version of the filename as
candidate. This makes ComputedHashes::Reader and VerifiedContents treat
this suffix behavior consistently.

This CL also adds unittest and browsertest for the fix.

Bug: 696208
Test: See bug for test repro.
Change-Id: I7c59add5c035815673259ef1f995c7b9ed01a75b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1772643
Commit-Queue: Istiaque Ahmed <lazyboy@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706493}