Fix web_test expectation handling on Mac10.14, 10.15.
When we added support for running on Mac 10.14 in
https://crrev.com/c/1333567 and 10.15 in
https://crrev.com/c/1776218, we didn't update the
global list of supported test configurations in
blinkpy/web_tests/port/base.py, which means that a
test expectation would have been ignored on those
machines *unless* you happened to be running on a
Retina machine (like a Macbook Pro), which is
treated as the "mac-retina" port and not "mac-10.14".

10.14 and 10.15 are still not fully supported, as
we don't support separate baselines for them and
don't support 10.14 or 10.15-specific expectations.

This CL just means that the tests will run and 'Mac'
expectations will work correctly.

Bug: 1014785
Change-Id: I43521c752ceee92f21ed4335ebe055c486cc8427
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1864002
Commit-Queue: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Rakib Hasan <rmhasan@google.com>
Reviewed-by: Robert Ma <robertma@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706555}