Adjust when WebXr timeouts run
There seems to be an issue with starting up the "timed out" UI when a
notifications prompt is visible or needs to be dismissed.  This modifies
the behavior of SetFramesThrottled to only stop/start the timeouts if a
webxr frame has not yet been submitted.  This is essentially the logic
that existed for these timeouts prior to the implementation of the
throttled logic.

Bug: 1014764
Change-Id: Ia9646799b8d54efc1f39e8800bf82acc5cf861a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1865810
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Klaus Weidner <klausw@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Klaus Weidner <klausw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706681}