Update uninstall dialog for ARC apps.
The ARC app uninstall dialog's layout is different with the
extension/Web App/Chrome App uninstall dialog, so use different layout
for the ARC app uninstall dialog. The icon for the ARC app is not the
window icon, but used for ImageView. So update the ARC app uninstall
dialog to keep the consistency with the current ARC app uninstall
dialog.

BUG=1009248

Change-Id: Ia624312b2b55dabfbbc43ea04aac0e35f8b44fed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1861794
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Bret Sepulveda <bsep@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706711}