Match location bar color better for find bar.
Because the location bar used to be light even in dark windows, the find bar had
a hack to use the non-incognito NativeTheme object.  Now that the location bar
is dark in incognito, this hack is incorrect.  Remove it.

This also means View::SetNativeTheme() is only called by test code, so this
renames it accordingly so people don't add additional non-test uses.

Bug: none
Change-Id: Id0a3363c0d3457cdfdff5ab81bff9be7ce34f15d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1864515
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706727}