Add 1 util method, test it. Add a Model abstract class with inline docs.
This CL could be split in two.

First, adds a method to Stroke for the biggest touch-major of that
stroke, and the corresponding unit test.

More importantly: Adds an abstract class for what a checked in model
looks like. All the configuration is accessible from the model when
instantiated, and no implementation is added. All the fields in
configuration are documented, and an explicit constructor for the
configuration is added since this counts as a complex struct.

Bug: 1009290
Change-Id: I7eeca918540e80a34133230c6d0a67c6b58c2af5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1864490
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Rob Schonberger <robsc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706743}