[directory-tree] Add custom cr.ui.Tree/TreeItem row style support
cr.ui.Tree assumes the rowElement of the TreeItem is one to indent via
style (padding-inline-start) to give the appearance of a tree.

The inner structure of TreeItem is known to the using app and it might
be that a child element the rowElement should be indented, and not the
rowElement itself, to give the appearance of a tree.

Allow cr.ui.Tree/TreeItem using apps to customise TreeItem's row style
using an optional rowElementDepthStyleHandler.

No change in behavior, no new tests.

Bug: 992819
Change-Id: Ifb5762ed97b265602df2e50e482045fd5d3f3151
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1859796
Reviewed-by: calamity <calamity@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706796}