Apply complete surface transform to rounded corners
This patch applies the complete surface transform to the rounded corner
bounds instead of just the translation and scaling component. This may
however fail if the transform being applied does not result in an axis
aligned output.

Since this is done at multiple code locations, a new API is added to
gfx::Transform to perform this.

Test: Added unittests for the new API.
Change-Id: I75b7234ca7b1abc918b35251b77b0a9f8aec1e2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1838980
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Malay Keshav <malaykeshav@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706813}