DL: Fix flaky getFullAXTree & wrongly updated accessibility tests
A devtoools display locking a11y test was marked as failing in MSAN bots
and is also flaky when tested locally, this CL updates that test and
fixes the cause of the failure (getFullAXTree might be called when
style/layout is dirty, which caused failing DCHECKs).

Some other display locking a11y tests were not updated correctly, and is
updated in this CL as well.

Bug: 996625
Change-Id: I32eac1385c54eb07b1e6dab800620a2485158437
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1846618
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Reviewed-by: Alice Boxhall <aboxhall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706834}