enterprise.platformKeys: Don't check EnabledForUser policy level
Temporarily remove the check of the AttestationEnabledForUser's
pref/policy level from the chrome.enteprise.platformKeys
implementation, since this check was not present before (until
r701979), and since this isn't compatible with the current corp
policies.

This is a temporary workaround that will need to be reverted in
the future, once the corp policies are fixed to have the
AttestationEnabledForUser policy configured at the "mandatory",
not "recommended", level.

Bug: 1000589, b:142616571
Test: set the AttestationEnabledForUser to the recommended level, open corp cert enrollment extension's background page in DevTools, run "chromecert.Request.testOnly.getEuk("")", check that it doesn't fail with "Remote attestation is not enabled for your account."
Change-Id: I0087364882728fdc63e84a789313e525c24a21d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1865340
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Auto-Submit: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Yves Arrouye <drcrash@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706854}