[cleanup] Avoid use of ArrayBufferContents::DataHandle
It seems to me that the additional step of first creating a DataHandle
was used to handle the case where allocation fails. However, this is
also possible by creating an ArrayBufferContents directly.

In addition I changed the allocation to avoid zero initialization,
since the whole memory is overwritten immediately anyways.

R=sammc@chromium.org

Bug: chromium:1008840
Change-Id: I43750c462e4959b3bd2b993bc177dcf56593fec9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1864785
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706872}