[TouchToFill][Android] Set correct peeking height
The Peeking state became customizable. With that, we can now show 1.5
suggestions initially and allow expanding the sheet. It is alsow
prerequisite for the footer which we only show in FULL state.

The flow of scrolling is now:
1) Keyboard opens in Peek state.
2) swipe up motion expands to Full state.
3) swipe up motion scrolls the list to the bottom.
4) swipe down motion scrolls the list to the top.
5) swipe down shrinks/dismisses the sheet.

Btw: the minimum height of the sheet could be smarter: instead of
using the fixed size of 340dp, the header size plus 1.5 suggestion
heights fit better. This smartness should be added AFTER the issue
1012219 is resolved.

Screenshots showing differences in states are in https://crbug.com/957532#c47.

Bug: 1012219, 1013257, 957532
Change-Id: I74b7209e447b474815f21a4aafed8672fc57889a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1857128
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706875}