Update the V8 version for Chromium branch 79
This is a roll-back to V8 7.9.317. This is the most stable candidate out
of the latest Canaries.

If this commit results in failures in Blink please contact the Blink
sheriff. They may need to make changes to Blink because of the
roll-back.

When in doubt please try to contact the committer and reviewers of this
CL before reverting it.

R=machenbach@chromium.org
CC=hablich@chromium.org

Change-Id: I18d9b5d0adf94660ea8590c2ac8c598b1814c938
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1866571
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#706897}