Ensure __HrLoadAllImportsForDll only runs once per dll.
If we patch a function in a delayloaded dll, __HrLoadAllImportsForDll is
called to load everything before the patch. If we then patch another
method in the same dll, __HrLoadAllImportsForDll is called again and
this will overwrite the original patch. This change adds tracking to
ensure the load only occurs once per dll.

Bug: 1012424
Change-Id: Ib20fe38461a7eab3735f851bdfd53f9cf99fc5d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1848362
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Cliff Smolinsky <cliffsmo@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#706944}