ozone/drm: Cache connector drm_display
There is some stuttering when viewing protected content with a DP MST
dongle attached.

Looking at a trace it looks like we spend a lot of time (200ms)
in DrmDisplay::GetHDCPState, in particular in drm_->GetConnector.

This patch caches the connector used to check if HDCP is available.

It additionally downgrades log severity when the HDCP properties
is not found to INFO, since it can happen every 2 seconds.

Bug: 1001211
Change-Id: I77eb86b2e55f0fd4640560ae2e38d64d5e2b2174
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1864354
Reviewed-by: Daniel Nicoara <dnicoara@chromium.org>
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Cr-Commit-Position: refs/heads/master@{#707065}