[omnibox] On-Focus Suggestions / ZeroSuggest - Fix revert to user text
Currently, the process of reverting to the user text is broken for
on-focus suggestions.

After using the arrow keys to choose a match, pressing Escape does not
work. Escape selects the top match, which normally works, but in
ZeroSuggest, the textfield contents could be a non-prefix of the first
suggestion.

This CL reverts the textfield to the input that was sent to the
Autocomplete controller, which should work for on-focus suggestions.

This also adds a test.

Bug: 1012990, 996516, 963173
Change-Id: Id19aaf961591dc0c81bf9083b639233c5f8e4420
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1865551
Reviewed-by: Kevin Bailey <krb@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#707068}