DevTools: Add alert to read out updates to device orientation in sensors tool
Fix accessibility bug: In the sensors tool, when the reset button under Device orientation was clicked, the screen reader did not read anything out to inform user that the orientation was changed.

Now whenever the device orientation updates, the new setting is read out. If the update was only read out when reset button was pressed, the experience would not be consistent, and the alert wouldn't be read out for subsequent resets, which is why the update is read out on every change.

Gif: https://imgur.com/a/1YeTX2l

Bug: 963183
Change-Id: Id4f82c4af8db1857af832d4b221d60cd1eea578c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1750194
Commit-Queue: Anubha Mathur <anubmat@microsoft.com>
Reviewed-by: Lorne Mitchell <lomitch@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#707069}