Update the in-page time element when an invalid value is selected from the popup.
This CL adds support for updating the in-page time element when an invalid value
is selected from the popup. A value can be invalid because it doesn't fit the rules
determined by step, min, max attribute.

For time picker the invalid values can't be disabled in the popup since a value
is composed from multiple columns (e.g. min=10:20, 21 minute cell should be disabled
when hour cell < 10).

Tests are added to validate the new behavior.

Bug: 1004997
Change-Id: Ia91c8b23b572915bf78d59166fd8aaf80ae9ab4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1866068
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Ionel Popescu <iopopesc@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#707084}