Move SetTrailingMargin to ToolbarButton base
ToolbarButton is already responsible for interior margins as insets on
both sides. Prior to this all ToolbarButtons could have a leading margin
but no trailing margin. In practice it just consolidates code for
consistency.

This change ends up removing duplication of GetAnchorBoundsInScreen()
and UpdateBorder() both of which are already handled in ToolbarButton.

As a bonus (!) this applies Fitts' law to toolbars shown in fullscreen
windows, which I needed to verify that extending trailing margins still
worked as expected on Mac.

Bug: None
Change-Id: I72bdd6be6663167226a8e9b8446d874859774cb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1866982
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#707166}