replace InvalidDevicePolicyFiles with InvalidDevicePolicyFilesStatus
The current UMA, InvalidDevicePolicyFiles, that logs the number of
invalid files doesn't allow to understand if all the files are broken
(which might suggest a disk corruption). This change introduces
InvalidDevicePolicyFilesStatus logs if no, some or all files are
invalid.

CL that implements the code change:
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1863798

Bug: 1014880
Change-Id: I79841e847a78f1b965f56e7b9616501b85966483
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1864771
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Commit-Queue: Victor-Gabriel Savu <vsavu@google.com>
Cr-Commit-Position: refs/heads/master@{#707275}