Adapt isolated-code-cache web tests to partitioned HTTP Cache
The not-site-per-process isolated-code-cache web tests currently assume
that cross-origin resources are cached when code cache site isolation is
not active. This causes them to fail when the HTTP Cache is partitioned
by Network Isolation Key. We thus disable the partitioned HTTP Cache for
those tests and add a new test for when site isolation is not active,
but the HTTP Cache is still partitioned.

Bug: 1009628
Change-Id: I0d5d438d9564d3f659b93484a5bcc8aa56d2cd26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1865421
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#707360}