Add the target url as a bundle property for Android links and images.
Any AccessibilityNodeInfo can include extra information in a Bundle of
key/value pairs. Some external developers requested that we expose the
target url for a link or image, so we can just include it in the bundle
of extras without needing a formal API for it.

Bug: 961865
Change-Id: I0edb75bc8320a5ee3bf166072a0a061e46786600
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1862862
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/master@{#707415}