Fix DCHECK in AXPlatformNodeTextRangeProviderWin::NormalizeTextRange
With Narrator and UIA enabled if the user navigates to the end of a
document, then the document is modified such that Narrator now has a
range whos |start_| or |end_| position is at the end or start of the
document respectively, |NormalizeTextRange| would have created a
NULL_POSITION for that range edge.

To avoid this, if there is no next/previous character for that edge
of the range, then that edge position will not be modified.

Bug: 1015593
Change-Id: If706057eaf533dfd0ab5e544eefd102d95ca767e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1869716
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Commit-Queue: Adam Ettenberger <adettenb@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#707573}