[JJI] Deflake two Js Java interaction tests
testJsAddAndRemoveEventListener and
testJsReplyProxyReplyToTheCorrectJsObject are flaky due to execute
JavaScript call and replying port are not in the same mojo channel, so
the sequence is not guaranteed, therefore there is a race between them.

These tests are following the same pattern:

// JavaScript posts a message to the app.

JsReplyProxy proxy = message.replyProxy;
proxy.postMessage(message);

// When the web page receives the message, it saves the message as
// JavaScript global variable say "JsVariable".

messageReceived = executeJavaScript("JsVariable");
assertEqual(expectedResult, messageReceived);

// Due to the race, when executeJavaScript the JsVariable may not be
// set yet, so the assert may fail.

This CL makes the replying channel to be associated with the legacy IPC
channel so the calling order reflects the execution order.

Bug: 1015023
Change-Id: I336d816d73096a3ec2c5cd35eaa205808359afa0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1865594
Reviewed-by: Changwan Ryu <changwan@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Commit-Queue: Shimi Zhang <ctzsm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#707620}