Refactor preconnect to its own WCO
This CL removes the preconnect logic from NavigationPredictor to
NavigationPredictorPreconnectClient. This will allow us to move from the
OnLoad event to a policy around the DidFinishNavigation event. The
rationale is that >10% of pages never reach OnLoad, so we may have a
better opportunity to preconnect from some delay after
DidFinishNavigation.

Change-Id: I7b8ee571acdc6a9c748c77ac9a7686e5b9c53b33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1861106
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Commit-Queue: Ryan Sturm <ryansturm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#707656}