Tighten the applicable condition of clickable property for ARC++ nodes
When android clickable nodes are nested, currently only the outer nodes
are clickable. In this case, even if there are multiple clickable inner
nodes, only the outer node was clickable.

This CL changes to set clickable property only when the node that is
clickable in Android and there is no clickable descendant node of it.

menu is accessible from ChromeVox.)

Bug: b:141969848
Test: unit_tests --gtest_filter="AXTreeSourceArcTest.*"
Test: manually (open PlayStore and confirm that search box and hamburger
Change-Id: I6f62658f8d620d650678162d3547d86af66efa75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1849548
Commit-Queue: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Sara Kato <sarakato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#707744}