[Autofill Assistant] Sending onboarding flag to server.
Before this change the server had no knowledge of whether or not the
onboarding was shown to the user in this run.
To be able to log this, a new flag is now sent to the server in the
requests for getting scripts.

Adding unit tests to check new behaviour.

Bug: b/142846407
Change-Id: Ic77bbc70c8add122beaff80a47e3cf7dca4454d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1866631
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#707763}