Rearrange swarming debug doc and add a concrete example.
- Move run-swarmed and mb.py to the top of the doc

This is a great doc but after working through a lot of steps, I found
that I could have just run mb.py. So this adds a "The easy way"
section at the top of the doc and moves the combined tools to that.

- Factor out the authentication section so that it's linkable.

- Add a concrete example and advice on how to construct $criteria.

Change-Id: I0e7dcbcf9ce38aeed08a614e35f3f1374d2fa8c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1871158
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Fergal Daly <fergal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#707793}