Enable CannotDecryptScryptKeyEncryptedDataWhenScryptDisabled test
The test was disabled (crrev.com/c/1849426) because it instantiates
FeatureLists after the browser threads have been created.

It was needed to inject NigoriSpecifics with scrypt key derivation
method, while test assumes that scrypt is disabled. The fix is just
initialization of NigoriSpecifics in test fixture constructor, where
it's safe to override features.

The test also used to inject an encrypted bookmark to the server, but
this part wasn't exercised further in test, so this CL just drop it.

Bug: 1013073
Change-Id: I8d2947484a0192cf99fd2b84e44325ea1d6220fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1872190
Commit-Queue: Maksim Moskvitin <mmoskvitin@google.com>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#707801}