webview: Move context wrapper to embedder_support.
Move ResourcesContextWrapperFactory to the embedder_support component so
that weblayer can share it. In the process, rename it to
ClassLoaderContextWrapperFactory, which better reflects what it actually
does: it has little to do with resources (which are always just
delegated directly to the context it is wrapping), and the primary thing
it overrides is getClassLoader().

Downstream code still refers to the old class, so leave a forwarding
version in place for now.

Change-Id: I3c22c390012289d7c91aeb6986f1bf44fd34dda1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1867230
Auto-Submit: Richard Coles <torne@chromium.org>
Reviewed-by: Tobias Sargeant <tobiasjs@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#707807}