[TouchToFill][Android] Show favicons for credentials with opaque origins
This CL addresses cases when no favicon is shown for some credentials.
This may happen when the credential has been saved in a different app,
as in this case credential origin is an opaque URL for which no favicon
can be obtained. To work around this, this CL adds plumbing to pass the
URL of the page for which TouchToFill has been triggered. If the
credential origin is opaque, TouchToFillController requests a favicon
for the frame origin.

Bug: 1012646, 1016486
Change-Id: I6d803734460279b3f079db77d05992c39a43ac02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1865338
Commit-Queue: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#707994}