Remove WebVR switches/UI Flags, Mark as "test" feature
This is the first step in removing WebVR. This change removes WebVR
from about://flags and also removes the separate command-line switch
that was used. In order to keep the tests running for the time being,
this change modifies the blink webvr feature to be set to "test." The
browser tests for WebVr have been updated to set the new flags that will
turn on WebVr.

This also removes the Origin Trial flag for WebVr, and as such deletes
the origin trial layout tests as well.

Bug: 960132
Change-Id: Ia625e3ae9986183ef2746dcb69e595cab49fd981
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1862855
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Klaus Weidner <klausw@chromium.org>
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#708005}