Make cancelling a webshare request more reliable.
When triggering a share intent chooser, use startActivityForResult as it
allows Chrome to receive a callback when the activity finishes and can
propagate the cancel much sooner.

This required swapping out Activity for WindowAndroid in the share API,
so it is largely a mechanical swap in the various places triggering share.

BUG=636274

Change-Id: Id3628e65757ca54713a57eefcd47efec24fc65a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1832589
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#708050}