Fix phone number regex whitespace matching.
We only handled some whitespace characters by using \s in the regex. One
example that is not handled by that is the non breaking space which is
quite common in HTML formatted text. Switching to \p{Z} to consider all
unicode characters for whitespace fixes this.

Bug: 1015189
Change-Id: Id3e6b3fe98a303e84d1b8427a227bc12a42d6664
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1872089
Commit-Queue: Alex Chau <alexchau@chromium.org>
Reviewed-by: Alex Chau <alexchau@chromium.org>
Cr-Commit-Position: refs/heads/master@{#708142}