Reland "Kill quiet permission requests if a new request is made"
This is a reland of 55a4bb68c86720dfb87d7116f2da2a6e4a2e3459

The original CL caused a crash (crbug.com/1016233)

The first patchset is the original CL, unmodified.

https://chromium-review.googlesource.com/c/chromium/src/+/1860015
introduced a crash because InfoBarService::RemoveInfoBar DCHECKS that
the infobar is part of the owned list of infobars. When the tab is
closed, the infobar is removed but then the PermissionPromptAndroid
object destructor attempts to remove it again. This CL ensures that
that this situation does not happen by listening for the infobar
remove event and ensuring we don't try to remove the infobar that
was already removed.

Original change's description:
> Kill quiet permission requests if a new request is made
>
> Bug: 1014026
> Change-Id: I5fe93b9bc7b39873900d9e15dcba66dfddd7c4ef
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1860015
> Commit-Queue: Andy Paicu <andypaicu@chromium.org>
> Reviewed-by: Balazs Engedy <engedy@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#706445}

Bug: 1014026, 1016233
Change-Id: Iaa857811c6617afd97af9708a51ea49ea98fe385
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1871698
Commit-Queue: Andy Paicu <andypaicu@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#708160}