Prepare ShortcutsProviderExtensionTest for imminent switching on of the HistoryServiceUsesTaskScheduler experiment.
The feature will soon be flipped to ENABLED_BY_DEFAULT. This will make HistoryServiceBackend's task runner
dispatch to the ThreadPool instead of a dedicated base::Thread. The tests need to be
prepared to stay functional when that happens.

This is done as the last mile of the chrome-wide effort to coalesce as much work as possible
to the ThreadPool.

Bug: 661143
Change-Id: Ic17ce4468b3de601763e1ebe5c36f1120313fb57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1873354
Reviewed-by: Kevin Bailey <krb@chromium.org>
Commit-Queue: Oliver Li <olivierli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#708224}