[Screen Capture] Update heuristic to determine system permission in Catalina
The old heuristic checked that the title of all windows was readable, but
it failed in some corner cases where an application creates a window whose
title cannot be read even with screen capture permissions.
The new heuristic checks that the title of at least one normal, dock or status
window running on another process is readable, which is less susceptible to
failures.

Drive-by: return SYSTEM_PERMISSION_DENIED instead of PERMISSION_DENIED when
the system permission check fails in getDisplayMedia().

Bug: 993692
Change-Id: Ie127d1e2d6bd29dd8885b0c461bdab838559fcb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1871885
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#708260}