[Switch Access] Major refactor of the navigation system
This change is an overhaul of how navigation between elements onscreen
is handled. The design doc can be found at go/cros-switch-navigation.

TL;DR each time we enter or exit a group, we construct an SARootNode
for the group and an SAChildNode for each of its children. This is
instead of dynamically calculating the elements of the group on each
navigation event.

Bug: 1011630
Change-Id: I264c1c3c9ef36fcb0d98327cd2e5fcc671b7f9df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1843214
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Cr-Commit-Position: refs/heads/master@{#708301}