Add ARC adb sideload option to Crostini settings
The adb sideloading setting is per device, and requires a reboot to
initiate the flow, such that the user can securely confirm the intention
while the device is in a clean state.

To opt out, the user needs to go through powerwash flow in order to
restore the device back to fully verified state.

Only the device owner can change this device setting, and the setting
applies to all users on the same device. On managed device, the feature
is currently disabled.

      entry does not show up

Test: Launch chrome without --enable-features=ArcAdbSideloading, new settings
Test: Launch chrome with --enable-features=ArcAdbSideloading
Test: Only when ARC++ is enabled in pref, the new setting sub menu shows
Test: Go through the enabling flow, and see the value in boot lockbox changed.
Test: Then, go to settings to disable the feature. Powerwash flow is initiated
Test: Non-device owner see the toggle disabled with a policy indicator
Test: On managed device, user see the toggle disabled with a policy indicator
Bug: chromium:893332
Change-Id: I2a709a9ae451d88b4bda87dead45772f63d03f3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1834896
Commit-Queue: Victor Hsieh <victorhsieh@chromium.org>
Auto-Submit: Victor Hsieh <victorhsieh@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Alexander Alekseev <alemate@chromium.org>
Reviewed-by: Michael Giuffrida <michaelpg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#708328}