Move URLRequestDataJob::BuildResponse logic to DataURL::BuildResponse.
Also move all tests of the method as well, and merge the fuzzers of
URLRequestDataJob and DataURL::Parser.

This is in preparation to remove URLRequestDataJob, which is only used
in tests (And those tests are only still around to test
URLRequestDataJob::BuildResponse()).

This test also adds a couple test cases to DataURL unit tests, since
that doesn't really seem worth a seprate CL, but noticed some cases
could be covered a bit better as I was moving stuff.

Bug: 934009
Change-Id: I4025e77e0bde27a271ed32c0d274b0e24604e842
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1869539
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Eric Roman <eroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#708410}