Control software input panel through TSF
This change adds mechanism to control the software input panel on
Windows using TSF input panel policy. It triggers the input panel
through a policy that is set in TSF status flag based on whether
user has interacted with the page or not. It uses the existing
TextInput state object to get that information. This is also a step
towards removing the TryShow/TryHide API calls which uses a legacy
IPC message and also don't work properly on some Windows touch
devices such as Surface Book. TryShow/TryHide input pane API calls
cannot co-exist with TSF input panel policy as it leads to a race
condition in input service that decides when to raise the virtual
keyboard which sometimes results in flickering of input panels.

Test: ui\base\ime\win\tsf_input_policy_unittest.cc

Bug: 1014365

Change-Id: I598a15bd16a90f578fbfc62fd06ef0c5724b243d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1862860
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Yohei Yukawa <yukawa@chromium.org>
Commit-Queue: Anupam Snigdha <snianu@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#708420}