Rename SiteInstanceImpl's HasWrongProcessForURL.
The HasWrongProcessForURL method on SiteInstanceImpl is a bit awkward to
read and returns negative of what we usually want to ask. The real
question this method answers is whether the input URL parameter is one
that can be hosted in the SiteInstance. As such a better name for this
would be IsSuitableForURL.

TBR=halliwell@chromium.org

Bug: 1015882
Change-Id: I4957b6fa62e1a5d81b3b6b67b64a45c20e452ca1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1869810
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#708599}