desktop-pwas: Get install mode from user-data dir
This change adds function GetUserDataSuffix(), which returns the install
suffix (e.g. "" for the primary install mode, " SxS", " Beta", or
" Dev") given a path in the User Data folder.

It is based on GetInstallSuffix(), which does the same given a path in
the Application folder.

This new function is the first part of finding the location of
chrome.exe from within the User Data directory, where Progressive Web
App (PWA) launchers are planned to be. The next step will be a function
that uses the install mode from FindInstallMode(GetUserDataSuffix()) to
get the full path of chrome.exe, which PWAs can then use to launch.

Change-Id: If69859fc5f36b7d2d2c4262a1955bd432d983597
Bug: 960245
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1866977
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Reviewed-by: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#708663}