Delay 60 seconds for the GPU watchdog kill to collect more data
Once GPU hang is detected in the GPU watchdog V2, the watchdog will try
to load/unload a file from the hard drive and create/destroy a thread
to see if the GPU has made any progress in the current task. This will
give a hint if the system is busy. If the GPU is still stuck, Wait until
60 seconds has passed and check again the GPU status. After this, the
watchdog kill the GPU main thread.

This is to see whether the GPU thread is killed too early and to
compare with V2 on the number of hangs. This piece of code can be deleted
after the data is collected.

Bug:949839

Change-Id: I3a6afeed21776553d7416682baf6eed2d521286c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1872166
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Maggie Chen <magchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#708758}