Add metal_util component
As part of the investigation of the hangs in the Metal shader compiler,
we will be running some Metal code in the browser process. Create a
stand-alone component for Metal utility functions that will be used
in the browser process and also in the gpu process.

For the moment this functionality includes the C++ Metal types and
the function to select the default Metal device (low-power if
available). Functionality to compile a test shader to determine if
it hangs will also be moved there (and called from the browser
process).

Bug: 974219
Change-Id: Ic6a696686e0c438ec339c6cc40b8200be6b0a689
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1869127
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Eric Karl <ericrk@chromium.org>
Reviewed-by: Cait Phillips <caitkp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#708798}