Split virtual test suites with different args
Virtual test suites under the same prefix with different args are
confusing and error-prone. As we'll combine virtual test suite entries
with the same prefix/arg into one suite in crrev.com/c/1873504, this
CL prepares for that. See that CL for the future format of
VirtualTestSuites.

Bug: 1014162
Change-Id: I707c67f7a5be925fde7fb252b5f7ceec50edb6f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1876740
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#709077}