Add insecure-broken security state to the DevTools protocol
Security UX is experimenting with using a grey triangle warning for
non-secure site connections (crbug.com/997972). When this icon is
shown, DevTools Security panel's security overview should match the
state shown in the omnibox. This cl sets up this change by adding an
insecure-broken state, so that the following states can be distinguished:
- neutral (info icon)
- insecure HTTP
- insecure broken HTTPS

This change will also make the security state naming more intuitive, since
insecure HTTP can be marked as 'insecure' rather than 'neutral' as part
of a follow-up cl.

Temporarily disables a web test since fixing it requires a frontend
change, which is in a separate repo. This will be re-enabled in a
follow-up.

Bug: 1008218
Change-Id: I628f2fe1168dfeead7874e36b210bcc0396b6e8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1874787
Commit-Queue: Livvie Lin <livvielin@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#709095}