Add a clang-tidy settings file for all of Chromium
We want to start running tidy as a part of Tricium. In order to do so,
we need to figure out the set of clang-tidy checks we want to run.

The checks here are sourced from the 'Y' column of:
https://docs.google.com/spreadsheets/d/16_z3pQFvTEzQocfWpch54GSykziTxVJMEcaDb5xW0YA/comment#gid=0

Ignoring google-explicit-constructor and modernize-use-auto (since the
former has code that relies on it, and the latter requires changes to
clang-tidy to meet Chromium's style guidelines).

We can add more (bugprone-*, ...) as time goes on, but those appear to
be the defaults that Chromium wants initially.

Please note that subdirectories can override all of this whenever they
want (e.g., v8 has a .clang-tidy file already, so any files inside of
v8/ will entirely obey that), so this is meant to be a generally
valuable and applicable set of checks.

Teaching `gn` about all of this might have some value eventually, but
for now, the goal is to see how far we can get with the simpler
solution.

Bug: 633856
Change-Id: I7b52e6a313ace8a91e9041313b222da737ef47fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1808532
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Cr-Commit-Position: refs/heads/master@{#709137}