Ash Tray: Remove shadows from system tray bubbles.
As part of the new UI redesign fo the system tray, we no longer render shadows
around the tray bubbles. This includes the quick settings, sliders, and
autoclick bubbles. Legacy bubbles (e.g. palette) are not affected.

Note that we need to add a new border to the message center bubble, which will
be done in a follow-up CL.

TEST=manually verified all bubbles
BUG=976952

Change-Id: I5e52b272262064044f52c79ababd14725111a735
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1877137
Commit-Queue: Tim Song <tengs@chromium.org>
Reviewed-by: Ahmed Mehfooz <amehfooz@chromium.org>
Reviewed-by: Tetsui Ohkubo <tetsui@chromium.org>
Reviewed-by: Michael Wasserman <msw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#709286}