Settings a11y: Provide link rows' icons with role description.
This is step one of the two part task of ensuring that settings subpage
buttons in cr-link-rows announce that they are subpage buttons. As is,
they are called 'link's [1], which almost always implies opening a new
tab in both browser and OS settings.

This change allows elements containing a cr-link-row to provide a
string to be passed down to the link row's nested cr-icon-button as
the button's role description.

As a practical test, this change also migrates the OS settings' main
page's 'Languages and input' row [2] to use a cr-link-row [3] with the
new role description "subpage button" injected [4].

----

Note: I initially missed an error resulting in a regression that caused
the act of routing to and then back from the Languages and input subpage
via this row to send DOM and a11y focus to the full page instead of back
the its previous location on the subpage button. This was fixed in
Patchset 7 by updating line 33 of os_languages_section to reflect the
change to the DOM caused by this CL. See [5] for a demonstration of the
correct routing behavior once Patchset 7 was implemented.

[1] https://drive.google.com/file/d/1tBGDSoeaRBRO3nJ4h7IMPtiiGMDGB10L/view?usp=sharing
[2] https://screenshot.googleplex.com/dsGjkF0e93c
[3] The migration to cr-link-row was planned anyway as part of
crbug.com/1012565 so it was a good trial case for the task of using the
"subpage button" language on all settings 'cr-link-row's, i.e.
crbug.com/1012104.
[4] https://drive.google.com/file/d/1Lsil-tybwB_cwSLN0TyabfCq_DFNpUSo/view?usp=sharing
[5] https://drive.google.com/file/d/13VYF-XPAryQYfI1paU8vi9KBC3J5fvhM/view?usp=sharing

Bug: 1012103, 1012104, 1012565
Change-Id: I7ac02205bbdd8a352fdbc7e920c1729acf634ba4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1846397
Commit-Queue: Jordy Greenblatt <jordynass@chromium.org>
Reviewed-by: Hector Carmona <hcarmona@chromium.org>
Cr-Commit-Position: refs/heads/master@{#709293}