[Hashed logging] Refactor API and internal logic.
This is the first change to convert the existing hashed logging code
into a general framework. This CL mostly removes code that was specific
to the search ranking use of hashed logging, and also changes the
API for logging.

Note that all hashed logging has been disabled for this refactor.

We need to delete this code now because there's a chain of
dependencies between the refactoring CLs. Roughly:

 1. Code referencing the current proto used to report logs must be
    deleted.

 2. The proto itself can then be changed internally, and re-exported
    to chromium. This has to be done as a unitary CL, with no other
    changes.

 3. Then, code can be re-added that references the new version of the
    proto.

Detailed changes:

 1. AppListLaunchRecorder::Record has become ::Log, which now takes
    lists of hashed and unhashed data. This is a tentative first try
    at the new API.

 2. The core hashed logic in AppListLaunchMetricsProvider has been
    deleted, because it will need to be generalized.

 3. The unit tests have all been disabled and the core EXPECT logic
    removed. This too must be rewritten for the generalized version.

Bug: 1016655
Change-Id: Iccc98771003c602b1a8375b4b2898e077f28def6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1874411
Reviewed-by: Thanh Nguyen <thanhdng@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#709329}