Create a new class which is dedicated to private property keys and add a new API using the class
Currently the API |GetSymbol| in V8PrivateProperty is used to get Symbol using HashMap with a static variable address as a key.
However, this key is not suitable for tracking private properties because the type is void*.

We can improve the tracking ability by using an instance of private property key dedicated class instead of void* type.

This CL creates a new class which is dedicated to private property keys, and adds a new API to get Symbol using an instance of the class.

Bug: 715418
Change-Id: Ieac862e8e2b5406a2c35c1a1f3368494833c3287
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1877490
Commit-Queue: Marina Sakai <marinasakai@google.com>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Hitoshi Yoshida <peria@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#709335}