AW UMA: integration tests for WebView metrics
This adds a test suite for WebView's metrics integration. This verifies
WebView correctly configures the MetricsService such that it logs values
for proto fields in the UMA record. Most of these tests won't validate
the values make sense (verifying this logic should be in unittests for
each MetricsProvider), but rather that the field itself is populated (as
the status quo is many fields are simply not populated).

This configures the MetricsService to upload the initial log record very
quickly, as an optimization in the test case (the current delay of 20sec
would not be acceptable for each test). This does not modify the interval
between subsequent uploads, as each test case runs in a separate
process, so we'll never reach subsequent uploads.

Bug: 1011991
Test: run_webview_instrumentation_test_apk --gtest_filter=AwMetricsMetadataTest.*
Change-Id: If8b0400ec9904258ba9ae653e98b9c6d53a36d1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1854507
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Tao Bai <michaelbai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#709337}