Rewrite addressspace-serviceworker-basic test
The previous version of this test had some issues and were the
source of flakiness:
* It used non-secure contexts and register() seemed not working.
* It used a mock object for service_worker_unregister_and_register()
  and the mock object silently consumed rejections of register().
* It didn't wait for iframe loading completion.
* It didn't clean up iframes.
* It didn't handle error cases correctly and the test passed
  in an unexpected way.

This CL rewrites the test to fix above issues:
* Use secure contexts.
* Use existing test helpers which provide better error handlings.

As a consequence we stop using addressspace-test.js in the test
but I think the new test is easier to read and understand.

This CL should de-flake the test.

Bug: 988074
Change-Id: I3d16c61742d819c08cadcff20a017b7336508efd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1877876
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#709384}