Add back the ShowInSearch checking.
In CL:1847698, the ShowInSearchChecking is removed because the
recommendable Internal/Built-in apps should be involved in the search
result. However, it causes the issue that more apps are list in search
result for the guest user.

Actually, only the two suggested apps should be search out for the
recommend searching result, other apps, if should not shown in search,
should still be filled out. So the ShowInSearch checking is added back
and add !IsSuggestionChip(update.AppId()) checking.


BUG=1017631

Change-Id: I1397e11ba5b8162e0d2986705b990fb054855c3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1880493
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#709502}