Local NTP, Realbox: Stops tracking the match element being deleted
Currently we track the match element being deleted. This is in order to
detect focusout events on the realbox wrapper as a result of matches being
replaced by new ones after deletion. This CL avoids that by stopping to
listen for 'focusout' just before the matches are replaced and starting to
listen for it immediately after.

Bug: 243926
Change-Id: I8e37fdc7d6a1f52f052de6ef3ec711ed7bc927bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1879660
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Dan Beam <dbeam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#709701}