Document that gfx::ColorSpace assumes an 8-bit range expansion.
For 10-bit and 12-bit formats this results in a relative error of ~2.9%
and ~3.7% respectively. As discussed offline, it's not worth adding a
bit depth to gfx::ColorSpace at this time for this.

Potentially at some point in the future someone will create the
equivalent of the full vs limited range images we have for 8-bit:
https://referencehometheater.com/2014/commentary/rgb-full-vs-limited/

At which point we may need to update our implementation to be more
accurate.

R=ccameron

Change-Id: I0cdc6a9023ec74b53da1620f84f64b531c323634
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1879720
Reviewed-by: ccameron <ccameron@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#709711}