All a11y methods in RenderFrameHost should check if the host is active before proceeding
When a RenderFrameHost is being swapped out, it will be in the process of being deleted.
Before any requests from accessibility should be serviced, we should
check if the RenderFrameHost is not being swapped out.
Otherwise, parts of the class instance, such as its RenderViewHost,
might have been destroyed, or the renderer associated with
the RenderFrameHost instance might not be able to
receive IPCs.

R=dmazzoni@chromium.org

Change-Id: I08a87cad7b789f0752adbcc1564b748456d5d338
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1881324
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#709917}