GMC: Fix separator lines on Mac
This CL fixes the separator line color on Mac by removing the
transparency of the MediaNotificationListView. The 12% opacity lines
against the transparent list view was part of the issue. The other part
is crbug.com/1012024, and while this change makes the lines work
properly on Mac, it does not fix crbug.com/1012024.

This change also switches from using a hard-coded SkColor value for
the line to using the MenuSeparatorColor from the NativeTheme.

Bug: 1016717
Change-Id: I60e132e6019722e59bdc0247fc19eedd9ae3d8d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1880570
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Commit-Queue: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#709947}