Ensure all safety tip heuristics always run.
This CL ensures that all safety tip heuristics are always run, in lieu
of the previous short-circuiting behavior, in preparation for recording
metrics. This CL also refactors the reputation check results out into a
wrapper struct, to cut down on the number of variables that need to be
passed down the call chain. Finally, this CL fixes a few miscellaneous
typos.

Bug: 1014598
Change-Id: Ibeda24f817f46d8d36f23fab60f30615821e544f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1881126
Commit-Queue: Aidan Beggs <beggs@google.com>
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#709953}