[Tab Groups] Remove manual inset calculations for the color picker.
The fixed insets makes it crash on the DCHECK with certain sizes.
Updating this to be managed by layout manager fixes this issue while
developing and fiddling with the final sizes. The insets might need to
be updated later on for the first and last elements.

Bug: 1015634
Change-Id: I947c1e6b01013fd353602e8e96739a21b5f4e7b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1884351
Commit-Queue: Charlene Yan <cyan@chromium.org>
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710087}