Componentize SSLErrorNavigationThrottle for reuse by WebLayer
This CL componentizes ssl_error_navigation_throttle.* in preparation
for using it within the implementation of SSL interstitials for
WebLayer. The only meaningful change is to move the concrete detection
of being within hosted apps to be implementation within
chrome_content_browser_client.cc and passed to
SSLErrorNavigationThrottle as a callback.

This CL leaves out the following, which we will undertake as followup:
- Componentization of the unittest. This requires extra work, notably
  the componentization of SSLBlockingPage
  (which we will be undertaking).
- Putting the moved files into the security_interstitials into their
  own namespace. This would just serve to add noise to this CL.

The (internal-only) design doc for this effort is here:
https://docs.google.com/document/d/1ab_zZ6TF6tMZE54IJ2dz8LSwsIp6XEwky9yzH7WPFqk/edit#

Code authored by blundell@chromium.org

Change-Id: I2e410f733cbdd7ed41bdbac0318f9dfe8d9c17a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1880032
Commit-Queue: Evan Stade <estade@chromium.org>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710171}