Do not output repeated tooltips
In Blink, the title attribute usually results in the accessibility node getting
a name or description with the title text. In addition, it always populates the
tooltip attribute.

Recently, the tooltip attribute was added to Output.hint_, but no suppression
was made when it is the same as either name or description. Make that so in this
change.

TBR=dcheng@chromium.org

Change-Id: I67d21f0206ce3cb17b691915db0b4dc5d127ff51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1872741
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Commit-Queue: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710223}