Added punctuation breaks during word movement
Ctrl+Break on Windows was busted. The algorithm
in |NextWordPositionInternal| never considered
punctuation as a word break character.
According to the Unicode spec for word break boundaries,
punctuation should be considered as word breaking characters.
Added this additional condition in both previous
and next word movement.
Also changed many spellchecking tests to match this behavior.
Insert a space after the misspelled word so it is spellchecked properly.
This is done to trigger spellchecking as the punctuation after the
misspelled word doesn't trigger spellchecking until the user has pressed
space bar or changed the selection.
This is how it should be done in real sites anyways so inserting the
space after misspelled word to trigger spellcheck  makes sense.

Test: third_party\blink\renderer\core\editing\visible_units_word_test.cc

Bug: 122304, 1004414

Change-Id: I88694d3cd2e275f06e3939e7e42181de0bd1b9c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1880602
Commit-Queue: Anupam Snigdha <snianu@microsoft.com>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710243}