Make V8PrivateProperty::SymbolKey class is_trivially_constructible
A class which is not trivially constructible might bloat out code size.

The member variable |desc_| of class V8PrivateProperty::SymbolKey is used for just a description of a private property, so we can remove the variable.

This CL guarantees that V8PrivateProperty::SymbolKey is is_trivially_constructible.

Bug: 715418
Change-Id: I7f44e4e70430b66275b001c99974326556d38477
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1880778
Commit-Queue: Marina Sakai <marinasakai@google.com>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Hitoshi Yoshida <peria@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710246}