Follow style information when using Skia font fallback
Retrieving fallback font information on Windows 8.1 and up using Skia's
matchFamilyStyleCharacter() did only pick the family name, but not
instantiate the fallback font from the additional SkFontStyle font style
information that Skia provides as well. This bug was fixed similarly
before for the Windows 8.0 and before fallback code path in issue
1008901.

This fixes an issue in which fallback for U+2758 LIGHT VERTICAL BAR
produces Segoe UI Semilight as fallback font, but the renderer was
unable to instantiate that font correctly and displayed a .notdef glyph
for this character.

Bug: 1018375
Change-Id: I18a37f8f432ff1c392dedbc868b8f882254b12d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1883515
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710252}