Rework SMIL timing update scheduling
This changes from using essentially a linear-scan for the "update
intervals" step to using a priority queue with the next interval time as
the key/priority.

Besides providing a dramatic reduction in the amount of actual interval
update steps being performed, this also shifts the responsibility for
the timing updates away from the sandwich structure and into the time
container. This will in turn allow future simplification of the setup
for animation value application/sandwich management.

To aid the switch to the priority queue scheduling mechanism, the update
of the "next interval time" is sunk out of and now performed as a
separate step after UpdateInterval().
Some code is consolidated around the new "reschedule" operation (the
SMILTimeContainer::Reschedule(...) method) which greatly simplifies some
code.

Bug: 998526
Change-Id: I1e22625eac022155607a3ccc66dd2d21f52cc7ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1876031
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#710265}