[Autofill Assistant] Moved overlay color parsing logic to native.
There are two benefits: First, we want to move all logic from Java to native anyway. Second, we are about to add more colors to the overlay (see bug), which exacerbates the problem.

This is a refactoring and does not contain any user-facing changes.

Bug: b/143452916
Change-Id: Ic418ce6993c1cbfd60746fa6df7f9ec493afdba3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1881560
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710319}