[iOS] Plumb Browser into SigninInteractionControllers
This is the fourth CL of several breaking crrev.com/c/1860025 into smaller
chunks.

This updates SigninInteractionController and Coordinator to accept a Browser
instead of a BrowserState. This necessitates updating a few other classes
(AccountsTableViewController, GoogleServicesSettingsCoordinator), and, more
significantly, SettingsNavigationController.

With this CL, SettingsNavigationController is initialized with a Browser
instead of a BrowserState, so it can initialize AccountsTableViewController
instances. Unfortunately, there is one init path for SettingsNavController
that can't yet be updated to use Browser -- the import data controller class
init is called from inside AuthenticationFlowPerformer, which is further
down in the stack of classes that Browser is being plumbed into.

As a temporary measure, internally SettingsNavigationController keeps an
initializer that accepts a BrowserState; calling it (which is only done from
the import data class init) will create a controller with a nullptr Browser
property. If this controller attempts to create an AccountsTableVC, it will
DCHECK(). This is rectified in crrev.com/c/1883724

Bug: 1018746
Change-Id: I58e82b6f2cd501d2ac90d48b4545b0eb6c836fb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1883519
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710324}