[session restore] Prevent restored windows from being completely white.
Since https://crrev.com/701411, restoring a window that was "last active"
and "minimized" when it was closed opens a window that is completely
white. This is because the window is activated but never shown.

This CL ensures that the window that gets activated when restoring a
session is also shown.

Bug: 1018885
Change-Id: I9d01bfb2cafa83df9744bf5c598570ff55252de0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1884952
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710397}