Portals: Factor out common test set up
Many of the portal browser tests have the same steps to run at the
beginning of the test: inject script to create a portal and to navigate
it to a given URL, and then wait for the navigation to finish. We factor
this out into a common function.

We also perform a few small cleanups. We rename the test
ReactivatePredecessor to AdoptPredecessor, since it doesn't actually
reactivate. We increase usage of JsReplace instead of StringPrintf.

Bug: None
Change-Id: Ia1ef41b8138df0189ad3291b76c34be53c291403
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1885171
Reviewed-by: Adithya Srinivasan <adithyas@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710402}