Fix tab hover card animation delay when moving quickly across tab strip.
When an animation is triggered while another is already in progress the animation was being stopped then restarted. This led to a 1 frame delay in the animation that was very noticeable when quickly scrubbing the tab strip. This change makes it so that the animation is never stopped and continues from where the interrupted animation left off. A linear animation also seems more fitting for this case since we are never reversing the animation.

Bug: 1017134, 1017029
Change-Id: I945f5c180c8758cc5e14d7d45ed792bc7c11ac89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1881949
Reviewed-by: Dana Fried <dfried@chromium.org>
Reviewed-by: Ian Vollick <vollick@chromium.org>
Commit-Queue: Caroline Rising <corising@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710420}