Do not report 0 LargestImagePaint and LargestTextPaint to UMA/UKM
This CL is a revert of the functional parts of commit ef51466. We
have decided that these should not be reported as they skew the values
in the histograms in undesirable ways. These 0 values correspond to
times during which the largest image is not yet fully loaded.

Bug: 1017335
Change-Id: I6bbae41f0e0357198f752291c50a439003cd9bc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1875285
Reviewed-by: Bryan McQuade <bmcquade@chromium.org>
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710544}