[sensors] Unship "ambient-light-sensor" Feature Policy
The AmbientLightSensor class has not been shipped and is still behind
the "SensorExtraClasses" flag. In feature_policy_features.json5 however
this policy was only behind the "Sensor" flag, which was marked stable
as part of shipping the motion and orientation subset of the Generic
Sensor family of APIs.

This change fixes the dependency, effectively unshipping this policy.

Change-Id: Idce46684545807cce6134021c357051520e6a713
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1762690
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710564}