[scheduler] Compositor priority experiments shouldn't overwrite use case based prioritization.
Previously, these experiments would not change the compositor priority if it was
set to highest or higher in one of the use cases. This precluded the compositor
gesture use case in which the priority would be set to low as a proxy for prioritizing
loading. We suspect this caused regressions in jankiness metrics during the experiment
as the low priority could be changed to very high or normal. This cl changes it so that
we will only overwrite compositor priority if it has not been set because of a use case.

Bug: 966177
Change-Id: If3ecfa898249f3f9ce52b1afbcfc7320dbbd96b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1877814
Commit-Queue: Katie Dillon <kdillon@chromium.org>
Reviewed-by: Scott Haseley <shaseley@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710586}