rendering benchmark: Run canvas/webgl tests for longer.
Instead of running the webgl/canvas animations for 5 seconds, run
them a little longer: 10 seconds. The web-animation tests are also
currently run for 10 seconds. So this change aligns these animations.

There are ~35 canvas stories, and ~15 webgl stories. So overall, this
would increase the runtime, if all tests were to run on the same bot,
by ~4 minutes (= 50 * 5seconds). If the stories are sharded, then the
increased runtime should not be very noticeable.

BUG=894287

Change-Id: I62e341cfd3b605ab4090af2fc185fa1648324459
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1882760
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710605}