Find in page: don't report all 0 match results as final
This was added in r422176 as a bug fix, but the original author doesn't
recall specifically what it fixed, speculating that it might be related to
a bug that caused searches not to terminate in some cases.

This interacts poorly with last year's refactor to use idle tasks for find,
causing a great deal of "final" updates per frame. I think it's likely that
the original bug was addressed by the same refactor, but if not, removing this
code will smoke it out so we can document *why* we need it.

Bug: 1015818
Change-Id: I8234aa3788b24ce54ae75120c8d1e00c6ed31d3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1888090
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: Leonard Grey <lgrey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710767}