bindings: Add ephemeron handling to heap snapshots
Up until now, heap snapshot could not handle ephemerons. Since heap snapshot
is always called after a GC, we can treat ephemerons as strong hash tables.
The visitor interface is extended with a method that get a KeyValuePair. This
is needed so that heap snapshot can determine the correct parent for the
value. The default visitor implementation ignores the key and only traces the
value. The DevTools visitor sets the key as the parent and then traces the
value. Since DOM worlds are ephemeron-based since r705062 this allows users to
debug those using DevTools as well.

Drive-by: remove redundant trace in Node::Trace.

Bug: 986235, 993415
Change-Id: Ifaf453b2d7c2883a929577825991a11fbb0cf835
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1834229
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710775}