[DeviceSync v2] Add CryptAuthDeviceNotifier class
Handles the BatchNotifyGroupDevices portion of the CryptAuth v2
DeviceSync protocol, which sends a GCM message--via CryptAuth--to
a subset of devices in the "DeviceSync:BetterTogether" group.

The implementation queues requests made via NotifyDevices() and
processes the requests sequentially. The implementation also handles
timeouts internally, so a callback passed to NotifyDevices() is always
guaranteed to be invoked.

CryptAuthDeviceNotifier::NotifyDevices() will replace the CryptAuth v1
method SoftwareFeatureManager::FindEligibleDevices(), which was used to
notify a phone that it was selected as the multi-device host.

Bug: 951969
Change-Id: I155e35847f5d6d9d99d8cf4477e70819d82d8e21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1888734
Commit-Queue: Josh Nohle <nohle@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710817}