Introduce appmenu.CustomViewBinder interface for update menu item
Introduce a new CustomViewBinder interface that allows
AppMenuPropertiesDelegates to handle binding for custom menu items.

This interface is used to provide a custom binder for the update menu
item, which relies on UpdateMenuItemHelper methods/classes that will not
be moved to the appmenu/ package when its modularized.

Bug: 966644
Change-Id: I6c49522775c1f0fca3cd64764a0eba6a5b82f3b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1872980
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#710895}