Avoid capture thrash due to Windows focus flicker on video capture.
Unfortunately, a side-effect of the capture code we're currently using
is that it triggers an instantaneous loss and then regaining of mouse
hover over the current controls; this results in capture starting and
stopping again in a tight loop.

This change allows existing capture to continue until the first time a
result comes in and either (a) the page is done loading [which is
existing behavior] or (b) we've stopped observing the thumbnail. It
prevents a momentary shakeup in whehter the thumbnail is being observed
from locking the browser into the tight loop described above.

We still want to follow up and determine *why* (and on what platforms -
it might just be Windows) starting capture causes mouse hover to be lost;
if we can't fix the signal we may have to debounce it in the hover card
system to prevent unintended visual results.

Bug: 1019303
Change-Id: Iec0cba64398891d589d3cbe6d7abeab0e1e230b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1891142
Commit-Queue: Dana Fried <dfried@chromium.org>
Commit-Queue: Collin Baker <collinbaker@chromium.org>
Reviewed-by: Collin Baker <collinbaker@chromium.org>
Auto-Submit: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711008}