[Telemetry] Provide abridged story sets for rendering and system health.
Note that this will have no impact on what is currently running
on the waterfall because all benchmarks are labeled as
abridged: False right now.

This change is now possible because:

1. Pinpoint was taught to use the --run-full-story-set flag by
https://chromium-review.googlesource.com/c/catapult/+/1874451

2. benchmark_smoke_unittest.py was fixed:
https://chromium-review.googlesource.com/c/chromium/src/+/1874821

Note that system_health_smoke_test.py is fixed as part of this
change.
I tested this fix by running
/run_tests benchmarks.system_health_smoke_test.SystemHealthBenchmarkSmokeTest.system_health.memory_desktop/load:game
and seeing that the stories, a mix of health_check stories
and stories without the health_check tag were all run.

Bug: 965158
Change-Id: I42ca28543155536da62d9b02d9bf31d8eb311ac3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1687783
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711017}