[remoting][android] Add gsf.permission.READ_GSERVICES
The "Send Feedback" button in the app no longer works and according to
the logs it seems we need to add gsf.permission.READ_GSERVICES
permission to the Android manifest.

I can't build and test the official build locally so I'd like to check
this in frst then verify it on the nightly build.

Bug: 1019866
Change-Id: Iac01e9db420d95a8c8a7aa952ad9d2ab1ac9b6e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1891464
Auto-Submit: Yuwei Huang <yuweih@chromium.org>
Reviewed-by: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Yuwei Huang <yuweih@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711028}