[serial] Prevent SerialPort from being GCed while open
This change makes SerialPort an ActiveScriptWrappable so that it can be
kept alive for as long as the port connection is open. This prevents
the streams created by the port from being closed if a page does not
keep a reference to the port itself.

Bug: 1017289
Change-Id: I23b4da0c5b27d16901d2d7eb7fd1c622a8404f07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1888513
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711032}