Set Delay by heuristic, flagged by configuration. Add thorough test.
This CL introduces behavior to delay touches that would be flagged as
palm by heuristic, until inference occurs. This is all flagged by a
configuration flag.

We add a thorough unittest to test behavior: that delay is set, and
kept, until Inference. Test that delay is not set when flag is off.

Bug: 1009290
Change-Id: I060db0a2dda207537a5cc1f61690d0fafcc2f4d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1888618
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Rob Schonberger <robsc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711052}