Use net::Error for BundledExchangesReader's BodyCompletionCallback
Currently the argument type of this completion callback is MojoResult.
The callback is called from BundledExchangesReader::ReadResponseBody() and the
argument is the result of mojo::DataPipeProducer::Write().
This argument is converted to net::OK or net::ERR_UNEXPECTED in
BundledExchangesURLLoaderFactory::EntryLoader::FinishReadingBody() and is passed
network::URLLoaderCompletionStatus.error_code.

This completion callback will be called also from MojoBlobReaderDelegate of
BundledExchangesBlobDataSource (https://crrev.com/c/1886271).
In this case, MojoBlobReaderDelegate receives the completion result as a
net::Error.
To avoid redundant conversion, this CL changes the argument type of
BundledExchangesReader's BodyCompletionCallback from MojoResult to net::Error.

Bug: 1018640
Change-Id: I99783eb4744784a09f838db8b1062f6a90e8e455
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1890452
Reviewed-by: Kunihiko Sakamoto <ksakamoto@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711086}