[Navigation] Update to NavigationSimulator same document commits.
Found a logic issue with the NavigationSimulator. CommitSameDocument
expects browser initiated same document navigation to be in a started
state and have the same_document_ flag set. The flag is set at commit
time so it is logically impossible. Just go through regular Commit() in
this case.

Change-Id: Id71d99311661f3e82a473267048ed84c1b63e3e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1724074
Reviewed-by: Camille Lamy <clamy@chromium.org>
Commit-Queue: Arthur Hemery <ahemery@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711298}