Consider substitute-ok to be a successful printer query.
We've noticed that some printers don't understand all the attributes we
are requesting and replying with subsitute-ok. In this case, we probably
cannot setup the printer as autoconf but can still perform make and model
matching.

I noticed that ipp-version-supported isn't required in IPP 1.0 but is
in 1.1.  So, we're assuming that if ipp-version-supported is missing,
it's a 1.0 printer.  In the past, this was recorded as a failed setup.

Bug: chromium:1013821
Change-Id: Id3d7d471e5f9d5cc336ce3fed41f2b1f1a0880a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1891765
Reviewed-by: Bailey Berro <baileyberro@chromium.org>
Commit-Queue: Sean Kau <skau@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711303}