Merge origin into help text of Click to Call dialogs.
If we show help text in Click to Call dialogs, we can just embed the
origin into the help text itself instead of showing it as a separate
line. We keep showing it as a standalone label in the footnote if there
is no help text.

Before: https://imgur.com/PSIfHmz
After: https://imgur.com/UXeUbmp

Bug: 1019242
Change-Id: Id59ff2e53d2f3a5e7ab9da9a1ed876fb5d45641d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1887617
Commit-Queue: Richard Knoll <knollr@chromium.org>
Reviewed-by: Himanshu Jaju <himanshujaju@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711361}