Clean up temporary logging iframe in automation_helper
For some sites they overload console commands, so we
create an iframe to log element info for debugging.

We need to make sure to remove that iframe afterwards, or
we build up a collection of them that might affect elements
on the page, particularly on lower resolutions.

Bug: 984664
Change-Id: I2f81cd8e563c4ca275f561efe94bb03027ee0d8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1884815
Auto-Submit: Ian Struiksma <ianstruiksma@google.com>
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711370}