Refactor the code to update all apps out of the updater main.
This is mechanical work before registering the app ids with the
updater lands. There is some common registration code that is
common on both install and update execution path. It is helpful
to have the update apps code in its own compilation unit, just as
install is.

R=waffles


Bug: 1020285
Change-Id: I06590cf1d73a5d22ae79d288de564c0040442678
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1893406
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711399}