[ios]  Fix BadgeViewController interpretation of badgeState
Since BadgeState is a mask, a direct equality comparison to
BadgeState values for the BadgeItem properties is not the correct
way to assess if a badge is accepted.

This is a cherry-pick of part of crrev.com/c/1881364 to minimize the amount of
additional cherry-picks necessary for this to work.

Bug: 976901, 1016793
Change-Id: Iadc7e3e42529ded0aa331da3845e8644e4cac9c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1881364
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Peter Lee <pkl@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#709675}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1894262
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/branch-heads/3945@{#385}
Cr-Branched-From: e4635fff7defbae0f9c29e798349f6fc0cce4b1b-refs/heads/master@{#706915}