[WebLayer] Fix file input after permission prompt
When a new permission is needed to show the chooser (e.g. camera
permission), the chooser was not showing up. This is because the
permission request was routed through the Activity instead of the
Fragment. This fix is similar to what was done to send the Intent
through the Fragment in http://crrev.com/c/1880072.

Bug: 1018955
Change-Id: Ie0226d388b8a9148afeeda2ccec8907f1b547dd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1891748
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711565}