bfcache: use the canonical function in blink.
To check whether the BackForwardCache is enabled or not, we choosed from
the beginning not to check the content::feature directly, but use a
function wrapping it.

It means everyone is checking against IsBackForwardCacheEnabled()
instead of using the feature::kBackForwardCache flag.

The goal was to be future-proof. If the decision should not be entirely
based of the feature flag, we would just have to update the
implementation of the function instead of every callers.

One future patch:
https://chromium-review.googlesource.com/c/chromium/src/+/1875091
will modify the implementation of this function.

So far so good!

The problem is that patch:
https://chromium-review.googlesource.com/c/chromium/src/+/1654739 landed
using the content::feature directly instead of the "canonical" function.

It means the blink/ part may be using the BackForwardCache while the
content/ isn't. Moreover, this might forces users with low-memory
devices to be reported in one of the enabled or control group.

Bug: 511340

Change-Id: I68da6c7ca4cd48716b9ba092a0c5d84bb59b7073
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1893262
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Kentaro Hara <haraken@chromium.org>
Auto-Submit: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711569}