WebApps: Fixes a crash upon sign in in Chrome OS.
This was caused because PendingAppManager's shortcut_manager_ was not
being initialized in SetSubsystems. It was not caught by existing tests
because we disable shortcut creation in tests, as it's hard to
clean up afterwards.

In a followup CL I'm going to try and push shortcuts being disabled
further down the stack.

Bug: 1019319
Change-Id: I91d8b22eb5cdba67a23481db778e0c4f9724e632
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1890394
Reviewed-by: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Commit-Queue: Jay Harris <harrisjay@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711611}