Remove the function CreateCachedV8Private and the macro V8_PRIVATE_PROPERTY_SYMBOL_STRING
The function CreateCachedV8Private, which is not recommended to use because of calling ForApi(), is called only in one place.
So, this CL writes the contents of the function directly in the place where uses the function, and removes the function.

The macro V8_PRIVATE_PROPERTY_SYMBOL_STRING is also called only in one place. So, we replace the macro with a constant string.

Bug: 715418
Change-Id: Ie136200f3b87d760aa79556f835c8438e8ad47ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1893602
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Hitoshi Yoshida <peria@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Marina Sakai <marinasakai@google.com>
Cr-Commit-Position: refs/heads/master@{#711649}