Android accessibility: Apply dip scale to text character bounds, too.
Change http://crrev.com/c/1161322 modified the Android accessibility
bounds computation to account for use-zoom-for-dsf. However, there's
another place in the code where this computation is needed - when
computing character bounds, for the API used by select-to-speak to
highlight individual words. This was still unscaled, so the bounds
were incorrect.

Bug: 870230, 933119
Change-Id: Ic0d80436895b9fb9b7586b73f55719c94621cdb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1894317
Reviewed-by: Ella Ge <eirage@chromium.org>
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711716}