AW UMA: ensure we don't log sensitive data
No change to production logic.

This adds a test to ensure we do *not* log bluetooth or usb data. We do
not currently log this, but it seems like a good idea to assert we don't
start logging this accidentally.

This leaves a comment to explain under which conditions it's OK to log
this data (and, by extension, modify the test case).

Bug: 1011991
Test: run_webview_instrumentation_test_apk -f AwMetricsMetadataTest.*
Change-Id: I7ba4fdd957c034eda405cbca10d07d1064eac66c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1889232
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Changwan Ryu <changwan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711861}