Replace DCHECK(not root) with plain conditional.
This DCHECK was first added in
http://crrev.com/19b1b85cd5c3227697d348a39cec89ddf5980690

But the code in question could be run as root, in the case when
curtain-mode is false. The host only aborts running as root if curtain
mode is enabled. The host service explicitly runs in both "Aqua" and
"LoginWindow" contexts, as specified in the launchd agent's plist.

Instead, this CL skips the permission-checks when host runs as root.

Change-Id: Idf07719d0780f3627290e96710c36d0fe98e8f7a
Bug: 1020739
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1896122
Commit-Queue: Lambros Lambrou <lambroslambrou@chromium.org>
Commit-Queue: Jamie Walch <jamiewalch@chromium.org>
Auto-Submit: Lambros Lambrou <lambroslambrou@chromium.org>
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711901}