Fix lost focus when resizing embedded Assistant size.
When the embedded Assistant is initially shown, it uses a small window.
When a big response arrives, its size is increased (to what is called
|ash::AppListViewState::kHalf|).  When this happened, the focus was lost
from the input text box (and moved to the - at this point not visible -
launcher search box).

This is fixed by updating the |AppListView| to not move the focus to the
launcher search box if the embedded Assistant UI is showing.

I also introduced a new |AssistantTextfield| which is used for inputting
text based queries. This makes things like logs and stacktraces easier
to parse as they will now mention |AssistantTextfield| instead of just
|Textfield| which is used everywhere.

test: Added unittests and manually verified.
bug: b/143715841
Change-Id: I4da2b363ec6c93b5bb166b415e45c60c08206810
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1896047
Commit-Queue: Jeroen Dhollander <jeroendh@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Tao Wu <wutao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711916}