Ignore subframe, same-doc navigations in LoginTabHelper
LoginTabHelper::DidFinishNavigation is responsible for observing when
a navigation that challenges for HTTP auth commits, and showing the
login prompt on top of it. This navigation failed to exclude subframe
or same-document navigations, meaning that after an HTTP auth prompt
triggered by a subframe was cancelled, we'd re-show another auth
prompt on top of it.

Bug: 1019969
Change-Id: Icef1544b8656d885eae593339fe50107e8e6f585
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1890733
Commit-Queue: Emily Stark <estark@chromium.org>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711981}