[LayoutNGFragmentPaint] Fixes background painting for table-cells.
This fixes the painting of table-cell backgrounds to paint in the
appropriate rect.

As the position of the table cells aren't sub-pixel aware (yet) we were
painting the background with the incorrect size, resulting in a the
table background bleeding through.

I suspect TablesNG won't have this issue (and won't need to use this
alternate constructor) as their position should be sub-pixel aware.

Bug: 988015
Change-Id: I4135070608f958a71aa7a843c4e11904323125bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1896049
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#711988}