bfcache: cleanup const in tests.
In the back_forward_cache_browsertest file, a lot test used "const" for
declaring GURL or Origin variables. This is unusual for a variable
scoped to a test function. "const" should be used for preventing
mutation, this is mainly used in class attribute or method return value.
In test, this is not really needed.

Adding "const" here is not harmful at all, but since things tend to be
copy pasted, it is better the cut the trend now, so that it won't spread
more.

Bug: None.
Change-Id: Ia100e61781cdadaa9b420fab75244d44acedf33b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1892957
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712078}