Add diagnostics and clean up chrome.release warnings
While working on some installer changes I realized that misspelled
patterns in chrome.release are silently ignored, which makes installer
errors harder to find. This change adds optional diagnostics (verbose
only) for patterns that don't match any files, and it cleans up the
errors found by this.

eventlog_provider.dll was not listed as a dependency and was therefore
not reliably placed in chrome.7z.

ffmpeg.dll, icudt.dll, and locales\*.dll no longer ship but were still
listed in chrome.release.

chrome_child.dll is missing but is retained for now so that we can go
back to multi-dll builds if we want to.

The default_apps and Chrome-branded logos were not in the GOOGLE_CHROME
section.

The only other warnings I have seen are from files that are
configuration specific.

We should also consider tagging which patterns are optional so that we
can fail on unexpected missing files, although there may be too many
files that are optional given config variations.

While cleaning up the Python script I made some print-commands
Python 3 compatible and changed how the archive command was printed
so that back-slashes wouldn't be doubled-up.

Bug: 920704, 445616
Change-Id: I96928ea1e1f18a4c52664cf19d0ccca0a10de1d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1894434
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712178}