mac: register with LaunchServices for protocol handler test
As of 10.15, an app must be registered with LaunchServices before
the corresponding bundle ID can become a protocol scheme handler. This
happens as part of app startup normally, but in tests where there is
no app, it needs to be done manually. Unfortunately these tests do not
have a bundle to register, so this change has them try to use the main
browser bundle when registering.

This is a less evil version of
https://chromium-review.googlesource.com/c/chromium/src/+/1890960

Bug: 998740
Change-Id: I96b2d21b1966c304dba3d61e4539308d96221d28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1893582
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712191}