Switch to will-change: transform for all compositing/ web tests
We are changing the behavior of trivial 3d transforms to not require
compositing and this means we lose test coverage of some compositing
codepaths in the compositing/ directory. This change replaces trivial
transforms that were likely intended to be compositing triggers with
will-change: transform.

Bug: 1018260
Change-Id: I6ec7edcbb418c530c7bf2e3b9477c4347c5eedc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1893513
Auto-Submit: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712203}