Fix URL bar background flash on swiping tabs and entering/exiting tab switcher
The background of URL bar is not displayed on swiping tabs or entering/existing
tab switcher. This CL is trying to correct location bar background color by
detecting whether it is an NTP and location bar is on top of the NTP or not.

Bug: 874101
Change-Id: I6749d0829839cf753a99e1663e6c0565bd9bf279
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1887655
Commit-Queue: Lijin Shen <lazzzis@google.com>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712207}