views: return correct AXRelativeBounds for table rows and cells
Despite the name, AXRelativeBounds are actually the absolute bounds
when used with Views views. The code in TableView was treating them
as though they were relative, which led to table rows and cells
having their virtual positions near the origin of the screen. After
this change, they are correctly positioned.

This change also implements AXVirtualView::GetNSWindow, which is
being called by VoiceOver when navigating these table cells. It
doesn't seem to matter whether it returns a correct value or not
but having it NOTREACHED() causes the browser to crash in dcheck
builds, which is bad.

Bug: 1009024
Change-Id: I67c0cd05786896de873fc8a36129cd2f0f3dff75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1894308
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712210}