extensions: constify most of ExtensionInstalledBubble
The primary goal of this change is to make it easier to reason about
the behavior of ExtensionInstalledBubbleViews. In aid of that, this
change makes most of the state of ExtensionInstalledBubble (all of it,
actually, modulo browser_ not actually being able to be const) const,
which guarantees client classes that that state does not change on a
given ExtensionInstalledBubble.

To achieve that, this change:
1) Defers construction of ExtensionInstalledBubble in
   ExtensionInstalledBubbleObserver until the extension is actually
   installed and ready for the bubble to be shown; to do this,
   the observer class now stores some state needed to construct the
   bubble later.
2) Factors away ExtensionInstalledBubble::Initialize, moving
   initialization of every member into a (private) constructor of
   ExtensionInstalledBubble, since const members can only be
   initialized this way.
3) Adds a few helpers in an anonymous namespace that compute various
   bits of derived state ExtensionInstalledBubble needs to store,
   for use in the constructor's initialization lists.
4) Modifies ExtensionInstalledBubble's browser test to no longer use
   the now-gone ::Initialize method.

One wrinkle is that much of the state of ExtensionInstalledBubble
depends on the extension's ActionInfo, but the bubble doesn't need to
and shouldn't store that ActionInfo. As such, this change adds a new
private constructor that accepts an ActionInfo, and has the public
constructor delegate to it, passing in the appropriate ActionInfo.

The next CL in this series will make use of this newfound const-ness
in ExtensionInstalledBubbleViews.

Bug: 1011446
Change-Id: Ib4677914ce74e6cc19570fe006c1bdc948781b9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1894879
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712264}