Fix display sizing when both overscan and zoom are enabled
Display sizes with zoom are wrong as the overscan insets are not scaled
according to the device scale factor that's actually in effect.

This caused ash to draw parts of the shell UI outside of the root window
whenever both overscan and zoom are active.

Fixed: 1020028

Change-Id: I832a8a8d2dc0c2a4f1702824124759e9d3844917

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1892056
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Michael Spang <spang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712271}