cc: Enable MSAA only for layers with slow paths.
Currently MSAA is controlled using a global switch for all layers which
is enabled based on whether currently painted content on any layer has
slow paths. Switching this state tears down all tile resources, both on
pending and active tree. As such this switching can be quite expensive
since we need to re-raster the entire visible viewport and causes jank,
because the active tree can not be redrawn until its tiles are ready.

For a long scrolling page where only some content of a layer has slow
paths, this switching can occur each time this content scrolls in and
out of the interest rect.

Avoid the above by allow MSAA to be applied on a per layer basis. Since
the decision changes only when the painted content of a layer is
updated, which should already be invalidated in that case, this avoids
any unnecessary raster invalidation.

R=ericrk@chromium.org

Bug: 1013758
Change-Id: I00678d21abb09b00c1246246076ee1a4e50d1822
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1891746
Auto-Submit: Khushal <khushalsagar@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Eric Karl <ericrk@chromium.org>
Commit-Queue: Khushal <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712297}