[buildbot] Reduce the amount of logging emitted by webkit_layout_tests
For some reason, outputting to stdout is quite slow on Mac. This likely
warrants a separate followup, but reducing the size of stdout to be more
information-dense seems generally valuable as well.

stdout goes from ~21MB to ~430KB, which saves ~7 minutes in on the Mac
bots, and ~1 minute on the Windows bots.

The output that's lost is generally of low value. One big loss is that
stdout from invocations of content_shell are no longer included in the
stdout of run_web_tests.py. This means that it's harder to get debug
info from LOG(ERROR) for passing tests... hopefully this isn't a common
use case.

Bug: 1014845
Change-Id: Ib7129ac9207c1641c1264a92b2e93a1be5c9b025
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1894105
Reviewed-by: Robert Ma <robertma@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712316}