Add DCHECK trace logging for event sequence in FrameSequenceTracker
This CL adds a trace string that represents frame reporting events and
will be output as part of the DCHECK message when the DCHECK fails. Each
letter in the trace string represent one reporting function call. The
length of the trace string is capped at 2,000 characters.

Bug: 1017291

Change-Id: I377737730b1f78a7ca87e53db6acba591ee00090
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1895379
Commit-Queue: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Xida Chen <xidachen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712424}