[TTS] Auto-adjust Caption and Context text widths.
Automatically adjusts the length of the Context and the Caption
text views as the panel expands open.  This fixes some issues
with Definitions which can have long captions and uses the
Context view to show the pronunciation.

The Caption was supposed to auto-adjust, but was not calling
super when expanding.  The Context just needed to pass the
adjustment parameters.

BUG=1016973

Change-Id: Ie331cde2cc08a059197ea1e7178099c865e4dbc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1898722
Auto-Submit: Donn Denman <donnd@chromium.org>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Commit-Queue: Donn Denman <donnd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712446}