Add tests for atomic inline+NBSP quirk in nowrap
This patch adds a test to check the atomic inline+NBSP
quirk[1] should not apply if `nowrap`.

This issue was originally introduced in r694788
crrev.com/c/1792465, but was accidentally fixed in r700578
crrev.com/c/1827097. The fix did not have a test because
it wasn't intended to fix this issue.

[1] https://www.w3.org/TR/css-text-3/#line-break-details

TBR=chrishtr@chromium.org, ikilpatrick@chromium.org

Bug: 1017465, 1018309
Change-Id: Id505b894c02d6521766e2cf354e7c9c8c4a114ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1888230
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712447}