Simplify SMIL animated value calculation and application
The ResetAnimatedType, ClearAnimatedType, ApplyResultsToTarget and
CalculateAnimatedValue methods that handles calculation and application
of animation values are only called on elements that has made it into a
sandwich - this means that they have passed the "has valid target"
check. Simplify the code accordingly by assuming there is a target and
that other checks that HasValidTarget() make holds. The latter for
example means dropping the null-check for the animateMotion transform
transform.
Also remove SVGAnimateElement::ShouldApplyAnimation() and always remove
the CSS property in SVGAnimateElement::ClearAnimatedType() when
animating CSS.

Bug: 1017723
Change-Id: I9d184325671fae820c5718d3145228b7bd8d284e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1895336
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#712531}