heap: Prepare concurrent marking for new jobs api
The concurrent marking tasks used to get an ids as an
argument. With the new jobs api we can no longer pass
the id that way.
This CL updates to tasks to get an id from a collection
of available ids. This way all tasks only need to get a
pointer to the ThreadState (so tasks don't get unique
arguments anymore).

Bug: 986235
Change-Id: Ibd7146840c74297dc4531a1b4c6c1475c647464d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1884850
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712553}