[WebLayer] Whitelist non-overridable SSL error strings
This CL whitelists the strings needed to show SSL error interstitials
for non-overridable SSL errors (i.e., errors where the user is not
allowed to proceed).

To test: Visit badssl.com in weblayer_shell on Android, tap on
"mitm-software", and test that an SSL error interstitial shows wherein
the buttons are "Reload" and "Advanced", and clicking on "Advanced"
does not present an option to proceed. In particular, the app should
not crash :).

Change-Id: I4cb8b4aa2289bc0830eaa7e035779e64e3498fb8
Binary-Size: Increase is due to translations and so cannot be avoided.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1899608
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Pavel Shmakov <pshmakov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712604}