X11 and Ozone: Fix LowerWindow feature.
The feature was accidentaly removed by crrev.com/c/1772537

This CL adds it back by adding necessary interfaces to
DWTHLinux, PlatformWindowLinux and X11Window.

The method is X11 specific and not in use for Wayland as long as
Wayland does not allow to manipulate clients' windows.

Bug: 1002358
Change-Id: Id43d5991f662f8986670ebef45f2443472806590
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1899464
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712692}