Fix message center left position offset
Correcting the x position calculation of the MessageCenterBubble to
correctly align with the quick settings bubble.
If shelf is left aligned we need to add the left inset of the
quick settings bubble to the horizontal position.
Otherwise we need to subtract the right inset of the quick settings
bubble from the horizontal position.

Bug: 1021632
Change-Id: I8d449d3c297858283f41bcc10163a0e15b2965b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1900161
Reviewed-by: Tim Song <tengs@chromium.org>
Commit-Queue: Ahmed Mehfooz <amehfooz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712756}