Migrate blink::mojom::UnhandledTapNotifier to BrowserInterfaceBroker
This CL migrates the GestureManager mojo client in Blink to the new
BrowserInterfaceBroker, and adapts the code in the browser process
to register the binder from PopulateChromeFrameBinders(), instead of
doing it right inside UnhandledTapWebContentsObserver.
UnhandledTapNotifierImpl.

Additionally it also adapts ShowUnhandledTapTest helper class so
that it properly set the mock overrides for BrowserInterfaceBroker,
as it was done before for InterfaceProvider.

Bug: 936482,1006361
Change-Id: I9f369e41cb8441043646b34fd67e1902f7010383
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1887615
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Reviewed-by: Oksana Zhuravlova <oksamyt@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Donn Denman <donnd@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712780}