RenderFrameHostImpl::delegate_ is not null. Stop checking it.
RenderFrameHost::delegate_ is never null.

New CLs are adding more and more null check for this, just to be sure,
because there are locations where this is checked.

This patch remove all the location where is checked to avoid this to
spread further.

Bug: None.
Change-Id: I5ba7903a199fdb3117df6adf514e187b0a36ea0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1899501
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Auto-Submit: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#712799}