Add braces to conditionals in accessibility component extensions
Previously, the accessibility component extensions did not follow the
bracket conventions for conditionals Chromium style guide, primarily due
to the history of the ChromeVox screen reader. This led to some
confusion, as well as competing requests from different reviewers.

This change modifies all of the accessibility component extensions to
use braces for all conditional statements.

This change is a pure refactor, there is no change to function or logic.

Bug: 706981
Change-Id: Iba6f4922332de7c4c09a0bc8fb14cbd3e45ba37d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1898434
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Cr-Commit-Position: refs/heads/master@{#712878}