ozone: Make CreateScreen pure virtual.
We have all the Ozone platforms implementing the PlatformScreen.

Now, it's the time to make that method pure virtual.

However, there is one point - Ozone/Drm and Cast do not support
PlatformScreen, but still implement it and expect that the
method is not reached.

Bug: 872339
Change-Id: I628878842294a8e99033893a9be42026066c29c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1869004
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#712927}