Add to rendering orphan handlers when destination is running.
Continue to add to the rendering orphan handlers list if the
destination is still rendering.

If the destination is requested to stop, set a flag so that the
rendering won't pull on the audio graph anymore, even if the
audio device is still running and requesting data.

In the test case from the bug, we were not adding the handlers to the
list because the context was closed.  However, the destination was
still running because the audio device wasn't stopped yet. In the test,
the handlers were being destroyed at the same time as the audio thread
was accessing the handlers to render audio.

(cherry picked from commit 8ce80308aed7ebdd9d74bbbba6e9438c68ab55fb)

Bug: 1003807, 1017961
Change-Id: I87052c8504c5698d71f041e6d4305cf42851f067
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1879784
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#712666}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1899197
Reviewed-by: Raymond Toy <rtoy@chromium.org>
Cr-Commit-Position: refs/branch-heads/3945@{#506}
Cr-Branched-From: e4635fff7defbae0f9c29e798349f6fc0cce4b1b-refs/heads/master@{#706915}