[omnibox] Cache stripped URLs of tabs for quick comparison
We currently re-strip each tab's URL prior to comparison with the
URL to be navigated to, in order to offer a tab switch instead.
This has grown to be non-trivially expensive. This CL caches the
stripped version of each tab's URL so that we at least don't have
to re-strip it (malloc, etc.)

Bug: 780835
Change-Id: Ifa56f289e20b66ed1979ed56b7d2211726b4a3e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1896508
Commit-Queue: Kevin Bailey <krb@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713067}