arcvm: Set androidboot.serialno in ArcVmClientAdapter
This CL adds a new ARC pref (prefs::kArcSerialNumber) which is NOT
network-synced (and therefore is unique per device and per profile.)
ArcSessionManager's Initialize() generates the number when the pref
value is empty, and pass it over to ArcVmClientAdapter to allow the
adapter to use the value as androidboot.serialno Android property.

This does not change ARC container's behavior.

BUG=b:143779997
TEST=ran unit tests. Start ARCVM twice, verify both are started with
 the same, non-empty serial number.

Change-Id: Ia436a722d13be58ba5ed59bb0053009c2352743d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1899732
Commit-Queue: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Yury Khmel <khmel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713077}