sync: Add settings UI to control sync for Chrome OS data types
When the SplitSettingsSync flag is set the Chrome OS "People" section
row for sync settings takes the user to a new settings page for OS
sync controls. For now only kOsPreferences and kPrinters are supported.

The WebUI for this follows the pattern of browser sync controls:
https://cs.chromium.org/chromium/src/chrome/browser/resources/settings/people_page/sync_controls.html

The JS handler follows the pattern of people_handler.cc:
https://cs.chromium.org/chromium/src/chrome/browser/ui/webui/settings/people_handler.cc

Screenshots:
http://screen/ijEUzJrqrLE
http://screen/faFhPmRgegL

Future CLs will add support for more data types and for passphrase
encryption.

Bug: 1013466
Test: added to unit_tests and browser_tests

Change-Id: Id33b9eb8ebb653d6b2d7fc103dc3154cfdaa1ca7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1894763
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713095}