Check scroll layer when updating composited scroll offset
ScrollingCoordinator::UpdateCompositedScrollOffset() assumes that the
scroll layer exists. We should not call the function when we haven't
created the scroll layer.

The check was not needed before crrev.com/c/1895081 because we created
the layers earilier (during CompositeUpdate). After the CL we created
the layers during PrePaint, but VisualViewport::DidSetScaleOrLocation()
can be called earlier.

Bug: 1020181
Change-Id: Ie9a341928ed13e02b4edbefdfbc513f1ecb22be0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1899598
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713141}