weblayer: Fix fullscreen video
The compositor surface is opaque, which means when video is lifted
into a separate surface during fullscreen below the compositor
surface, it is never showing through.

Always set the compositor surface to translucent for now, which will
use more memory even when not in fullscreen.

Bug: 1022042
Change-Id: I2ff368dfa2a04ba0e62891039244c0bcb148e6fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1899901
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713178}