Move SetWebUIProperty from RenderViewHost to RenderFrameHost.
WebUI has moved to be a per-frame object, but the SetWebUIProperty
method is still on the RenderViewHost object. This CL moves it to be
per-frame, therefore on the RenderFrameHost object, but does not make
it available to subframes. The goal is to be a purely code move CL.

Bug: 763548, 1002276, 1019739
Change-Id: Ie857a253f750316fe090ac2daa6686376cdb1833
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1901587
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713231}