[Files ranking] Fix bug: zero state result persist to search
There is a bug in the provider for zero state Drive file results that
means it returns results even when a search query is issued, messing
up the results for a search. This is caused by the result cache being
cleared at the wrong time.

Bug: 1022150
Change-Id: Ic1382e37a7991284c492963c57681d63f2f02a35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1902167
Reviewed-by: Thanh Nguyen <thanhdng@chromium.org>
Reviewed-by: Jia Meng <jiameng@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713233}