weblayer: Load implementation using WebViewFactory.
Load the implementation of WebLayer using WebViewFactory by default.
This requires that the currently selected WebView implementation is the
WebLayer implementation you want to use.

If the client APK specifies a package name in the metadata tag, it will
be used as the implementation instead, but this is only intended for
testing and does not use the exact same loading mechanisms.

Change-Id: Id12d13a1b49f718bb7c6bc6dbedb848a37f74355
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1869296
Commit-Queue: Richard Coles <torne@chromium.org>
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713255}