Add model config, radius_polynomial_resize for polynomial radius update.
Adds a configuration into model configuration to allow a model to have
the radius transformed. Intention is that a model is customized per
device depending on the characteristics of the device being executed
on.

Updates unit test appropriately, and also remove the former
approximate for radius conversion.

Bug: 1009290
Change-Id: I80e0160302a98ad8564ac73b0bb07f1ab8327467
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1900516
Commit-Queue: Rob Schonberger <robsc@chromium.org>
Reviewed-by: Michael Spang <spang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713260}