WebUI Tab Strip: Fix flaky tests
Previously, there was a race condition between when the
IntersectionObserver's callback would get fired for the first time
each TabElement was added into the view, and when the callback
would be expected to have been called for an individual test's
setup.

Ran 100 times with no failures.

Fixed: 1022008
Change-Id: I5e418cb8fc691fd4ecd32c3d88a9805f1bdc2f0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1902070
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: John Lee <johntlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713314}