Auto-open bubble after for recently accepted media permissions
If a media (camera/mic/camera+mic) permission was recently accepted on a
site level, but it was blocked on a system level, auto open bubble
that leads users to system permissions.

Situation this will be triggered:
1) Site asks for media permission.
2) Permission prompt shows on the left side of the omnibox.
3) User accepts the permission prompt.
4) Blocked icon appears on the right side of the omnibox & the bubble
which gives the user the option to "Open system preferences" auto-opens.

Note that if mic and camera are requested individually, the bubble will
auto-open for each of them once.

Change-Id: I611b4f256ca4f0793d2e97bb8bc8bc4cd12babe1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1897659
Commit-Queue: Kamila Hasanbega <hkamila@google.com>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713354}