Adjust DCHECK to include inactive documents for scrollingElement.
The DCHECK already covers the stopped/stopping states, so the only states it
doesn't cover are kInactive and kUninitialized. Since JS is calling
this API, the kUninitialized state is impossible. So the only state that
this occurs in is kInactive. The method can return null in quirks mode
so adjusting the DCHECK for inactive documents seems best.

BUG=1020787, 1002472

Change-Id: Ief51eb7783ea7ed506107646a258f24e274127b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1902007
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713427}