Reland "Define guest view custom elements synchronously"
This reverts commit 331504cf86e8b148315ab69703b7631cc5f7e651.

Define guest view custom elements synchronously

Previously, we've had to wait for readystatechange before defining
the guest view custom elements because
1) there were circular dependencies that would cause errors if we
tried to perform the definition immediately, and
2) the Custom Elements V0 registration context was scoped to the
document rather than the window, so if we defined the elements
while the document was still on about:blank, the definition would
be available on about:blank, but not for the extension.

Now that the circular dependencies are fixed and guest view is
migrated to Custom Elements V1 (whose registration context is
scoped to the window), it is no longer necessary to wait for
readystatechange.

We now perform the definition immediately.

Bug: 810012
Change-Id: I009dbc7710316dd9f4b7fbe5a12ad091e6b597bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1869115
Reviewed-by: James MacLean <wjmaclean@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713584}