Ash Tray: Introduce status area CollapseState and update visibility of trays.
The CollapseState indicates whether the tray is collapsed or expanded or not
collapsable (default behaviour) at all. Currently, the status area will only be
collapsable in tablet mode and in-app.

The visibility logic for individual tray items is also being cleaned up in this
CL, as this logic is fairly complex. We rename SetVisible() to
SetVisiblePreferred() because the effective visibility also depends on factors
such as the virtual keyboard and status area collapse state.

Future CLs will properly update the CollapseState as well as calculate which
tray items should be visible when collapsed.

TEST=unit tests
BUG=982511

Change-Id: I6f6fed7836d0b1dc4991319fb4ec8e58a522da37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1903399
Commit-Queue: Tim Song <tengs@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Tetsui Ohkubo <tetsui@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713694}