Uncompress terminal system app assets
In order to save space on rootfs, we will store crosh
and terminal assets gzipped.

This CL allows the data source to serve non-compressed or
gzipped files.  E.g. on a request for
chrome://terminal/js/terminal.js, we first search for
/usr/share/chromeos-assets/crosh_builtin/js/terminal.js,
and serve it if it exists, else we search for
/usr/share/chromeos-assets/crosh_builtin/js/terminal.js.gz
and serve the uncompressed contents.

Bug: 1022250
Change-Id: I3060d86d64b43673ad7c75606083fad2dc3b33fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1903170
Reviewed-by: calamity <calamity@chromium.org>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713700}