[omnibox] Encapsulate an accessibility call
After a previous change, it was apparent that OmniboxEditModel was
exposing more members than it needed to, and that the code was a
little more complicated than necessary. This CL encapsulates the
call, so that no data needs to be public, and the code is simpler.

Change-Id: Ibb3dcc23148d1e2568f54a9645de4c4267e3d2b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1903619
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Kevin Bailey <krb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713703}