Add getTransform function to offscreen canvases
It seems like maybe this was just overlooked? The fix is as easy as
adding it to the idl file.

It seems like this is already in the spec as:
https://html.spec.whatwg.org/multipage/canvas.html#offscreencanvasrenderingcontext2d
contains the line
OffscreenCanvasRenderingContext2D includes CanvasTransform;
And https://html.spec.whatwg.org/multipage/canvas.html#canvastransform lists the
getTransform() function.

Bug: 1016454
Change-Id: Icb5405d716524e2a0b6935db81b18ff53cd31d4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1872562
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Commit-Queue: Aaron Krajeski <aaronhk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#713886}