[NetworkDetails] Prevent jank when collapsing sections of the page.
There was a 10px margin at the top of each collapsible section which
was not included in the animation used when collapsing.  This lead to
a smooth animation for most of the collapse, followed by an awkward
jump for the end of it.

Fixed: 1009978
Change-Id: I04b94a4c8f497fe0d6e7574f201fae389b626d00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1907198
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Jon Mann <jonmann@chromium.org>
Cr-Commit-Position: refs/heads/master@{#714022}