Remove compositing query in PaintLayer::PaintWithFilters
In pre-CAP code, if a filter is on a paint layer that has a
composited layer, it is determined to not paint with filters. In
which case, the filter is not taken into account for things like
computing the visual rect. However in the CAP world there is no
compositing state to consult so any PaintLayer that has filters
is considered to have paint with filters.

Bug: 1007989
Change-Id: Ibd5aae9a34f855a47fb863633bfe8d2319c3d08b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1902727
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#714055}