[Files ranking] Do not load saved models on dev for debugging.
We still have one user report that isn't showing zero state files, and
this is suspected to be because of a bad model that was saved to disk
early in development. This CL prevents loading models on dev, which
will help debug the issue.

This code will be removed as soon as we have finished debugging.

Bug: 955893
Change-Id: Ifef902e98919d166135127c2d8c140743ca5094e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1898896
Reviewed-by: Jia Meng <jiameng@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#714132}