[Overlay] Cleanup some shadow-visibility booleans.
Removes some booleans and methods for visibility of shadows that are
now always visible.

The Bar shadow was probably switched so we could draw a banner (which
we no longer ever do).  The Panel shadow was switched due to not
being needed on the new layout, but now we do use it.

Change-Id: Idf90fbe52841ba49e7760425768795b378c24e52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1906316
Auto-Submit: Donn Denman <donnd@chromium.org>
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#714250}