GMC: Add ability to drag a notification out of the dialog
This CL allows the user to click and drag a media notification out of
a MediaDialogView and into an always-on-top overlay. This allows the
user to continue to control media without having to go back to the
dialog.

This feature is put behind a disabled-by-default flag.

Bug: 1022452
Change-Id: Ifd4b837c5b2407c886a9b16403d15f710e38db0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1904158
Commit-Queue: Tommy Steimel <steimel@chromium.org>
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#714382}