CrOS home launcher: Rename OnHomeLauncher{Target,}PositionChanged
This is used (and will be used more) to give observers updates about
the home launcher position (within the animation / drag) and not just
updates about what the expected target position is, so the new name
is a little more informative.

It's hard to predict what the "target" position is anyway, since
during a drag the user can easily go back and forth.

This is a pure rename (plus some comment updates) and has no expected
user-visible impact.

Change-Id: I8eb8d039e3cc33a43a4fa9760dcb332e15a83e24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1910768
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#714390}