Add HIDE_IN_OVERVIEW window type in wayland protocol
We want to control the visibility of a window in overview mode
from the Android side.

The actual scenario we have so far is:
We want to hide the Netflix home (non-video) window if the video
is in PIP mode because selecting the Netflix home window in the
alt-tab cycler kills PIP, which is a bit confusing and may not be
the user's intention. See b/129367828 for more details.

BUG=b:143789289
TEST=PIP'd Netflix and confirmed the netflix home window is not listed in the alt-tab cycler.

Change-Id: Iee4aac66c54646972a030cf3a00a138cf5762896
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1905302
Auto-Submit: Kazuki Takise <takise@chromium.org>
Commit-Queue: Kazuki Takise <takise@chromium.org>
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#714411}