[partitionalloc] Change OOM_CRASH macro for easier crash classification
Crashes caused by the OOM_CRASH macro are currently not classified as
OOM crashes by Chrome Crash at the moment. The reason is that Chrome
Crash analyzes the stack trace to do the classification, and OOM_CRASH
does not show up in the stack trace at the moment.

This CL changes OOM_CRASH to call the NOINLINE function OnNoMemory,
which then causes the crash. Thereby the OOM crash is visible on the
stack trace and can thus be classified correctly by Chrome Crash.

R=rmargold@chromium.org, yuzus@chromium.org

Bug: chromium:1019661
Change-Id: Id1a755076db4464b5de9a36aaf4ddeaa9a7dbc16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1906148
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Yuzu Saijo <yuzus@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#714448}