Add PRESUBMIT.py warning on adding new images to repo
This cl adds a presubmit warning on upload if the CL contains images.
The warning points the author to
//docs/speed/binary_size/optimization_advice.md#optimizing-images so as
to ensure they have read the optimization tips and tried applying them
to their images. There is no easy way to check if the image is optimized
thus the presubmit always fires if there are images but does not block
CQ.

Bug: 921227
Change-Id: Ib8dd144bd2f728b166ca5b1c9d3882df24e0216b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1899645
Commit-Queue: Mohamed Heikal <mheikal@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#714507}