mac newsdk: don't depend on definition of IMP
As of the 10.15 SDK IMP has changed from:
  id (*IMP)(id receiver, SEL selector, ...);
to:
  void (*IMP)(void);
thus requiring a cast to call it. Rather than having this cast at
every place that invokes an IMP, this change adds a new method to
ScopedObjCClassSwizzler to allow invoking the original method from
the swizzler, which accounts for every invocation of IMP in the tree.

TBR=avi@chromium.org

Bug: 973128
Change-Id: I13eab9a79bab23e9a2c3854fb95e42de61e04d3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1906334
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#714568}