Add new Chrome OS OOBE sync consent dialog for SplitSettingsSync
As part of go/cros-sync and go/split-settings-sync we're creating a new
OOBE sync consent dialog. The main new feature is that the user can turn
OS sync off in the dialog and the OS sync system will never start.
OS sync is tied to a user profile pref that is set only when the user
closes the dialog.

This version of the dialog has placeholder strings and layout. Strings
won't be available until M81 and I suspect the layout might change.

See bug for screenshot and link to mocks.

Bug: 1023854
Test: added to browser_tests
Change-Id: I3e86a827eae1d80b8784b89b433d9aebb92eb5d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1912899
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/master@{#714719}