[CrOS Wallpaper] Fix 'My Image' thumbnail image distortion.
Currently, the thumbnail image for the current wallpaper is distorted
if the wallpaper image is the user's file (not provided by Google).
Instead of resizing the image to a distorted state for the current
wallpaper thumbnail, this CL ensures that the thumbnail image is scaled
properly and center cropped.

Before screenshot--
https://screenshot.googleplex.com/mBZziQFOf3W

After screenshot--
https://screenshot.googleplex.com/OjYfOhyfFSK

Fixed: 985542
Change-Id: Idf698141a7ac6b38e797d9bd91d256ce5ea02db4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1911229
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#714729}