Fix AXPosition Paragraph navigation can pick Ignored node endpoints
This works is a continuation of improvements that Jacques Newman
had recently merged with this CL :
https://crrev.com/c/1810244

The problem with Paragraph navigation was that both the methods
|AtStartOfParagraph| and |AtEndOfParagraph| allowed an AXPosition
anchored to an Ignored node to be the start or end of a paragraph.

Adding a browsertest to exercise paragraph navigation when there
are Ignored nodes interleaved in the document which fails without
this patch.

The changes to |Create{Next|Previous}LeafTreePosition| does not change
behavior, there's no reason to copy |const base::RepeatingCallback| or
its bound parameters in this case.

Bug: 928948
Change-Id: I2cfee5481ffbea9a9a821cae6bcfc8e4652e847c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1885021
Commit-Queue: Adam Ettenberger <adettenb@microsoft.com>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#714739}