AW: implement stub UI for flipping flags
No change to production behavior.

This adds a UI for toggling flags (similar to chrome://flags), although
nothing actually happens yet when toggling the UI. This CL only builds
the visual component of the UI, and adds a single flag to the list to
demonstrate how the UI will look. In follow up work, we'll plumb this
toggled list of flags to the WebViews embedded in other applications.

Bug: 981143
Test: Manual - open the DevUI, observe the flags UI works as intended
Change-Id: I00e8e3c5472e71af3cdb3b70e7992d25a38a16b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1910811
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#714740}