[WebLayer] Bring in WebView's support for displaying network errors
This CL provides an initial implementation of displaying network errors
in WebLayer on Android. This implementation is copied from that of
WebView. Note that there is no attempt made to share the code as this CL
will need to be merged to 79.

This CL does not copy WebView's support for suppressing the showing of
error pages (AWContentRendererClient::ShouldSuppressErrorPage()). This
support is present in WebView for apps that want to display custom
error pages, which is not a requirement in WebLayer at this time.

To test: put your device into airplane mode and visit a webpage in
WebLayer. You should get an error page with the upside-down Android
icon.

Change-Id: I4e63cb577e433e3a34db126110ada9b91c851e51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1906349
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/master@{#714999}