AW UMA: wake up MetricsService like Chrome does
This configures AwMetricsServiceClient to "wake up" the MetricsService
the same way Chrome's client does. This listens to the content-layer
events for page load start/stop, renderer hangs, and closed renderer
processes.

This also adds an integration test to verify this works as expected (for
the page load start case), and adds plumbing to configure the upload
interval in the test (because the commandline switch is rate-limited to
a 20 second minimum to avoid DOS'ing the server).

Bug: 1003204
Test: run_webview_instrumentation_test_apk -f AwMetricsIntegrationTest.*
Test: Manual - verify renderer hang case by using the AndroidX demo app
Test: (which supports triggering and recovering from renderer hangs)
Change-Id: I1f19f89cbc3625d4911755a2b17dd82908c2bb62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1900613
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Changwan Ryu <changwan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#715132}