Update screen sharing system permisison heuristic for macOS Catalina.
Now we consider all windows instead of the visible ones.
Restricting to visible ones caused problems when Chrome was maximized
and some windows required for the heuristic to work were not visible.
Also, this patch removes the check for status windows, since it is
common to have one whose title can be read even with permissions
disabled.

Bug: 1023028
Change-Id: I877bfe518c93961ee3e8e5cc4e768a1ca990eb80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1917159
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Marina Ciocea <marinaciocea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#715274}