Notify observers of node delete at end of AXTree::Unserialize
Previously, we notify an AXTreeObserver of a node removed from AXTree
before the node is actually removed from the tree. In this change, we
notify an AXTreeObserver of a node removed from AXTree after the tree
is stable and its nodes have been updated.

Previously, when in AXTree::Unserialize and before tree modification
takes place, we call NotifyNodeWillBeReparentedOrDeleted on the node
to be deleted. AXTreeObserver/
BrowserAccessibilityManager::OnNodeWillBeDeleted removes its
reference to the AxNode's BrowserAccessibility object and destroy it
as well. This takes place before the actual ax node is deleted from
the tree. The subsequent action, also before tree modification, such
as a call to BrowserAccessibility::InternalGetFirstChild() will reveal
the inconsistent state--retrieving the child node from ax tree,
however when we query through
BrowserAccessibilityManager::id_wrapper_map_ for its
BrowserAccessibility, it is not found because it is deleted in the
prior action. This inconsistent state can cause a null dereferencing
in BrowserAccessibility::PlatformIsLeafIncludingIgnored for
role::kButton.

In the following scenario, a crash would occur.
AXTree:
kRootWebArea #1
++kButton #5
++++kCheckBox #2
++++kGenericContainer #3 IGNORED
++++++kStaticText #4 IGNORED

AXTreeUpdate:
1. remove node kCheckBox #2
2. remove subtree kGenericContainer #3 IGNORED
---
AXTree::Unserailize: {
  NotifyNodeWillBeReparentedOrDeleted(kCheckBox#2)
    OnNodeWillBeDeleted(kCheckBox#2)
      id_wrapper_map_.erase(kCheckBox#2)
      delete BrowserAccessibility of kCheckBox#2
  NotifySubTreeWillBeReparentedOrDeleted(#kGenericContainer#3)
    FireWinAccessibilityEvent(kGenericContainer#3)
      ...
      PlatformIsLeafIncludingIgnored(kButton#5)
        InternalGetFirstChild(kButton#5)
          GetUnignoredChildAtIndex(0) returns kCheckBox ax node
          id_wrapper_map.get(kCheckBox#2) -> null
        Dereferencing null, exception occurs.

  //--- then tree modification/ax node deletion happens --
}

Bug: 1020187
Change-Id: I8e83b3450ca6253e837f4854a1c90a4779b3d435
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1907774
Commit-Queue: Victor Fei <vicfei@microsoft.com>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Adam Ettenberger <adettenb@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#715341}