android: list pending uploads before done uploads
The upload list is capped at 50 crashes. Previously, the already
completed uploads were listed before the pending uploads. This
patch places the pending uploads first so that it is possible for the
user to see them and manually* trigger an upload.

*The "send now" button does not actually send the report now. It
schedules the dump for upload at a later time of the Job Scheduler's
choosing.

Bug: 1023836
Change-Id: Ided52abc451f4ac167dc9eb0db93d4ea32bdc090
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1917689
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Joshua Peraza <jperaza@chromium.org>
Cr-Commit-Position: refs/heads/master@{#715385}