Remove unnecessary stop on Settings networking subpages
All networking sub-pages have an unnecessary navigation stop on
the enabled-state of the feature. This is particularly bad on the wi-fi
sub page because the enabled-state text and the toggle are separated by
the add network button.

Fixed: 1024597, 1007058
Change-Id: I12d6233090dc3c5ff1b37526be9bf67211f78886
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1914827
Commit-Queue: Bailey Berro <baileyberro@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Auto-Submit: Bailey Berro <baileyberro@chromium.org>
Cr-Commit-Position: refs/heads/master@{#715404}