WebUI: cr-icon-button, use stroke color as ripple color if set
For example cr-camera's videocam-shutter has a red fill enclosed by a
white stroke. In this case, the ripple should be white matching the
outer region of the icon. Currently, --cr-icon-button-color to white
for all cr-icon-buttons in cr-camera. There should be no change in
behavior due to this CL.

This CL is needed for a follow-up CL that removes
--cr-icon-button-color.

Bug: 1022752
Change-Id: Ib22042d8719aa8d4d54cea9c87ce1641e7f7994b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1914515
Commit-Queue: Esmael Elmoslimany <aee@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#715433}