Add metrics for hotseat gestures
In this CL, metrics are added for the following hotseat gestures:
(1) The gesture to enter overview mode from home page.
(2) The gesture to show/hide the in-app shelf.
(3) Fling up the in-app shelf to show the home page.
Test cases for the metrics are also created.

Bug: 1018921
Change-Id: I86bdee0e28f28d222a03d1b474fb47b74e5b23f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1906301
Reviewed-by: Nik Bhagat <nikunjb@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#715442}