media: Support default Renderer in PipelineImpl
Currently PipelineImpl always uses the CreateRendererCB to create
the Renderer asynchronously, which involves a post to the render main
thread and another post back to the media thread. This has caused
regressions in start-to-play-time especially on lower end devices.

This CL partially restores the previous behavior by creating a default
Renderer on the main thread before pipeline start. The default Renderer
will be used in normal cases so we avoid the round trip to the main
thread to create the Renderer. In special cases where we need a
different Renderer, the CreateRendererCB will be used to create one.

Bug: 1018372,1023579
Test: Tests updated
Change-Id: I71627a959cdfe12cd0f5ef0ee358f461323f5a7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1915226
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#715550}