[webview] Add dark mode support to webview shell
This adds force dark setting menu items, and modifies the shell theme
sufficiently to allow FORCE_DARK_AUTO to work. Apps must opt in to force
darkening, and must have a theme that is not already "dark", so the
AppCompat.DayNight theme must be configured as such.

Bug: 947324
Change-Id: Ib40af5e3817d91fd497925e58689d9c0fc4f55a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1899613
Commit-Queue: Tobias Sargeant <tobiasjs@chromium.org>
Reviewed-by: Tim Volodine <timvolodine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#715731}