Don't resize native view when detaching from NativeViewHostAura
When NativeViewHostAura is detaching an aura::Window, it sets the
window's bounds to that of the host's clipping window. This results in a
resize of the hosted window if part of it is out of the host's bounds.

This fix simply removes the bounds change.

Fixed: 1024046
Change-Id: If9ef2b15e05985428b7f648f0508ee224a5de13a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1919427
Commit-Queue: Collin Baker <collinbaker@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Auto-Submit: Collin Baker <collinbaker@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#715784}