Only call DiscoveryStopped once when discovery stops
The call to BluetoothTaskManagerWin::DiscoverDevices runs on a loop
until either discovery stops or we find the device we are searching for.
If discovery stops this function called
BluetoothTaskManagerWin::OnStopDiscovery.  However, we call
BluetoothTaskManagerWin::OnStopDiscovery already when we stop. The
second call causes a crash. This CL removes that call.

Bug: 1022149
Change-Id: I74d977abdfd69cbde2e525a024b1f6016148ed7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1915300
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#715843}