Settings a11y: Give subpage routing cr-link-rows precise a11y label
The cr-link-row element is currently described by ChromeVox as a 'link'
despite the fact when it is used as a subpage route, it stays in the
current tab instead of opening a new one like the other links in
settings. The alternative aria-roledescription 'subpage button'
preferred by the a11y team remove that ambiguity.

Examples:
https://drive.google.com/open?id=1su58VJ_AksPs44HFoNWambjqcSLIQPdv

Bug: 1012104
Change-Id: I7d591bb74b8b5828ca44e7270d7edf6ab61755c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1881816
Reviewed-by: Hector Carmona <hcarmona@chromium.org>
Commit-Queue: Jordy Greenblatt <jordynass@chromium.org>
Cr-Commit-Position: refs/heads/master@{#715879}