Clear LayoutObject in NGPhysicalFragment when it is destroyed
Because we're replacing |NGPaintFragment| by |NGFragmentItem|,
this patch moves |is_layout_object_destroyed_| flag to
|NGPhysicalFragment| by clearning its |layout_object_| to
|nullptr|.

Though |NGPhysicalFragment| is immutable, keeping a pointer
to a destroyed object isn't good for anything.

Note, this is done only for inline |LayoutObject| as it is
called through |NGPaintFragment| today. We may revisit this
in future.

Bug: 982194
Change-Id: If89881cd9493526307106345db0549c0e9c8a6fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1914580
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Aleks Totic <atotic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#715954}