Don't update match icons when the popup isn't open.
This is a speculative fix for http://crbug.com/1025168 .  If there's an open
IME, we could skip adding child views, but the callback to get icons won't be
canceled.  In this case, the popup itself should be closed, so the code can
check for that.  (Clearly, if the popup is closed, it's pointless to update
match icons anyway, so this seems reasonable.)

Also makes tiny tweaks I noticed while thinking about this (int/size_t mismatch,
impossible '<' case).

Bug: 1025168
Change-Id: Ic84967f5376ad4dc19787b878cbe6f473b3b7d74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1920100
Reviewed-by: Kevin Bailey <krb@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#715966}