Menu cleanup: Remove never-overridden function.
SimpleMenuModel::Delegate::GetMinorTextForCommandId() was never implemented by
anyone, so just remove it.

Technically, this is a potential behavior change: if an item previously set a
minor text AND was marked as dynamic, we'd hide the minor text, whereas after
this patch we'd show it.  This is never actually used and the old behavior
seems buggy anyway.

Bug: none
Change-Id: I00d1d77bbef42d1e8d53ac0e16f6cb6600b0176d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1918388
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#716250}