webauthn: make webauthn dialog back button restart request per default
Make the WebAuthn UI "back" button invoke the request restart logic on
every dialog that it is shown by changing the default implementation
in AuthenticatorRequestSheetModel::OnBack(). Previously, the default
implementation was to return to the transport selection UI *without*
restarting the request, which is not a particularly useful
implementation and was rarely used.

This also fixes a bug that was caused by the account selection back
button using said default implementation.

Fixed: 1020757
Change-Id: Id3af91ad2eab756f3dae7cf55410f1b46b18810b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1902646
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Nina Satragno <nsatragno@chromium.org>
Cr-Commit-Position: refs/heads/master@{#716288}