[imagebitmap] Remove another use of ArrayBuffer
In this CL we remove another use of ArrayBuffer in image_bitmap.cc,
similar to https://crrev.com/c/1835659. However, this time it cannot
be avoided to create a SkPixmap object, due to scaling and color
settings. Therefore we still create the SkPixmap object to call
scalePixels and setColorSpace on it, but then we only use info()
and rowBytes() from it to create the SkImage from the initially
allocated SkData.


R=jbroman@chromium.org

Bug: chromium:1008840
Change-Id: I82ad357db16f11139e799fbee4423a777d5c2a33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1916476
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#716302}