[AW] Always schedule variations seed download jobs when a seed is requested
We don't currently schedule a seed download in WebView if the previous
download occurred less than 24 hours ago, which can add an additional 24
hours to the time between seed downloads. This CL reimplements the 24
hour throttle to be a delay in the scheduled job, meaning a job will
always be scheduled when a new seed is requested, but Android will only
run the job once 24 hours have past since the last download.

This CL also adds a missing hamcrest dependency to
//aw/test:webview_instrumentation_test_apk

Test: ninja -C out/aw webview_instrumentation_test_apk
Test: out/aw/bin/run_webview_instrumentation_test_apk -f 'AwVariationsSeedFetcherTest#*'
Bug: 979075
Change-Id: I0ea1cc8197f66b77634e8de2b145fe4228c74067
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1906121
Reviewed-by: Changwan Ryu <changwan@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Cr-Commit-Position: refs/heads/master@{#716305}