Migrate the Chrome OS assistant service off the mojo pref service
The assistant service runs in the browser process on the UI thread.
It can directly use the Profile's PrefService* to access prefs.
This simplifies the code and is a step toward deleting the mojo
pref service entirely.

The only non-trivial change is that the AssistantServiceConnection
object needs to destroy itself earlier in Profile teardown.
AssistantStateBase has a PrefChangeRegistrar that needs to be
destroyed before the Profile's PrefService is destroyed.

I manually tested on linux-chromeos under ASAN and on device.

TBR=battre@chromium.org

Bug: 977637
Test: chromeos_unittests
Change-Id: I6dec70a346b611c0825746cd1c39321d0dce466c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1919790
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Yue Li <updowndota@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#716399}