Fix a crash when label and value attribute are the same in an option element
After finishing the OnionSoup on the DateTimeChooser, there is
a crash when an option element doesn't have a label or value
attribute.

The crash has caused by setting a null value to the label attribute
when the label and the value are the same in
HTMLInputElement::SetupDateTimeChooserParameters. Because Mojo
message pipe can't get a null value at the receiving side.

This CL makes SetupDateTimeChooserParameters set an empty string
instead of a null string when the attributes are the same.

TEST: ExternalDateTimeChooserTest.OpenDateTimeChooserShouldNotCrashWhenLabelIsNone

Bug: 1022302
Change-Id: If0be3f1cb0d4f66d7791dbc36a9c60e60a1f072e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1923778
Commit-Queue: Gyuyoung Kim <gyuyoung@igalia.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/master@{#716541}