Split WebLayer#create() into createSync() and createAsync()
Currently WebLayer#create() returns a Future that is not a proper
implementation of Future interface: e.g. it requires get() to be called
only on UI thread. There seems to be no way to properly implement
Future interface that supports forcing sync load via get() (at least
as long as BrowserStartupController triggers StartupCallback on UI
thread). So it's better to split create() into createSync() and
createAsync().

This work needs to be done regardless of the outcome of
ListenableFuture discussion, because no proper Future
implementation seems possible in this particular case
(whether as supertype of guava's ListenableFuture or
our own ListenableFuture).

This CL removes our ListenbleFuture implementation:
if we don't go with Guava's ListenableFuture, then
ListenableResult is all we need, if we do -
ListenableResult will be straightforwardly replaced
with ListenableFuture everywhere.

Change-Id: I8a76bb69cc75b1a83d3daa07756f04c1bd5c4df6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1917162
Commit-Queue: Pavel Shmakov <pshmakov@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#716587}