Extend enable-portals flag expiry to M90.
Also use the portal/OWNERS file as the owners list, so that this
list of people is kept up to date (for instance, to include mcnee@).

Left myself as a backup, which seems common for groups listed in
this file, but this should be redundant with the OWNERS file.

Bug: 1026251
Change-Id: I44f8b20fe0afb9013b88c06bfc7ed337b2e109c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1924657
Auto-Submit: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#716660}