Correct resizer hit testing in overlay overflow controls reordered PaintLayers.
This hit testing issue can be divided into two aspects:
1. In a hit testing operation, the resizer was checked after positive z-order
and normal flow children. It may be obscured by child paint layers.

2. Resizer should be displayed above all DOM children's content, but the
inclusion relationship of the paint layers may not be consistent with the DOM
tree. That is to say, if one DOM Node A is the parent of Node B, their
correspond paint layers may not form a parent-child relationship. So find the
hit test result in paint layer's order may skip the resizer.

For example:
<div id="A" style="width: 200px; height: 200px; overflow: scroll">
  <div id="B" style="width: 200px; height: 200px; position: relative">
  </div>
</div>
This will create two sibling paint layers. When hit test the resizer, Node B
will be the result.

Bug: 1006173
Change-Id: I0484bc0947792f73237260097f5613b423a00dba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1921928
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#716662}