Migrate AppService off the mojo pref service
The app service runs in the browser process on the UI thread.
It can directly use the Profile's PrefService* to access prefs.
This simplifies the code, in particular avoiding some cases
where the connection to the pref service might not be available
yet.

AppService is also the last remaining customer of the mojo pref
service, so this is also a step toward deleting the mojo pref
service entirely.

Bug: 977637
Test: unit_tests

Change-Id: Ia6cb9377ed1dbbd72b16e8ac2bda162a199c8fa8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1920003
Commit-Queue: James Cook <jamescook@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Maggie Cai <mxcai@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Dominic Battré <battre@chromium.org>
Cr-Commit-Position: refs/heads/master@{#716676}