[Translate] Remove identical language logic for the new Tab UI
Playing around with this, it was very confusing when choosing an
identical language and then UI wouldn't change. Even if you went back into
the language picker UI, it would show the old language you had chosen.
It was also confusing as to which language I was actually trying to translate
to. Just allow the user to choose identical languages. They may want to do
this if they're changing the target before going and then changing
the source.

Bug: 976352
Change-Id: Icf15c6c11e9af9a708184082e800d783f9b93bdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1918655
Reviewed-by: anthonyvd <anthonyvd@chromium.org>
Commit-Queue: Megan Jablonski <megjablon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#716745}