Adds back support for the phoneticCharacters tts property
This property was removed back in
https://chromium-review.googlesource.com/c/chromium/src/+/1801749
but should not have been.

The purpose of the phoneticCharacters property is to specify when single letters should be pronounced phonetically. In general, we do not want to do so. For example, a
<p>a</p>
would have phonetics pronounced even if the user navigates to it via Search+Right.

Any text that happens to be split into single letters would be pronounced phonetically and then trigger a delay.

Change-Id: I71ef6033c2bdb4dc01926444dfa86a7cdf6698f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1919903
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Commit-Queue: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#716811}