[Extensions Menu] Update on tab changes
Extension items can change in response to tab changes (such as the
active tab, if a tab changes title or location, etc). Update the
extensions menu in response to tab changes.

Though we currently repopulate the entire menu by removing and
re-adding all child views, this does not *appear* to cause visible
UI jank (on a release or debug build), as long as the items aren't
moving sections. However, as we add more of these, we will likely
want to update items in-place.

Bug: 1021356

Change-Id: Iac3b208c338698d70dcec7ea77022650f659d3a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1920002
Reviewed-by: Caroline Rising <corising@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#716830}