DwoC in NetworkService should log url and initiator in crash keys.
This CL changes NetworkService, so that when a DumpWithoutCrashing
(DwoC) is called during a call to CreateLoaderAndStart, then the crash
report will include crash keys that capture the request url and the
request initiator.

This CL helps prepares for starting to enforce
request_initiator_site_lock in a follow-up CL (enforcing = treating as a
bad message and calling DwoC if the lock doesn't match the actual
request initiator).  The enforcement might need to be reverted (there is
still a very low volume of uninvestigated cases where the request
initiator doesn't match the lock).  The url and request initiator crash
keys are desirable for the long term, so landing them in this, separate
CL prevents unnecessary churn when working/landing/reverting/relanding
the lock enforcement.

Bug: 920634
Change-Id: I3900bf6a8f4a07d04166102dd78bfac53c3d0d96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1922698
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/master@{#716861}