Use unicode max codepoint for delimiter instead of ORC, and skip buffers with null NGOffsetMapping
It's possible to try to find the Object Replacement Character (ORC), so
we should not use that as a delimiter of invalid elements as we might
wrongfully think that the delimiters are an actual match, causing
crashes.

Additionally in some cases layout might fail causing the FindBuffer to
have null NGOffsetMapping, causing crashes. In this case we should skip
the entire block as we can't get the ranges correctly.

Bug: 1020105, 1002753, 1024256
Change-Id: I4c4cc886a84a919a617789db6d3bb0351a5d7477
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1924031
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/master@{#716882}