WebGPU: Implement returning adapter properties from GPU process
This patch implements sending the adapter properties from GPU process to
the renderer process in the call of adapter.RequestAdapter().

At the client side of WebGPU command buffer, before sending the request
of RequestAdapter() to the server side, a callback is saved in a map
whose key is the serial number of the request. After selecting the
preferred adapter, the GPU process will send the id and properties of
the adapter to the rendering process, the related callback will be
called to create the GPUAdapter object in Blink with the information
about this adapter from the GPU process.

BUG=chromium:996713

Change-Id: I81e252559a7d256213f4bf93405d63cd0e6fa02b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1897475
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Austin Eng <enga@chromium.org>
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>
Cr-Commit-Position: refs/heads/master@{#716963}