Refactor device name from c/b/sharing to components/STTS
For all sharing features in the case where the device name is not
useful we should fall back to using a common naming scheme. This CL
adopts that common naming scheme and moves it and the required helper
functions into components for cross platform access.

Bug: 1025362
Change-Id: Ibb20dc9db3a99b24f9688fd51838febab0419694
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1919738
Commit-Queue: Jeffrey Cohen <jeffreycohen@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Alex Chau <alexchau@chromium.org>
Reviewed-by: Richard Knoll <knollr@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#717208}