Prepare MediaEngagementServiceTest for HistoryServiceUsesTaskScheduler activation.
The feature will soon be flipped to ENABLED_BY_DEFAULT. This will make
HistoryServiceBackend's task runner dispatch to the ThreadPool instead
of a dedicated base::Thread. The tests need to be prepared to stay
functional when that happens.

This is done as the last mile of the chrome-wide effort to coalesce as
much work as possible to the ThreadPool.

Bug: 661143
Change-Id: I3793de26ba54c7818ef3f710dd88c389e5d01297
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1924624
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Commit-Queue: Oliver Li <olivierli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#717217}