ChromeVox: Add command to reset pitch, rate, and volume to defaults.
This change adds support for a new command to reset the ChromeVox
text to speech pitch, rate, and volume back to their default settings.
This also adds a SpokenFeedbackTest to ensure correct behavior and adds
the new command to the ChromeVox menu so users can discover it.

Bug: 820693
Change-Id: Icaa4de13aaba7ab2fb00abc5dd632713a5afa4d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1902266
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#717341}