[CrOS Settings] Display correct tooltip on account image if its a video.
Currently, if the user hovers over the thumbnail of the last taken
image/video, the tooltip always indicates that the account image is an
image even if it's a video.  For account images or videos taken in the
future, this CL ensures that the tooltip shown when hovering over the
thumbnail correctly describes the thumbnail as an image or video.

Fixed: 819115
Change-Id: I04b3d08cc9f6d4fb5c4b01a254047197694ae300
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1915460
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#717362}