Merge OptionalFromDoubleWithNull() with ValueOrUnresolved()
The latter was the first one to be added months ago in crrev.com/c/1615314. Then
I added also the former in crrev.com/c/1913418 just a few days ago without
noticing that both of them were basically the same.

This merges both functions in a single one named ValueOrUnresolved() but
implemented as OptionalFromDoubleWithNull() was because it's more compact (a
oneline ternary operator). Apart from that it's moved from animation.cc to
timing.h because it needs to be used outside animation.cc.

Bug: 737867
Change-Id: I800a8c1ab8a2f238f94d0f5972705a7144103805
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1926948
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Commit-Queue: Sergio Villar <svillar@igalia.com>
Cr-Commit-Position: refs/heads/master@{#717375}