[Extensions Menu] Fix TestIconImageObserver
TestIconImageObserver::Wait() waits for an icon image to be fully loaded
(which is an asynchronous operation). However, this image load happens
on-demand. If the extension is in the extensions menu, and not visible
on the toolbar, it won't load.

Force the image to load by requesting a bitmap.

Bug: 984654
Change-Id: I90651c2ec1f851c91bdc34f3787c8a50471070a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1927026
Reviewed-by: Caroline Rising <corising@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#717380}