Always try Chrome OS shortcuts when ChromeVox is enabled.
This is a short-term workaround for the issue that Chrome OS system
shortcuts are never invoked while ChromeVox is running.
By this change, a key event are sent to both of a wayland client and
Chrome OS's event flow regardless of whether the wayland client sends
ack to the event.
As a result, Chrome OS system shortcuts like Ctrl-w always work
immediately while ChromeVox is enabled.

Bug: b/130312917
Test: exo_unittest
Change-Id: I5eb4c13c02f0a48eee10795a94a19abf814e536f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1924029
Reviewed-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Mitsuru Oshima (OOO - back on Dec 2nd) <oshima@chromium.org>
Commit-Queue: Yuichiro Hanada <yhanada@chromium.org>
Cr-Commit-Position: refs/heads/master@{#717406}