headless: Fix how language is set.
https://crrev.com/491708 tried to fix this by reading --lang and passing
it to HeadlessBrowserContext::Builder::SetAcceptLanguage(). However,
this produces an inconsistent state were JS navigator.languages has the
right language, but JS Date.toLocaleDateString() ignores --lang on
platforms like Linux, where --lang is typically not honored.

Looking into this, one clue is there exists another piece of code,
HeadlessContentMainDelegate::InitializeResourceBundle(), that also reads
--lang. InitializeResourceBundle() will indirectly trigger a call to
l10n_util::GetApplicationLocale(), and l10n_util is the component that
knows how to read the system locale, and how to apply --lang
appropriately. Given this, the to-be-used locale is already stored
somewhere, so it just needs to be read out and passed to
SetAcceptLanguage().

https://crrev.com/717313 adds InitApplicationLocale() to use l10n_util
code to set the ICU locale. With that in place, this CL retrieves the
locale with base::i18n::GetConfiguredLocale().

Now, headless mode will read the OS locale setting, and also respect
--lang in the same way as the browser would normally. This means
headless mode will use $LANG on Linux, and --lang on some other
platforms. See GetApplicationLocaleInternal() to better understand the
platform specific behavior with respect to |pref_locale|, which
represents the --lang value.

Bug: 747953, 755338
Change-Id: I724320a093dcfe9ffc3739e321115c9e48b52eeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1916739
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#717412}