[DeviceSync v2] Process v2 devices in RemoteDevice provider
The new implementation can handle v1 and/or v2 DeviceSync data,
depending on the state of the feature flags to enable v2 DeviceSync and
deprecate v1 DeviceSync.

If v1 and v2 DeviceSync are running in parallel, we consider the
following facts:
 - Devices returned from a v2 DeviceSync should be a subset of the
   devices returned from a v1 DeviceSync.
 - The public key is the only consistent identifier between v1 and v2
   devices.
 - V2 devices might not have a decrypted public key or decrypted beacon
   seeds.
 - V1 and v2 device data might have different beacon seeds.
Given these facts, we merge the v1 and v2 devices in the following way
so as to avoid having duplicate device records and to ensure that v2
device data is preferred if it has decrypted metadata:
 - Initially, populate the synced-device list with all v1 devices.
 - If a v2 device does not have a public key, ignore it.
 - If a v2 device has a public key that matches that of a v1 device,
   replace the v1 device with the v2 device in the synced-device list.
 - If a v2 device has a public key that does not match any v1 device,
   append it to the synced-device list.

See http://go/cros-devicesync-v2-remote-device-id for more details.


Bug: 951969, 1019206
Change-Id: I2a79b87da8111d9fc215bb17f1c72b89b95ed7d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1907322
Commit-Queue: Josh Nohle <nohle@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#717416}