Language parity: batch six
This CL is part of a group that add the .xtb files to give Chrome and
Android language parity.

The first CL can be found here: https://crrev.com/c/1898148
It added a single locale to all .grd files.  The next group of CLs
adds the remaining languages to a group of files.

The three files modified in this CL are:
	- chrome/app/chromium_strings.grd
	- chrome/app/google_Chrome_strings.grd
	- chrome/credential_provider/gaiacp/gaia_resources.grd

They are unique in that use the generate_embedded_i18n[1] which requires
a list of locales to match exactly with the .xtb files created.  If more
.xtb files then expected are found the build will fail.

All new languages are added to the xtb_only_locales variable in
build/config/locales.gni which was added in c/1898148.

I'm not exactly sure why these three files are different.
generate_embedded_i18n creates .rc and .h files and should not increase
binary size.


Bug: 812837
Change-Id: Ic1d72f4c8236f9c541a7c0a2c1cac569585167b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1919495
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Trevor  Perrier <perrier@chromium.org>
Cr-Commit-Position: refs/heads/master@{#717447}