Language parity: batch two
This CL is part of a group that add the .xtb files to give Chrome and
Android language parity.

The first CL can be found here: https://crrev.com/c/1898148
It added a single locale to all .grd files.  The next group of CLs
adds the remaining languages to a group of files.

In batch two the following 7 .grd files and associated .xtb files are
updated:

	- chrome/app/google_chrome_strings.grd
	- chrome/browser/resources/chromeos/camera/src/strings/
	  camera_strings.grd
	- chrome/browser/resources/chromeos/chromevox/strings/
	  chromevox_strings.grd
	- chrome/browser/resources/chromeos/select_to_speak/strings/
	  select_to_speak_strings.grd
	- chrome/browser/resources/chromeos/switch_access/strings/
	  switch_access_strings.grd
	- chrome/browser/touch_to_fill/android/internal/java/strings/
	  android_touch_to_fill_strings.grd
	- chrome/browser/ui/android/strings/android_chrome_strings.grd

Bug: 812837
Change-Id: I08d8bebcb50f5ba006af491a2245306d88a359d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1918838
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Trevor  Perrier <perrier@chromium.org>
Cr-Commit-Position: refs/heads/master@{#717450}