[Network details] Fix proxy server info persisting after forgetting.
This CL fixes issue with manual proxy server input fields retaining
their old values even after the network was removed from known
networks. This was caused because the element caches the field values
so that they can be restored between proxy type switches, but the cache
was not being cleared when the page updated to a different network.

Fixed: 1025008
Change-Id: Ie18418116e5f3b2079f54c02725717a688fd0894
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1927173
Commit-Queue: Azeem Arshad <azeemarshad@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#717458}