Reland: Suppress beforeunload and in-progress download prompts for forced restarts.
This change introduces RelaunchIgnoreUnloadHandlers and uses it in
RelaunchNotificationController when the deadline for a forced relaunch
is reached. This new function is to AttemptRelaunch as
ExitIgnoreUnloadHandlers is to AttemptExit.

This also introduces browser_shutdown::ShouldIgnoreUnloadHandlers(),
which is a convenience for code that needs to know whether or not an
in-progress shutdown should ignore unload handlers.

(cherry picked from commit 2c98cab9372aa3bcadfc3f32bc2acd21b11d2156)

BUG=958865
TBR=treib@chromium.org

Change-Id: I291543b924d7b249093297a952ffa5c48f5343fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1924197
Commit-Queue: Greg Thompson <grt@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#717631}