[Passwords] Add checkbox to save to remote store
With this CL, the PasswordPendingView shows an additional checkbox in
test and debug builds. When the checkbox is selected/unselected, it
changes the store inside the pending_form_ of the
ManagePasswordsBubbleModel. So far, this field is unused (since
pending_form_ is a copy and only username and password are passed to
the delegate).

To see the checkbox:
* Build with is_debug=true or check_always_on=true
* set the flag \#passwords-account-storage-saving-ui
(or check out the screenshots in the linked bug)

Bug: 1018701
Change-Id: Ie26931fb375677cffb5de3c0b5dff3b0a1544375
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1926587
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#717685}