[net] Replace EmbeddedTestServer::Start() with managed variant.
It is common for tests to Start() the EmbeddedTestServer and to rely
on the server being shutdown on destruction. This creates scope for
use-after-free of objects held on the stack, or in test-case classes,
by request monitor & handler callbacks, which may continue to be run
on the EmbeddedTestServer's IO thread until it is shutdown.

Change-Id: I967caf71828269aad089b4dc5ab4aac39b24446d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1922687
Commit-Queue: Wez <wez@chromium.org>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#717917}