Fix crash after entering correct Parent Access Code in tablet mode
If the user is in tablet mode and enters the correct Parent Access
Code (PAC) on the first try, triggering auto-submit, then ChromeOS
crashes. The crash is caused by a race condition. This CL fixes the
root cause of the crash while maintaining the auto-submit behavior.

Bug: 1025075
Change-Id: I9fbed86993173951111a5ea7ed3cafddaea006bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1926529
Commit-Queue: Toby Huang <tobyhuang@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#717976}