Have ATK scrolling APIs scroll even when the node is already visible
Previously, when a node was already visible, ATK scrolling APIs would
not trigger a scroll. This change modifies the platform interface to
include an argument determining the behavior when the node is visible.
It is used to modify the behavior of the ATK API implementation.

Bug: 1022075
Change-Id: I31a70c4bb0086910768c439337f39d070933d6af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1911781
Reviewed-by: dsinclair <dsinclair@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Martin Robinson <mrobinson@igalia.com>
Cr-Commit-Position: refs/heads/master@{#718093}