Add highlights in PdfAccessibilityTree
Highlights are sent in a vector from the plugin layer. This CL uses the
vector to create highlight nodes and inserts them in the
PdfAccessibilityTree.

Test has been added to validate creation of PdfAccessibilityTree with
highlight as one of the nodes in the tree.

Bug: 1008775
Change-Id: Ia540b4421c7042a4da44a23676893a5916510d39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1903786
Commit-Queue: Ankit Kumar 🌪️ <ankk@microsoft.com>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#718167}