SuperSize: Caspian: Use lightweight GroupPath for TreeNode id_path
An id_path in the Tiger viewer is either a source path or a source path
prepended with some grouping information (a '>'-separated component like
"Blink>JavaScript>Runtime", or a template name like "std::vector<>").

Previously these concatenations were constructed and allocated, which
was extremely costly if the source path was very long. This change adds
a lightweight GroupPath object, which is basically a pair of
string_views with some utility methods for obtaining parent paths. This
also cleans up a lot of thorny logic for path-parsing: now, source_paths
are '/'-split, and the group can have a custom separator ('>' for "group
by component", '/' for other/group-by-template, since we don't want to
split on '>' characters in templates.

Unfortunately, now that id_paths are stored as this lightweight
GroupedPath object, TreeBuilder::Open(string path) can no longer do a
hashmap lookup by id_path key. In hindsight we could have maintained a
hashmap from GroupedPath -> TreeNode* and try all brute-force
string-splits of the path passed to Open(), but what I did instead was
split the input path into tokens and recurse down the tree looking for
nodes with matching names. This regresses tree performance for nodes
with many, many children, since tree traversal is doing a linear scan
on each level. Definitely not worth addressing in this CL, though.

Tests were added in a previous CL to ensure this change is safe.

Change-Id: I377d880ca1b8861a88d54bc2a208824fb4b8fb02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1929533
Commit-Queue: Jasper Chapman-Black <jaspercb@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Samuel Huang <huangs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#718260}