Use FakeInstanceIDWithSubtype in chrome unit tests.
These tests do not have network access, but InstanceID attempts to make
network queries when asked for a token (via getToken()). The prefetch
service posts a task during startup that ultimately asks for a token,
which causes the task to fail with a network timeout eventually (after
30 seconds). This causes tests that use a base::TaskEnvironment to be
very slow as their destruction waits for the 30 second task to
complete.

This halves the time spent in each unit_tests shard on the bots.

R=avi@chromium.org

Change-Id: I84815b07d167446e827df8777506f3d171bcfba5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1928297
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Filip Gorski <fgorski@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#718296}