Clarify inner and outer scroll node existence
This is a small refactoring to make it clearer that the inner and outer
viewport scroll nodes' existence is linked. That is, if either one is
non-null, then the other one must be as well.

Viewport scroll nodes (and others like page scale, elastic transform,
etc.) are only created for the renderer's main-frame compositor. In an
OOPIF renderer or the UI compositor, these property nodes are never
registered and will always return null. In the main-frame renderer, we
can always assume they're non-null.

Bug: 1016229
Change-Id: Ie022d3d9cb937848a0ceb6560e936ce1ff06142f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1871329
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#718301}