WebUI Tab Strip: Reorganize spacing for tabs
This CL sets overflow of the tabs container to hidden so that when the
last tab animates out and its max-width goes to 0, the container's
width also shrinks.

Moving the gap internally to TabElement makes more sense in terms of
consistency since it's also included internally in the animation of
slideIn and slideOut. Also, previously, having 16px margin on all tabs
except the last tab caused the 16px to snap into place when the last
tab slides out and a new tab becomes the last-child.

Change-Id: Ia0bec1c6c86ab0aa2bfd769fc488429419ff9613
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1924619
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#718310}