Disable extension installs for child users when pref is toggled off
When the "Permissions for sites and apps" toggle in Chrome family
dashboard is disabled, we should prevent child users from installing
new extensions and approving additional permissions for existing
extensions. This CL focuses on preventing child users from installing
new extensions. The second functionality will be implemented in a
future CL. When the toggle is disabled, pref
kSupervisedUserExtensionsMayRequestPermissions becomes false, and this
is used to implement the desired effects in this CL.

Bug: 1018956,1007011
Change-Id: Ida3c53d87da55fd266edc9d021211f5cdc641d0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1904502
Commit-Queue: Toby Huang <tobyhuang@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Dan S <danan@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Cr-Commit-Position: refs/heads/master@{#718389}