Restore the GPU watchdog V2 kInitFactor to 2 in Windows.
After adding more thread time and 60 seconds to the watchdog timeout, we
don't need full 60 seconds for initialization timeout. 30 seconds
(kInitFactor=2) should be a good number. We can adjust it later when
Timeout histogram data is available.

(gpu/OWNERS is added to the owner list of gpu_watchdog_timeout.h. The
constants defined in this file are for watchdog internal use and there is
no security concern.)


Bug: 988344, 949839
Change-Id: I9c5148f961e5db53c1b7123a7d5c7e2b13de6b89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1930143
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Maggie Chen <magchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#718451}