[Reland] Remove MIN_INTERSECT_FOR_REVEAL in ScrollIntoView
This threshold goes all the way back to:
https://crrev.com/ee5bc3acd0986e8fa87df7f804c49c120b8c9bc2 and was
initially added to avoid superfluous scrolling of form controls on
horizontally scrollable pages. If an element horizontally overlaps the
scrollport by at least this threshold it's considered "fully visible" as
per the scroll into view algorithm:
https://drafts.csswg.org/cssom-view/#scroll-an-element-into-view

Horizontal scrollable pages and partially visible form controls should
be rather uncommon and we should remove magical numbers from platform
behavior. If this behaivor is truly required for form controls we can
special case those but lets first try removing this altogether.

This does have some small impact on web-facing APIs in that
scrollIntoView (as well as fragment navigations) will now align an
element that's only partially on screen with the left edge. This matches
the behavior in FireFox and the spec. I've modified the
input-placeholder-paint-order.html and the ShowUnhandledTapUIIfNeeded
tests to avoid scrolling and anchor-frames-same-origin.html to expect
horizontal scrolling for this reason.

[Reland Note]
When first landed this caused
SitePerProcessHitTestBrowserTest.ScrolledNestedPopupMenuTest/0 to start
failing on linux-chromeos-rel with viz disabled. It now passes when
relanded on ToT since https://crrev.com/c/1928895 turned Viz on by
default on ChromeOS so this CL is a simple reland.

TBR=szager@chromium.org

Bug: 916631,1027403
Change-Id: Ic830083defa3684b0f2570edc5ae92249a6d9f18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1931381
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#718454}