Make SourceStream::ClearDataReadyCallback() cancel scheduled DownloadFileImpl::StreamActive() callback.
DownloadFileImpl::StreamActive() will post a task to call itself to
yield the task runner to others.
However, if before the task is posted, a Pause() and a Resume() is
also posted, the Pause() and Resume() will be executed before the
StreamActive(). Resume() will call StreamActive(), and possibly clearing
the stream handle in the SourceStream object if all the body is read.
This causes the StreamActive() call that is posted after the Resume()
to crash.

This CL fixes the issue by making the Task posted by StreamActive()
cancelable. So that in Pause(), the callback will be canceled and no
longer be executed after Resume() is called.

An alternative solution is in DownloadFileImpl::Resume(), we don't call
StreamActive() if the method is already posted.

BUG=1009839

Change-Id: I97d5c2be4fc18b69afbcee29500677c84376aef2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1933633
Commit-Queue: Min Qin <qinmin@chromium.org>
Reviewed-by: Xing Liu <xingliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#718728}