extensions: fix error bubble crash
This crash is caused by the following code in GenerateMessages():

  std::vector<base::string16> message;
  std::transform(external.begin(), external.end(), message.begin(),
                 [](const auto& extension) { ... });
  std::transform(forbidden.begin(), forbidden.end(), message.begin(),
                 [](const auto& extension) { ... });

Here, 'message' is empty so message.begin() doesn't point to a valid
location, which causes a crash when std::transform() tries to assign to
it. There is another bug too - the second std::transform() invocation
reuses message.begin(), so it will overwrite messages produced by the
first invocation.

This change:
1) Fixes both bugs mentioned above
2) Adds unit test coverage of this logic

Bug: 1028003
Change-Id: I2cbd660a5d5beb56a910a8c1a8ba239818c235f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1933952
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#718758}