RenderSubtree: Early out in PaintLayer::UpdateLayerPositionAfterLayout if locked.
This patch puts an early out for cases where the layout (for position)
or paint (for pagination) is blocked by the display lock.

This takes the growing update time on the referenced bug (example
in comment 17) from ~8ms per frame to ~1ms per frame on my linux desktop.

R=chrishtr@chromium.org

Bug: 999486
Change-Id: I1c7a2a1948371c61565c079d30341c0ee4c7e8b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1934509
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: vmpstr <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#718776}