Include raw view elements in UI Automation tree dumps
While working on tuning which nodes we include in the UI Automation
control and content views, I discovered that our tree dumper was only
dumping nodes in the control view. This CL applies a filter to dump the
raw view, which is a superset of the control view, giving us dumps that
are more complete.

R=iapres@microsoft.com

Bug: 1028345
Change-Id: I98d53285abb77eb8490903f85a2c21e63e1f8c33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1935027
Reviewed-by: Ian Prest <iapres@microsoft.com>
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#718915}