Do not dispatch mouse events to ChromeVox when ARC++ window is active
Mouse events should not be forwarded to ChromeVox,if called from ARC++.
This is what leads to "jumpy" and unexpected focus behaviour, as
described in the bug.

This CL checks if an arc window is active, and does this by adding an
activation observer to wm_helper.

TEST=manual (follow repro steps in bug, and observed expected behaviour)
BUG=b:141597151

Change-Id: I3d1da3801c1b826164e73a97fd639523b21202ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1905038
Commit-Queue: Sara Kato <sarakato@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Sara Kato <sarakato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#718923}