shelf a11y: Correct the ChromeVox hotseat navigation order
This change ensures that ShelfItems have the same ordering in both the
ShelfView's |view_model_| and the ShelfView's child views. This is
done to keep the child views in correct traversal order for ChromeVox.

This change also ensures that views that are removed from the
|view_model_| are also removed as child views from ShelfView. This
is done in ShelfItemRemoved() by initializing a unique_ptr to the
removed view. Once the unique_ptr goes out of scope the removed view
is deleted and removed as a child view.

Bug: 1024085
Change-Id: I4b4d6c3a855b41fc315a1d76f39c56ca0db380d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1925783
Reviewed-by: Manu Cornet <manucornet@chromium.org>
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#718959}