File types CSS: add files-ng file type CSS rules
In the files-ng case, the active / inactive icon of a directory tree
.tree-row will be handled with a -webkit-mask-image. Since .tree-row
are directly followed by a .file-row child element in files-ng, then
the -webkit-mask-image rules are written as follows:

   .tree-row > .file-row > blah ... {
     -webkit-mask-image: url(../some/source/image.svg);
   }

so they only match in the files-ng case. Next, the pre-existing

   .tree-row[selected] blah ...
   .tree-row blah ...

rules can match in both normal and files-ng files app. Rewrite those
rules with direct descendant selectors:

   .tree-row[selected] > blah ...
   .tree-row > blah ..

so they continue to match in the normal files app case (no change in
behavior) but do not match in the files-ng case.

Bug: 992819
Change-Id: Idba611834f12f951becfc45bdad6db9ac6090bfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1934132
Reviewed-by: Noel Gordon <noel@chromium.org>
Reviewed-by: Alex Danilo <adanilo@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#718960}