[OS Settings] Apply Chrome OS semantic colors to cr-toggles.
This CL plumbs the Chrome OS semantic colors into cr-toggles in OS
Settings by adding a shared style module to Settings on Chrome OS which
is active only on OS Settings and not Browser Settings.

To test, go to chrome://os-settings and type

$$('html')[0].style.setProperty('--cros-default-button-color', 'red')

into DevTools which should change the color of all toggles on the page.

Bug: 1018654
Change-Id: I7d680a9bb5c188c44639e59748f96845c83e67d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1925834
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: calamity <calamity@chromium.org>
Cr-Commit-Position: refs/heads/master@{#719014}