Add a retry in the JS code for initial focus.
Since the test fails even when the embedder WebContents and the iframe
for the text area both have focus (in the browser process), then next
theory is that focus hasn't had time to propagate to the DOM in some
cases.

This CL adds a retry mechanism in the JS code that tests for initial
focus in the iframe containing the text area.

TBR=alexmos@chromium.org

Bug: 1024672
Change-Id: I4a35a6e1c7858b207b65d41d954b3e5a618a2af8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1935023
Reviewed-by: James MacLean <wjmaclean@chromium.org>
Commit-Queue: James MacLean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/master@{#719155}