Convert Callbacks to OnceCallbacks for //content/app/
Use OnceCallback where possible, and BindRepeating where it is
meant to be called more than once.

The Closure here was explicitly a RepeatingClosure in one place,
but it really should be a OnceClosure. The closure is used to
exit the RunLoop, which can't occur more than once. It is
currently not used at all.

R=jam@chromium.org

Bug: 953861
Change-Id: I8a6b2a36b10c8fe523d163cf26fccaa8b155a319
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1934785
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#719186}