Make domain mixing metrics time lengths more explicit.
The test part of this change is necessary to support running HistoryBackend on
the ThreadPool.(Feature:HistoryServiceUsesTaskScheduler) On the ThreaPoool  we cannot
FastForwardUntilNoTasksRemain() for tasks using repeating timers.

Regardless of the new feature this change is intended to be a no-op and has
the added benefit of being more explicit in the time related expectations.

Bug: 661143
Change-Id: I1dcd2f3e2d1e5a5455953b0555c6475b2d0fdafd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1932417
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Cait Phillips <caitkp@chromium.org>
Auto-Submit: Oliver Li <olivierli@chromium.org>
Commit-Queue: Oliver Li <olivierli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#719223}