[Android] Move password check toggle next to Safe Browsing
The password check switch can only be toggled when the Safe Browsing
toggle is enabled.

The subtitle is displayed when the user is signed out and safe browsing
is enabled, since that is the moment when leak detection will be enabled
when signing in.

Note that disabling the Safe Browsing toggle only disabled the switch in
the UI and doesn't change the pref. The code which prevents the leak
detection logic to be run if safe browsing is disabled is on the
native side in https://crrev.com/c/1937175.

Bug: 1028095
Change-Id: I84ff30b8245417e8255024cea2e7679437a98636
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1932785
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#719248}