Android: Support pure-32-bit version codes on 64-bit build configs
This is an unchanged reland of 0c8f086794b4d8183f1094b2a43b0c9fbd41286d.

See the associated bug for reasons why this work is happening.

Refactor the version number generator script to be more explicit, and do
away with the hack that was in place to support multiple version numbers
in 64-bit config.

Also, add the new pure 32-bit Mono and Trichrome numbers, as well as an
alias 32_64 version number, which is equivalent to the ambiguous
"Monochrome" version.

TBR=agrieve@chromium.org

Bug: 1027573
Change-Id: Ib1d32b02c9db6e57eba49fba602486503071fce6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1936890
Reviewed-by: Christopher Grant <cjgrant@chromium.org>
Commit-Queue: Christopher Grant <cjgrant@chromium.org>
Cr-Commit-Position: refs/heads/master@{#719274}