[Fuchsia] Fix FilteredServiceDirectory to copy service name from char*.
Previously FilteredServiceDirectory was taking service names as char*
and then retaining that pointer to use it later. This didn't work
correctly when AddService() was called with name that's not stored in a
static buffer. Updated it to take StringPiece and then copy it to
std::string internally.

Change-Id: I382e3929eabb9806f4a74664ed355da886dcc61a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1935723
Reviewed-by: Wez <wez@chromium.org>
Commit-Queue: Wez <wez@chromium.org>
Auto-Submit: Sergey Ulanov <sergeyu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#719417}