Re-focus View on tab strip close
When the WebUI tab strip's WebContents gains focus, the focused
TextInputClient changes while the focused View does not. This causes
text input to stop working in the omnibox despite it having Views-focus.

This blurs and re-focuses the currently focused View. For the Omnibox,
this makes it re-install itself as the focused TextInputClient.

This is a workaround and should be removed once bug 994350 is fixed.

Bug: 1027375, 994350
Change-Id: I2be055110535dc6164a7d31a64a3cf503fae7a7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1935729
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#719473}