Add run_simpleperf.py
Adding a simple script to run simpleperf for WebView.
Unfortunately for WebView, there is no easy way to symbolize the native
stack output.

This is a somewhat quick and dirty workaround to use stack script to do
the post-processing on the original report HTML file that simpleperf
produces.

Bug: 1015236
Change-Id: Ica49b8af3af809162d2f895663c37f913c646028
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1929765
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Andrew Luo <aluo@chromium.org>
Commit-Queue: Changwan Ryu <changwan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#719485}