Add fill_into_edit to clipboard matches for iOS
On iOS, if there is no text in the Omnibox text field, the "Go" button
and the return key (on bluetooth keyboards) are disabled. This means
that when using the keyboard to select matches, the return key doesn't
work for the clipboard matches because no text is filled into the
omnibox.

This adds the following fill_into_edit:
URL match -> URL
text match -> Text
image match -> Search Copied Image (the same as the match's
  description)

Fixed: 1026595
Change-Id: I04686bc4b0cd38651d33b54113c97001a3a30eea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1937074
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#719541}