GMC: Only count click to go back to tabs as interactions
This CL updates the auto-dismiss logic to only count click-to-tabs as
interactions instead of counting when the tab is focused by any means.
The only reason we had it that way was to accommodate the
"pre-refactor" world where it wasn't feasible to always detect the
difference.

Bug: 1026079
Change-Id: I8806f64ade7019716413ef54c07693574f192dfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1937141
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Commit-Queue: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#719586}