Avoid retrieving surrounding text and its offsets for detached frames
Mojo receivers for the mojom::blink::Frame interface are not yet reset
when a frame gets detached on this stable branch (M79), so we need to
protect remote calls to GetTextSurroundingSelection() by making sure
that we retrieve the requested information for attached frames only.

Note that this is not an issue on the master branch due to other
changes present in there (crrev.com/c/1863718, crrev.com/c/1894918),
but this fix is definitely needed for the stable branch.

Bug: 1018124
Change-Id: Ifff8dfe52e19bd9243d93aad2c6a9ec8006f814f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1939490
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/branch-heads/3945@{#807}
Cr-Branched-From: e4635fff7defbae0f9c29e798349f6fc0cce4b1b-refs/heads/master@{#706915}