Consistently prevent replacement in NavigateFromFrameProxy
In NavigationControllerImpl::NavigateFromFrameProxy, there is a check
that prevents replacement if there are no existing entries. However,
this is only done for the browser's NavigationEntry. The original value
provided by the caller is used for the navigation request. The
renderer sees the initial navigation being done with replacement which
causes it to DCHECK in RenderFrameImpl::UpdateNavigationHistory.

We now prevent replacement for the navigation request to be consistent
with the navigation entry.

Bug: 1029036
Change-Id: I34ee5bc0b4b9c44c531f2b70fba24133eba23064
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1877129
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#719709}