[Extensions Cleanup] Remove ExtensionView::GetBrowser()
ExtensionView::GetBrowser() is used by the ExtensionViewHost and
ExtensionPopup to retrieve the Browser object the ExtensionView is
hosted in (if any). However, the Browser is a platform-agnostic concept,
and is passed through the ExtensionViewHost when creating the view.
Instead, have ExtensionViewHost hold the reference to the browser when
the view is created.

This cuts down on virtual methods and redirection, and allows the
ExtensionViewHost (which was 3/4 of the callers) to use the Browser
directly, eliminating it from the ExtensionView (which never actually
needed it) API surface.

Also a smattering of drive-by cleanups: NULL->nullptr, make_unique, etc

Bug: None
Change-Id: Ibef4d5c9685f026c44640d01cb948b16405c0e43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1938030
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#719746}