Fill in the second pass of the parser to analyze hprof files.
The main purpose of the java heap profiler is to obtain a heap dump in
the form of an hprof file and then parse the hprof file. Currently,
there exists a parser that does one pass and keeps track of all
the different instances and fields associated with them such as
object_id, name, etc.

With this CL, we pass over all the instances we just stored. The purpose
if that is that we are able to connect references between two different
instances.


Bug: 1012072
Change-Id: I03180731ab5ce347383f64d1103d5dd750326a96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1907281
Commit-Queue: ssid <ssid@chromium.org>
Reviewed-by: ssid <ssid@chromium.org>
Cr-Commit-Position: refs/heads/master@{#719942}