Allow powerwash by default, but forbid revert for enrolled devices
The powerwash was forbidden for enrolled devices initially to fix an
enrollment escape bug where users could revert to older Chrome OS
version and apply the weaknesses from the old version.

After this change lands, powerwash will be allowed for enrolled devices
in the following cases:
1 - Admin didn't set any value for the policy powerwash_allowed
2 - Admin set the powerwash allowed policy value to true
3 - Admin set the powerwash allowed policy value to false and the value
of policy allow_tpm_firmware_update to true and the device needs to update
the tpm firmware.

In case of point 3, the powerwash is allowed only to explicitly update the
tpm firmware.

At the same time, the option to powerwash and revert remains disabled
for enrolled devices.

The CL from session manager to allow this on Chrome OS side:
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1924409

BUG=chromium:923296
TEST=Check that ctrl+shift+R brings powerwash screen, and second press of
the keys doesn't bring the "powerwash and revert" screen.

Change-Id: I870a2c59c6989d041ad58dddf2c8ae40f20d94d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1924193
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Igor <igorcov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#719966}