ServiceWorker: Register only ServiceWorkerContainerHost for clients with ServiceWorkerContextCore
This is a follow-up CL for the previous review comment:
https://chromium-review.googlesource.com/c/chromium/src/+/1939157/5/content/browser/service_worker/service_worker_container_host.cc#70

ServiceWorkerContextCore::container_host_by_uuid_ contains the pairs of service
worker client UUID and a rawptr to ServiceWorkerContainerHost. This map is used
only for service worker client APIs, so we don't have to put
ServiceWorkerContainerHost for service worker execution contexts into the map.
This CL removes the case.

Also, the client UUID is specific to service worker clients, not for service
worker execution contexts. This CL stops assigning the client UUID to
ServiceWorkerContainerHost for service worker execution contexts.

Bug: 931087
Change-Id: Ia4cba861add33083c977845d60ec41c51db685e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1941532
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#719988}