[base] Disallow nested base::FieldTrialList instances
Such instantiations were temporarily allowed in
https://chromium-review.googlesource.com/c/chromium/src/+/1883567 before
migrating all tests that depended on it.

Now that no tests rely on such mechanism, the API can be removed, and
the absence of nested instances enforced via DCHECKs.

The only exception to this was SearchProviderFeatureTestComponent which
was recently changed to exercise nested FieldTrialList instances. Since
it's simple to fix, and before additional tests run into the same issue,
the test fixture is updated as part of this patch.

TBR=krb@chromium.org

Bug: 1011088,891959
Change-Id: I92b58ca5361351247567c512084aebc17d7348bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1904200
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Auto-Submit: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#720006}