[WebRTC] Restore event log generation for webrtc.
Since onion souping, the event logs generated for peer connection were
translated from std::string to WTF::String. However, the data being
stored is binary and WTF::String attempts to appropriately format the string,
resulting in corrupted binary data.

This CL changes back WTF::String to std::string where possible and
converts the binary data into a vector<uint8> to be sent via Mojo. This
way the data is correctly stored in the output file.

This is the first of 3 steps that will convert the data from string to
vector<uint8>, which is more suitable for representing binary data. The
next step will be to convert the data representation to vectors in the
renderer code and the last in the browser code.

Fixed: 1028536
Change-Id: If470069c45da24e298e9a4ceab926b8d7c08d05c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1942652
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Armando Miraglia <armax@chromium.org>
Cr-Commit-Position: refs/heads/master@{#720043}