SessionRestorePolicy - preparation for the LocalDB move to PM
This CL removes the |SiteCharacteristicsDataReader| object from the
SessionRestorePolicy::TabData struct. The goal is to make it easier to
switch to the Graph owned local DB once it's ready.

The |used_in_bg| bit is still set asynchronously and we need a way to
cancel the initialization callback when it goes out of scope. It used
to be done indirectly by releasing the SiteCharacteristicsDataReader
but it now has to be explicit, this is achieved by a adding a
CancelableOnceClosure member to SessionRestore::TabData. This callback
will also be useful once we have the PM based version of the LocalDB.

Bug: 961336
Change-Id: I443277dc711f2c22558fd76ec328b3623777b0d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1927168
Commit-Queue: Sébastien Marchand <sebmarchand@chromium.org>
Reviewed-by: Chris Hamilton <chrisha@chromium.org>
Cr-Commit-Position: refs/heads/master@{#720049}