Delay hiding search box until mouse click
Selecting a file, followed by opening the search box and attempting to
click 'delete' or 'open' buttons to the left of the search box failed
due to a mousedown listener that removed the search box and in turn,
moved the position of the 'delete' or 'open' buttons mid-click causing
the user action to fail.

Adds a new class to the search box 'hide-pending' that defers the
shrinking of the search box until a 'click' event is received on any
part of the action bar.

Noted during debugging that aria-disabled wasn't being set correctly, so
replace the use of 'hidden' for the <cr-input> element with 'disabled'
to control show/hide of the input text area, that updates aria-hidden
true || false correctly in the <cr-input>.

Change the Ctrl-F/Escape open and close of the search box to set the
'disabled' property instead of 'hidden'.

Bug: 668427
Tests: browser_tests --gtest_filter="*HidingTextEntryField"
Change-Id: Iac5f1cc0bec31f024f397797dd9c0397a6d703e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1940325
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#720096}