cc: Add image type to raster delay UMA.
This CL expands the raster task scheduling delay UMA to count raster
work that depends on decodes that could be hardware accelerated
(regardless of the status of the hardware decode acceleration feature
flags).

This is useful to measure the benefit of hardware decode acceleration on
raster work that would be directly affected by it.

Bug: 995155
Test: Added test coverage for the UMA expansion.
Change-Id: Ic70443558f026308e8334061829dd05813fbe9fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1794064
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Reviewed-by: Eric Karl <ericrk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#720296}