MacPWAs: Open PWA URL in window on failure
PWA app shims can become sad zombies that sit around and can't be
opened. I think we've plugged the holes through app shims can forget
to be deleted, but this can still happen if the user manually deletes
their user data dir, among other uncontrollable things.

Should we find that the shim provided is not for an app installed on
any profile, then open the URL specified in the shim in a browser
window, preferring the profile specified in the shim (if one is
specified and it exists), and then falling back to the last active
profile.

Also LOG(ERROR) the reason for failure in the app shim process (not
seeing this has made debugging hard in the past).

Bug: 1001213
Change-Id: I6fa396686525f3e3efbfd5b115b6439c714101ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1940495
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#720334}