ServiceWorker: Reduce dependencies on ServiceWorkerProviderHost from navigation
This is a part of the effort to split ServiceWorkerProviderHost into
ServiceWorkerHost and ServiceWorkerContainerHost.

This CL reduces dependencies on ServiceWorkerProviderHost from navigation in
favor of ServiceWorkerContainerHost. Conceptually, navigation for service
worker clients don't have to depend on ServiceWorkerProviderHost to be renamed
to ServiceWorkerHost. See the design doc for details.

Design doc:
https://docs.google.com/document/d/1epWIgelE-7uwxJHrYPKlbwqMRP9in2xLUR6mpiU_afY/edit?usp=sharing

Bug: 931087
Change-Id: I51c3bcb2e96149a576e58c98dceaa40ff88355fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1942673
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#720380}