[Android][TouchToFill] Increase half-state height to show 2+ suggestions
This cl increases the height of the touch to fill sheet to show a second
suggestion by default. To keep the height consistent, the peeking height
for the third suggestion was reduced.

It now looks according to the mocks linked in the linked bug.

Bug: 1029704
Change-Id: I52bbb15c3ee6cf9bc423a4745168addced02d69d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1945172
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#720506}