ConvertToBaseCallback to ConvertToBaseRepeatingCallback.
There are two CrossThreadFunction implementations, one that accepts
a base::OnceCallback, and another that accepts a base::Callback
(which is implicitly repeating). This change switches to explicitly
using base::RepeatingCallback to eliminate the use of base::Callback.

Also, renaming ConvertToBaseCallback to ConvertToBaseRepeatingCallback
for symmetry with CrossThreadBindRepeating and base::RepeatingCallback.

Bug: 1007835
Change-Id: I0b0a231bf74259194a78af9870960dbba0cc1f49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1919941
Commit-Queue: Chris Mumford <cmumford@google.com>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/master@{#720574}