Global Media Controls: stop observing WebContents when the session is destroyed.
The session objects stops observer the WebContents as part of its destructor but
the default destructor runs the inherited classes' destructors after the items'
destructor. Because an item ends up calling a WebContentsObserver callback, the
re-entrancy ends up in a crash. The solution is to stop observing the
WebContents as soon as the session is destroyed.

Bug: 1029311
Change-Id: Ic841eaeaa75f46aa591a3c0b9954688e87d3746e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1944887
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/branch-heads/3945@{#839}
Cr-Branched-From: e4635fff7defbae0f9c29e798349f6fc0cce4b1b-refs/heads/master@{#706915}