Make SourceStream::ClearDataReadyCallback() cancel scheduled DownloadFileImpl::StreamActive() callback.
DownloadFileImpl::StreamActive() will post a task to call itself to
yield the task runner to others.
However, if before the task is posted, a Pause() and a Resume() is
also posted, the Pause() and Resume() will be executed before the
StreamActive(). Resume() will call StreamActive(), and possibly clearing
the stream handle in the SourceStream object if all the body is read.
This causes the StreamActive() call that is posted after the Resume()
to crash.

This CL fixes the issue by making the Task posted by StreamActive()
cancelable. So that in Pause(), the callback will be canceled and no
longer be executed after Resume() is called.

An alternative solution is in DownloadFileImpl::Resume(), we don't call
StreamActive() if the method is already posted.

BUG=1009839

(cherry picked from commit eab211bb64bc9954e34fe49bc68e265223f2a9b3)

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1933633
Commit-Queue: Min Qin <qinmin@chromium.org>
Reviewed-by: Xing Liu <xingliu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#718728}
Change-Id: I3e06a5c0aa2a11767529fc13c117712ba483a5f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1946938
Reviewed-by: Min Qin <qinmin@chromium.org>
Cr-Commit-Position: refs/branch-heads/3945@{#841}
Cr-Branched-From: e4635fff7defbae0f9c29e798349f6fc0cce4b1b-refs/heads/master@{#706915}