cbui: don't have two cancel buttons in error bubble
This change fixes a bug in the extension error bubble in which that
bubble shows two "Details" buttons.

This was ultimately caused by the fact that
GlobalErrorWithStandardBubble defined this method:

  virtual bool ShouldUseExtraView();

When this method returned true, GlobalErrorBubbleView was supposed to:

1) Not have a Cancel button, even if there was Cancel button text
2) Add a button in the extra view slot whose text was the Cancel button
   text
3) Call the GlobalErrorWithStandardBubble back via the cancel pressed
   method when that extra button was pressed

This was used to add a "Details" button to extension errors.
Unfortunately, the condition in (1) led to code like this:

  if (!ShouldUseExtraView() && has_cancel_text)
    AddCancelButton();

which looks like a bug, and so in 74eb4979 I "corrected" it in the
obvious manner. That led to there being two cancel buttons in this
instance. :(

This change:
1) Adds first-class support to GlobalErrorWithStandardBubble for a
   "details" button slot, independent of the cancel & accept buttons
2) Migrates ExtensionErrorUIDefault to use that mechanism
3) Removes GlobalErrorWithStandardBubble::ShouldUseExtraView

Bug: 1024995
Change-Id: I31fe6a7a1d64f0bca2d7abd524650954b4579286
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1933205
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#720611}