[MERGE] weblayer patches
This merges two patches from WebLayer:

Change WebLayer's DownloadCallback interface to return a boolean if a download is intercepted.

This will make it easier in the future to add other methods to allow WebLayer to handle the download so that corner cases of downloading (e.g. depend on cookies/auth, POSTs) work. By adding a return value, we would know in future iterations of the API when WebLayer can handle the download.

Change-Id: Id933a14a39793aa8f00346ae73557fba5163fa43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1941117
Cr-Commit-Position: refs/heads/master@{#719823}

Replace url with uri to be consistent in the API.

Change-Id: I29fa0411f7abe815bc51463c91763d64be54c00e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1924602
Cr-Commit-Position: refs/heads/master@{#717746}

Bug: 1025603, 1024913
Change-Id: Ie94671db4789ebfbabb89ce058e53d06682674e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1946988
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/branch-heads/3945@{#842}
Cr-Branched-From: e4635fff7defbae0f9c29e798349f6fc0cce4b1b-refs/heads/master@{#706915}