Remove DCHECKs asserting deterministic URL rewrites during navigation.
As noted in the comment next to one of the DCHECKs,
RewriteUrlForNavigation has non-deterministic behavior in practice.

This is particularly relevant to Reader Mode, which includes a random
string into the distiller's URL, causing each call to the URL-
generating method to be different.

Bug: 992073
Change-Id: I78eb582adc48c9fe90a460adeeab9efc66d9fea9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1873957
Commit-Queue: Aran Gilman <gilmanmh@google.com>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#720710}