Keeping Mobile.DefaultBrowser.State
This metric was set to expire in M77, but is used in conjunction with
the others that do not expire to understand the browser landscape and
Chrome's default usage on the device.  Removing M77 because this is an
ongoing thing we'd like to track.

Change-Id: I263a73f145fb6bd44e496bb2afbb42295eb808f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1625270
Commit-Queue: David Trainor <dtrainor@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Auto-Submit: David Trainor <dtrainor@chromium.org>
Cr-Commit-Position: refs/heads/master@{#720737}