Honoring DPI changes in PdfAccessibilityTree
This change fixes the incorrect scale that we apply to root transform
in PdfAccessibilityTree. At present, we remove the dpi factor from root
transform. This results in only zoom factor being applied when
calculating bounds. For Windows, device scaling is also handled by
application. This is not the case with Mac. As a result, conditional
scale calculation, based on OS, should be done.

Fix uses content::IsUseZoomForDSFEnabled() to find out the correct
scale that needs to be applied.

Bug: 1007169
Change-Id: I4cda488b8acf3823178b4ff68503b081e12aeaa2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1820697
Commit-Queue: Virender Singh <virens@microsoft.com>
Reviewed-by: Ian Prest <iapres@microsoft.com>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#720834}