Add metrics for phone number regex
This adds metrics for the new phone number regex compared to the
previous one. These are logged when we show the context menu and when a
user selects a device from the menu. With those we can determine the
false positive and false negative rate compared to the current regex.

Bug: 1001033
Change-Id: Ibacfb095305eb106ec3dd21bce7c0e161172824c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1942655
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Michael van Ouwerkerk <mvanouwerkerk@chromium.org>
Commit-Queue: Richard Knoll <knollr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#720904}