Rename SVGMarkerData to SVGMarkerDataBuilder
Since SVGMarkerData is essentially a builder class for creating the list
of marker positions, add a Builder suffix to the class name.

Also move the "driver" parts into new method Build(...) on the class and
make the old "entrypoints" private. Move MarkerForType(...) to
MarkerPosition and rename it to SelectMarker(...).

Bug: 674797, 633012
Change-Id: I7191434047c9227ee6555a318afa5b1829a4bfdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1947653
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/master@{#720973}