Fix check for mouse event in occluded window.
WindowOcclusionTracker tracks WebContentsViewAura’s window_, not it's own
window_. See WebContentsViewAura::CreateAuraWindow().
The WebContentsViewAura’s window_ appears to be a great grandchild of
DesktopWindowTreeHostWin()'s window.

It's unclear why this stat was ever non-zero,
but it fired 15 times in a week on stable.

It would be nice to have a test for this, but I'm not sure where a
meaningful test would live.

Bug: 813093
Change-Id: I65c175a2ee2343742c4e9196d950015a84154234
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1941113
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: David Bienvenu <davidbienvenu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#720990}