[CV] Move hashes computing to ComputedHashes
Historically we've computed extension resources' hashes in
ContentHash::CreateHashes, but logically it's a job for ComputedHashes
class, so now we do it in ComputedHashes::Compute method.

In the future this will allow us to reuse ComputedHashes::Create for
other purposes (eg. testing or computing hashes upon installation for
self-hosted extensions).

Bug: 796395, 958794
Change-Id: Ie16c53ee682a4b8066dd2e929e13cbbc5be2d6ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1913415
Commit-Queue: Oleg Davydov <burunduk@chromium.org>
Reviewed-by: Nikita Podguzov <nikitapodguzov@chromium.org>
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#721051}