[Translate] Don't go back to before translating when confirming advanced options
In the Tab UI, we shouldn't go back to the pre-translating state on
confirming advanced option when the languages stayed the same. We should
stay in the after translate state.

In addition, this cl fixes a bug for the button to confirm advanced
options. In the Tab UI, both the source and target language views own
their own done buttons. They both need to be stored and updated with
'Translate' and 'Done' in UpdateAdvancedView. Previously just the most
recently created button that is stored in advanced_done_button_ was
getting updated, since advanced_done_button_ was overridden on the
second call to CreateViewAdvancedTabUi.

Bug: 976352
Change-Id: Ie16534717f5f8fdf0a608bd654b9e26db82d3a51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1935725
Commit-Queue: Megan Jablonski <megjablon@chromium.org>
Reviewed-by: anthonyvd <anthonyvd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#721092}