Make ToolbarButton border align to background
The background is inset visually to not necessarily fill the bounds of
the view. These insets are non-zero in touch mode and at the edges of
the screen. This splits the insets of the border into painted insets and
additional insets for positioning of elements.

Bug: None
Change-Id: Ia290aa0948e24c40f3fccb80861973f3d53981c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1948490
Reviewed-by: Collin Baker <collinbaker@chromium.org>
Commit-Queue: Collin Baker <collinbaker@chromium.org>
Commit-Queue: Peter Boström <pbos@chromium.org>
Auto-Submit: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#721093}