[omnibox] ZeroSuggestProvider: classify search results w/ input text
Currently, when we get search suggest results back from
ZeroSuggestProvider, nothing is bolded. This is inconsistent with
as-you-type search suggestions.

The root cause is that we purposely set AutocompleteInput::input_text
to an empty string for all ZeroSuggest queries, to avoid leaking the
user's current URL. (The URL is separately included in eligible).

The lack of bolding is an unintended side effect.

This CL fixes that, by re-classifying all the search suggest results
that come back with the real input text on the client side.

This is primarily intended to improve On Focus Query Refinements,
but this will affect all ZeroSuggest modes that return search results
from a non-empty omnibox. (Should not affect NTP).

Note nothing is bolded at all when the input text is empty (on NTP).

Navigational suggestions are not affected by this CL. They are
currently not bolded at all.

Bug: 963173
Change-Id: I7982c780f4b03f3d2b4992051d291fe1cd86f4b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1947765
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#721114}