Ephemeral tab : Fixed destroy call
When ephemeral tab is gone, we should destroy the WebContents. Earlier
destroy was called when the sheet is hidden, which causes an issue if
we go to tab switcher and come back to the same tab. This CL moves it
to the onSheetContentChanged method, which is a better way of knowing
if the sheet has been closed.

Bug: 1029291
Change-Id: I1008355aa9ed4fc407f4d827ab3d0eecf2cc98b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1949079
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#721169}